### 2025-09-05 #8938 @editorialbot
[PRE REVIEW]: multimodars: A Rust-powered toolkit for multi-modality cardiac image fusion and registration
**Submitting author:** @yungselm (<a href="http://orcid.org/0000-0002-5861-3753">Anselm W.  Stark</a>)
**Repository:** https://github.com/yungselm/multimoda-rs
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.6
**Managing EiC:** Rachel Kurchin
