### 2025-09-05 #8938 @rkurchin
Hi @yungselm and thanks for your submission! I am looking for some specific items to make sure your submission fits our requirements at a high level (not at the more detailed review level) before moving on to finding an editor or putting this on our waitlist if no relevant editors are available.

- [ ] [OSI-approved](https://opensource.org/licenses) license -> I see you have a license file, but it's not currently being auto-identified as one of the standard OSI ones, please take a look at this
- [ ] Check paper (make sure compiles, read through, check for length [250-1000 words](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain)) -> paper is a bit long, please see if it can be trimmed down, e.g. if some content should be moved to software documentation

In the meantime, please take a look at the comments above ⬆️ from the editorialbot to address any DOI, license, or paper issues if you're able (there may not be any), or suggest reviewers.

For reviewers, please suggest 5 reviewers from the [database listed above](https://reviewers.joss.theoj.org/lookup) or your own ([non-conflicted](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)) extended network. Their GitHub handles are most useful to receive but **please** do not use "@" to reference them since it will prematurely ping them, and do not go asking folks to review yourself.
