### 2025-09-08 26437b8
shorten paper, update license, minor changes documentation
