+++ paper/paper.md
@@ -43,39 +43,34 @@ bibliography: bibliography.bib
 ---
 
 # Summary
-Coronary artery anomalies (CAAs) and coronary artery disease (CAD) require precise morphological and functional assessment for diagnosis and treatment planning. Cardiac computed tomography angiography (CCTA) provides comprehensive 3D coronary anatomy but lacks the sub-millimetre resolution and dynamic tissue detail available from intravascular imaging (intravascular ultrasound [IVUS], optical coherence tomography [OCT]). Developed initially to study dynamic lumen changes in CAAs, `multimodars` is a general-purpose toolkit that registers high-resolution intravascular pullbacks to CCTA-derived centerlines, resulting in locally enhanced fusion 3D vessel representations for visualization, geometric analysis, and patient-specific modelling. The toolkit supports four alignment paradigms (full, double-pair, single-pair, single) essential for comparing pullbacks acquired under different haemodynamic states (e.g., rest and stress in CAAs) or at different timepoints (e.g., pre- and post-stenting in CAD) [@stark2025true; @stark2025ivus] (see Figure 1).
+Coronary artery anomalies (CAAs) and coronary artery disease (CAD) require precise morphological and functional assessment for diagnosis and treatment planning. Cardiac computed tomography angiography (CCTA) provides comprehensive 3D coronary anatomy but lacks the sub-millimetre resolution and dynamic tissue detail available from intravascular imaging (intravascular ultrasound [IVUS], optical coherence tomography [OCT]). Developed initially to quantify dynamic lumen changes in CAAs, `multimodars` is a general-purpose toolkit that registers high-resolution intravascular pullbacks to CCTA-derived centerlines, producing locally enhanced fusion 3D vessel representations for visualization, geometric analysis, and patient-specific modelling. The toolkit implements four alignment paradigms (full, double-pair, single-pair, single) to compare pullbacks acquired under different haemodynamic states (e.g., rest vs. pharmacologic stress) or at different timepoints (e.g., pre- vs. post-stenting). `multimodars` targets deterministic, reproducible multimodal fusion for CAA research and general CAD applications. [@stark2025true; @stark2025ivus] (see Figure @fig:aim).
 
 # Statement of need
-Combining complementary imaging modalities is critical for developing reliable 3D models and improving plaque and lumen assessment, particularly in complex conditions like coronary artery anomalies (CAAs) where dynamic vessel deformation under stress is a key diagnostic feature. Intravascular imaging offers unmatched resolution for assessing these dynamic changes but provides limited 3D context; CCTA provides 3D anatomy but suffers from lower resolution and artefacts such as blooming. The need to quantify haemodynamic-induced lumen changes in our CAA research was the primary driver for developing `multimodars`. Prior studies demonstrated intravascular/CCTA fusion [@van20103d; @kilic2020evolution; @wu20203d; @bourantas2013new] for general coronary imaging, yet no open, flexible toolkit exists that is specifically designed for multi-state analysis (e.g., rest/stress, pre/post intervention) and balances high performance, deterministic behaviour, and easy integration into research workflows. `multimodars` addresses this gap by providing a Python package with a high-performance Rust backend that links to numpy arrays, enabling reproducible experiments and scalable processing for both CAAs and general CAD applications.
[…]
