+++ docs/tutorial.rst
@@ -14,7 +14,7 @@ This step-by-step tutorial demonstrates how to:
 
 1. Workflow csv files
 ^^^^^^^^^^^^^^^^^^^^^^
-After pip installing or locally building the package install it in the familiar way.
+After pip installing or locally building the package, install it in the familiar way.
 
 .. code-block:: python
 
@@ -38,9 +38,9 @@ Every file should be structured in the following way (no headers):
 +-------+---------+---------+---------+
 
 To acquire meaningful measurement data, the coordinates should be provided in mm or another SI unit instead of pixel values.
-Optionally a record file can be provided `combined_sorted_manual.csv`, which should have the following structure, here the first column should contain the desired frame order and measurement_1 
-represent the thickness of the wall between aorta and coronary and measurement_2 for the thickness between pulmonary artery and coronary (position just for demonstration). This is based on the 
-output of the AIVUS-CAA software _a link: https://github.com/AI-in-Cardiovascular-Medicine/AIVUS-CAA/:
+Optionally a record file can be provided `combined_sorted_manual.csv`, which should have the following structure, here the first column should contain the desired frame order and "measurement_1" 
+represent the thickness of the wall between aorta and coronary and "measurement_2" for the thickness between pulmonary artery and coronary (position just for demonstration). This is based on the 
+output of the `AIVUS-CAA software <https://github.com/AI-in-Cardiovascular-Medicine/AIVUS-CAA/>`_:
 
 +-----------------+---------------+---------------+---------------+---------------+
 | frame           | (position)    |   phase       | measurement_1 | measurement_2 |
@@ -70,7 +70,7 @@ This simplifies the workflow, by just providing a directory to automatically pro
         stress_output_path="output/stress",
     )
 
-However the preferred more flexible array is from numpy arrays.
+However the preferred more flexible way is from numpy arrays.
 
 2. Workflow numpy arrays and Finetuning
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -124,7 +124,7 @@ and number of points to represent the catheter. If no walls are provided a defau
     )
 
 This ``from_array`` function automatically aligns the frames within a pullback and then between pullbacks. The algorithm translates contours to the same centroid as the most proximal contour,
-and then finds the best rotation based on contour **AND** contour points.
+and then finds the best rotation based on contour **AND** catheter points.
 
 .. image:: ../paper/figures/Figure3.jpg
    :alt: Example figure
@@ -132,7 +132,7 @@ and then finds the best rotation based on contour **AND** contour points.
    :width: 400px
 
 The number of catheter points (``n_points``) therefore influences how much weight is given to the original image center. For mostly round contours, where Hausdorff distances are similar in different angles,
-this image center can increase accuarcy of the right rotation. For stenotic section or coronary artery anomalies, where the vessel has distinct shape difference, this number can be kept
+this image center can increase accuracy of the correct rotation. For stenotic sections or coronary artery anomalies, where the vessel has distinct shape difference, this number can be kept
 rather small (default 20 points compared to 500 for the contour).
 
 ``range_rotation_deg`` and ``step_rotation_deg`` define the +/- degree range where the rotation is tested (default 90° so full range) and step_rotation_deg in what step sizes (default 0.5°).
@@ -140,7 +140,12 @@ This algorithm is optimized and where it downsamples the original contour to 200
 in 0.1° steps and so on until the desired acccuracy). If bruteforce is set to 'True' the complete range is sweeped with the provided acccuracy (not recommended O(n^3)).
 
[…]
