### 2025-09-08 cc7031a
fix figure references
