### 2025-09-08 6d809d4
fix affiliations, bibliography and license
