+++ paper/bibliography.bib
@@ -6,7 +6,7 @@
   number={22},
   pages={104741},
   year={2025},
-  publisher={American College of Cardiology Foundation Washington DC}
+  publisher={American College of Cardiology Foundation Washington DC},
   doi={10.1016/j.jaccas.2025.104741}
 }
 
@@ -15,7 +15,7 @@
   author={Stark, Anselm Walter and Bigler, Marius Reto and Kakizaki, Ryota and R{\"a}ber, Lorenz and Gr{\"a}ni, Christoph},
   journal={Catheterization and Cardiovascular Interventions},
   year={2025},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi={10.1002/ccd.31660}
 }
 
@@ -27,7 +27,7 @@
   number={7},
   pages={781--796},
   year={2010},
-  publisher={Springer}
+  publisher={Springer},
   doi={10.1007/s10554-009-9546-y}
 }
 
@@ -38,7 +38,7 @@
   volume={7},
   pages={33},
   year={2020},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
   doi={10.3389/fcvm.2020.00033}
 }
 
@@ -50,7 +50,7 @@
   number={1},
   pages={18049},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
   doi={10.1038/s41598-020-74264-w}
 }
 
@@ -61,7 +61,7 @@
   volume={9},
   number={5},
   pages={582--93},
-  year={2013}
+  year={2013},
   doi={10.4244/EIJV9I5A94}
 }
 
@@ -71,6 +71,6 @@
   journal={medRxiv},
   pages={2025--02},
   year={2025},
-  publisher={Cold Spring Harbor Laboratory Press}
+  publisher={Cold Spring Harbor Laboratory Press},
   doi={10.1101/2025.02.18.25322450}
 }
