### 2025-09-08 3f9f87c
fix affiliations, bibliography and license
