+++ paper/bibliography.bib
@@ -7,6 +7,7 @@
   pages={104741},
   year={2025},
   publisher={American College of Cardiology Foundation Washington DC}
+  doi={10.1016/j.jaccas.2025.104741}
 }
 
 @article{stark2025ivus,
@@ -15,6 +16,7 @@
   journal={Catheterization and Cardiovascular Interventions},
   year={2025},
   publisher={Wiley Online Library}
+  doi={10.1002/ccd.31660}
 }
 
 @article{van20103d,
@@ -26,6 +28,7 @@
   pages={781--796},
   year={2010},
   publisher={Springer}
+  doi={10.1007/s10554-009-9546-y}
 }
 
 @article{kilic2020evolution,
@@ -36,6 +39,7 @@
   pages={33},
   year={2020},
   publisher={Frontiers Media SA}
+  doi={10.3389/fcvm.2020.00033}
 }
 
 @article{wu20203d,
@@ -47,6 +51,7 @@
   pages={18049},
   year={2020},
   publisher={Nature Publishing Group UK London}
+  doi={10.1038/s41598-020-74264-w}
 }
 
 @article{bourantas2013new,
@@ -57,6 +62,7 @@
   number={5},
   pages={582--93},
   year={2013}
+  doi={10.4244/EIJV9I5A94}
 }
 
 @article{stark2025automated,
@@ -66,4 +72,5 @@
   pages={2025--02},
   year={2025},
   publisher={Cold Spring Harbor Laboratory Press}
+  doi={10.1101/2025.02.18.25322450}
 }
