+++ src/centerline_align/preprocessing.rs
@@ -4,7 +4,26 @@ use crate::io::input::{Centerline, CenterlinePoint, ContourPoint};
 use crate::io::Geometry;
 use crate::processing::align_between::GeometryPair;
 
-pub fn ensure_descending_z(centerline: &mut Centerline) {
+/// Resample `centerline` along its arc-length so that adjacent points are spaced at the
+/// mean Euclidean distance between consecutive contour centroids in `ref_mesh`.
+///
+/// Precondition (expected caller behavior):
+/// - `centerline` should be trimmed so the first point corresponds to the aortic start
+///   (you already call `remove_leading_points_cl` before this).
+/// - `centerline` should be in decreasing z-order if that matters (you call `ensure_descending_z`).
+pub fn preprocess_centerline(
+    centerline: Centerline,
+    reference_point: &(f64, f64, f64),
+    ref_mesh: &Geometry,
+) -> Result<Centerline, &'static str> {
+    let mut cl = centerline;
+    ensure_descending_z(&mut cl);
+    let cl = remove_leading_points_cl(cl, reference_point);
+    let cl = resample_centerline_by_contours(&cl, ref_mesh)?;
+    Ok(cl)
+}
+
+fn ensure_descending_z(centerline: &mut Centerline) {
     if !centerline.points.is_empty() {
         let first_z = centerline.points[0].contour_point.z;
         let last_z = centerline.points.last().unwrap().contour_point.z;
@@ -14,7 +33,7 @@ pub fn ensure_descending_z(centerline: &mut Centerline) {
     }
 }
 
-pub fn remove_leading_points_cl(
+fn remove_leading_points_cl(
     mut centerline: Centerline,
     reference_point: &(f64, f64, f64),
 ) -> Centerline {
@@ -70,19 +89,176 @@ fn distance_sq(a: &ContourPoint, b: &(f64, f64, f64)) -> f64 {
     dx * dx + dy * dy + dz * dz
 }
 
-/// Resample `centerline` along its arc-length so that adjacent points are spaced at the
-/// mean Euclidean distance between consecutive contour centroids in `ref_mesh`.
-///
-/// Precondition (expected caller behavior):
-/// - `centerline` should be trimmed so the first point corresponds to the aortic start
-///   (you already call `remove_leading_points_cl` before this).
-/// - `centerline` should be in decreasing z-order if that matters (you call `ensure_descending_z`).
-pub fn resample_centerline_by_contours(centerline: &Centerline, ref_mesh: &Geometry) -> Centerline {
-    // If no centerline -> nothing to do
+fn resample_centerline_by_contours(centerline: &Centerline, ref_mesh: &Geometry) -> Result<Centerline, &'static str> {
     if centerline.points.is_empty() {
-        return Centerline { points: Vec::new() };
+        return Err("Centerline is empty");
+    }
+    if ref_mesh.contours.is_empty() {
+        return Err("Reference mesh has no contorus");
     }
 
+    let (centroids, mean_spacing_opt) = calculate_mean_spacing(ref_mesh);
+
+    let cum = cumulative_arc_length(centerline);
+    let total_length = *cum.last().unwrap_or(&0.0);
+
+    // Use n_segments = original segments count
+    let n_segments = centerline.points.len().saturating_sub(1);
+
+    let spacing = match decide_spacing(mean_spacing_opt, total_length, n_segments) {
+        Some(s) => s,
+        None => {
+            eprintln!("resample_centerline_by_contours: invalid spacing computed, returning original centerline");
+            return Ok(centerline.clone());
+        }
+    };
+
+    eprintln!(
[…]
