### 2025-10-20 d16b4ec
move intravascular logic into intravascular module
