### 2025-10-20 dbbc2b5
add stl reader for ccta mesh
