### 2025-10-21 14a8465
create a labeled mesh after reading in
