+++ src/ccta/io/input.rs
@@ -2,16 +2,72 @@ use anyhow::{anyhow, Context, Result};
 use std::fs::OpenOptions;
 use std::path::Path;
 use stl_io::IndexedMesh;
+use std::collections::{HashMap, HashSet};
 
-fn read_stl_ccta<P: AsRef<Path>>(path: P) -> Result<IndexedMesh> {
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub enum Label {
+    LCA,
+    RCA,
+    Intravascular,
+    Aorta,
+    Unknown,
+}
+
+#[derive(Debug, Clone)]
+pub struct LabeledMesh {
+    pub mesh: IndexedMesh,
+    pub labels: Vec<Option<Label>>,
+    label_index: HashMap<Label, HashSet<usize>>,
+}
+
+impl LabeledMesh {
+    /// Create a new LabeledMesh with all vertices unlabeled
+    pub fn from_mesh(mesh: IndexedMesh) -> Self {
+        let labels = vec![None; mesh.vertices.len()];
+        Self { 
+            mesh, 
+            labels,
+            label_index: HashMap::new(),
+         }
+    }
+
+    /// Assign a label to a vertex
+    pub fn set_label(&mut self, vertex_idx: usize, label: Label) {
+        if vertex_idx < self.labels.len() {
+            self.labels[vertex_idx] = Some(label);
+        } else {
+            eprintln!("Warning: vertex index {} out of bounds", vertex_idx);
+        }
+    }
+
+    /// Retrieve the label (if any) for a vertex
+    pub fn get_label(&self, vertex_idx: usize) -> Option<&Label> {
+        self.labels.get(vertex_idx).and_then(|o| o.as_ref())
+    }
+
+    /// Count how many vertices have a given label
+    pub fn count_label(&self, label: &Label) -> usize {
+        self.labels.iter().filter(|l| l.as_ref() == Some(label)).count()
+    }
+
+    /// Retrieve all vertices for a given label
+    pub fn vertices_with_label(&self, label: &Label) -> Vec<usize> {
+        self.labels.iter()
+            .enumerate()
+            .filter_map(|(i, l)| if l.as_ref() == Some(label) { Some(i) } else { None })
+            .collect()
+    }
+}
+
+fn read_stl_ccta<P: AsRef<Path>>(path: P) -> Result<LabeledMesh> {
     let path_ref = path.as_ref();
 
     if !path_ref.exists() {
-        return Err(anyhow!("File does not exist: {:?}", path_ref));
+        return Err(anyhow!("File does not exist: {:?}", path_ref.display()));
     }
 
     if !path_ref.is_file() {
-        return Err(anyhow!("Path is not a file: {:?}", path_ref));
+        return Err(anyhow!("Path is not a file: {:?}", path_ref.display()));
     }
 
     let mut file = OpenOptions::new()
@@ -21,15 +77,15 @@ fn read_stl_ccta<P: AsRef<Path>>(path: P) -> Result<IndexedMesh> {
 
     let stl = stl_io::read_stl(&mut file)
         .with_context(|| format!("Failed to parse STL file: {:?}", path_ref.display()))?;
-    
-    // check if is valid mesh
+
     if stl.validate().is_err() {
         return Err(anyhow!("STL file is not a valid mesh: {:?}", path_ref.display()));
     }
 
-    Ok(stl)
+    Ok(LabeledMesh::from_mesh(stl))
 }
 
+#[allow(dead_code)]
 fn read_obj_ccta() -> () {
     todo!()
 }
@@ -41,22 +97,7 @@ mod input_ccta_tests {
     use std::path::PathBuf;
[…]
