### 2025-10-22 92595a1
write watertight meshes, update full documentation, add trimesh as dependency
