+++ src/intravascular/binding/mod.rs
@@ -44,23 +44,23 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
 ///       v                         v
 ///    systole  -------------->  systole
 ///
-/// Arguments:
-///
-/// - ``rest_input_path`` – Path to REST input file
-/// - ``stress_input_path`` – Path to STRESS input file
-/// - ``step_rotation_deg`` (default 0.5°) – Rotation step in degree
-/// - ``range_rotation_deg`` (default 90°) – Rotation (+/-) range in degree, for 90° total range 180°
-/// - ``image_center`` (default (4.5mm, 4.5mm)) in mm
-/// - ``radius`` (default 0.5mm) in mm for catheter
-/// - ``n_points`` (default 20) number of points for catheter, more points stronger influence of image center
-/// - ``write_obj`` (default true)
-/// - ``rest_output_path`` (default "output/rest")
-/// - ``stress_output_path`` (default "output/stress")
-/// - ``diastole_output_path`` (default "output/diastole")
-/// - ``systole_output_path`` (default "output/systole")
-/// - ``interpolation_steps`` (default 28)
-/// - ``bruteforce`` (default false)
-/// - ``sample_size`` (default 200) number of points to downsample to
+/// Args:
+///     rest_input_path: Path to REST input file
+///     stress_input_path: Path to STRESS input file
+///     step_rotation_deg (default 0.5°): Rotation step in degree
+///     range_rotation_deg (default 90°): Rotation (+/-) range in degree, for 90° total range 180°
+///     image_center (default (4.5mm, 4.5mm)): in mm
+///     radius (default 0.5mm): in mm for catheter
+///     n_points (default 20): number of points for catheter, more points stronger influence of image center
+///     write_obj (default true): Wether to write OBJ files
+///     watertight (default true): Wether to write shell or watertight mesh to OBJ.
+///     rest_output_path (default "output/rest"): 
+///     stress_output_path (default "output/stress"):
+///     diastole_output_path (default "output/diastole"):
+///     systole_output_path (default "output/systole"):
+///     interpolation_steps (default 28): Number of interpolated meshes
+///     bruteforce (default false): Wether to use bruteforce alignment (comparison for every step size)
+///     sample_size (default 200): number of points to downsample to
 ///
 /// CSV format:
 ///
@@ -71,31 +71,27 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
 ///    ...
 ///
 /// Returns:
-///
-/// A ``PyGeometryPair`` for rest, stress, diastole, systole.
-/// A 4-tuple of ``Vec<id, matched_to, rel_rot_deg, total_rot_deg, tx, ty, centroid_x, centroid_y>``
-/// for (diastole logs, systole logs, diastole stress logs, systole stress logs).
+///     ``PyGeometryPair`` for rest, stress, diastole, systole.
+///     A 4-tuple of ``Vec<id, matched_to, rel_rot_deg, total_rot_deg, tx, ty, centroid_x, centroid_y>``
+///     for (diastole logs, systole logs, diastole stress logs, systole stress logs).
 ///
 /// Example:
-///
-/// .. code-block:: python
-///
-///    import multimodars as mm
-///    rest, stress, dia, sys, _ = mm.from_file_full(
-///        "data/ivus_rest", "data/ivus_stress"
-///    )
+///     >>> import multimodars as mm
+///     >>> rest, stress, dia, sys, _ = mm.from_file_full(
+///     ...     "data/ivus_rest", "data/ivus_stress"
+///     ... )
 #[pyfunction]
 #[pyo3(
     signature = (
         rest_input_path,
[…]
