### 2025-10-23 fd58185
read in any 3d file with trimesh
