+++ multimodars/__init__.py
@@ -12,7 +12,7 @@ from ._wrappers import (
     to_centerline,
 )
 from .io.write_geometries import centerline_to_obj
-from .io.read_geometrical import read_stl_geometry
+from .io.read_geometrical import read_mesh
 
 __all__ = [
     # rust‑backend functions
@@ -59,5 +59,5 @@ __all__ = [
     "to_obj",
     # io
     "centerline_to_obj",
-    "read_stl_geometry",
+    "read_mesh",
 ]
