### 2025-10-23 88f326f
simplify requirements
