+++ docs/conf.py
@@ -14,9 +14,9 @@ sys.path.insert(0, os.path.abspath("../src"))
 
 # Project info
 project = "multimodars"
-copyright = "2024, Anselm W. Stark"
+copyright = "2025, Anselm W. Stark"
 author = "Anselm W. Stark"
-release = "0.0.6"
+release = "0.0.7"
 
 # Extensions
 extensions = [
