+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "multimodars"
-version = "0.0.6"
+version = "0.0.7"
 edition = "2021"
 license = "MIT"
 
