+++ src/ccta/io/input.rs
@@ -94,114 +94,114 @@ fn read_obj_ccta() -> () {
     todo!()
 }
 
-#[cfg(test)]
-mod input_ccta_tests {
-    use super::*;
-    use std::fs;
-    use std::path::PathBuf;
-    use std::time::{SystemTime, UNIX_EPOCH};
+// #[cfg(test)]
+// mod input_ccta_tests {
+//     use super::*;
+//     use std::fs;
+//     use std::path::PathBuf;
+//     use std::time::{SystemTime, UNIX_EPOCH};
 
-    const STL_TEST_CONTENTS: &str = r#"solid tetra
-        facet normal 0 0 1
-            outer loop
-            vertex 0 0 0
-            vertex 1 0 0
-            vertex 0 1 0
-            endloop
-        endfacet
-        facet normal 0 1 0
-            outer loop
-            vertex 0 0 0
-            vertex 0 0 1
-            vertex 1 0 0
-            endloop
-        endfacet
-        facet normal 1 0 0
-            outer loop
-            vertex 0 0 0
-            vertex 0 1 0
-            vertex 0 0 1
-            endloop
-        endfacet
-        facet normal -1 -1 -1
-            outer loop
-            vertex 1 0 0
-            vertex 0 0 1
-            vertex 0 1 0
-            endloop
-        endfacet
-        endsolid tetra
-        "#;
+//     const STL_TEST_CONTENTS: &str = r#"solid tetra
+//         facet normal 0 0 1
+//             outer loop
+//             vertex 0 0 0
+//             vertex 1 0 0
+//             vertex 0 1 0
+//             endloop
+//         endfacet
+//         facet normal 0 1 0
+//             outer loop
+//             vertex 0 0 0
+//             vertex 0 0 1
+//             vertex 1 0 0
+//             endloop
+//         endfacet
+//         facet normal 1 0 0
+//             outer loop
+//             vertex 0 0 0
+//             vertex 0 1 0
+//             vertex 0 0 1
+//             endloop
+//         endfacet
+//         facet normal -1 -1 -1
+//             outer loop
+//             vertex 1 0 0
+//             vertex 0 0 1
+//             vertex 0 1 0
+//             endloop
+//         endfacet
+//         endsolid tetra
+//         "#;
 
-    fn write_temp_stl(contents: &str) -> PathBuf {
-        let mut path = std::env::temp_dir();
-        let nanos = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .expect("time went backwards")
-            .as_nanos();
-        let filename = format!("test_stl_{}_{}.stl", std::process::id(), nanos);
-        path.push(filename);
-        fs::write(&path, contents).expect("failed to write temp stl");
-        path
-    }
+//     fn write_temp_stl(contents: &str) -> PathBuf {
+//         let mut path = std::env::temp_dir();
+//         let nanos = SystemTime::now()
+//             .duration_since(UNIX_EPOCH)
+//             .expect("time went backwards")
+//             .as_nanos();
+//         let filename = format!("test_stl_{}_{}.stl", std::process::id(), nanos);
+//         path.push(filename);
+//         fs::write(&path, contents).expect("failed to write temp stl");
+//         path
+//     }
 
-    #[test]
-    fn test_read_stl_success() {
-        // a minimal ASCII STL content with a single triangle
-    let stl_contents = STL_TEST_CONTENTS;
+//     #[test]
+//     fn test_read_stl_success() {
+//         // a minimal ASCII STL content with a single triangle
+//     let stl_contents = STL_TEST_CONTENTS;
 
[…]
