+++ src/intravascular/binding/mod.rs
@@ -62,11 +62,12 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
 ///     bruteforce (default false): Wether to use bruteforce alignment (comparison for every step size)
 ///     sample_size (default 200): number of points to downsample to
 ///
-/// CSV format:
+/// .. warning::
+/// 
+///    The CSV must have **no header**. Each row is (frame index, x-coord (mm), y-coord (mm), z-coord (mm)):
 ///
 /// .. code-block:: text
 ///
-///    Frame Index, X-coord (mm), Y-coord (mm), Z-coord (mm)
 ///    185, 5.32, 2.37, 0.0
 ///    ...
 ///
@@ -178,8 +179,8 @@ pub fn from_file_full(
 ///
 ///    Rest:                    Stress:
 ///    diastole                  diastole
-///       |                         |
-///       v                         v
+///        │                         │
+///        ▼                         ▼
 ///    systole                   systole
 ///
 /// Args:
@@ -198,11 +199,12 @@ pub fn from_file_full(
 ///     bruteforce (default false): Wether to use bruteforce alignment (comparison for every step size)
 ///     sample_size (default 200): number of points to downsample to
 ///
-/// CSV format:
+/// .. warning::
+/// 
+///    The CSV must have **no header**. Each row is (frame index, x-coord (mm), y-coord (mm), z-coord (mm)):
 ///
 /// .. code-block:: text
 ///
-///    Frame Index, X-coord (mm), Y-coord (mm), Z-coord (mm)
 ///    185, 5.32, 2.37, 0.0
 ///    ...
 ///
@@ -307,18 +309,18 @@ pub fn from_file_doublepair(
 ///     step_rotation_deg (default 0.5°) – Rotation step in degree
 ///     range_rotation_deg (default 90°) – Rotation (+/-) range in degree, for 90° total range 180°
 ///     image_center (default (4.5mm, 4.5mm)): Center coordinates (x, y).  
-///     radius (default 0.5mm) Processing radius.  
-///     n_points (default 20) Number of boundary points.  
-///     write_obj (default true)
+///     radius (default 0.5mm): in mm for catheter
+///     n_points (default 20): number of points for catheter, more points stronger influence of image center
+///     write_obj (default true): Wether to write OBJ files
 ///     watertight (default true): Wether to write shell or watertight mesh to OBJ.
 ///     output_path: Path to write the processed geometry. 
-///     interpolation_steps (default 28) Number of interpolation steps.  
-///     bruteforce (default false)
-///     sample_size (default 200) number of points to downsample to
+///     interpolation_steps (default 28): Number of interpolated meshes
+///     bruteforce (default false): Wether to use bruteforce alignment (comparison for every step size)
+///     sample_size (default 200): number of points to downsample to
 ///
-/// CSV Format:
-///
-///     The CSV must have **no header**. Each row is:
+/// .. warning::
+/// 
+///    The CSV must have **no header**. Each row is:
 ///
 /// .. code-block:: text
 ///
@@ -331,7 +333,7 @@ pub fn from_file_doublepair(
 ///     for (diastole logs, systole logs).
 ///
 /// Raises:
-///     ``RuntimeError`` if the Rust pipeline fails.
+///     **RuntimeError** if the Rust pipeline fails.
 ///
 /// Example:
 ///     >>> import multimodars as mm
@@ -339,9 +341,6 @@ pub fn from_file_doublepair(
 ///     ...     "data/ivus_rest.csv",
 ///     ...     "output/rest"
 ///     ... )
-///
-/// .. note::
[…]
