+++ src/intravascular/io/geometry.rs
@@ -1,0 +1,355 @@
+use super::input::{ContourPoint, Record};
+use std::collections::HashMap;
+use std::f64::consts::PI;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum ContourType {
+    Lumen,
+    Eem,
+    Calcification,
+    Sidebranch,
+    Catheter,
+    Wall,
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct Contour {
+    pub id: u32,
+    pub original_frame: u32,
+    pub points: Vec<ContourPoint>,
+    pub centroid: Option<(f64, f64, f64)>,
+    pub aortic_thickness: Option<f64>,
+    pub pulmonary_thickness: Option<f64>,
+    pub kind: ContourType,
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct Frame {
+    pub id: u32,
+    pub centroid: (f64, f64, f64),
+    // groundtruth, must exist!
+    pub lumen: Contour,
+    pub extras:  HashMap<ContourType, Contour>,
+    pub reference_point: Option<ContourPoint>,
+}
+
+#[derive(Debug, Clone)]
+pub struct Geometry {
+    pub frames: Vec<Frame>,
+    pub label: String,
+}
+
+impl Contour {
+    pub fn new(&mut self,
+        points: Vec<ContourPoint>,  
+        records: Option<Vec<Record>>, 
+        label: ContourType) -> Self {
+        if records.is_some() {
+            Self::build_contour_records(self)
+        } else {
+            Self::build_contour(self)
+        }
+        todo!()
+    }
+
+    fn build_contour_records(&mut self) -> () {
+        todo!()
+    }
+
+    fn build_contour(&mut self) -> () {
+        todo!()
+    }
+
+    pub fn compute_centroid(&self) -> Self {
+        let (sum_x, sum_y, sum_z) = self.points.iter().fold((0.0, 0.0, 0.0), |(sx, sy, sz), p| {
+            (sx + p.x, sy + p.y, sz + p.z)
+        });
+        let n = self.points.len() as f64;
+        Self {
+            id: self.id,
+            original_frame: self.original_frame,
+            points: self.points.clone(),
+            centroid: Some((sum_x / n, sum_y / n, sum_z / n)),
+            aortic_thickness: self.aortic_thickness,
+            pulmonary_thickness: self.pulmonary_thickness,
+            kind: self.kind,
+        }
+    }
+
+    /// Finds the pair of farthest points in the current contour.
+    pub fn find_farthest_points(&self) -> ((&ContourPoint, &ContourPoint), f64) {
+        let mut max_dist = 0.0;
+        let mut farthest_pair = (&self.points[0], &self.points[0]);
+
+        for i in 0..self.points.len() {
+            for j in i + 1..self.points.len() {
+                let dx = self.points[i].x - self.points[j].x;
+                let dy = self.points[i].y - self.points[j].y;
+                let dist = (dx * dx + dy * dy).sqrt();
+                if dist > max_dist {
+                    max_dist = dist;
+                    farthest_pair = (&self.points[i], &self.points[j]);
+                }
+            }
+        }
+
+        (farthest_pair, max_dist)
+    }
+
+    /// Find the pair of points whose coord is the smallest diameter,
+    /// by matching each point to the one whose angle (about the centroid)
+    /// differs by as close to π radians as possible.
+    pub fn find_closest_opposite(&self) -> ((&ContourPoint, &ContourPoint), f64) {
+        let n = self.points.len();
+        assert!(n > 2, "Need at least 3 points");
+
+        // 1) Compute centroid (x0,y0). If missing, compute it first.
[…]
