### 2025-10-24 69a83b3
populate blueprint with existing functions
