+++ src/intravascular/io/wall.rs
@@ -1,0 +1,216 @@
+use std::ops::RangeInclusive;
+
+use crate::intravascular::io::input::ContourPoint;
+use crate::intravascular::io::geometry::{ContourType, Contour, Frame};
+
+pub fn create_wall_frames(frames: &Vec<Frame>, with_pulmonary: bool) -> Vec<Frame> {
+    let mut new_frames = Vec::new();
+
+    for frame in frames.iter() {
+        let new_contour = if with_pulmonary {
+            create_wall_contour_with_pulmonary(&frame.lumen)
+        } else {
+            create_wall_contour_aortic_only(&frame.lumen)
+        };
+
+        // clone extras and insert the wall contour under ContourType::Wall
+        let mut new_extras = frame.extras.clone();
+        new_extras.insert(ContourType::Wall, new_contour);
+
+        let new_frame = Frame {
+            id: frame.id,
+            centroid: frame.centroid,
+            lumen: frame.lumen.clone(),
+            extras: new_extras,
+            reference_point: frame.reference_point.clone(),
+        };
+        new_frames.push(new_frame)
+    }
+
+    new_frames
+}
+
+fn create_wall_contour_aortic_only(contour: &Contour) -> Contour {
+    if contour.aortic_thickness.is_none() {
+        let new_contour = offset_contour(contour, 1.0, None);
+        new_contour
+    } else {
+        let new_contour = create_aortic_wall(contour);
+        new_contour
+    }
+}
+
+#[allow(dead_code)]
+fn create_wall_contour_with_pulmonary(_contour: &Contour) -> Contour {
+    todo!("not yet implemented");
+}
+
+/// Offsets every point away from the centroid by exactly `distance` units,
+/// only for those whose `point_index` is in the optional `point_range`.
+/// If `point_range` is `None`, all points get offset.
+pub fn offset_contour(
+    contour: &Contour,
+    distance: f64,
+    point_range: Option<RangeInclusive<u32>>,
+) -> Contour {
+    let (cx, cy, cz) = match contour.centroid {
+        Some(c) => c,
+        None => {
+            let computed = contour.compute_centroid();
+            computed.centroid.unwrap()
+        }
+    };
+
+    let new_points = contour.points.iter().map(|pt| {
+        let mut p = pt.clone();
+
+        let do_offset = point_range
+            .as_ref()
+            .map(|r| r.contains(&(pt.point_index as u32)))
+            .unwrap_or(true);
+
+        if do_offset {
+            // vector from centroid → point
+            let dx = pt.x - cx;
+            let dy = pt.y - cy;
+            let dz = pt.z - cz;
+
+            let len = (dx * dx + dy * dy + dz * dz).sqrt();
+            if len > std::f64::EPSILON {
+                // build a unit vector, then add `distance` along it
+                let ux = dx / len;
+                let uy = dy / len;
+                let uz = dz / len;
+
+                p.x += ux * distance;
+                p.y += uy * distance;
+                p.z += uz * distance;
+            }
+        }
+
+        p
+    });
+
+    Contour {
+        id: contour.id,
+        original_frame: contour.original_frame,
+        points: new_points.collect(),
+        centroid: contour.centroid,
+        aortic_thickness: contour.aortic_thickness,
+        pulmonary_thickness: contour.pulmonary_thickness,
+        kind: ContourType::Wall,
+    }
+}
+
+/// Create aortic wall contour with the thickness from IVUS images
+/// assumption: using preexisting point indices since the contour is already aligned
+/// 0 is the highest point, 250 is the lowest point, point 375 used to build the thickness
[…]
