+++ src/intravascular/io/mod.rs
@@ -1,6 +1,7 @@
 pub mod input;
 pub mod output;
 pub mod geometry;
+pub mod wall;
 
 use anyhow::Context;
 use input::{read_records, Contour, ContourPoint, Record};
