+++ src/intravascular/io/geometry.rs
@@ -1,4 +1,5 @@
 use super::input::{ContourPoint, Record};
+use super::wall;
 use std::collections::HashMap;
 use std::f64::consts::PI;
 
@@ -37,19 +38,22 @@ pub struct Frame {
 pub struct Geometry {
     pub frames: Vec<Frame>,
     pub label: String,
+    // pointer to proximal end
 }
 
 impl Contour {
-    pub fn new(&mut self,
-        points: Vec<ContourPoint>,  
-        records: Option<Vec<Record>>, 
-        label: ContourType) -> Self {
-        if records.is_some() {
-            Self::build_contour_records(self)
-        } else {
-            Self::build_contour(self)
-        }
-        todo!()
+    // pub fn new(
+    //     points: Vec<ContourPoint>,  
+    //     records: Option<Vec<Record>>, 
+    //     label: ContourType) -> Self {
+    //     Self {
+    //         id,
+    //         original_frame,
+
+    //     }
+    // }
+    pub fn create_contour() {
+        
     }
 
     fn build_contour_records(&mut self) -> () {
@@ -60,6 +64,7 @@ impl Contour {
         todo!()
     }
 
+    /// Compute centroid by calculating mean of x-, y- and z-coordinates
     pub fn compute_centroid(&self) -> Self {
         let (sum_x, sum_y, sum_z) = self.points.iter().fold((0.0, 0.0, 0.0), |(sx, sy, sz), p| {
             (sx + p.x, sy + p.y, sz + p.z)
@@ -103,7 +108,7 @@ impl Contour {
         let n = self.points.len();
         assert!(n > 2, "Need at least 3 points");
 
-        // 1) Compute centroid (x0,y0). If missing, compute it first.
+        // 1) Compute centroid (x0,y0) if missing
         let (cx, cy, _) = match self.centroid {
             Some(c) => c,
             None => {
@@ -291,52 +296,69 @@ impl Frame {
         }
     }
 
-    pub fn create_catheter_points(
-        points: &Vec<ContourPoint>,
+    fn create_catheter_contour(
+        contour: &Contour,
         image_center: (f64, f64),
         radius: f64,
         n_points: u32,
-    ) -> Vec<ContourPoint> {
-        // Map to store unique frame indices and one associated z coordinate per frame.
-        let mut frame_z: HashMap<u32, f64> = HashMap::new();
-        for point in points {
-            // Use the first encountered z-coordinate for each frame index.
-            frame_z.entry(point.frame_index).or_insert(point.z);
-        }
+    ) -> Contour {
+        let centroid = match contour.centroid {
+            Some(c) => c,
+            None => {
+                let computed = contour.compute_centroid();
+                computed.centroid.unwrap()
+            }
+        };
 
         let mut catheter_points = Vec::new();
-        // Sort the frame indices to ensure a predictable order.
-        let mut frames: Vec<u32> = frame_z.keys().cloned().collect();
-        frames.sort();
 
-        // Parameters for the catheter circle.
         let center_x = image_center.0;
         let center_y = image_center.1;
-        let radius = radius;
-        let num_points = n_points;
 
-        // For each unique frame, generate 20 catheter points around a circle.
-        for frame in frames {
-            let z = frame_z[&frame];
-            for i in 0..num_points {
-                let angle = 2.0 * PI * (i as f64) / (num_points as f64);
-                let x = center_x + radius * angle.cos();
-                let y = center_y + radius * angle.sin();
-                catheter_points.push(ContourPoint {
-                    frame_index: frame,
-                    point_index: i,
-                    x,
-                    y,
-                    z,
-                    aortic: false,
-                });
-            }
+        // Generate n_points around a circle
+        for i in 0..n_points {
[…]
