### 2025-10-27 eefa332
save progress refactoring
