+++ src/intravascular/io/geometry.rs
@@ -1,7 +1,7 @@
-use super::input::{ContourPoint, Record};
-use super::wall;
+use super::input::{ContourPoint, Record, InputData};
 use std::collections::HashMap;
 use std::f64::consts::PI;
+use std::collections::HashSet;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum ContourType {
@@ -38,33 +38,64 @@ pub struct Frame {
 pub struct Geometry {
     pub frames: Vec<Frame>,
     pub label: String,
-    // pointer to proximal end
 }
 
 impl Contour {
-    // pub fn new(
-    //     points: Vec<ContourPoint>,  
-    //     records: Option<Vec<Record>>, 
-    //     label: ContourType) -> Self {
-    //     Self {
-    //         id,
-    //         original_frame,
-
-    //     }
-    // }
-    pub fn create_contour() {
-        
-    }
-
-    fn build_contour_records(&mut self) -> () {
+    pub fn new(&mut self,
+        points: Vec<ContourPoint>,  
+        records: Option<Vec<Record>>, 
+        label: ContourType) -> Self {
         todo!()
     }
+    
+    fn build_contour(
+        points: Vec<ContourPoint>,
+        records: Option<Vec<Record>>,
+        kind: ContourType,
+    ) -> anyhow::Result<Vec<Contour>> {
+        let mut groups: HashMap<u32, Vec<ContourPoint>> = HashMap::new();
+        for p in points {
+            groups.entry(p.frame_index).or_default().push(p);
+        }
 
-    fn build_contour(&mut self) -> () {
-        todo!()
+        let mut contours = Vec::new();
+
+        // Only process measurements for Lumen contour type
+        let measurements = if kind == ContourType::Lumen {
+            let mut meas = HashMap::new();
+            if let Some(records) = records {
+                for record in records {
+                    meas.insert(record.frame, (record.measurement_1, record.measurement_2));
+                }
+            }
+            Some(meas)
+        } else {
+            None
+        };
+
+        for (frame_idx, points) in groups {
+            let (aortic, pulmonary) = if let Some(ref meas) = measurements {
+                meas.get(&frame_idx)
+                    .map(|&(m1, m2)| (Some(m1), Some(m2)))
+                    .unwrap_or((None, None))
+            } else {
+                (None, None)
+            };
+
+            contours.push(Contour {
+                id: frame_idx,
+                original_frame: frame_idx,
+                points,
+                centroid: None,
+                aortic_thickness: aortic.flatten(),
+                pulmonary_thickness: pulmonary.flatten(),
+                kind,
+            });
+        }
+
+        Ok(contours)
     }
 
-    /// Compute centroid by calculating mean of x-, y- and z-coordinates
     pub fn compute_centroid(&self) -> Self {
         let (sum_x, sum_y, sum_z) = self.points.iter().fold((0.0, 0.0, 0.0), |(sx, sy, sz), p| {
             (sx + p.x, sy + p.y, sz + p.z)
@@ -108,7 +139,7 @@ impl Contour {
         let n = self.points.len();
         assert!(n > 2, "Need at least 3 points");
 
-        // 1) Compute centroid (x0,y0) if missing
+        // 1) Compute centroid (x0,y0). If missing, compute it first.
         let (cx, cy, _) = match self.centroid {
             Some(c) => c,
             None => {
@@ -239,6 +270,71 @@ impl Frame {
         todo!()
     }
 
+    /// Sets a value for all contours and reference point in the frame.
+    /// You can set `id`, `original_frame`, `points`, or `centroid` for all contours and reference point.
+    pub fn set_value(
+        &mut self,
+        id: Option<u32>,
[…]
