+++ src/intravascular/io/integrity_check.rs
@@ -1,0 +1,14 @@
+
+
+pub fn check_mandatory_data() {
+    todo!("Check that everything required is present")
+}
+
+pub fn check_point_number() {
+    todo!("Checks if every contour type has same number of points")
+}
+
+pub fn check_ids() {
+    todo!("Check that ids are in continuous order")
+}
+
