+++ src/intravascular/io/input.rs
@@ -1,9 +1,11 @@
+use super::geometry::ContourType;
+
 use anyhow::{anyhow, Context, Result};
 use csv::ReaderBuilder;
 use nalgebra::Vector3;
 use serde::Deserialize;
+
 use std::collections::HashMap;
-use std::f64::consts::PI;
 use std::fs::File;
 use std::io::{BufRead, BufReader};
 use std::path::Path;
@@ -23,24 +25,168 @@ pub struct InputData {
     sidebranch: Option<Vec<ContourPoint>>,
     record: Option<Vec<Record>>,
     ref_point: ContourPoint,
+    label: &str,
 }
 
 impl InputData {
-    pub fn new() {
+    pub fn new(
+        lumen: Vec<ContourPoint>,
+        eem: Option<Vec<ContourPoint>>,
+        calcification: Option<Vec<ContourPoint>>,
+        sidebranch: Option<Vec<ContourPoint>>,
+        record: Option<Vec<Record>>,
+        ref_point: ContourPoint,
+        label: &str,
+    ) -> anyhow::Result<InputData> {
+        let input = InputData {
+            lumen,
+            eem,
+            calcification,
+            sidebranch,
+            record,
+            ref_point,
+            label,
+        };
+        Ok(input)
+    }
+
+    pub fn process_directory<P: AsRef<Path>>(
+        path: P,
+        names: HashMap<ContourType, &str>,
+        diastole: bool,
+    ) -> anyhow::Result<InputData> {
+        let path = path.as_ref();
+
+        let mut lumen: Option<Vec<ContourPoint>> = None;
+        let mut eem: Option<Vec<ContourPoint>> = None;
+        let mut calcification: Option<Vec<ContourPoint>> = None;
+        let mut sidebranch: Option<Vec<ContourPoint>> = None;
+        let mut record: Option<Vec<Record>> = None;
+        let mut ref_point: Option<ContourPoint> = None;
+
+        let phase = if diastole { "diastolic" } else { "systolic" };
+
+        // iterate over provided names and attempt to build/read the corresponding files
+        for (_ctype, raw_name) in names.iter() {
+            let name = raw_name.trim().to_lowercase();
+            match name.as_str() {
+                "" | "lumen" => {
+                    // main contour files: "<phase>_contours.csv" and "<phase>_reference_points.csv"
+                    let contours_fname = format!("{}_contours.csv", phase);
+                    let contours_path = path.join(&contours_fname);
+                    if contours_path.exists() {
+                        lumen = Some(
+                            ContourPoint::read_contour_data(&contours_path)
+                                .with_context(|| format!("reading {}", contours_path.display()))?,
+                        );
+                    } else {
+                        return Err(anyhow::anyhow!(
+                            "expected lumen contours file missing: {:?}",
+                            contours_path
+                        ));
+                    }
+
+                    let ref_fname = format!("{}_reference_points.csv", phase);
+                    let ref_path = path.join(&ref_fname);
+                    if ref_path.exists() {
+                        ref_point = Some(
+                            ContourPoint::read_reference_point(&ref_path)
+                                .with_context(|| format!("reading {}", ref_path.display()))?,
+                        );
+                    } else {
+                        return Err(anyhow::anyhow!(
+                            "expected reference-point file missing: {:?}",
+                            ref_path
+                        ));
+                    }
+                }
+
+                "branch" | "sidebranch" => {
+                    let fname = format!("{}_{}_contours.csv", "branch", phase);
+                    let p = path.join(&fname);
+                    if p.exists() {
+                        sidebranch = Some(
+                            ContourPoint::read_contour_data(&p)
[…]
