+++ src/intravascular/io/mod.rs
@@ -4,313 +4,155 @@ pub mod geometry;
 pub mod wall;
 mod integrity_check;
 
-use anyhow::Context;
-use input::{read_records, ContourPoint, Record, InputData};
-use geometry::{Contour, Frame, Geometry};
+use input::{InputData, ContourPoint};
+use geometry::{Contour, Frame, Geometry, ContourType};
 use std::path::Path;
+use std::collections::HashMap;
 
 pub fn build_geometry_from_inputdata(
     input_data: Option<InputData>,
+    path: Option<&Path>,
     label: &str,
-    // both have default values
-    n_points: u32,
+    diastole: bool,
+    image_center: (f64, f64),
     radius: f64,
+    n_points: u32,
 ) -> anyhow::Result<Geometry> {
-    // for every contourtype get an Vec<Contour> by using the build_contour
-    // only passing records if kind is lumen.
+    let input_data = if let Some(input_data) = input_data {
+        input_data
+    } else if let Some(path) = path {
+        // Default mapping for typical AIVUS-CAA structure
+        let mut names = HashMap::new();
+        names.insert(ContourType::Lumen, "lumen");
+        names.insert(ContourType::Eem, "eem");
+        names.insert(ContourType::Calcification, "calcium");
+        names.insert(ContourType::Sidebranch, "branch");
+        names.insert(ContourType::Catheter, "catheter");
+        
+        InputData::process_directory(path, names, diastole)?
+    } else {
+        return Err(anyhow::anyhow!("Either input_data or path must be provided"));
+    };
 
-    // reorder by Record if [Record] is Some (should assign new id?)
+    let lumen_contours = Contour::build_contour(
+        input_data.lumen,
+        input_data.record.clone(),
+        ContourType::Lumen,
+    )?;
 
-    // create catheter contours for every frame
-    
-    todo!()
-}
+    let eem_contours = if let Some(eem_points) = input_data.eem {
+        Contour::build_contour(eem_points, None, ContourType::Eem)?
+    } else {
+        Vec::new()
+    };
 
-fn split_points_unique_id() {
-    todo!()
-}
+    let calcification_contours = if let Some(calc_points) = input_data.calcification {
+        Contour::build_contour(calc_points, None, ContourType::Calcification)?
+    } else {
+        Vec::new()
+    };
 
-fn assign_thickness() {
-    todo!()
-}
+    let sidebranch_contours = if let Some(side_points) = input_data.sidebranch {
+        Contour::build_contour(side_points, None, ContourType::Sidebranch)?
+    } else {
+        Vec::new()
+    };
 
-#[derive(Debug, Clone)]
-pub struct Geometry {
-    pub contours: Vec<Contour>,
-    pub catheter: Vec<Contour>,
-    pub walls: Vec<Contour>,
-    pub reference_point: ContourPoint, // needs to be set on aortic wall ostium!
-    pub label: &str,
-}
+    let mut frame_map: HashMap<u32, Frame> = HashMap::new();
 
-impl Geometry {
-    /// Creates a new Geometry instance by loading all required data files
-    pub fn new(
-        input_dir: &str,
-        label: String,
-        diastole: bool,
-        image_center: (f64, f64),
-        radius: f64,
-        n_points: u32,
-    ) -> anyhow::Result<Self> {
-        let label = if diastole {
-            format!("{}_diastole", label)
-        } else {
-            format!("{}_systole", label)
+    // Process lumen contours (mandatory)
+    for contour in lumen_contours {
+        let frame_id = contour.id;
+        let mut frame = Frame {
+            id: frame_id,
+            centroid: contour.centroid.unwrap_or((0.0, 0.0, 0.0)),
+            lumen: contour.compute_centroid(),
[…]
