+++ src/intravascular/io/output.rs
@@ -1,14 +1,13 @@
 use anyhow::{anyhow, bail, Context};
 use std::fs::File;
 use std::io::{BufWriter, Write};
-
-use crate::intravascular::io::input::Contour;
-use crate::intravascular::io::Geometry;
-use rayon::prelude::*;
 use std::path::Path;
 
+use super::geometry::{Contour, Geometry, ContourType};
+use rayon::prelude::*;
+
 pub fn write_obj_mesh(
-    contours: &Vec<Contour>,
+    contours: &[Contour],
     uv_coords: &[(f64, f64)],
     filename: &str,
     mtl_filename: &str,
@@ -67,7 +66,7 @@ pub fn write_obj_mesh(
 
     // Compute and write normals for original vertices
     for contour in &sorted_contours {
-        let centroid = &contour.centroid;
+        let centroid = contour.centroid.unwrap_or((0.0, 0.0, 0.0));
         for point in &contour.points {
             let dx = point.x - centroid.0;
             let dy = point.y - centroid.1;
@@ -105,20 +104,22 @@ pub fn write_obj_mesh(
 
     if watertight {
         let proximal_centroid_index = current_offset;
+        let first_centroid = sorted_contours[0].centroid.unwrap_or((0.0, 0.0, 0.0));
         writeln!(
             writer,
             "v {} {} {}",
-            sorted_contours[0].centroid.0, sorted_contours[0].centroid.1, sorted_contours[0].centroid.2
+            first_centroid.0, first_centroid.1, first_centroid.2
         )?;
         writeln!(writer, "vt 0.5 0.5")?;
         writeln!(writer, "vn 0.0 0.0 -1.0")?; // Pointing inward
 
         let distal_centroid_index = current_offset + 1;
         let last_contour = &sorted_contours[sorted_contours.len() - 1];
+        let last_centroid = last_contour.centroid.unwrap_or((0.0, 0.0, 0.0));
         writeln!(
             writer,
             "v {} {} {}",
-            last_contour.centroid.0, last_contour.centroid.1, last_contour.centroid.2
+            last_centroid.0, last_centroid.1, last_centroid.2
         )?;
         writeln!(writer, "vt 0.5 0.5")?;
         writeln!(writer, "vn 0.0 0.0 1.0")?; // Pointing inward
@@ -165,7 +166,7 @@ fn close_end(
 }
 
 pub fn write_obj_mesh_without_uv(
-    contours: &Vec<Contour>,
+    contours: &[Contour],
     filename: &str,
     mtl_filename: &str,
     watertight: bool,
@@ -175,35 +176,51 @@ pub fn write_obj_mesh_without_uv(
         .map_err(|e| anyhow!("Failed to write OBJ mesh without UV: {}", e))
 }
 
-#[derive(Copy, Clone)]
-pub enum GeometryType {
-    Contour,
-    Catheter,
-    Wall,
-}
-
-impl GeometryType {
+impl ContourType {
     // Get the contour data from a Geometry based on the enum variant
-    pub fn get_contours<'a>(&self, geometry: &'a Geometry) -> &'a Vec<Contour> {
+    pub fn get_contours(&self, geometry: &Geometry) -> Vec<Contour> {
         match self {
-            GeometryType::Contour => &geometry.contours,
-            GeometryType::Catheter => &geometry.catheter,
-            GeometryType::Wall => &geometry.walls,
+            ContourType::Lumen => geometry.frames.iter()
+                .map(|frame| frame.lumen.clone())
+                .collect(),
+            ContourType::Catheter => geometry.frames.iter()
+                .filter_map(|frame| frame.extras.get(&ContourType::Catheter))
+                .cloned()
+                .collect(),
+            ContourType::Wall => geometry.frames.iter()
[…]

[…]
