### 2025-10-27 b58ad0d
finish blueprint integrity_check
