+++ src/intravascular/io/integrity_check.rs
@@ -1,14 +1,459 @@
+use anyhow::{anyhow, Result};
+use std::collections::{HashMap, HashSet};
 
+use super::geometry::{Geometry, ContourType};
+use super::input::ContourPoint;
 
-pub fn check_mandatory_data() {
-    todo!("Check that everything required is present")
+/// Performs integrity checks on a Geometry structure
+pub fn check_geometry_integrity(geometry: &Geometry) -> Result<()> {
+    if geometry.frames.is_empty() {
+        return Err(anyhow!("Geometry has no frames"));
+    }
+
+    check_frame_ids_consecutive(geometry)?;
+    check_centroids_match(geometry)?;
+    check_lumen_presence(geometry)?;
+    check_reference_point(geometry)?;
+    check_contour_point_counts(geometry)?;
+    check_original_frame_consistency(geometry)?;
+
+    Ok(())
 }
 
-pub fn check_point_number() {
-    todo!("Checks if every contour type has same number of points")
+/// Check that frame IDs are consecutive (0, 1, 2, ...)
+fn check_frame_ids_consecutive(geometry: &Geometry) -> Result<()> {
+    for (index, frame) in geometry.frames.iter().enumerate() {
+        if frame.id != index as u32 {
+            return Err(anyhow!(
+                "Frame IDs are not consecutive. Expected ID {}, found ID {}",
+                index,
+                frame.id
+            ));
+        }
+    }
+    Ok(())
 }
 
-pub fn check_ids() {
-    todo!("Check that ids are in continuous order")
+/// Check that frame centroid matches lumen centroid
+fn check_centroids_match(geometry: &Geometry) -> Result<()> {
+    for (frame_index, frame) in geometry.frames.iter().enumerate() {
+        let frame_centroid = frame.centroid;
+        
+        // Compute lumen centroid from points if not present
+        let lumen_centroid = match frame.lumen.centroid {
+            Some(centroid) => centroid,
+            None => compute_centroid_from_points(&frame.lumen.points),
+        };
+
+        if !points_approximately_equal(frame_centroid, lumen_centroid) {
+            return Err(anyhow!(
+                "Frame centroid does not match lumen centroid in frame {} (ID {}). Frame: {:?}, Lumen: {:?}",
+                frame_index,
+                frame.id,
+                frame_centroid,
+                lumen_centroid
+            ));
+        }
+
+        // Also check that the stored lumen centroid matches computed if present
+        if let Some(stored_lumen_centroid) = frame.lumen.centroid {
+            if !points_approximately_equal(stored_lumen_centroid, lumen_centroid) {
+                return Err(anyhow!(
+                    "Stored lumen centroid does not match computed centroid in frame {} (ID {}). Stored: {:?}, Computed: {:?}",
+                    frame_index,
+                    frame.id,
+                    stored_lumen_centroid,
+                    lumen_centroid
+                ));
+            }
+        }
+    }
+    Ok(())
 }
 
+/// Check that lumen is present in every frame
+fn check_lumen_presence(geometry: &Geometry) -> Result<()> {
+    for (frame_index, frame) in geometry.frames.iter().enumerate() {
+        if frame.lumen.points.is_empty() {
+            return Err(anyhow!(
+                "Lumen contour has no points in frame {} (ID {})",
+                frame_index,
+                frame.id
+            ));
+        }
+
+        if frame.lumen.kind != ContourType::Lumen {
+            return Err(anyhow!(
+                "Lumen contour has incorrect type in frame {} (ID {}). Expected Lumen, found {:?}",
+                frame_index,
+                frame.id,
+                frame.lumen.kind
+            ));
+        }
+    }
+    Ok(())
+}
+
[…]
