+++ src/intravascular/binding/classes.rs
@@ -1,7 +1,7 @@
 // File: src/python_bind.rs
 use super::entry_arr::refine_ordering;
-use crate::intravascular::io::input::{Centerline, CenterlinePoint, Contour, ContourPoint, Record};
-use crate::intravascular::io::Geometry;
+use crate::intravascular::io::input::{Centerline, CenterlinePoint, ContourPoint, Record};
+use crate::intravascular::io::geometry::{Contour, Frame, Geometry};
 use crate::intravascular::processing::align_between::GeometryPair;
 use nalgebra::Vector3;
 use pyo3::prelude::*;
