+++ src/intravascular/utils/utils.rs
@@ -4,8 +4,7 @@ use std::fs::File;
 use std::io::{BufWriter, Write};
 use std::path::Path;
 
-use crate::intravascular::io::input::Contour;
-use crate::intravascular::io::Geometry;
+use crate::intravascular::io::geometry::{Contour, Geometry};
 
 #[allow(dead_code)]
 pub fn write_geometry_to_csv<P: AsRef<Path>>(
