+++ src/intravascular/binding/classes.rs
@@ -1,10 +1,11 @@
 // File: src/python_bind.rs
 use super::entry_arr::refine_ordering;
 use crate::intravascular::io::input::{Centerline, CenterlinePoint, ContourPoint, Record};
-use crate::intravascular::io::geometry::{Contour, Frame, Geometry};
+use crate::intravascular::io::geometry::{Contour, Frame, Geometry, ContourType};
 use crate::intravascular::processing::align_between::GeometryPair;
 use nalgebra::Vector3;
 use pyo3::prelude::*;
+use std::collections::HashMap;
 
 /// Python representation of a 3D contour point
 ///
@@ -133,9 +134,17 @@ pub struct PyContour {
     #[pyo3(get, set)]
     pub id: u32,
     #[pyo3(get, set)]
+    pub original_frame: u32,
+    #[pyo3(get, set)]
     pub points: Vec<PyContourPoint>,
     #[pyo3(get, set)]
     pub centroid: (f64, f64, f64),
+    #[pyo3(get, set)]
+    pub aortic_thickness: Option<f64>,
+    #[pyo3(get, set)]
+    pub pulmonary_thickness: Option<f64>,
+    #[pyo3(get, set)]
+    pub kind: String, // String representation of ContourType
 }
 
 #[pymethods]
@@ -146,29 +155,39 @@ impl PyContour {
     ///     id (int): Contour identifier
     ///     points (List[PyContourPoint]): List of contour points
     #[new]
-    fn new(id: u32, points: Vec<PyContourPoint>) -> Self {
-        let mut contour = Self {
+    fn new(
+        id: u32,
+        original_frame: u32,
+        points: Vec<PyContourPoint>,
+        centroid: (f64, f64, f64),
+        aortic_thickness: Option<f64>,
+        pulmonary_thickness: Option<f64>,
+        kind: String,
+    ) -> Self {
+        Self {
             id,
+            original_frame,
             points,
-            centroid: (0.0, 0.0, 0.0),
-        };
-        contour.compute_centroid();
-        contour
+            centroid,
+            aortic_thickness,
+            pulmonary_thickness,
+            kind,
+        }
     }
 
-    /// Returns human-readable representation of contour
     fn __repr__(&self) -> String {
         format!(
-            "Contour(id={}, points={}, centroid=({:.2}, {:.2}, {:.2}))",
+            "Contour(id={}, frame={}, points={}, centroid=({:.2}, {:.2}, {:.2}), kind={})",
             self.id,
+            self.original_frame,
             self.points.len(),
             self.centroid.0,
             self.centroid.1,
-            self.centroid.2
+            self.centroid.2,
+            self.kind
         )
     }
 
-    /// Returns the len of points
     fn __len__(&self) -> usize {
         self.points.len()
     }
@@ -211,7 +230,7 @@ impl PyContour {
     pub fn find_farthest_points(&self) -> PyResult<((PyContourPoint, PyContourPoint), f64)> {
         let rust_contour = self.to_rust_contour()?;
         let ((p1, p2), distance) = rust_contour.find_farthest_points();
-        Ok(((p1.into(), p2.into()), distance))
+        Ok(((PyContourPoint::from(&p1), PyContourPoint::from(&p2)), distance))
     }
 
     /// Finds closest points on opposite sides of the contour
@@ -224,7 +243,7 @@ impl PyContour {
     pub fn find_closest_opposite(&self) -> PyResult<((PyContourPoint, PyContourPoint), f64)> {
         let rust_contour = self.to_rust_contour()?;
         let ((p1, p2), distance) = rust_contour.find_closest_opposite();
-        Ok(((p1.into(), p2.into()), distance))
+        Ok(((PyContourPoint::from(&p1), PyContourPoint::from(&p2)), distance))
     }
 
[…]
