### 2025-10-28 63bcb88
add blueprint new processing module
