### 2025-10-29 a967deb
add contour rotation update blueprit align_within.rs
