### 2025-10-31 c34a652
add to align within blueprint
