### 2025-11-01 ec885b4
finish core align within logic
