+++ src/intravascular/neo_processing/align_within.rs
@@ -5,22 +5,21 @@ use std::f64::consts::PI;
 use std::sync::{Arc, Mutex};
 
 use crate::intravascular::io::input::ContourPoint;
-use crate::intravascular::io::geometry::{Contour, Frame, Geometry};
+use crate::intravascular::io::geometry::{ContourType, Frame, Geometry};
 use crate::intravascular::neo_processing::process_utils::downsample_contour_points;
 
 #[derive(Debug)]
 pub struct AlignLog {
     pub contour_id: u32,
     pub matched_to: u32,
-    pub rel_rot_deg: f64,
-    pub total_rot_deg: f64,
+    pub rot_deg: f64,
     pub tx: f64,
     pub ty: f64,
     pub centroid: (f64, f64),
 }
 
 pub fn align_frames_in_geometry(
-    geometry: Geometry,
+    geometry: &mut Geometry,
     step_deg: f64,
     range_deg: f64,
     smooth: bool,
@@ -28,12 +27,122 @@ pub fn align_frames_in_geometry(
     sample_size: usize,    
 ) -> (Geometry, Vec<AlignLog>) {
     let ref_idx = geometry.find_ref_frame_idx().unwrap_or(geometry.find_proximal_end_idx());
+    let sample_ratio = sample_size as f64 / geometry.frames[0].lumen.points.len() as f64;
+    let sample_size_catheter = if geometry.frames[0].extras.contains_key(&ContourType::Catheter) {
+        Some((geometry.frames[0].extras[&ContourType::Catheter].points.len() as f64 * sample_ratio).ceil() as usize)
+    } else {
+        None
+    };
 
+    let logger = Arc::new(Mutex::new(Vec::<AlignLog>::new()));
+
+    for i in 1..geometry.frames.len() {
+        let (prev_frames, curr_frames) = geometry.frames.split_at_mut(i);
+        let current = &mut curr_frames[0];
+        let previous = &prev_frames[i-1];
+
+        // TODO: Later maybe add option to move first contour to (0.0, 0.0, 0.0)
+        let translation = (
+            previous.centroid.0 - current.centroid.0,
+            previous.centroid.1 - current.centroid.1,
+            0.0,
+        );
+
+        current.translate_frame(translation);
+
+        let _testing_points = catheter_lumen_vec_from_frames(
+            current, 
+            sample_size, 
+            sample_size_catheter);
+        let _reference_points = catheter_lumen_vec_from_frames(
+            &previous, 
+            sample_size, 
+            sample_size_catheter);
+        
+        let best_rotation = if bruteforce {
+            search_range(
+                &_reference_points, 
+                &_testing_points,
+                step_deg, 
+                range_deg, 
+                &current.centroid, 
+                None, 
+                range_deg)
+        } else {
+            find_best_rotation(
+                &_reference_points, 
+                &_testing_points, 
+                step_deg, 
+                range_deg, 
+                &current.centroid)
+        };
+        current.rotate_frame(best_rotation);
+
+        let new_log = AlignLog {
+            contour_id: current.id,
+            matched_to: previous.id,
+            rot_deg: best_rotation.to_degrees(),
+            tx: translation.0,
+            ty: translation.1,
+            centroid: (current.centroid.0, current.centroid.1),
+        };
+        logger.lock().unwrap().push(new_log);
+    };
     todo!()
 }
 
-fn catheter_lumen_contourvec_from_frames() {
-    todo!()
+fn catheter_lumen_vec_from_frames(
+    frame: &Frame, 
+    sample_size_lumen: usize, 
+    sample_size_catheter: Option<usize>
+) -> Vec<ContourPoint> {
+    let mut lumen_points = downsample_contour_points(&frame.lumen.points, sample_size_lumen);
[…]
