+++ src/intravascular/neo_processing/align_within.rs
@@ -1,9 +1,9 @@
+use anyhow::anyhow;
 use rayon::prelude::*;
-use std::collections::HashMap;
-use std::collections::HashSet;
 use std::f64::consts::PI;
 use std::sync::{Arc, Mutex};
 
+use super::wall::create_wall_frames;
 use crate::intravascular::io::input::ContourPoint;
 use crate::intravascular::io::geometry::{ContourType, Frame, Geometry};
 use crate::intravascular::neo_processing::process_utils::downsample_contour_points;
@@ -25,8 +25,20 @@ pub fn align_frames_in_geometry(
     smooth: bool,
     bruteforce: bool,
     sample_size: usize,    
-) -> (Geometry, Vec<AlignLog>) {
-    let ref_idx = geometry.find_ref_frame_idx().unwrap_or(geometry.find_proximal_end_idx());
+) -> anyhow::Result<(Geometry, Vec<AlignLog>)> {
+    if geometry.frames.is_empty() {
+        return Err(anyhow!("Geometry contains no frames"));
+    }
+    if geometry.frames[0].lumen.points.is_empty() {
+        return Err(anyhow!("Lumen contours have no points"));
+    }
+    if sample_size == 0 {
+        return Err(anyhow!("sample_size must be > 0"));
+    }
+
+    let ref_idx = geometry.
+        find_ref_frame_idx().
+        unwrap_or(geometry.find_proximal_end_idx()) as usize;
     let sample_ratio = sample_size as f64 / geometry.frames[0].lumen.points.len() as f64;
     let sample_size_catheter = if geometry.frames[0].extras.contains_key(&ContourType::Catheter) {
         Some((geometry.frames[0].extras[&ContourType::Catheter].points.len() as f64 * sample_ratio).ceil() as usize)
@@ -88,7 +100,38 @@ pub fn align_frames_in_geometry(
         };
         logger.lock().unwrap().push(new_log);
     };
-    todo!()
+    let anomalous_bool = is_anomalous_coronary(&geometry.frames[ref_idx]);
+    let additional_rotation = angle_ref_point_to_right(
+        &geometry.frames[ref_idx],
+        anomalous_bool)?;
+    for frame in geometry.frames.iter_mut() {
+        frame.rotate_frame(additional_rotation);
+        frame.sort_frame_points();
+    }
+
+    let mut final_geometry = if anomalous_bool {
+        assign_aortic(geometry.clone())
+    } else {
+        geometry.clone()
+    };
+
+    let wall_frames = create_wall_frames(&final_geometry.frames, false);
+    final_geometry = Geometry {
+        frames: wall_frames,
+        label: final_geometry.label,
+    };
+
+    if smooth {
+        final_geometry = final_geometry.smooth_frames();
+    }
+
+    let logs = Arc::try_unwrap(logger)
+        .expect("No other Arc references to logger exist")
+        .into_inner()
+        .expect("Logger mutex was poisoned");
+    dump_table(&logs);
+
+    Ok((final_geometry, logs))
 }
 
 fn catheter_lumen_vec_from_frames(
@@ -270,8 +313,88 @@ fn directed_hausdorff(contour_a: &[ContourPoint], contour_b: &[ContourPoint]) ->
 //     max_sq.sqrt()
 // }
 
+fn is_anomalous_coronary(ref_frame: &Frame) -> bool {
+    // clinical definition is >1.3 but use 2.0 here to avoid false positives
+    ref_frame.lumen.elliptic_ratio() > 2.0
+        || ref_frame.lumen.aortic_thickness.is_some()
+        || ref_frame.lumen.pulmonary_thickness.is_some()
+}
+
[…]
