+++ src/intravascular/neo_processing/mod.rs
@@ -1,4 +1,5 @@
 mod preprocessing;
 pub mod align_within;
 pub mod align_between;
+mod wall;
 mod process_utils;
