+++ src/intravascular/neo_processing/align_within.rs
@@ -4,8 +4,8 @@ use std::f64::consts::PI;
 use std::sync::{Arc, Mutex};
 
 use super::wall::create_wall_frames;
-use crate::intravascular::io::input::ContourPoint;
 use crate::intravascular::io::geometry::{ContourType, Frame, Geometry};
+use crate::intravascular::io::input::ContourPoint;
 use crate::intravascular::neo_processing::process_utils::downsample_contour_points;
 
 #[derive(Debug)]
@@ -24,7 +24,7 @@ pub fn align_frames_in_geometry(
     range_deg: f64,
     smooth: bool,
     bruteforce: bool,
-    sample_size: usize,    
+    sample_size: usize,
 ) -> anyhow::Result<(Geometry, Vec<AlignLog>)> {
     if geometry.frames.is_empty() {
         return Err(anyhow!("Geometry contains no frames"));
@@ -36,12 +36,21 @@ pub fn align_frames_in_geometry(
         return Err(anyhow!("sample_size must be > 0"));
     }
 
-    let ref_idx = geometry.
-        find_ref_frame_idx().
-        unwrap_or(geometry.find_proximal_end_idx()) as usize;
+    let ref_idx = geometry
+        .find_ref_frame_idx()
+        .unwrap_or(geometry.find_proximal_end_idx()) as usize;
     let sample_ratio = sample_size as f64 / geometry.frames[0].lumen.points.len() as f64;
-    let sample_size_catheter = if geometry.frames[0].extras.contains_key(&ContourType::Catheter) {
-        Some((geometry.frames[0].extras[&ContourType::Catheter].points.len() as f64 * sample_ratio).ceil() as usize)
+    let sample_size_catheter = if geometry.frames[0]
+        .extras
+        .contains_key(&ContourType::Catheter)
+    {
+        Some(
+            (geometry.frames[0].extras[&ContourType::Catheter]
+                .points
+                .len() as f64
+                * sample_ratio)
+                .ceil() as usize,
+        )
     } else {
         None
     };
@@ -51,7 +60,7 @@ pub fn align_frames_in_geometry(
     for i in 1..geometry.frames.len() {
         let (prev_frames, curr_frames) = geometry.frames.split_at_mut(i);
         let current = &mut curr_frames[0];
-        let previous = &prev_frames[i-1];
+        let previous = &prev_frames[i - 1];
 
         // TODO: Later maybe add option to move first contour to (0.0, 0.0, 0.0)
         let translation = (
@@ -62,31 +71,29 @@ pub fn align_frames_in_geometry(
 
         current.translate_frame(translation);
 
-        let _testing_points = catheter_lumen_vec_from_frames(
-            current, 
-            sample_size, 
-            sample_size_catheter);
-        let _reference_points = catheter_lumen_vec_from_frames(
-            &previous, 
-            sample_size, 
-            sample_size_catheter);
-        
+        let _testing_points =
+            catheter_lumen_vec_from_frames(current, sample_size, sample_size_catheter);
+        let _reference_points =
+            catheter_lumen_vec_from_frames(&previous, sample_size, sample_size_catheter);
+
         let best_rotation = if bruteforce {
             search_range(
-                &_reference_points, 
+                &_reference_points,
                 &_testing_points,
-                step_deg, 
-                range_deg, 
-                &current.centroid, 
-                None, 
-                range_deg)
+                step_deg,
+                range_deg,
+                &current.centroid,
+                None,
+                range_deg,
+            )
         } else {
             find_best_rotation(
-                &_reference_points, 
-                &_testing_points, 
-                step_deg, 
-                range_deg, 
-                &current.centroid)
+                &_reference_points,
+                &_testing_points,
[…]
