+++ src/intravascular/io/geometry.rs
@@ -745,6 +745,18 @@ impl Geometry {
         self.frames = smoothed_frames;
         self
     }
+
+    pub fn rotate_geometry(mut self, angle_rad: f64) {
+        for frame in self.frames.iter_mut() {
+            frame.rotate_frame(angle_rad);
+        }
+    }
+
+    pub fn translate_geometry(mut self, translation: (f64, f64, f64)) {
+        for frame in self.frames.iter_mut() {
+            frame.translate_frame(translation);
+        }
+    }
 }
 
 #[cfg(test)]
