### 2025-11-02 6d0f6ec
fix inplace modification
