### 2025-11-03 e75567c
make search_range generic, finish simple align_between
