+++ src/intravascular/neo_processing/postprocessing.rs
@@ -1,0 +1,5 @@
+// TODO: trim to same number of contours, by matching on the reference contour
+
+// TODO: Check sample_rate of the two geoms, if different interpolate between frames
+    
+// TODO: equalize spacing between frames to be same in both geoms
