+++ src/intravascular/neo_processing/align_between.rs
@@ -1,10 +1,13 @@
-use anyhow::{Result, Context, anyhow};
+use anyhow::Result;
 use rayon::prelude::*;
-use std::f64::consts::PI;
 
 use crate::intravascular::io::input::ContourPoint;
-use crate::intravascular::io::geometry::{Contour, ContourType, Frame, Geometry};
-use crate::intravascular::neo_processing::process_utils::downsample_contour_points;
+use crate::intravascular::io::geometry::Geometry;
+use crate::intravascular::neo_processing::process_utils::{
+    downsample_contour_points,
+    search_range,
+    hausdorff_distance,
+};
 
 #[derive(Clone, Debug)]
 pub struct GeometryPair {
@@ -30,18 +33,16 @@ pub fn align_between_geometries(
     step_rot_deg: f64,
     sample_size: usize,
 ) -> Result<GeometryPair> {
+    println!("Aligning geometry '{}' to geometry '{}'", geom_b.label, geom_a.label);
     let ref_frame_a_idx = geom_a
         .find_ref_frame_idx()
         .unwrap_or(geom_a.find_proximal_end_idx()) as usize;
     let ref_frame_b_idx = geom_b
         .find_ref_frame_idx()
         .unwrap_or(geom_b.find_proximal_end_idx()) as usize;
-    let ref_frame_a = geom_a.frames[ref_frame_a_idx].clone();
-    let ref_frame_b = geom_b.frames[ref_frame_b_idx].clone();
-    let n_frames_a = geom_a.frames.len();
-    let n_frames_b = geom_b.frames.len();
-    let n_frames = (n_frames_a + n_frames_b) / 2;
-    let internal_sample_size = sample_size * n_frames; 
+    
+    let ref_frame_a = &geom_a.frames[ref_frame_a_idx];
+    let ref_frame_b = &geom_b.frames[ref_frame_b_idx];
 
     let translation = (
         ref_frame_a.centroid.0 - ref_frame_b.centroid.0,
@@ -50,31 +51,120 @@ pub fn align_between_geometries(
     );
     geom_b.translate_geometry(translation);
 
-    let mut test_geom_a: Vec<ContourPoint> = Vec::new();
-    for frame in &geom_a.frames {
-        for point in frame.lumen.points.iter() {
-            test_geom_a.push(point.clone());
-        }
+    let test_geom_a = extract_geometry_points_with_frame_info(geom_a, sample_size);
+    let test_geom_b = extract_geometry_points_with_frame_info(geom_b, sample_size);
+
+    let best_rotation = find_best_rotation_between(
+        &test_geom_a,
+        &test_geom_b,
+        step_rot_deg,
+        rot_deg,
+    );
+
+    for frame in &mut geom_b.frames {
+        frame.rotate_frame(best_rotation);
     }
 
-    let mut test_geom_b: Vec<ContourPoint> = Vec::new();
-    for frame in &geom_b.frames {
-        for point in frame.lumen.points.iter() {
-            test_geom_b.push(point.clone());
-        }
-    }
+    println!(
+        "Best rotation to align '{}' to '{}' is {:.2} degrees",
+        geom_b.label,
+        geom_a.label,
+        best_rotation.to_degrees(),
+    );
 
-    
-    // TODO: trim to same number of contours, by matching on the reference contour
-    
-    // TODO: Check sample_rate of the two geoms, if different interpolate between frames
-    
-    // TODO: equalize spacing between frames to be same in both geoms
-
-    // TODO: create two dummy geometries for finding best rotation, downsample both leave only
-    // lumen contours, find minimal hausdorffdistance over all frames, return best angle
[…]
