### 2025-11-03 bee4c69
finish boilerplate for InputData and add conversion function
