+++ src/intravascular/binding/classes.rs
@@ -1,11 +1,229 @@
-// File: src/python_bind.rs
+use anyhow::{anyhow, Result};
 use super::entry_arr::refine_ordering;
 use crate::intravascular::io::geometry::{Contour, ContourType, Frame, Geometry};
-use crate::intravascular::io::input::{Centerline, CenterlinePoint, ContourPoint, Record};
+use crate::intravascular::io::input::{Centerline, CenterlinePoint, ContourPoint, Record, InputData};
 use crate::intravascular::processing::align_between::GeometryPair;
 use nalgebra::Vector3;
 use pyo3::prelude::*;
 use std::collections::HashMap;
+use std::convert::TryFrom;
+
+/// Python representation of InputData
+/// 
+/// Attributes:
+///    lumen (List[PyContour]): Vessel lumen contours
+///    eem (List[PyContour] | None): Vessel EEM contours
+///    calcification (List[PyContour] | None): Vessel calcification contours
+///    sidebranch (List[PyContor] | None): Vessel sidebranch contours
+///    record (PyRecord): Metadata about the input data
+///    ref_point (PyContourPoint): Reference point for alignment
+///    diastole (bool): Flag indicating if data is diastolic
+///    label (str): label for the input data
+/// Example:
+///     >>> input_data = PyInputData(
+///     ...     lumen=[lumen_contour1, lumen_contour2, ...],
+///     ...     eem=[eem_contour1, eem_contour2, ...],
+///     ...     calcification=[],
+///     ...     sidebranch=[],
+///     ...     record=record,
+///     ...     diastole=True,
+///     ...     lablel="Pat00_diastole_rest"
+///     ... )
+#[pyclass]
+#[derive(Debug, Clone)]
+pub struct PyInputData {
+    #[pyo3(get, set)]
+    pub lumen: Vec<PyContour>,
+    #[pyo3(get, set)]
+    pub eem: Option<Vec<PyContour>>,
+    #[pyo3(get, set)]
+    pub calcification: Option<Vec<PyContour>>,
+    #[pyo3(get, set)]
+    pub sidebranch: Option<Vec<PyContour>>,
+    #[pyo3(get, set)]
+    pub record: Option<Vec<PyRecord>>,
+    #[pyo3(get, set)]
+    pub ref_point: PyContourPoint,
+    #[pyo3(get, set)]
+    pub diastole: bool,
+    #[pyo3(get, set)]
+    pub label: String,
+}
+
+#[pymethods]
+impl PyInputData {
+    #[new]
+    fn new(
+        lumen: Vec<PyContour>,
+        eem: Option<Vec<PyContour>>,
+        calcification: Option<Vec<PyContour>>,
+        sidebranch: Option<Vec<PyContour>>,
+        record: Option<Vec<PyRecord>>,
+        ref_point: PyContourPoint,
+        diastole: bool,
+        label: String,
+    ) -> Self {
+        Self {
+            lumen,
+            eem,
+            calcification,
+            sidebranch,
+            record,
+            ref_point,
+            diastole,
+            label,
+        }
+    }
+
+    fn __repr__(&self) -> String {
+        format!(
+            "InputData(lumen={}, eem={}, calcification={}, sidebranch={}, record={}, ref_point={}, diastole={}, label='{}')",
+            self.lumen.len(),
+            self.eem.as_ref().map_or(0, |v| v.len()),
+            self.calcification.as_ref().map_or(0, |v| v.len()),
+            self.sidebranch.as_ref().map_or(0, |v| v.len()),
+            self.record.as_ref().map_or(0, |v| v.len()),
+            self.ref_point.__repr__(),
+            self.diastole,
+            self.label,
+        )
[…]
