+++ multimodars/_converters.py
@@ -1,6 +1,6 @@
 import numpy as np
 from typing import Union, Tuple
-from multimodars import PyGeometry, PyCenterline
+from multimodars import PyGeometry, PyCenterline, PyInputData
 
 
 def to_array(generic) -> Union[np.ndarray, dict, Tuple[dict, dict]]:
@@ -213,3 +213,184 @@ def numpy_to_centerline(
         )
 
     return PyCenterline.from_contour_points(pts)
+
+
+def array_to_pyinputdata(
+    lumen=None,
+    eem=None,
+    calcification=None,
+    sidebranch=None,
+    records=None,
+    reference=None,
+    diastole: bool = True,
+    label: str = "",
+) -> PyInputData:
+    """
+    Create a PyInputData from either Py* objects (no-op) or NumPy arrays.
+
+    Parameters mirror PyInputData fields. For layer arrays each row must be
+    (frame_index, x, y, z). `records` accepts structured array or list/array
+    of rows (frame, phase, m1, m2) or existing PyRecord instances. `reference`
+    is (1,4) or (4,) row with frame,x,y,z.
+
+    Returns
+    -------
+    PyInputData
+    """
+    from multimodars import (
+        PyContour,
+        PyContourPoint,
+        PyRecord,
+        PyInputData,
+    )
+
+    # Reuse the internal helpers already present in your file.
+    # If you keep the functions _to_numeric_array and build_layer, use them.
+    # Otherwise include the minimal versions below:
+
+    def _to_numeric_array(arr, layer_name: str):
+        if arr is None:
+            return np.zeros((0, 4), dtype=float)
+        if isinstance(arr, (list, tuple)):
+            arr = np.asarray(arr, dtype=object)
+        if isinstance(arr, np.ndarray) and arr.dtype.names:
+            try:
+                arr = np.vstack([arr[name] for name in arr.dtype.names]).T
+            except Exception as e:
+                raise ValueError(f"Could not convert structured array for {layer_name}: {e}")
+        arr = np.asarray(arr)
+        if arr.size == 0:
+            return np.zeros((0, 4), dtype=float)
+        if arr.ndim == 1:
+            if arr.shape[0] == 4:
+                arr = arr[np.newaxis, :]
+            else:
+                raise ValueError(f"{layer_name} 1D array must have length 4, got {arr.shape}")
+        return arr.astype(object)
+
+    def build_layer_from_array(arr, layer_name: str):
+        """Return list[PyContour] from a numeric array (frame,x,y,z) grouped by frame."""
+        arr = _to_numeric_array(arr, layer_name)
+        if arr.size == 0:
+            return []
+
+        if arr.ndim != 2 or arr.shape[1] < 4:
+            raise ValueError(f"{layer_name} must be (N,4)-like, got shape {arr.shape}")
+
+        frames = np.unique(arr[:, 0].astype(int))
+        contours = []
+        for frame in frames:
+            mask = arr[:, 0].astype(int) == frame
+            pts_arr = arr[mask]
+            pts = []
+            for i, row in enumerate(pts_arr):
+                fr = int(row[0])
+                x = float(row[1])
+                y = float(row[2])
+                z = float(row[3])
+                pts.append(PyContourPoint(frame_index=fr, point_index=i, x=x, y=y, z=z, aortic=False))
+            contours.append(PyContour(frame, pts))
+        return contours
+
+    def ensure_contours(maybe):
+        """Accept a list of PyContour already, or numpy arrays, or None."""
+        if maybe is None:
+            return []
+        if isinstance(maybe, list) and maybe and hasattr(maybe[0], "points") and hasattr(maybe[0], "id"):
+            return maybe
[…]
