+++ src/intravascular/neo_processing/postprocessing.rs
@@ -3,3 +3,5 @@
 // TODO: Check sample_rate of the two geoms, if different interpolate between frames
     
 // TODO: equalize spacing between frames to be same in both geoms
+
+// TODO: Adjust walls in geometrypair.
