+++ src/intravascular/binding/mod.rs
@@ -2,6 +2,7 @@ pub mod align;
 pub mod classes;
 pub mod entry_arr;
 pub mod entry_file;
+pub mod entry;
 
 use crate::intravascular::io::{
     input::{Contour, ContourPoint, Record},
