+++ src/intravascular/neo_processing/mod.rs
@@ -1,5 +1,5 @@
 pub mod align_between;
 pub mod align_within;
-mod preprocessing;
+pub mod preprocessing;
 mod process_utils;
 mod wall;
