+++ src/intravascular/neo_processing/mod.rs
@@ -1,5 +1,6 @@
 pub mod align_between;
 pub mod align_within;
 pub mod preprocessing;
+pub mod postprocessing;
 mod process_utils;
 mod wall;
