### 2025-11-03 8f1e122
provide blueprint postprocessing
