+++ src/intravascular/neo_processing/postprocessing.rs
@@ -1,29 +1,57 @@
+use anyhow::Context;
+
 use super::align_between::GeometryPair;
 use crate::intravascular::io::input::ContourPoint;
 use crate::intravascular::io::geometry::{Contour, ContourType, Frame, Geometry};
+use std::collections::HashMap;
 
-// TODO: Check sample_rate of the two geoms, if different interpolate between frames
 pub fn postprocess_geom_pair(geom_pair: &GeometryPair, tol: f64) -> anyhow::Result<GeometryPair> {
     let (same_sample_rate, avg_diff_a, avg_diff_b) = check_same_sample_rate_geompair(geom_pair, tol);
+    let ref_idx_a = geom_pair.geom_a.find_ref_frame_idx()?;
+    let ref_idx_b = geom_pair.geom_b.find_ref_frame_idx()?;
+    let ref_z_a = geom_pair.geom_a.frames[ref_idx_a].centroid.2;
+    let ref_z_b = geom_pair.geom_b.frames[ref_idx_b].centroid.2;
 
     let geom_pair_resampled = if same_sample_rate {
         let mean_diff = (avg_diff_a + avg_diff_b) / 2.0;
         let geom_a = resample_by_diff(&geom_pair.geom_a, mean_diff);
         let geom_b = resample_by_diff(&geom_pair.geom_b, mean_diff);
         GeometryPair {
-            geom_a: geom_a,
-            geom_b: geom_b,
+            geom_a,
+            geom_b,
+            label: geom_pair.label.clone(),
+        }
+    } else if avg_diff_a < avg_diff_b {
+        let n = geom_pair.geom_a.frames.len();
+        let end_zero = geom_pair.geom_a.frames[0].centroid.2;
+        let end_n = geom_pair.geom_a.frames[n - 1].centroid.2;
+        let (start, stop) = if end_zero < end_n { (end_zero, end_n) } else { (end_n, end_zero) };
+
+        let z_coords = predict_z_positions(ref_z_a, start, stop, avg_diff_b);
+        let new_geom = new_frames_by_sample_rate(&geom_pair.geom_a, z_coords);
+        let _new_resampled = resample_by_diff(&new_geom, avg_diff_b);
+        GeometryPair {
+            geom_a: new_geom,
+            geom_b: geom_pair.geom_b.clone(),
             label: geom_pair.label.clone(),
         }
     } else {
-        // TODO:
-        // if avg_diff_a < avg_diff_b {
-        //     new_frames_by_sample_rate(geom_a)
-        //     resample_by_diff(geom_a)
-            geom_pair.clone()
-        };
-        
-    Ok(geom_pair_resampled)
+        let n = geom_pair.geom_b.frames.len();
+        let end_zero = geom_pair.geom_b.frames[0].centroid.2;
+        let end_n = geom_pair.geom_b.frames[n - 1].centroid.2;
+        let (start, stop) = if end_zero < end_n { (end_zero, end_n) } else { (end_n, end_zero) };
+
+        let z_coords = predict_z_positions(ref_z_b, start, stop, avg_diff_b);
+        let new_geom = new_frames_by_sample_rate(&geom_pair.geom_b, z_coords);
+        let _new_resampled = resample_by_diff(&new_geom, avg_diff_b);
+        GeometryPair {
+            geom_a: geom_pair.geom_a.clone(),
+            geom_b: new_geom,
+            label: geom_pair.label.clone(),
+        }
+    };
+    let trimmed_geom_pair = trim_geom_pair(&geom_pair_resampled);
+    Ok(trimmed_geom_pair)
 }
 
[…]
