+++ src/intravascular/binding/entry.rs
@@ -1,11 +1,13 @@
+use crate::intravascular::io::geometry::Geometry;
+use crate::intravascular::io::input::InputData;
+use crate::intravascular::neo_processing::align_between::{align_between_geometries, GeometryPair};
+use crate::intravascular::neo_processing::align_within::{align_frames_in_geometry, AlignLog};
+use crate::intravascular::neo_processing::preprocessing::{
+    prepare_n_geometries, ProcessingOptions,
+};
 use anyhow::{anyhow, Context, Result};
 use crossbeam::thread;
 use rand::seq::index::sample;
-use crate::intravascular::io::input::InputData;
-use crate::intravascular::io::geometry::Geometry;
-use crate::intravascular::neo_processing::preprocessing::{ProcessingOptions, prepare_n_geometries};
-use crate::intravascular::neo_processing::align_within::{AlignLog, align_frames_in_geometry};
-use crate::intravascular::neo_processing::align_between::{GeometryPair, align_between_geometries};
 use std::path::Path;
 
 pub fn full_processing_rs(
@@ -33,17 +35,23 @@ pub fn full_processing_rs(
         image_center,
         radius,
         n_points,
-        Some(vec![input_data_a, input_data_b, input_data_c, input_data_d]
-            .into_iter()
-            .flatten()
-            .collect()),
+        Some(
+            vec![input_data_a, input_data_b, input_data_c, input_data_d]
+                .into_iter()
+                .flatten()
+                .collect(),
+        ),
         input_path_a.map(Path::new),
         input_path_b.map(Path::new),
         ProcessingOptions::Full,
-    ).context("Failed to prepare geometries for full processing")?;
+    )
+    .context("Failed to prepare geometries for full processing")?;
 
     if geometries.len() != 4 {
-        return Err(anyhow!("Full processing requires exactly 4 geometries, got {}", geometries.len()));
+        return Err(anyhow!(
+            "Full processing requires exactly 4 geometries, got {}",
+            geometries.len()
+        ));
     }
 
     let mut geom_a = geometries.remove(0);
@@ -51,132 +59,135 @@ pub fn full_processing_rs(
     let mut geom_c = geometries.remove(0);
     let mut geom_d = geometries.remove(0);
 
-    let (mut geom_a, mut geom_b, mut geom_c, mut geom_d, logs_a, logs_b, logs_c, logs_d) = thread::scope(|s| -> Result<(
-        Geometry,
-        Geometry,
-        Geometry,
-        Geometry,
-        Vec<AlignLog>,
-        Vec<AlignLog>,
-        Vec<AlignLog>,
-        Vec<AlignLog>,
-    )> {
-        let geom_a_handle = s.spawn(|_| -> anyhow::Result<_> {
-            let (geom, logs) = align_frames_in_geometry(
-                &mut geom_a,
-                step_deg,
-                range_deg,
-                smooth,
-                bruteforce,
-                sample_size,
-            )
-            .context("Failed to align frames within geometry A")?;
-            Ok((geom, logs))
-        });
+    let (mut geom_a, mut geom_b, mut geom_c, mut geom_d, logs_a, logs_b, logs_c, logs_d) =
+        thread::scope(
+            |s| -> Result<(
+                Geometry,
+                Geometry,
+                Geometry,
+                Geometry,
+                Vec<AlignLog>,
+                Vec<AlignLog>,
+                Vec<AlignLog>,
+                Vec<AlignLog>,
+            )> {
+                let geom_a_handle = s.spawn(|_| -> anyhow::Result<_> {
+                    let (geom, logs) = align_frames_in_geometry(
+                        &mut geom_a,
+                        step_deg,
+                        range_deg,
+                        smooth,
+                        bruteforce,
[…]
