+++ src/intravascular/processing/align_within.rs
@@ -1,403 +1,170 @@
+use anyhow::anyhow;
 use rayon::prelude::*;
-use std::collections::HashMap;
-use std::collections::HashSet;
-use std::f64::consts::PI;
 use std::sync::{Arc, Mutex};
 
-use crate::intravascular::io::geometry::{Contour, Frame, Geometry};
+use super::wall::create_wall_frames;
+use crate::intravascular::io::geometry::{Contour, ContourType, Frame, Geometry};
 use crate::intravascular::io::input::ContourPoint;
-use crate::intravascular::processing::align_between::GeometryPair;
-use crate::intravascular::processing::process_utils::downsample_contour_points;
+use crate::intravascular::processing::process_utils::{
+    downsample_contour_points, hausdorff_distance, search_range,
+};
 
 #[derive(Debug)]
 pub struct AlignLog {
     pub contour_id: u32,
     pub matched_to: u32,
-    pub rel_rot_deg: f64,
-    pub total_rot_deg: f64,
+    pub rot_deg: f64,
     pub tx: f64,
     pub ty: f64,
     pub centroid: (f64, f64),
 }
 
-pub fn align_frames_in_geometries(
-    geom_pair: GeometryPair,
-    step_deg: f64,
-    range_deg: f64,
-    smooth: bool,
-    bruteforce: bool,
-    sample_size: usize,
-) -> anyhow::Result<(GeometryPair, (Vec<AlignLog>, Vec<AlignLog>))> {
-    let (diastole, dia_logs) = align_frames_in_geometry(
-        geom_pair.dia_geom,
-        step_deg,
-        range_deg,
-        smooth,
-        bruteforce,
-        sample_size,
-    );
-    let (mut systole, sys_logs) = align_frames_in_geometry(
-        geom_pair.sys_geom,
-        step_deg,
-        range_deg,
-        smooth,
-        bruteforce,
-        sample_size,
-    );
-
-    GeometryPair::translate_contours_to_match(&diastole, &mut systole);
-    GeometryPair::apply_z_transformation(&diastole, &mut systole);
-    let geom_pair_clean = GeometryPair {
-        dia_geom: diastole,
-        sys_geom: systole,
-    };
-
-    Ok((geom_pair_clean, (dia_logs, sys_logs)))
-}
-
 pub fn align_frames_in_geometry(
-    geometry: Geometry,
+    geometry: &mut Geometry,
     step_deg: f64,
     range_deg: f64,
     smooth: bool,
     bruteforce: bool,
     sample_size: usize,
-) -> (Geometry, Vec<AlignLog>) {
-    let (mut geometry, reference_index, reference_pos, ref_contour) = prep_data_geometry(geometry);
-
-    let (p1, p2, updated_ref) = assign_aortic(ref_contour.clone(), &geometry);
-    let ref_contour = updated_ref.clone();
-
-    let (_line_angle, rotation_to_y, rotated_ref, aortic_flag) =
-        rotate_reference_contour(p1, p2, ref_contour.clone());
-
-    // Update the reference contour in geometry with rotated version
-    geometry.contours.insert(reference_pos, rotated_ref.clone());
-
-    // prepare reference catheter
-    for catheter in geometry.catheter.iter_mut() {
-        if catheter.id == reference_index {
-            catheter.rotate_contour_around_point(
-                rotation_to_y,
-                (ref_contour.centroid.0, ref_contour.centroid.1),
-            );
-            catheter.sort_contour_points();
-            break;
-        }
+) -> anyhow::Result<(Geometry, Vec<AlignLog>)> {
+    if geometry.frames.is_empty() {
+        return Err(anyhow!("Geometry contains no frames"));
     }
[…]
