+++ src/intravascular/mod.rs
@@ -4,5 +4,3 @@ pub mod io;
 pub mod processing;
 pub mod texture;
 pub mod utils;
-
-pub mod neo_processing;
