+++ src/intravascular/binding/entry.rs
@@ -1,14 +1,32 @@
 use crate::intravascular::io::geometry::Geometry;
 use crate::intravascular::io::input::InputData;
-use crate::intravascular::processing::align_between::{align_between_geometries, GeometryPair};
-use crate::intravascular::processing::align_within::{align_frames_in_geometry, AlignLog};
 use crate::intravascular::processing::preprocessing::{
     prepare_n_geometries, ProcessingOptions,
 };
+use crate::intravascular::processing::align_within::{align_frames_in_geometry, AlignLog};
+use crate::intravascular::processing::align_between::{align_between_geometries, GeometryPair};
+use crate::intravascular::processing::postprocessing::postprocess_geom_pair;
 use anyhow::{anyhow, Context, Result};
 use crossbeam::thread;
 use rand::seq::index::sample;
 use std::path::Path;
+
+// tolerance of distance between frames [mm], that counts as 0
+const TOLERANCE: f64 = 0.03;
+
+fn maybe_postprocess(
+    pair: &GeometryPair,
+    tolerance: f64,
+    anomalous: bool,
+    postprocessing: bool,
+) -> anyhow::Result<GeometryPair> {
+    if postprocessing {
+        postprocess_geom_pair(pair, tolerance, anomalous)
+            .with_context(|| format!("Failed postprocessing of {}", pair.label))
+    } else {
+        Ok(pair.clone())
+    }
+}
 
 pub fn full_processing_rs(
     label: String,
@@ -23,12 +41,16 @@ pub fn full_processing_rs(
     input_data_d: Option<InputData>,
     write_obj: bool,
     watertight: bool,
-    output_path: &str,
+    output_path_a: &str,
+    output_file_b: &str,
+    output_file_c: &str,
+    output_file_d: &str,
     step_deg: f64,
     range_deg: f64,
     smooth: bool,
     bruteforce: bool,
     sample_size: usize,
+    postprocessing: bool,
 ) -> Result<()> {
     let mut geometries = prepare_n_geometries(
         &label,
@@ -59,7 +81,7 @@ pub fn full_processing_rs(
     let mut geom_c = geometries.remove(0);
     let mut geom_d = geometries.remove(0);
 
-    let (mut geom_a, mut geom_b, mut geom_c, mut geom_d, logs_a, logs_b, logs_c, logs_d) =
+    let (mut geom_a, mut geom_b, mut geom_c, mut geom_d, logs_a, logs_b, logs_c, logs_d, bool_a, bool_b, bool_c, bool_d) =
         thread::scope(
             |s| -> Result<(
                 Geometry,
@@ -70,9 +92,13 @@ pub fn full_processing_rs(
                 Vec<AlignLog>,
                 Vec<AlignLog>,
                 Vec<AlignLog>,
+                bool,
+                bool,
+                bool,
+                bool,
             )> {
                 let geom_a_handle = s.spawn(|_| -> anyhow::Result<_> {
-                    let (geom, logs) = align_frames_in_geometry(
+                    let (geom, logs, anomalous_bool) = align_frames_in_geometry(
                         &mut geom_a,
                         step_deg,
                         range_deg,
@@ -81,11 +107,11 @@ pub fn full_processing_rs(
                         sample_size,
                     )
                     .context("Failed to align frames within geometry A")?;
-                    Ok((geom, logs))
+                    Ok((geom, logs, anomalous_bool))
                 });
 
                 let geom_b_handle = s.spawn(|_| -> anyhow::Result<_> {
-                    let (geom, logs) = align_frames_in_geometry(
+                    let (geom, logs, anomalous_bool) = align_frames_in_geometry(
                         &mut geom_b,
                         step_deg,
                         range_deg,
[…]
