+++ src/intravascular/io/mod.rs
@@ -144,7 +144,7 @@ pub fn build_geometry_from_inputdata(
         frame.sort_frame_points();
     }
 
-    todo!("Ensure that proximal frame has id 0");
+    todo!("Ensure that proximal frame has id 0 and lowest z");
     check_geometry_integrity(&geometry)?;
 
     Ok(geometry)

[…]
