### 2025-11-08 024013c
finish interpolations.rs
