+++ src/intravascular/mod.rs
@@ -2,5 +2,5 @@ pub mod binding;
 pub mod centerline_align;
 pub mod io;
 pub mod processing;
-pub mod obj_preparation;
+pub mod to_object;
 pub mod utils;
