+++ src/intravascular/to_object/interpolation.rs
@@ -1,0 +1,545 @@
+use anyhow::bail;
+
+use crate::intravascular::io::input::{ContourPoint};
+use crate::intravascular::io::geometry::{Contour, ContourType, Frame, Geometry};
+use std::collections::HashMap;
+
+/// Interpolates between two aligned Geometry configurations with number of steps
+/// used to visualize deformation over a cardiac cycle.
+pub fn interpolate_contours(
+    start: &Geometry,
+    end: &Geometry,
+    steps: usize,
+    contour_types: &[ContourType],
+) -> anyhow::Result<Vec<Geometry>> {
+    use std::cmp::min;
+
+    // Find matching number of frames
+    let n_frames = min(start.frames.len(), end.frames.len());
+
+    let mut geoms = Vec::with_capacity(steps + 2);
+    geoms.push(start.clone());
+
+    for step in 0..steps {
+        let t = step as f64 / (steps - 1) as f64;
+
+        let mut frames = Vec::with_capacity(n_frames);
+
+        for i in 0..n_frames {
+            let start_frame = &start.frames[i];
+            let end_frame = &end.frames[i];
+
+            // Interpolate lumen contour
+            let lumen = interpolate_contour(&start_frame.lumen, &end_frame.lumen, t)?;
+
+            // Interpolate extras based on requested contour types
+            let mut extras = HashMap::new();
+            for contour_type in contour_types {
+                match contour_type {
+                    ContourType::Lumen => {}, // Already handled above
+                    _ => {
+                        if let (Some(start_contour), Some(end_contour)) = (
+                            start_frame.extras.get(contour_type),
+                            end_frame.extras.get(contour_type),
+                        ) {
+                            let interp_contour = interpolate_contour(start_contour, end_contour, t)?;
+                            extras.insert(*contour_type, interp_contour);
+                        }
+                    }
+                }
+            }
+
+            // Interpolate reference point
+            let reference_point = match (&start_frame.reference_point, &end_frame.reference_point) {
+                (Some(start_rp), Some(end_rp)) => Some(interpolate_contour_point(start_rp, end_rp, t)),
+                (Some(start_rp), None) => Some(*start_rp),
+                (None, Some(end_rp)) => Some(*end_rp),
+                (None, None) => None,
+            };
+
+            // Interpolate centroid
+            let centroid = (
+                start_frame.centroid.0 * (1.0 - t) + end_frame.centroid.0 * t,
+                start_frame.centroid.1 * (1.0 - t) + end_frame.centroid.1 * t,
+                start_frame.centroid.2 * (1.0 - t) + end_frame.centroid.2 * t,
+            );
+
+            frames.push(Frame {
+                id: start_frame.id,
+                centroid,
+                lumen,
+                extras,
+                reference_point,
+            });
+        }
+
+        geoms.push(Geometry {
+            frames,
+            label: format!("{}_inter_{}", start.label, step),
+        });
+    }
+
+    geoms.push(end.clone());
+    Ok(geoms)
+}
+
+/// Interpolate between two contours
+fn interpolate_contour(start: &Contour, end: &Contour, t: f64) -> anyhow::Result<Contour> {
+    if start.points.len() != end.points.len() {
+        bail!("Contour point counts do not match between start and end");
+    }
+
+    let points = start.points.iter()
+        .zip(&end.points)
+        .map(|(ps, pe)| interpolate_contour_point(ps, pe, t))
+        .collect();
+
+    let centroid = match (start.centroid, end.centroid) {
[…]
