### 2025-11-08 4ed710f
finish writing of GeometryPair
