+++ src/intravascular/to_object/write_mtl.rs
@@ -1,181 +1,317 @@
 use crate::intravascular::io::geometry::{Contour, ContourType, Geometry};
 use ::std::fs::File;
 use ::std::io::Write;
+use std::collections::HashMap;
 use std::path::Path;
 use super::texture::{
     compute_displacements, compute_uv_coordinates, create_black_texture,
     create_displacement_texture, create_transparent_texture,
 };
 
-/// Creates a UV-map based on displacement between geometries and saves the map
-/// in png format. Additionally creates a .mtl file with material parameter
-/// 1 1 1 1 for the moment.
+/// Creates UV-maps and textures for different contour types based on their requirements:
+/// - Lumen and EEM: displacement texture
+/// - Catheter and Calcification: black texture  
+/// - Wall and Sidebranch: transparent texture
 pub fn write_mtl_geometry(
     geometries_to_process: &Vec<Geometry>,
     output_dir: &str,
     case_name: &str,
-) -> (
-    Vec<Vec<(f64, f64)>>,
-    Vec<Vec<(f64, f64)>>,
-    Vec<Vec<(f64, f64)>>,
-) {
-    let reference_contours = geometries_to_process[0].clone();
+    contour_types: &[ContourType],
+) -> HashMap<ContourType, Vec<Vec<(f64, f64)>>> {
+    let mut uv_coords_map = HashMap::new();
 
-    // calculate max displacements, since all meshes in between are interpolated
-    // meshes, it is enough to compare first and last element of Vec<Geometry> and
-    // then extract the max value needed to normalize.
-    let disp = compute_displacements(
-        &geometries_to_process[0],
-        geometries_to_process.last().unwrap(),
-    );
-    let max_disp = disp.iter().cloned().reduce(f64::max).unwrap();
+    for &contour_type in contour_types {
+        let uv_coords = write_mtl_for_contour_type(
+            geometries_to_process,
+            output_dir,
+            case_name,
+            contour_type,
+        );
+        uv_coords_map.insert(contour_type, uv_coords);
+    }
 
-    let mut uv_coords_contours = Vec::new();
+    uv_coords_map
+}
 
-    for (i, mesh) in geometries_to_process.into_iter().enumerate() {
-        let mut mesh_contours = Vec::new();
-        for frame in mesh.frames.clone() {
-            mesh_contours.push(frame.lumen)
+/// Writes MTL and texture files for a specific contour type
+fn write_mtl_for_contour_type(
+    geometries_to_process: &Vec<Geometry>,
+    output_dir: &str,
+    case_name: &str,
+    contour_type: ContourType,
+) -> Vec<Vec<(f64, f64)>> {
+    match contour_type {
+        ContourType::Lumen | ContourType::Eem => {
+            // For lumen and EEM: displacement texture
+            write_displacement_texture(
+                geometries_to_process,
+                output_dir,
+                case_name,
+                contour_type,
+            )
         }
-        let uv_coords = compute_uv_coordinates(&mesh_contours);
+        ContourType::Catheter | ContourType::Calcification => {
+            // For catheter and calcification: black texture
+            write_black_texture(
+                geometries_to_process,
+                output_dir,
+                case_name,
+                contour_type,
+            )
+        }
+        ContourType::Wall | ContourType::Sidebranch => {
+            // For wall and sidebranch: transparent texture
+            write_transparent_texture(
+                geometries_to_process,
+                output_dir,
+                case_name,
+                contour_type,
+            )
+        }
+    }
+}
 
-        let texture_height = mesh_contours.len() as u32;
+/// Writes displacement texture for contour types that need displacement mapping
+fn write_displacement_texture(
+    geometries_to_process: &Vec<Geometry>,
[…]
