+++ src/intravascular/binding/mod.rs
@@ -3,22 +3,22 @@ pub mod classes;
 pub mod entry;
 
 use crate::intravascular::io::{
-    input::{InputData, ContourPoint, Record},
+    input::InputData,
     output::write_obj_mesh_without_uv,
 };
 use crate::intravascular::processing::align_within::AlignLog;
-use classes::{PyContour, PyFrame, PyGeometry, PyGeometryPair, PyRecord, PyInputData};
+use classes::{PyContourType, PyGeometry, PyGeometryPair, PyInputData};
 use entry::*;
 use pyo3::prelude::*;
+use std::path::Path;
 
-fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64, f64)> {
+fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64)> {
     logs.into_iter()
         .map(|l| {
             (
                 l.contour_id,
                 l.matched_to,
-                l.rel_rot_deg,
-                l.total_rot_deg,
+                l.rot_deg,
                 l.tx,
                 l.ty,
                 l.centroid.0,
@@ -28,17 +28,16 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
         .collect()
 }
 
-/// Processes four geometries in parallel.
-///
-/// Pipeline:
+/// Processes four geometries in parallel from folders.
 ///
 /// .. code-block:: text
 ///
-///    Rest:                    Stress:
-///    diastole  -------------->  diastole
-///       |                         |
-///       v                         v
-///    systole  -------------->  systole
+///    Rest           Stress
+///    diastole ──▶ diastole
+///       │            │
+///       ▼            ▼
+///    systole ──▶ systole
+///
 ///
 /// Args:
 ///     rest_input_path: Path to REST input file
@@ -82,6 +81,8 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
     signature = (
         rest_input_path,
         stress_input_path,
+        label = "full",
+        _diastole = true,
         step_rotation_deg = 0.5f64,
         range_rotation_deg = 90.0f64,
         image_center = (4.5f64, 4.5f64),
@@ -96,11 +97,16 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
         interpolation_steps = 28usize,
         bruteforce = false,
         sample_size = 500,
+        contour_types = vec![PyContourType::Lumen, PyContourType::Catheter, PyContourType::Wall],
+        smooth = true,
+        postprocessing = true,
     )
 )]
 pub fn from_file_full(
     rest_input_path: &str,
     stress_input_path: &str,
+    label: &str,
+    _diastole: bool,
     step_rotation_deg: f64,
     range_rotation_deg: f64,
     image_center: (f64, f64),
@@ -115,55 +121,75 @@ pub fn from_file_full(
     interpolation_steps: usize,
     bruteforce: bool,
     sample_size: usize,
+    contour_types: Vec<PyContourType>,
+    smooth: bool,
+    postprocessing: bool,
 ) -> PyResult<(
     PyGeometryPair,
     PyGeometryPair,
     PyGeometryPair,
     PyGeometryPair,
     (
-        Vec<(u32, u32, f64, f64, f64, f64, f64, f64)>,
-        Vec<(u32, u32, f64, f64, f64, f64, f64, f64)>,
[…]
