+++ src/intravascular/binding/entry.rs
@@ -1,9 +1,12 @@
 use anyhow::{anyhow, Context, Result};
 use crossbeam::thread;
 use std::path::Path;
+use std::fs::File;
+use std::io::Write;
 
-use crate::intravascular::io::geometry::{ContourType, Geometry};
+use crate::intravascular::io::geometry::{Contour, ContourType, Geometry};
 use crate::intravascular::io::input::InputData;
+use crate::intravascular::io::output::write_obj_mesh_without_uv;
 use crate::intravascular::processing::preprocessing::{
     prepare_n_geometries, ProcessingOptions,
 };
@@ -40,6 +43,7 @@ pub fn full_processing_rs(
     input_data_b: Option<InputData>,
     input_data_c: Option<InputData>,
     input_data_d: Option<InputData>,
+    _diastole: bool,
     write_obj: bool,
     interpolation_steps: usize,
     contour_types: Vec<ContourType>,
@@ -66,6 +70,7 @@ pub fn full_processing_rs(
                 .flatten()
                 .collect(),
         ),
+        _diastole,
         input_path_a.map(Path::new),
         input_path_b.map(Path::new),
         ProcessingOptions::Full,
@@ -264,6 +269,7 @@ pub fn double_pair_processing_rs(
     input_data_b: Option<InputData>,
     input_data_c: Option<InputData>,
     input_data_d: Option<InputData>,
+    _diastole: bool,
     write_obj: bool,
     interpolation_steps: usize,
     contour_types: Vec<ContourType>,
@@ -288,6 +294,7 @@ pub fn double_pair_processing_rs(
                 .flatten()
                 .collect(),
         ),
+        _diastole,
         input_path_a.map(Path::new),
         input_path_b.map(Path::new),
         ProcessingOptions::Full,
@@ -441,6 +448,7 @@ pub fn pair_processing_rs(
     input_path_a: Option<&str>,
     input_data_a: Option<InputData>,
     input_data_b: Option<InputData>,
+    _diastole: bool,
     write_obj: bool,
     interpolation_steps: usize,
     contour_types: Vec<ContourType>,
@@ -464,6 +472,7 @@ pub fn pair_processing_rs(
                 .flatten()
                 .collect(),
         ),
+        _diastole,
         input_path_a.map(Path::new),
         None,
         ProcessingOptions::Pair,
@@ -544,7 +553,9 @@ pub fn single_processing_rs(
     n_points: u32,
     input_path_a: Option<&str>,
     input_data_a: Option<InputData>,
+    _diastole: bool,
     write_obj: bool,
+    watertight: bool,
     contour_types: Vec<ContourType>,
     output_path: &str,
     step_deg: f64,
@@ -552,13 +563,14 @@ pub fn single_processing_rs(
     smooth: bool,
     bruteforce: bool,
     sample_size: usize,
-) -> Result<()> {
+) -> Result<(Geometry, Vec<AlignLog>)> {
     let mut geom = prepare_n_geometries(
         &label,
         image_center,
         radius,
         n_points,
         Some(vec![input_data_a].into_iter().flatten().collect()),
+        _diastole,
         input_path_a.map(Path::new),
         None,
         ProcessingOptions::Single,
@@ -584,7 +596,102 @@ pub fn single_processing_rs(
 
     // TODO: postprocessing geometry (resample)
 
-    // TODO: writing to obj
+    if write_obj {
+        // Create output directory
+        std::fs::create_dir_all(output_path)?;
+
+        // Write each contour type to OBJ
+        for contour_type in &contour_types {
+            let contours = extract_contours_by_type(&geom, *contour_type);
+            
+            if contours.is_empty() {
+                eprintln!("Warning: No contours found for type {:?}, skipping", contour_type);
+                continue;
+            }
+
[…]
