+++ src/intravascular/old_processing/align_between.rs
@@ -1,834 +1,0 @@
-use anyhow::Context;
-use rayon::prelude::*;
-use std::f64::consts::PI;
-
-use crate::intravascular::io::input::{Contour, ContourPoint};
-use crate::intravascular::io::Geometry;
-use crate::intravascular::old_processing::align_within::hausdorff_distance;
-use crate::intravascular::old_processing::process_utils::downsample_contour_points;
-
-#[derive(Clone, Debug)]
-pub struct GeometryPair {
-    pub dia_geom: Geometry,
-    pub sys_geom: Geometry,
-}
-
-pub fn get_geometry_pair(
-    case_name: String,
-    input_dir: &str,
-    image_center: (f64, f64),
-    radius: f64,
-    n_points: u32,
-) -> anyhow::Result<GeometryPair> {
-    let mut geometries =
-        GeometryPair::new(input_dir, case_name.clone(), image_center, radius, n_points)
-            .context(format!("Creating GeometryPair({}) failed", case_name))?;
-    geometries = geometries.adjust_z_coordinates();
-    geometries = geometries.trim_geometries_same_length();
-
-    Ok(geometries)
-}
-
-impl GeometryPair {
-    pub fn new(
-        input_dir: &str,
-        label: String,
-        image_center: (f64, f64),
-        radius: f64,
-        n_points: u32,
-    ) -> anyhow::Result<GeometryPair> {
-        let dia_geom = Geometry::new(
-            input_dir,
-            label.clone(),
-            true,
-            image_center,
-            radius,
-            n_points,
-        )?;
-        println!("geometry pair: diastolic geometry generated");
-        let sys_geom = Geometry::new(input_dir, label, false, image_center, radius, n_points)?;
-        println!("geometry pair: systolic geometry generated");
-        Ok(GeometryPair { dia_geom, sys_geom })
-    }
-
-    /// Translates systolic contours to the diastolic contours, aligns z-coordinates and
-    /// finds the best rotation and trims them to same length.    
-    pub fn align_between_geometries(
-        mut self,
-        steps_rotation_deg: f64,
-        range_rotation_deg: f64,
-    ) -> anyhow::Result<GeometryPair> {
-        Self::translate_contours_to_match(&self.dia_geom, &mut self.sys_geom);
-
-        // Adjust the z-coordinates of systolic contours. (later replaceed by adjust_z_coordinates)
-        Self::apply_z_transformation(&self.dia_geom, &mut self.sys_geom);
-
-        let best_rotation_angle = find_best_rotation_all(
-            &self.dia_geom,
-            &self.sys_geom,
-            steps_rotation_deg, // number of candidate steps (e.g. 200 or 400)
-            range_rotation_deg, // rotation range (e.g. 1.05 for ~±60°)
-        );
-
-        for ref mut contour in self
-            .sys_geom
-            .contours
-            .iter_mut()
-            .chain(self.sys_geom.catheter.iter_mut())
-        {
-            contour.rotate_contour(best_rotation_angle);
-        }
-        Ok(GeometryPair {
-            dia_geom: self.dia_geom,
-            sys_geom: self.sys_geom,
-        })
-    }
-
-    pub fn translate_contours_to_match(dia: &Geometry, sys: &mut Geometry) {
-        let dia_ref = dia.contours.last().unwrap().centroid;
-        let sys_ref = sys.contours.last().unwrap().centroid;
-        let offset = (dia_ref.0 - sys_ref.0, dia_ref.1 - sys_ref.1);
-
-        for contour in &mut sys.contours {
-            contour.translate_contour((offset.0, offset.1, 0.0));
-        }
-
-        for catheter in &mut sys.catheter {
[…]
