+++ src/intravascular/centerline_align/align_algorithms.rs
@@ -1,4 +1,4 @@
-use crate::intravascular::io::geometry::{Contour, Geometry};
+use crate::intravascular::io::geometry::{ContourType, Contour, Geometry};
 use crate::intravascular::io::input::ContourPoint;
 use crate::intravascular::io::input::{Centerline, CenterlinePoint};
 use nalgebra::{Point3, Rotation3, Unit, Vector3};
@@ -34,46 +34,55 @@ impl FrameTransformation {
     }
 }
 
-pub fn get_transformations(mesh: Geometry, centerline: &Centerline) -> Vec<FrameTransformation> {
+pub fn get_transformations(geometry: Geometry, centerline: &Centerline) -> Vec<FrameTransformation> {
     let mut transformations = Vec::new();
 
-    for mut contour in mesh.contours.into_iter() {
-        if let Some(cl_point) = centerline.get_by_frame(contour.id) {
-            let transformation = align_frame(&mut contour, cl_point);
-            transformations.push(transformation)
+    for frame in geometry.frames.into_iter() {
+        if let Some(cl_point) = centerline.get_by_frame(frame.id) {
+            // Align the lumen contour
+            let transformation = align_frame(&frame.lumen, cl_point);
+            transformations.push(transformation);
+            
+            // Optionally align extra contours if needed
+            for (contour_type, contour) in frame.extras {
+                if matches!(contour_type, ContourType::Eem | ContourType::Wall) {
+                    let extra_transformation = align_frame(&contour, cl_point);
+                    transformations.push(extra_transformation);
+                }
+            }
         }
     }
     transformations
 }
 
-/// Modified align_frame returns the transformation applied.
-fn align_frame(frame: &mut Contour, cl_point: &CenterlinePoint) -> FrameTransformation {
-    if frame.id != cl_point.contour_point.frame_index {
+/// Modified align_frame returns the transformation applied (doesn't mutate input)
+fn align_frame(frame: &Contour, cl_point: &CenterlinePoint) -> FrameTransformation {
+    if frame.original_frame != cl_point.contour_point.frame_index {
         panic!(
             "Frame Index {} does not match Centerline Point Frame Index {}",
-            frame.id, cl_point.contour_point.frame_index
+            frame.original_frame, cl_point.contour_point.frame_index
         );
     }
+
+    // Get centroid or compute if None
+    let centroid = frame.centroid.unwrap_or_else(|| {
+        let x_avg = frame.points.iter().map(|p| p.x).sum::<f64>() / frame.points.len() as f64;
+        let y_avg = frame.points.iter().map(|p| p.y).sum::<f64>() / frame.points.len() as f64;
+        let z_avg = frame.points.iter().map(|p| p.z).sum::<f64>() / frame.points.len() as f64;
+        (x_avg, y_avg, z_avg)
+    });
 
     // === Translation Step ===
     // Compute the translation vector to bring the frame's centroid to the centerline point.
     let translation_vec = Vector3::new(
-        cl_point.contour_point.x - frame.centroid.0,
-        cl_point.contour_point.y - frame.centroid.1,
-        cl_point.contour_point.z - frame.centroid.2,
+        cl_point.contour_point.x - centroid.0,
+        cl_point.contour_point.y - centroid.1,
+        cl_point.contour_point.z - centroid.2,
     );
-    for point in frame.points.iter_mut() {
-        point.x += translation_vec.x;
-        point.y += translation_vec.y;
-        point.z += translation_vec.z;
-    }
-    frame.centroid.0 += translation_vec.x;
-    frame.centroid.1 += translation_vec.y;
-    frame.centroid.2 += translation_vec.z;
 
     // === Rotation Step ===
[…]
