+++ src/intravascular/binding/align.rs
@@ -1,7 +1,6 @@
 use pyo3::prelude::*;
-
 use crate::intravascular::{
-    binding::classes::{PyCenterline, PyGeometryPair},
+    binding::classes::{PyContourType, PyCenterline, PyGeometryPair},
     centerline_align::{align_manual_rs, align_three_point_rs},
 };
 
@@ -47,6 +46,7 @@ use crate::intravascular::{
         watertight=true,
         interpolation_steps=28usize,
         output_dir="output/aligned",
+        contour_types=vec![PyContourType::Lumen, PyContourType::Catheter, PyContourType::Wall],
         case_name="None",
     )
 )]
@@ -61,8 +61,10 @@ pub fn align_three_point(
     watertight: bool,
     interpolation_steps: usize,
     output_dir: &str,
+    contour_types: Vec<PyContourType>,
     case_name: &str,
 ) -> PyResult<(PyGeometryPair, PyCenterline)> {
+    let rust_contour_types: Vec<crate::intravascular::io::geometry::ContourType> = contour_types.iter().map(|ct| ct.into()).collect();
     let cl_rs = centerline.to_rust_centerline();
     let geom_pair_rs = geometry_pair.to_rust_geometry_pair();
     let angle_step = angle_step_deg.to_radians();
@@ -78,8 +80,9 @@ pub fn align_three_point(
         watertight,
         interpolation_steps,
         output_dir,
+        rust_contour_types,
         case_name,
-    );
+    ).map_err(|e| pyo3::exceptions::PyRuntimeError::new_err(e.to_string()))?;
 
     let aligned_py: PyGeometryPair = geom_pair.into();
     let resampled_cl: PyCenterline = cl.into();
@@ -120,11 +123,12 @@ pub fn align_three_point(
         centerline,
         geometry_pair,
         rotation_angle,
-        start_point,
+        ref_point,
         write=false,
         watertight=true,
         interpolation_steps=28usize,
         output_dir="output/aligned",
+        contour_types=vec![PyContourType::Lumen, PyContourType::Catheter, PyContourType::Wall],
         case_name="None",
     )
 )]
@@ -132,13 +136,15 @@ pub fn align_manual(
     centerline: PyCenterline,
     geometry_pair: PyGeometryPair,
     rotation_angle: f64,
-    start_point: usize,
+    ref_point: (f64, f64, f64),
     write: bool,
     watertight: bool,
     interpolation_steps: usize,
     output_dir: &str,
+    contour_types: Vec<PyContourType>,
     case_name: &str,
 ) -> PyResult<(PyGeometryPair, PyCenterline)> {
+    let rust_contour_types: Vec<crate::intravascular::io::geometry::ContourType> = contour_types.iter().map(|ct| ct.into()).collect();
     let cl_rs = centerline.to_rust_centerline();
     let geom_pair_rs = geometry_pair.to_rust_geometry_pair();
 
@@ -146,13 +152,14 @@ pub fn align_manual(
         cl_rs,
         geom_pair_rs,
         rotation_angle,
-        start_point,
+        ref_point,
         write,
         watertight,
         interpolation_steps,
         output_dir,
+        rust_contour_types,
         case_name,
-    );
+    ).map_err(|e| pyo3::exceptions::PyRuntimeError::new_err(e.to_string()))?;
 
     let aligned_py: PyGeometryPair = geom_pair.into();
     let resampled_cl: PyCenterline = cl.into();
