### 2025-11-10 66c1706
finish tests for io module
