+++ src/intravascular/io/mod.rs
@@ -6,7 +6,7 @@ pub mod output;
 use geometry::{Contour, ContourType, Frame, Geometry};
 use input::{ContourPoint, InputData};
 use integrity_check::check_geometry_integrity;
-use std::collections::HashMap;
+use std::collections::{HashSet, HashMap};
 use std::path::Path;
 
 pub fn build_geometry_from_inputdata(
@@ -35,27 +35,72 @@ pub fn build_geometry_from_inputdata(
             "Either input_data or path must be provided"
         ));
     };
+    let print_input_data = input_data.clone();
 
-    let lumen_contours = Contour::build_contour(
+    // First, collect all unique original frame IDs from ALL contour types
+    let mut all_original_frames: HashSet<u32> = HashSet::new();
+
+    // Add lumen frames
+    for point in &input_data.lumen {
+        all_original_frames.insert(point.frame_index);
+    }
+
+    // Add eem frames if present
+    if let Some(eem_points) = &input_data.eem {
+        for point in eem_points {
+            all_original_frames.insert(point.frame_index);
+        }
+    }
+
+    // Add calcification frames if present
+    if let Some(calc_points) = &input_data.calcification {
+        for point in calc_points {
+            all_original_frames.insert(point.frame_index);
+        }
+    }
+
+    // Add sidebranch frames if present
+    if let Some(side_points) = &input_data.sidebranch {
+        for point in side_points {
+            all_original_frames.insert(point.frame_index);
+        }
+    }
+
+    // Add reference point frame
+    all_original_frames.insert(input_data.ref_point.frame_index);
+
+    // Create sorted list of original frames and mapping to sequential IDs
+    let mut sorted_original_frames: Vec<u32> = all_original_frames.into_iter().collect();
+    sorted_original_frames.sort();
+
+    let frame_mapping: HashMap<u32, u32> = sorted_original_frames
+        .iter()
+        .enumerate()
+        .map(|(i, &original_id)| (original_id, i as u32))
+        .collect();
+
+    // Now build contours using the shared mapping
+    let lumen_contours = Contour::build_contour_with_mapping(
         input_data.lumen,
         input_data.record.clone(),
         ContourType::Lumen,
+        &frame_mapping,
     )?;
 
     let eem_contours = if let Some(eem_points) = input_data.eem {
-        Contour::build_contour(eem_points, None, ContourType::Eem)?
+        Contour::build_contour_with_mapping(eem_points, None, ContourType::Eem, &frame_mapping)?
     } else {
         Vec::new()
     };
 
     let calcification_contours = if let Some(calc_points) = input_data.calcification {
-        Contour::build_contour(calc_points, None, ContourType::Calcification)?
+        Contour::build_contour_with_mapping(calc_points, None, ContourType::Calcification, &frame_mapping)?
     } else {
         Vec::new()
     };
 
     let sidebranch_contours = if let Some(side_points) = input_data.sidebranch {
-        Contour::build_contour(side_points, None, ContourType::Sidebranch)?
+        Contour::build_contour_with_mapping(side_points, None, ContourType::Sidebranch, &frame_mapping)?
     } else {
         Vec::new()
     };
@@ -63,7 +108,7 @@ pub fn build_geometry_from_inputdata(
     let mut frame_map: HashMap<u32, Frame> = HashMap::new();
 
     // Process lumen contours (mandatory)
-    for contour in lumen_contours {
+    for mut contour in lumen_contours {
         contour.compute_centroid();
         let frame_id = contour.id;
         let mut frame = Frame {
@@ -74,35 +119,35 @@ pub fn build_geometry_from_inputdata(
             reference_point: None,
         };
 
-        if input_data.ref_point.frame_index == frame_id {
-            frame.reference_point = Some(input_data.ref_point);
[…]
