+++ src/intravascular/io/input.rs
@@ -18,6 +18,7 @@ use std::path::Path;
 ///     frame_idx, x, y, z
 ///
 /// not automatically kept in sync with geometry.rs
+#[derive(Debug, Clone)]
 pub struct InputData {
     pub lumen: Vec<ContourPoint>,
     pub eem: Option<Vec<ContourPoint>>,
@@ -69,39 +70,41 @@ impl InputData {
 
         let phase = if diastole { "diastolic" } else { "systolic" };
 
-        // iterate over provided names and attempt to build/read the corresponding files
+        // Read required files - these will crash if missing
+        let contours_fname = format!("{}_contours.csv", phase);
+        let contours_path = path.join(&contours_fname);
+        if contours_path.exists() {
+            lumen = Some(
+                ContourPoint::read_contour_data(&contours_path)
+                    .with_context(|| format!("reading {}", contours_path.display()))?,
+            );
+        } else {
+            return Err(anyhow::anyhow!(
+                "required contours file missing: {:?}",
+                contours_path
+            ));
+        }
+
+        let ref_fname = format!("{}_reference_points.csv", phase);
+        let ref_path = path.join(&ref_fname);
+        if ref_path.exists() {
+            ref_point = Some(
+                ContourPoint::read_reference_point(&ref_path)
+                    .with_context(|| format!("reading {}", ref_path.display()))?,
+            );
+        } else {
+            return Err(anyhow::anyhow!(
+                "required reference-point file missing: {:?}",
+                ref_path
+            ));
+        }
+
+        // Process optional files from names mapping - skip if missing
         for (_ctype, raw_name) in names.iter() {
             let name = raw_name.trim().to_lowercase();
             match name.as_str() {
                 "" | "lumen" => {
-                    // main contour files: "<phase>_contours.csv" and "<phase>_reference_points.csv"
-                    let contours_fname = format!("{}_contours.csv", phase);
-                    let contours_path = path.join(&contours_fname);
-                    if contours_path.exists() {
-                        lumen = Some(
-                            ContourPoint::read_contour_data(&contours_path)
-                                .with_context(|| format!("reading {}", contours_path.display()))?,
-                        );
-                    } else {
-                        return Err(anyhow::anyhow!(
-                            "expected lumen contours file missing: {:?}",
-                            contours_path
-                        ));
-                    }
-
-                    let ref_fname = format!("{}_reference_points.csv", phase);
-                    let ref_path = path.join(&ref_fname);
-                    if ref_path.exists() {
-                        ref_point = Some(
-                            ContourPoint::read_reference_point(&ref_path)
-                                .with_context(|| format!("reading {}", ref_path.display()))?,
-                        );
-                    } else {
-                        return Err(anyhow::anyhow!(
-                            "expected reference-point file missing: {:?}",
-                            ref_path
-                        ));
-                    }
+                    // Already handled above, skip
                 }
 
                 "branch" | "sidebranch" => {
@@ -113,7 +116,7 @@ impl InputData {
                                 .with_context(|| format!("reading {}", p.display()))?,
                         );
                     } else {
-                        return Err(anyhow::anyhow!("expected file missing: {:?}", p));
+                        eprintln!("sidebranch file not found, skipping: {:?}", p);
                     }
                 }
 
@@ -126,7 +129,7 @@ impl InputData {
                                 .with_context(|| format!("reading {}", p.display()))?,
                         );
[…]
