+++ Cargo.toml
@@ -20,6 +20,4 @@ rayon = "1.7.1"
 crossbeam = "0.8.4"
 anyhow = "1.0.97"
 approx = "0.5"
-serde_json = "1.0"
-thiserror = "2.0"
 stl_io = "0.8.5"

[…]
