### 2025-11-10 89c3167
finish io module, fix reading record data
