+++ src/intravascular/io/mod.rs
@@ -40,36 +40,30 @@ pub fn build_geometry_from_inputdata(
     // First, collect all unique original frame IDs from ALL contour types
     let mut all_original_frames: HashSet<u32> = HashSet::new();
 
-    // Add lumen frames
     for point in &input_data.lumen {
         all_original_frames.insert(point.frame_index);
     }
 
-    // Add eem frames if present
     if let Some(eem_points) = &input_data.eem {
         for point in eem_points {
             all_original_frames.insert(point.frame_index);
         }
     }
 
-    // Add calcification frames if present
     if let Some(calc_points) = &input_data.calcification {
         for point in calc_points {
             all_original_frames.insert(point.frame_index);
         }
     }
 
-    // Add sidebranch frames if present
     if let Some(side_points) = &input_data.sidebranch {
         for point in side_points {
             all_original_frames.insert(point.frame_index);
         }
     }
 
-    // Add reference point frame
     all_original_frames.insert(input_data.ref_point.frame_index);
 
-    // Create sorted list of original frames and mapping to sequential IDs
     let mut sorted_original_frames: Vec<u32> = all_original_frames.into_iter().collect();
     sorted_original_frames.sort();
 
@@ -107,7 +101,6 @@ pub fn build_geometry_from_inputdata(
 
     let mut frame_map: HashMap<u32, Frame> = HashMap::new();
 
-    // Process lumen contours (mandatory)
     for mut contour in lumen_contours {
         contour.compute_centroid();
         let frame_id = contour.id;
@@ -119,7 +112,6 @@ pub fn build_geometry_from_inputdata(
             reference_point: None,
         };
 
-        // Map reference point using the same mapping
         if let Some(&mapped_frame_id) = frame_mapping.get(&input_data.ref_point.frame_index) {
             if mapped_frame_id == frame_id {
                 frame.reference_point = Some(input_data.ref_point);
@@ -129,7 +121,6 @@ pub fn build_geometry_from_inputdata(
         frame_map.insert(frame_id, frame);
     }
 
-    // Add other contour types to frames using the same mapping
     for mut contour in eem_contours {
         contour.compute_centroid();
         if let Some(frame) = frame_map.get_mut(&contour.id) {
@@ -151,7 +142,6 @@ pub fn build_geometry_from_inputdata(
         }
     }
 
-    // Create catheter contours if requested
     if n_points > 0 {
         let all_points: Vec<ContourPoint> = frame_map
             .values()
@@ -161,7 +151,6 @@ pub fn build_geometry_from_inputdata(
         let catheter_points =
             Frame::create_catheter_points(&all_points, image_center, radius, n_points);
         
-        // Use the same mapping for catheter contours
         let catheter_contours =
             Contour::build_contour_with_mapping(catheter_points, None, ContourType::Catheter, &frame_mapping)?;
 
@@ -173,7 +162,6 @@ pub fn build_geometry_from_inputdata(
         }
     }
 
-    // Convert frame map to sorted vector
     let mut frames: Vec<Frame> = frame_map.into_values().collect();
     frames.sort_by_key(|f| f.id);
 
@@ -262,7 +250,6 @@ mod input_tests {
         let mut orig_map: HashMap<ContourType, Vec<u32>> = HashMap::new();
 
         for (idx, frame) in geometry.frames.iter().enumerate() {
-            // Lumen
             ids_map
                 .entry(ContourType::Lumen)
                 .or_default()
@@ -272,7 +259,6 @@ mod input_tests {
                 .or_default()
                 .push(frame.lumen.original_frame);
 
[…]
