+++ Cargo.toml
@@ -13,7 +13,6 @@ pyo3 = "0.25.0"
 csv = "1.1"
 serde = { version = "1.0.219", features = ["derive"]}
 image = "0.25.5"
-rand = "0.9.2"
 nalgebra = "0.34.0"
 toml = "0.9.5"
 rayon = "1.7.1"
