### 2025-11-11 422831e
create dummy geometry to check alignment
