+++ src/intravascular/utils/test_utils.rs
@@ -1,51 +1,204 @@
-use crate::intravascular::io::geometry::{Contour, ContourType};
+use crate::intravascular::io::geometry::{Contour, ContourType, Frame, Geometry};
 use crate::intravascular::io::input::ContourPoint;
+use std::collections::HashMap;
 use std::f64::consts::PI;
 
-/// Generates ellipse contour points for testing
-pub fn generate_ellipse_points(
-    major: f64,
-    minor: f64,
-    num_points: usize,
-    rotation: f64,
-    translation: (f64, f64),
-    frame_idx: u32,
-) -> Vec<ContourPoint> {
-    let mut points = Vec::with_capacity(num_points);
-    for i in 0..num_points {
-        let theta = 2.0 * PI * (i as f64) / (num_points as f64);
-        let x = major * theta.cos();
-        let y = minor * theta.sin();
-        let (x_rot, y_rot) = rotate_point((x, y), rotation);
-        points.push(ContourPoint {
-            frame_index: frame_idx,
-            point_index: i as u32,
-            x: x_rot + translation.0,
-            y: y_rot + translation.1,
-            z: 0.0,
-            aortic: false,
-        });
-    }
-    points
-}
+pub fn dummy_geometry() -> Geometry {
+    let points_a = vec![
+            ContourPoint {
+                frame_index: 1,
+                point_index: 0,
+                x: 0.0,
+                y: 0.0,
+                z: 0.0,
+                aortic: false,
+            },
+            ContourPoint {
+                frame_index: 1,
+                point_index: 1,
+                x: 0.0,
+                y: 2.0,
+                z: 0.0,
+                aortic: false,
+            },
+            ContourPoint {
+                frame_index: 1,
+                point_index: 2,
+                x: 3.0,
+                y: 1.0,
+                z: 0.0,
+                aortic: false,
+            },
+            ContourPoint {
+                frame_index: 1,
+                point_index: 3,
+                x: 2.0,
+                y: 2.0,
+                z: 0.0,
+                aortic: false,
+            },
+            ContourPoint {
+                frame_index: 1,
+                point_index: 4,
+                x: 2.0,
+                y: 0.0,
+                z: 0.0,
+                aortic: false,
+            },           
+    ];
+    let points_b = vec![
+            ContourPoint {
+                frame_index: 2,
+                point_index: 0,
+                x: 0.0,
+                y: 0.0,
+                z: 1.0,
+                aortic: false,
+            },
+            ContourPoint {
+                frame_index: 2,
+                point_index: 1,
+                x: 0.0,
+                y: 2.0,
+                z: 1.0,
+                aortic: false,
+            },
+            ContourPoint {
+                frame_index: 2,
+                point_index: 2,
+                x: 3.0,
+                y: 1.0,
+                z: 1.0,
+                aortic: false,
+            },
+            ContourPoint {
+                frame_index: 2,
+                point_index: 3,
+                x: 2.0,
+                y: 2.0,
+                z: 1.0,
+                aortic: false,
+            },
+            ContourPoint {
+                frame_index: 2,
+                point_index: 4,
+                x: 2.0,
+                y: 0.0,
+                z: 1.0,
+                aortic: false,
+            },           
+    ];
+    let points_c = vec![
+            ContourPoint {
+                frame_index: 3,
+                point_index: 0,
+                x: 0.0,
+                y: 0.0,
+                z: 2.0,
+                aortic: false,
+            },
+            ContourPoint {
+                frame_index: 3,
[…]
