### 2025-11-11 cf5580f
start testing dummy_geometry
