### 2025-11-11 1b03950
major bug fix in alignment algorithm, add tests
