+++ src/intravascular/processing/process_utils.rs
@@ -27,49 +27,44 @@ where
 {
     let range_rad = range_deg.to_radians();
     let step_rad = step_deg.to_radians();
+    
+    // Handle edge case: zero or negative step
     if step_rad <= 0.0 {
-        return 0.0;
+        return center_angle.unwrap_or(0.0);
     }
 
     let center = center_angle.unwrap_or(0.0);
     let limes = limes_deg.to_radians();
-    let lower_limes = -limes;
 
-    let mut start_angle = center - range_rad;
-    let stop_angle = center + range_rad;
-    start_angle = start_angle.max(lower_limes);
-    let stop_angle = stop_angle.min(limes);
+    let start_angle = (center - range_rad).max(-limes);
+    let stop_angle = (center + range_rad).min(limes);
 
     if stop_angle <= start_angle {
-        return ((center + PI).rem_euclid(2.0 * PI)) - PI;
+        return center;
     }
 
     let steps = (((stop_angle - start_angle) / step_rad).ceil() as usize).max(1);
+    
+    let mut best_angle = center;
+    let mut best_cost = f64::INFINITY;
 
-    let mut angle_cost_pairs = Vec::with_capacity(steps);
     for i in 0..=steps {
-        let angle_lin = start_angle + (i as f64) * step_rad;
-        if angle_lin > stop_angle {
+        let angle = start_angle + (i as f64) * step_rad;
+        if angle > stop_angle {
             break;
         }
 
-        let angle = angle_lin.rem_euclid(2.0 * PI);
+        // Map angle to [-π, π] range
         let mapped_angle = ((angle + PI).rem_euclid(2.0 * PI)) - PI;
+        let cost = cost_fn(mapped_angle);
 
-        let cost = cost_fn(angle);
-        angle_cost_pairs.push((mapped_angle, cost));
+        if cost < best_cost {
+            best_cost = cost;
+            best_angle = mapped_angle;
+        }
     }
 
-    if angle_cost_pairs.is_empty() {
-        return ((center + PI).rem_euclid(2.0 * PI)) - PI;
-    }
-
-    let (min_angle, _min_cost) = angle_cost_pairs
-        .iter()
-        .min_by(|a, b| a.1.partial_cmp(&b.1).unwrap_or(std::cmp::Ordering::Equal))
-        .unwrap();
-
-    *min_angle
+    best_angle
 }
 
 /// Computes the Hausdorff distance between two point sets.
@@ -139,3 +134,404 @@ fn directed_hausdorff(contour_a: &[ContourPoint], contour_b: &[ContourPoint]) ->
 }
 
 // TODO: Move the interpolation to process_utils
+
+#[cfg(test)]
+mod process_utils_tests {
+    use super::*;
+    use crate::intravascular::utils::test_utils::dummy_geometry;
+    use approx::assert_relative_eq;
+
+    #[test]
+    fn test_downsample_geometry() {
+        let dummy = dummy_geometry();
+        let cont_points = dummy.frames[0].lumen.points.clone();
+        let downsampled_contour = downsample_contour_points(&cont_points, 3);
+
+        assert_eq!(downsampled_contour.len(), 3);
+        assert_eq!(downsampled_contour[0].point_index, 0);
+        assert_eq!(downsampled_contour[1].point_index, 2);
+
+        let downsampled_contour = downsample_contour_points(&cont_points, 6);
+
+        assert_eq!(downsampled_contour.len(), 6);
+        assert_eq!(downsampled_contour[0].point_index, 0);
+        assert_eq!(downsampled_contour[1].point_index, 1);
+
+        let downsampled_contour = downsample_contour_points(&cont_points, 5);
+        let n = downsampled_contour.len();
[…]
