+++ src/intravascular/processing/align_within.rs
@@ -58,48 +58,43 @@ pub fn align_frames_in_geometry(
 
     let logger = Arc::new(Mutex::new(Vec::<AlignLog>::new()));
     
-    let mut cumulative_rotation = 0.0;
+    let mut cumulative_rotation: f64 = 0.0;
 
-    // Helper function to wrap rotation to [-PI, PI]
-    fn wrap_rotation(angle: f64) -> f64 {
-        use std::f64::consts::PI;
-        let two_pi = 2.0 * PI;
-        let mut wrapped = angle % two_pi;
-        if wrapped > PI {
-            wrapped -= two_pi;
-        } else if wrapped < -PI {
-            wrapped += two_pi;
-        }
-        wrapped
-    }
-
+    // Use sequential alignment (frame to previous frame)
     for i in 1..geometry.frames.len() {
-        let (prev_frames, curr_frames) = geometry.frames.split_at_mut(i);
-        let current = &mut curr_frames[0];
-        let previous = &prev_frames[i - 1];
+        let prev_frame = geometry.frames[i - 1].clone(); // Clone to avoid borrowing issues
+        let current = &mut geometry.frames[i];
 
+        println!("Aligning Frame {} to previous Frame {}", current.id, prev_frame.id);
+
+        // 1) Apply cumulative rotation accumulated from earlier frames
+        if cumulative_rotation != 0.0 {
+            // rotate around current centroid (preserves the frame's own local center)
+            current.rotate_frame(cumulative_rotation);
+        }
+
+        // 2) Translate current frame to match previous frame centroid
         let translation = (
-            previous.centroid.0 - current.centroid.0,
-            previous.centroid.1 - current.centroid.1,
+            prev_frame.centroid.0 - current.centroid.0,
+            prev_frame.centroid.1 - current.centroid.1,
             0.0,
         );
-
         current.translate_frame(translation);
 
-        let _testing_points =
-            catheter_lumen_vec_from_frames(current, sample_size, sample_size_catheter);
-        let _reference_points =
-            catheter_lumen_vec_from_frames(&previous, sample_size, sample_size_catheter);
+        // 3) Prepare points for alignment (after rotation+translation)
+        let testing_points = catheter_lumen_vec_from_frames(current, sample_size, sample_size_catheter);
+        let reference_points = catheter_lumen_vec_from_frames(&prev_frame, sample_size, sample_size_catheter);
 
-        current.rotate_frame(cumulative_rotation);
+        // 4) Find best rotation — IMPORTANT: rotate testing_points around current.centroid
+        // (not prev_frame.centroid). That matches the old behaviour.
         let best_rotation = if bruteforce {
             search_range(
                 |angle: f64| {
-                    let rotated: Vec<ContourPoint> = _testing_points
+                    let rotated: Vec<ContourPoint> = testing_points
                         .par_iter()
                         .map(|p| p.rotate_point(angle, (current.centroid.0, current.centroid.1)))
                         .collect();
-                    hausdorff_distance(&_reference_points, &rotated)
+                    hausdorff_distance(&reference_points, &rotated)
                 },
                 step_deg,
                 range_deg,
@@ -108,19 +103,22 @@ pub fn align_frames_in_geometry(
             )
         } else {
             find_best_rotation(
-                &_reference_points,
-                &_testing_points,
+                &reference_points,
+                &testing_points,
                 step_deg,
                 range_deg,
-                &current.centroid,
[…]

[…]
