### 2025-11-12 8a028b7
fix translation bug, add test for alignment and one hole fix
