+++ src/intravascular/processing/align_within.rs
@@ -62,18 +62,15 @@ pub fn align_frames_in_geometry(
 
     // Use sequential alignment (frame to previous frame)
     for i in 1..geometry.frames.len() {
-        let prev_frame = geometry.frames[i - 1].clone(); // Clone to avoid borrowing issues
+        let prev_frame = geometry.frames[i - 1].clone();
         let current = &mut geometry.frames[i];
 
         println!("Aligning Frame {} to previous Frame {}", current.id, prev_frame.id);
 
-        // 1) Apply cumulative rotation accumulated from earlier frames
         if cumulative_rotation != 0.0 {
-            // rotate around current centroid (preserves the frame's own local center)
             current.rotate_frame(cumulative_rotation);
         }
 
-        // 2) Translate current frame to match previous frame centroid
         let translation = (
             prev_frame.centroid.0 - current.centroid.0,
             prev_frame.centroid.1 - current.centroid.1,
@@ -81,12 +78,9 @@ pub fn align_frames_in_geometry(
         );
         current.translate_frame(translation);
 
-        // 3) Prepare points for alignment (after rotation+translation)
         let testing_points = catheter_lumen_vec_from_frames(current, sample_size, sample_size_catheter);
         let reference_points = catheter_lumen_vec_from_frames(&prev_frame, sample_size, sample_size_catheter);
 
-        // 4) Find best rotation — IMPORTANT: rotate testing_points around current.centroid
-        // (not prev_frame.centroid). That matches the old behaviour.
         let best_rotation = if bruteforce {
             search_range(
                 |angle: f64| {
@@ -107,15 +101,13 @@ pub fn align_frames_in_geometry(
                 &testing_points,
                 step_deg,
                 range_deg,
-                &current.centroid, // <-- use current frame centroid
+                &current.centroid,
             )
         };
 
-        // 5) Apply best rotation to current frame (around its centroid) and accumulate
         current.rotate_frame(best_rotation);
         cumulative_rotation += best_rotation;
 
-        // 6) Log
         let new_log = AlignLog {
             contour_id: current.id,
             matched_to: prev_frame.id,
@@ -132,6 +124,7 @@ pub fn align_frames_in_geometry(
 
     let anomalous_bool = is_anomalous_coronary(&geometry.frames[ref_idx]);
     let additional_rotation = angle_ref_point_to_right(&geometry.frames[ref_idx], anomalous_bool)?;
+
     for frame in geometry.frames.iter_mut() {
         frame.rotate_frame(additional_rotation);
         frame.sort_frame_points();
@@ -338,7 +331,7 @@ fn detect_holes(geometry: &Geometry) -> (bool, f64) {
     }
     let avg_diff: f64 = z_diffs.iter().sum::<f64>() / z_diffs.len() as f64;
     for &diff in &z_diffs {
-        if diff > 1.5 * avg_diff {
+        if diff >= 1.5 * avg_diff {
             return (true, avg_diff);
         }
     }
@@ -358,7 +351,7 @@ pub fn fill_holes(geometry: &mut Geometry) -> anyhow::Result<Geometry> {
         return Ok(geometry.clone());
     }
 
-    println!("⚠️ Hole detected! Attempting to fix using Geometry::insert_frame(...)");
+    println!("⚠️\tHole detected! Attempting to fix using Geometry::insert_frame(...)");
 
     // Walk through frames; insert when we see a gap
     let mut i: usize = 1;
[…]
