### 2025-11-12 3963e35
finish tests align_within and geometry
