+++ src/intravascular/io/geometry.rs
@@ -922,6 +922,7 @@ mod geometry_tests {
     use super::*;
     use std::f64::consts::PI;
     use crate::intravascular::utils::test_utils::dummy_geometry;
+    use approx::assert_relative_eq;
 
     // test for contours
     #[test]
@@ -2014,5 +2015,339 @@ mod geometry_tests {
         assert!(geom.frames[0].reference_point.is_some());
         assert_eq!(geom.frames[0].reference_point.as_ref().unwrap().z, 0.0);
         assert_eq!(geom.frames[0].reference_point.as_ref().unwrap().frame_index,678);
+    }
+
+    #[test]
+    fn test_smoothing_effect() -> anyhow::Result<()> {
+        use crate::intravascular::io::geometry::{Contour, ContourType, Frame, Geometry};
+        use std::collections::HashMap;
+
+        // Create 3 frames with square contours of different sizes
+        let mut frames = Vec::new();
+        
+        // Frame 0: Square with edge length 2, centered at (0,0)
+        let frame0_points = vec![
+            ContourPoint { frame_index: 0, point_index: 0, x: -1.0, y: -1.0, z: 0.0, aortic: false },
+            ContourPoint { frame_index: 0, point_index: 1, x: 1.0, y: -1.0, z: 0.0, aortic: false },
+            ContourPoint { frame_index: 0, point_index: 2, x: 1.0, y: 1.0, z: 0.0, aortic: false },
+            ContourPoint { frame_index: 0, point_index: 3, x: -1.0, y: 1.0, z: 0.0, aortic: false },
+        ];
+        
+        let mut lumen0 = Contour {
+            id: 0,
+            original_frame: 0,
+            points: frame0_points,
+            centroid: None,
+            aortic_thickness: None,
+            pulmonary_thickness: None,
+            kind: ContourType::Lumen,
+        };
+        lumen0.compute_centroid();
+        
+        let frame0 = Frame {
+            id: 0,
+            centroid: lumen0.centroid.unwrap(),
+            lumen: lumen0,
+            extras: HashMap::new(),
+            reference_point: None,
+        };
+        
+        // Frame 1: Square with edge length 1, centered at (0,0)  
+        let frame1_points = vec![
+            ContourPoint { frame_index: 1, point_index: 0, x: -0.5, y: -0.5, z: 1.0, aortic: false },
+            ContourPoint { frame_index: 1, point_index: 1, x: 0.5, y: -0.5, z: 1.0, aortic: false },
+            ContourPoint { frame_index: 1, point_index: 2, x: 0.5, y: 0.5, z: 1.0, aortic: false },
+            ContourPoint { frame_index: 1, point_index: 3, x: -0.5, y: 0.5, z: 1.0, aortic: false },
+        ];
+        
+        let mut lumen1 = Contour {
+            id: 1,
+            original_frame: 1,
+            points: frame1_points,
+            centroid: None,
+            aortic_thickness: None,
+            pulmonary_thickness: None,
+            kind: ContourType::Lumen,
+        };
+        lumen1.compute_centroid();
+        
+        let frame1 = Frame {
+            id: 1,
+            centroid: lumen1.centroid.unwrap(),
+            lumen: lumen1,
+            extras: HashMap::new(),
+            reference_point: None,
+        };
+        
+        // Frame 2: Square with edge length 2, centered at (0,0) - same as frame 0
+        let frame2_points = vec![
+            ContourPoint { frame_index: 2, point_index: 0, x: -1.0, y: -1.0, z: 2.0, aortic: false },
+            ContourPoint { frame_index: 2, point_index: 1, x: 1.0, y: -1.0, z: 2.0, aortic: false },
+            ContourPoint { frame_index: 2, point_index: 2, x: 1.0, y: 1.0, z: 2.0, aortic: false },
[…]
