### 2025-11-12 757a149
add postprocessing tests
