+++ src/intravascular/processing/postprocessing.rs
@@ -80,6 +80,9 @@ fn check_same_sample_rate_geompair(geom_pair: &GeometryPair, tol: f64) -> (bool,
 }
 
 fn get_avg_z_diff(geometry: &Geometry) -> f64 {
+    if geometry.frames.len() < 2 {
+        return 0.0;
+    }
     let mut diffs_geom = Vec::new();
     for i in 1..geometry.frames.len() {
         let curr = &geometry.frames[i];
@@ -436,5 +439,443 @@ fn adjust_walls_anomalous_geom_pair(geom_pair: &GeometryPair) -> GeometryPair {
             label: geom_pair.geom_b.label.clone(),
         },
         label: geom_pair.label.clone(),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::intravascular::io::geometry::{Contour, ContourType, Frame, Geometry};
+    use std::collections::HashMap;
+
+    // Helper function to create test contours
+    fn create_test_contour(id: u32, z: f64, thickness: Option<f64>, kind: ContourType) -> Contour {
+        let points = vec![
+            ContourPoint {
+                x: 1.0,
+                y: 2.0,
+                z,
+                frame_index: id,
+                point_index: 0,
+                aortic: false,
+            },
+            ContourPoint {
+                x: 3.0,
+                y: 4.0,
+                z,
+                frame_index: id,
+                point_index: 1,
+                aortic: false,
+            },
+        ];
+        
+        Contour {
+            id,
+            original_frame: id,
+            points,
+            centroid: Some((2.0, 3.0, z)),
+            aortic_thickness: thickness,
+            pulmonary_thickness: None,
+            kind,
+        }
+    }
+
+    // Helper function to create test frames with reference points
+    fn create_test_frame(id: u32, z: f64, lumen_thickness: Option<f64>, set_ref: bool) -> Frame {
+        let lumen = create_test_contour(id, z, lumen_thickness, ContourType::Lumen);
+        let eem = create_test_contour(id, z, None, ContourType::Eem);
+        
+        let mut extras = HashMap::new();
+        extras.insert(ContourType::Eem, eem);
+        
+        let reference_point = if set_ref {
+            Some(ContourPoint {
+                x: 0.0, y: 0.0, z, frame_index: id, point_index: 0, aortic: false
+            })
+        } else {
+            None
+        };
+        
+        Frame {
+            id,
+            centroid: (2.0, 3.0, z),
+            lumen,
+            extras,
+            reference_point,
+        }
+    }
+
+    // Helper function to create test geometry with reference frame
+    fn create_test_geometry(label: &str, z_values: Vec<f64>, thicknesses: Vec<Option<f64>>) -> Geometry {
+        let mut frames: Vec<Frame> = z_values.iter().enumerate().map(|(i, &z)| {
+            // Set reference point on the middle frame
+            let set_ref = i == z_values.len() / 2;
+            create_test_frame(i as u32, z, *thicknesses.get(i).unwrap_or(&None), set_ref)
+        }).collect();
+        
+        // Ensure at least one frame has a reference point
+        if frames.iter().all(|f| f.reference_point.is_none()) && !frames.is_empty() {
+            frames[0].reference_point = Some(ContourPoint {
+                x: 0.0, y: 0.0, z: frames[0].centroid.2, frame_index: 0, point_index: 0, aortic: false
+            });
+        }
+        
+        Geometry {
+            frames,
+            label: label.to_string(),
+        }
+    }
+
[…]
