### 2025-11-13 0b3dd0f
fix wrappers and converters for from_file and from_array variants
