+++ src/lib.rs
@@ -39,8 +39,10 @@ fn multimodars(_py: Python, m: pyo3::prelude::Bound<'_, PyModule>) -> PyResult<(
     m.add_function(wrap_pyfunction!(to_obj, m.clone())?)?;
 
     // Updated class registration
+    m.add_class::<PyInputData>()?;
     m.add_class::<PyContourPoint>()?;
     m.add_class::<PyContour>()?;
+    m.add_class::<PyContourType>()?;
     m.add_class::<PyGeometry>()?;
     m.add_class::<PyGeometryPair>()?;
     m.add_class::<PyCenterlinePoint>()?;
