+++ multimodars/_wrappers.py
@@ -1,24 +1,4 @@
 from typing import Literal, Any, Dict, Tuple, Union
-from multimodars import (
-    from_file_full,
-    from_file_doublepair,
-    from_file_singlepair,
-    from_file_single,
-)
-
-from multimodars import (
-    geometry_from_array,
-    from_array_full,
-    from_array_doublepair,
-    from_array_singlepair,
-)
-
-from multimodars import (
-    align_three_point,
-    align_manual,
-)
-
-from .multimodars import PyGeometry, PyGeometryPair, PyCenterline
 
 Mode = Literal["full", "doublepair", "singlepair", "single"]
 
@@ -29,10 +9,10 @@ def from_file(
 ) -> Union[
     # full returns 4 geometry pairs + 4 log-lists
     Tuple[
-        PyGeometryPair,
-        PyGeometryPair,
-        PyGeometryPair,
-        PyGeometryPair,
+        "PyGeometryPair",
+        "PyGeometryPair",
+        "PyGeometryPair",
+        "PyGeometryPair",
         list,
         list,
         list,
@@ -40,8 +20,8 @@ def from_file(
     ],
     # doublepair returns 2 geom + 4 log-lists
     Tuple[
-        PyGeometryPair,
-        PyGeometryPair,
+        "PyGeometryPair",
+        "PyGeometryPair",
         list,
         list,
         list,
@@ -49,14 +29,14 @@ def from_file(
     ],
     # singlepair returns 1 geom‐pair + 2 log-lists
     Tuple[
-        PyGeometryPair,
-        PyGeometryPair,
+        "PyGeometryPair",
+        "PyGeometryPair",
         list,
         list,
     ],
     # single returns 1 geom + 1 log-list
     Tuple[
-        PyGeometryPair,
+        "PyGeometryPair",
         list,
     ],
 ]:
@@ -120,7 +100,11 @@ def from_file(
     ValueError
         If an unsupported `mode` is passed.
     """
+    from . import PyContourType, PyGeometry, PyGeometryPair
+    from . import from_file_full, from_file_doublepair, from_file_singlepair, from_file_single
+
     defaults = {
+        "_diastole": True,
         "step_rotation_deg": 0.01,
         "range_rotation_deg": 30,
         "image_center": (4.5, 4.5),
@@ -128,20 +112,25 @@ def from_file(
         "n_points": 20,
         "write_obj": True,
         "watertight": True,
-        "rest_output_path": "output/rest",
-        "stress_output_path": "output/stress",
-        "diastole_output_path": "output/diastole",
-        "systole_output_path": "output/systole",
+        "output_path_a": "output/rest",
+        "output_path_b": "output/stress",
+        "output_path_c": "output/diastole",
+        "output_path_d": "output/systole",
         "interpolation_steps": 28,
         "bruteforce": False,
         "sample_size": 500,
+        "contour_types": [PyContourType.Lumen, PyContourType.Catheter, PyContourType.Wall],
+        "smooth": True,
+        "postprocessing": True,
     }
     merged = {**defaults, **kwargs}
 
     if mode == "full":
         required = (
-            "rest_input_path",
-            "stress_input_path",
+            "input_path_a",
+            "input_path_b",
+            "label",
+            "_diastole",
             "step_rotation_deg",
             "range_rotation_deg",
             "image_center",
@@ -149,18 +138,23 @@ def from_file(
             "n_points",
             "write_obj",
             "watertight",
-            "rest_output_path",
-            "stress_output_path",
-            "diastole_output_path",
-            "systole_output_path",
+            "output_path_a",
+            "output_path_b",
+            "output_path_c",
+            "output_path_d",
             "interpolation_steps",
             "bruteforce",
             "sample_size",
+            "contour_types",
+            "smooth",
[…]
