+++ src/intravascular/binding/mod.rs
@@ -40,8 +40,8 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
 ///
 ///
 /// Args:
-///     rest_input_path: Path to REST input file
-///     stress_input_path: Path to STRESS input file
+///     input_path_a: Path to e.g. REST input file
+///     input_path_b: Path to e.g. STRESS input file
 ///     step_rotation_deg (default 0.5°): Rotation step in degree
 ///     range_rotation_deg (default 90°): Rotation (+/-) range in degree, for 90° total range 180°
 ///     image_center (default (4.5mm, 4.5mm)): in mm
@@ -49,13 +49,16 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
 ///     n_points (default 20): number of points for catheter, more points stronger influence of image center
 ///     write_obj (default true): Wether to write OBJ files
 ///     watertight (default true): Wether to write shell or watertight mesh to OBJ.
-///     rest_output_path (default "output/rest"):
-///     stress_output_path (default "output/stress"):
-///     diastole_output_path (default "output/diastole"):
-///     systole_output_path (default "output/systole"):
+///     output_path_a (default "output/rest"):
+///     output_path_b (default "output/stress"):
+///     output_path_c (default "output/diastole"):
+///     output_path_d (default "output/systole"):
 ///     interpolation_steps (default 28): Number of interpolated meshes
 ///     bruteforce (default false): Wether to use bruteforce alignment (comparison for every step size)
 ///     sample_size (default 200): number of points to downsample to
+///     contour_type (default [PyContourType.Lumen, PyContourType.Catheter, PyContourType.Wall])
+///     smooth (default true): bool smooth after alignment with 3-point moving average
+///     postprocessing (default true): adjusts spacing within/between geometry/geometries to have equal offsets
 ///
 /// .. warning::
 ///
@@ -79,8 +82,8 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
 #[pyfunction]
 #[pyo3(
     signature = (
-        rest_input_path,
-        stress_input_path,
+        input_path_a,
+        input_path_b,
         label = "full",
         _diastole = true,
         step_rotation_deg = 0.5f64,
@@ -90,10 +93,10 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
         n_points = 20u32,
         write_obj = true,
         watertight = true,
-        rest_output_path = "output/rest",
-        stress_output_path = "output/stress",
-        diastole_output_path = "output/diastole",
-        systole_output_path = "output/systole",
+        output_path_a = "output/rest",
+        output_path_b = "output/stress",
+        output_path_c = "output/diastole",
+        output_path_d = "output/systole",
         interpolation_steps = 28usize,
         bruteforce = false,
         sample_size = 500,
@@ -103,8 +106,8 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
     )
 )]
 pub fn from_file_full(
-    rest_input_path: &str,
-    stress_input_path: &str,
+    input_path_a: &str,
+    input_path_b: &str,
     label: &str,
     _diastole: bool,
     step_rotation_deg: f64,
@@ -114,10 +117,10 @@ pub fn from_file_full(
     n_points: u32,
     write_obj: bool,
     watertight: bool,
-    rest_output_path: &str,
[…]
