### 2025-11-15 f6c8aa9
update example file and format all files
