+++ src/intravascular/utils/test_utils.rs
@@ -4,182 +4,182 @@ use std::collections::HashMap;
 
 pub fn dummy_geometry() -> Geometry {
     let points_a = vec![
-            ContourPoint {
-                frame_index: 1,
-                point_index: 0,
-                x: 1.0,
-                y: 3.0,
-                z: 0.0,
-                aortic: false,
-            },
-            ContourPoint {
-                frame_index: 1,
-                point_index: 1,
-                x: 0.0,
-                y: 2.0,
-                z: 0.0,
-                aortic: false,
-            },
-            ContourPoint {
-                frame_index: 1,
-                point_index: 2,
-                x: 0.0,
-                y: 0.0,
-                z: 0.0,
-                aortic: false,
-            },
-            ContourPoint {
-                frame_index: 1,
-                point_index: 3,
-                x: 1.0,
-                y: 0.0,
-                z: 0.0,
-                aortic: false,
-            },
-            ContourPoint {
-                frame_index: 1,
-                point_index: 4,
-                x: 2.0,
-                y: 0.0,
-                z: 0.0,
-                aortic: false,
-            },
-            ContourPoint {
-                frame_index: 1,
-                point_index: 5,
-                x: 2.0,
-                y: 2.0,
-                z: 0.0,
-                aortic: false,
-            },
+        ContourPoint {
+            frame_index: 1,
+            point_index: 0,
+            x: 1.0,
+            y: 3.0,
+            z: 0.0,
+            aortic: false,
+        },
+        ContourPoint {
+            frame_index: 1,
+            point_index: 1,
+            x: 0.0,
+            y: 2.0,
+            z: 0.0,
+            aortic: false,
+        },
+        ContourPoint {
+            frame_index: 1,
+            point_index: 2,
+            x: 0.0,
+            y: 0.0,
+            z: 0.0,
+            aortic: false,
+        },
+        ContourPoint {
+            frame_index: 1,
+            point_index: 3,
+            x: 1.0,
+            y: 0.0,
+            z: 0.0,
+            aortic: false,
+        },
+        ContourPoint {
+            frame_index: 1,
+            point_index: 4,
+            x: 2.0,
+            y: 0.0,
+            z: 0.0,
+            aortic: false,
+        },
+        ContourPoint {
+            frame_index: 1,
+            point_index: 5,
+            x: 2.0,
+            y: 2.0,
+            z: 0.0,
+            aortic: false,
+        },
     ];
     let points_b = vec![
-            ContourPoint {
-                frame_index: 2,
-                point_index: 0,
-                x: 1.0,
-                y: 3.0,
-                z: 1.0,
-                aortic: false,
-            },
-            ContourPoint {
-                frame_index: 2,
-                point_index: 1,
-                x: 0.0,
-                y: 2.0,
-                z: 1.0,
-                aortic: false,
-            },
-            ContourPoint {
-                frame_index: 2,
-                point_index: 2,
-                x: 0.0,
-                y: 0.0,
-                z: 1.0,
-                aortic: false,
-            },
-            ContourPoint {
-                frame_index: 2,
-                point_index: 3,
-                x: 1.0,
-                y: 0.0,
-                z: 1.0,
-                aortic: false,
-            },
-            ContourPoint {
-                frame_index: 2,
-                point_index: 4,
-                x: 2.0,
-                y: 0.0,
-                z: 1.0,
-                aortic: false,
-            },
[…]
