+++ multimodars/_wrappers.py
@@ -104,7 +104,7 @@ def from_file(
     from . import from_file_full, from_file_doublepair, from_file_singlepair, from_file_single
 
     defaults = {
-        "_diastole": True,
+        "diastole": True,
         "step_rotation_deg": 0.01,
         "range_rotation_deg": 30,
         "image_center": (4.5, 4.5),
@@ -130,7 +130,7 @@ def from_file(
             "input_path_a",
             "input_path_b",
             "label",
-            "_diastole",
+            "diastole",
             "step_rotation_deg",
             "range_rotation_deg",
             "image_center",
@@ -154,7 +154,7 @@ def from_file(
             args["input_path_a"],
             args["input_path_b"],
             args["label"],
-            args["_diastole"],
+            args["diastole"],
             args["step_rotation_deg"],
             args["range_rotation_deg"],
             args["image_center"],
@@ -179,7 +179,7 @@ def from_file(
             "input_path_a",
             "input_path_b",
             "label",
-            "_diastole",
+            "diastole",
             "step_rotation_deg",
             "range_rotation_deg",
             "image_center",
@@ -201,7 +201,7 @@ def from_file(
             args["input_path_a"],
             args["input_path_b"],
             args["label"],
-            args["_diastole"],
+            args["diastole"],
             args["step_rotation_deg"],
             args["range_rotation_deg"],
             args["image_center"],
@@ -223,7 +223,7 @@ def from_file(
         required = (
             "input_path",
             "label",
-            "_diastole",
+            "diastole",
             "step_rotation_deg",
             "range_rotation_deg",
             "image_center",
@@ -243,7 +243,7 @@ def from_file(
         return from_file_singlepair(
             args["input_path"],
             args["label"],
-            args["_diastole"],
+            args["diastole"],
             args["step_rotation_deg"],
             args["range_rotation_deg"],
             args["image_center"],
@@ -388,7 +388,7 @@ def from_array(
     from . import from_array_full, from_array_doublepair, from_array_singlepair, from_array_single
 
     defaults = {
-        "_diastole": True,
+        "diastole": True,
         "step_rotation_deg": 0.01,
         "range_rotation_deg": 30,
         "image_center": (4.5, 4.5),
@@ -419,7 +419,7 @@ def from_array(
             "input_data_c",
             "input_data_d",
             "label",
-            "_diastole",
+            "diastole",
             "step_rotation_deg",
             "range_rotation_deg",
             "image_center",
@@ -445,7 +445,7 @@ def from_array(
             args["input_data_c"],
             args["input_data_d"],
             args["label"],
-            args["_diastole"],
+            args["diastole"],
             args["step_rotation_deg"],
             args["range_rotation_deg"],
             args["image_center"],
@@ -472,7 +472,7 @@ def from_array(
             "input_data_c",
             "input_data_d",
             "label",
-            "_diastole",
+            "diastole",
             "step_rotation_deg",
             "range_rotation_deg",
             "image_center",
@@ -496,7 +496,7 @@ def from_array(
             args["input_data_c"],
             args["input_data_d"],
             args["label"],
-            args["_diastole"],
+            args["diastole"],
             args["step_rotation_deg"],
             args["range_rotation_deg"],
             args["image_center"],
@@ -519,7 +519,7 @@ def from_array(
             "input_data_a",
             "input_data_b",
             "label",
-            "_diastole",
+            "diastole",
             "step_rotation_deg",
             "range_rotation_deg",
             "image_center",
@@ -540,7 +540,7 @@ def from_array(
             args["input_data_a"],
             args["input_data_b"],
             args["label"],
-            args["_diastole"],
+            args["diastole"],
             args["step_rotation_deg"],
             args["range_rotation_deg"],
             args["image_center"],
