### 2025-11-16 2422974
fix resampling logic
