+++ src/intravascular/processing/postprocessing.rs
@@ -19,7 +19,7 @@ pub fn postprocess_geom_pair(
     let ref_z_a = geom_pair.geom_a.frames[ref_idx_a].centroid.2;
     let ref_z_b = geom_pair.geom_b.frames[ref_idx_b].centroid.2;
 
-    let geom_pair_resampled = if same_sample_rate {
+    let mut geom_pair_resampled = if same_sample_rate {
         let mean_diff = (avg_diff_a + avg_diff_b) / 2.0;
         let geom_a = resample_by_diff(&geom_pair.geom_a, mean_diff);
         let geom_b = resample_by_diff(&geom_pair.geom_b, mean_diff);
@@ -32,6 +32,7 @@ pub fn postprocess_geom_pair(
         let n = geom_pair.geom_b.frames.len();
         let end_zero = geom_pair.geom_b.frames[0].centroid.2;
         let end_n = geom_pair.geom_b.frames[n - 1].centroid.2;
+        // should never be the case... just to be extra sure
         let (start, stop) = if end_zero < end_n {
             (end_zero, end_n)
         } else {
@@ -41,10 +42,9 @@ pub fn postprocess_geom_pair(
         let z_coords = predict_z_positions(ref_z_b, start, stop, avg_diff_a);
         let new_geom = new_frames_by_sample_rate(&geom_pair.geom_b, z_coords);
         let new_resampled_a = resample_by_diff(&geom_pair.geom_a, avg_diff_a);
-        let new_resampled_b = resample_by_diff(&new_geom, avg_diff_a);
         GeometryPair {
             geom_a: new_resampled_a,
-            geom_b: new_resampled_b,
+            geom_b: new_geom,
             label: geom_pair.label.clone(),
         }
     } else {
@@ -57,16 +57,21 @@ pub fn postprocess_geom_pair(
             (end_n, end_zero)
         };
 
-        let z_coords = predict_z_positions(ref_z_a, start, stop, avg_diff_a);
+        let z_coords = predict_z_positions(ref_z_a, start, stop, avg_diff_b);
         let new_geom = new_frames_by_sample_rate(&geom_pair.geom_a, z_coords);
-        let new_resampled_a = resample_by_diff(&new_geom, avg_diff_b);
         let new_resampled_b = resample_by_diff(&geom_pair.geom_b, avg_diff_b);
         GeometryPair {
-            geom_a: new_resampled_a,
+            geom_a: new_geom,
             geom_b: new_resampled_b,
             label: geom_pair.label.clone(),
         }
     };
+    // ensure again that ref points are on same position before trimming
+    let ref_idx_a_resample = geom_pair_resampled.geom_a.find_ref_frame_idx()?;
+    let ref_idx_b_resample = geom_pair_resampled.geom_b.find_ref_frame_idx()?;
+    let translation = geom_pair.geom_a.frames[ref_idx_a_resample].centroid.2 - geom_pair.geom_b.frames[ref_idx_b_resample].centroid.2;
+    geom_pair_resampled.geom_a.translate_geometry((0.0, 0.0, translation));
+
     let mut trimmed_geom_pair = trim_geom_pair(&geom_pair_resampled);
     trimmed_geom_pair = if anomalous {
         adjust_walls_anomalous_geom_pair(&trimmed_geom_pair)
@@ -187,8 +192,14 @@ fn predict_z_positions(ref_z: f64, start_z: f64, stop_z: f64, z_diff: f64) -> Ve
 fn new_frames_by_sample_rate(geometry: &Geometry, mut z_coords: Vec<f64>) -> Geometry {
     let mut new_frames = Vec::new();
     z_coords.sort_by(|a, b| a.partial_cmp(b).unwrap_or(Ordering::Equal));
[…]
