+++ src/intravascular/processing/align_within.rs
@@ -317,76 +317,99 @@ fn assign_aortic(mut geometry: Geometry) -> Geometry {
     geometry
 }
 
+fn median(values: &mut [f64]) -> f64 {
+    if values.is_empty() {
+        return 0.0;
+    }
+    values.sort_by(|a, b| a.partial_cmp(b).unwrap());
+    let n = values.len();
+    if n % 2 == 1 {
+        values[n / 2]
+    } else {
+        (values[n / 2 - 1] + values[n / 2]) / 2.0
+    }
+}
+
+/// Detect whether there's any gap that's substantially larger than typical spacing.
+/// Returns (has_hole, baseline_spacing)
 fn detect_holes(geometry: &Geometry) -> (bool, f64) {
-    let mut z_diffs = Vec::new();
+    let mut z_diffs: Vec<f64> = Vec::new();
     for i in 1..geometry.frames.len() {
         let z_prev = geometry.frames[i - 1].centroid.2;
         let z_curr = geometry.frames[i].centroid.2;
-        let diff = (z_curr - z_prev).abs();
-        z_diffs.push(diff);
+        z_diffs.push((z_curr - z_prev).abs());
     }
     if z_diffs.is_empty() {
         return (false, 0.0);
     }
-    let avg_diff: f64 = z_diffs.iter().sum::<f64>() / z_diffs.len() as f64;
-    for &diff in &z_diffs {
-        if diff >= 1.5 * avg_diff {
-            return (true, avg_diff);
-        }
+
+    let mut sorted = z_diffs.clone();
+    let baseline = median(&mut sorted);
+
+    // avoid divide-by-zero later; if baseline is zero treat as "no hole"
+    if baseline <= std::f64::EPSILON {
+        return (false, baseline);
     }
-    (false, avg_diff)
+
+    let has_hole = z_diffs.iter().any(|&d| d >= 1.5 * baseline);
+
+    (has_hole, baseline)
 }
 
 /// Fill holes by inserting averaged / interpolated frames using Geometry::insert_frame.
-/// Uses z thresholds:
-///   < 1.5 => OK, do nothing
+/// Uses ratio thresholds against the baseline spacing (median):
+///   ratio < 1.5 => OK, do nothing
 ///  [1.5,2.5) => one missing frame -> insert averaged frame
 ///  [2.5,3.5) => two missing frames -> insert two interpolated frames
-///  >=3.5 => error (too big to auto-fix)
+///  >= 3.5 => error (too big to auto-fix)
 pub fn fill_holes(geometry: &mut Geometry) -> anyhow::Result<Geometry> {
-    let (hole, _avg_diff) = detect_holes(geometry);
+    let (hole, baseline) = detect_holes(geometry);
 
     if !hole {
         return Ok(geometry.clone());
     }
 
-    println!("⚠️\tHole detected! Attempting to fix using Geometry::insert_frame(...)");
+    if baseline <= std::f64::EPSILON {
+        return Err(anyhow!("Baseline spacing is zero or too small to decide."));
+    }
 
-    // Walk through frames; insert when we see a gap
+    println!("⚠️\tHole detected! Attempting to fix using Geometry::insert_frame(...) (baseline spacing = {:.3})", baseline);
+
     let mut i: usize = 1;
     while i < geometry.frames.len() {
-        let prev = &geometry.frames[i - 1].clone();
-        let curr = &geometry.frames[i].clone();
+        let prev = geometry.frames[i - 1].clone();
+        let curr = geometry.frames[i].clone();
 
         let diff = (curr.centroid.2 - prev.centroid.2).abs();
+        let ratio = diff / baseline;
 
-        if diff < 1.5 {
+        if ratio < 1.5 {
[…]
