+++ src/intravascular/centerline_align/align_algorithms.rs
@@ -41,27 +41,29 @@ pub fn get_transformations(
 ) -> Vec<FrameTransformation> {
     let mut transformations = Vec::new();
 
-    let ref_id_cl = centerline.find_reference_cl_point_idx(ref_pt) as u32;
-
-    for frame in geometry.frames.into_iter() {
-        if let Some(cl_point) = centerline.get_by_frame(frame.id + ref_id_cl) {
-            // Create one transformation per frame using the lumen as reference
+    // Find the reference point in the centerline
+    let ref_idx_cl = centerline.find_reference_cl_point_idx(ref_pt);
+    
+    // The geometry frames are ordered, and we assume they correspond to centerline points
+    // starting from the reference point and moving in the same direction
+    for (i, frame) in geometry.frames.into_iter().enumerate() {
+        // Calculate which centerline point corresponds to this geometry frame
+        // We start from the reference centerline point and move through the centerline
+        // based on the geometry frame's position relative to the reference frame
+        let cl_index = ref_idx_cl as isize + (i as isize); // or - i depending on your coordinate system direction
+        
+        if cl_index >= 0 && cl_index < centerline.points.len() as isize {
+            let cl_point = &centerline.points[cl_index as usize];
             let transformation = align_frame(&frame.lumen, cl_point);
             transformations.push(transformation);
+        } else {
+            eprintln!("Centerline index {} out of bounds for geometry frame {}", cl_index, frame.id);
         }
     }
     transformations
 }
 
-/// Modified align_frame returns the transformation applied (doesn't mutate input)
 fn align_frame(frame: &Contour, cl_point: &CenterlinePoint) -> FrameTransformation {
-    if frame.original_frame != cl_point.contour_point.frame_index {
-        panic!(
-            "Frame Index {} does not match Centerline Point Frame Index {}",
-            frame.original_frame, cl_point.contour_point.frame_index
-        );
-    }
-
     // Get centroid or compute if None
     let centroid = frame.centroid.unwrap_or_else(|| {
         let x_avg = frame.points.iter().map(|p| p.x).sum::<f64>() / frame.points.len() as f64;
@@ -71,7 +73,6 @@ fn align_frame(frame: &Contour, cl_point: &CenterlinePoint) -> FrameTransformati
     });
 
     // === Translation Step ===
-    // Compute the translation vector to bring the frame's centroid to the centerline point.
     let translation_vec = Vector3::new(
         cl_point.contour_point.x - centroid.0,
         cl_point.contour_point.y - centroid.1,
@@ -79,7 +80,6 @@ fn align_frame(frame: &Contour, cl_point: &CenterlinePoint) -> FrameTransformati
     );
 
     // === Rotation Step ===
-    // Compute the rotation needed to align the frame's normal with the centerline normal.
     let current_normal = calculate_normal(&frame.points, &centroid);
     let desired_normal = cl_point.normal;
     let angle = current_normal.angle(&desired_normal);
@@ -95,16 +95,15 @@ fn align_frame(frame: &Contour, cl_point: &CenterlinePoint) -> FrameTransformati
         }
     };
 
-    // Define the pivot as the centerline point.
+    // Define the pivot as the centerline point
     let pivot = Point3::new(
         cl_point.contour_point.x,
         cl_point.contour_point.y,
         cl_point.contour_point.z,
     );
 
-    // Return the transformation details for later use
     FrameTransformation {
-        frame_index: frame.original_frame,
[…]
