+++ src/intravascular/processing/align_between.rs
@@ -32,52 +32,101 @@ pub fn align_between_geometries(
     step_rot_deg: f64,
     sample_size: usize,
 ) -> Result<GeometryPair> {
-    println!(
-        "Aligning geometry '{}' to geometry '{}'",
-        geom_b.label, geom_a.label
-    );
-    let ref_frame_a_idx = geom_a
-        .find_ref_frame_idx()
-        .unwrap_or(geom_a.find_proximal_end_idx());
-    let ref_frame_b_idx = geom_b
-        .find_ref_frame_idx()
-        .unwrap_or(geom_b.find_proximal_end_idx());
+    println!("Aligning geometry '{}' to '{}'", geom_b.label, geom_a.label);
+    
+    // Find reference frames
+    let ref_frame_a_idx = geom_a.find_proximal_end_idx();
+    let ref_frame_b_idx = geom_b.find_proximal_end_idx();
 
-    let ref_frame_a = &geom_a.frames[ref_frame_a_idx];
-    let ref_frame_b = &geom_b.frames[ref_frame_b_idx];
+    let ref_frame_a_centroid = geom_a.frames[ref_frame_a_idx].centroid;
+    let ref_frame_b_centroid = geom_b.frames[ref_frame_b_idx].centroid;
 
+    // Apply translation to align centroids
     let translation = (
-        ref_frame_a.centroid.0 - ref_frame_b.centroid.0,
-        ref_frame_a.centroid.1 - ref_frame_b.centroid.1,
-        ref_frame_a.centroid.2 - ref_frame_b.centroid.2,
+        ref_frame_a_centroid.0 - ref_frame_b_centroid.0,
+        ref_frame_a_centroid.1 - ref_frame_b_centroid.1,
+        ref_frame_a_centroid.2 - ref_frame_b_centroid.2,
     );
     geom_b.translate_geometry(translation);
 
-    let test_geom_a = extract_geometry_points_with_frame_info(geom_a, sample_size);
-    let test_geom_b = extract_geometry_points_with_frame_info(geom_b, sample_size);
+    // Extract points for alignment
+    let test_geom_a = extract_geometry_points_with_frame_info(geom_a, sample_size.max(500));
+    let test_geom_b = extract_geometry_points_with_frame_info(geom_b, sample_size.max(500));
 
     let best_rotation =
         find_best_rotation_between(&test_geom_a, &test_geom_b, step_rot_deg, rot_deg);
 
-    for frame in &mut geom_b.frames {
-        frame.rotate_frame(best_rotation);
-    }
-
     println!(
-        "Best rotation to align '{}' to '{}' is {:.2} degrees",
-        geom_b.label,
-        geom_a.label,
+        "Applied rotation of {:.2}° to align '{}' to '{}'",
         best_rotation.to_degrees(),
+        geom_b.label,
+        geom_a.label
     );
 
+    // Apply the rotation to the ENTIRE geometry around the reference frame A centroid
+    rotate_geometry_around_point(geom_b, best_rotation, ref_frame_a_centroid);
+
     Ok(GeometryPair::new(geom_a.clone(), geom_b.clone())?)
+}
+
+/// Rotate entire geometry around a single reference point
+fn rotate_geometry_around_point(geometry: &mut Geometry, angle_rad: f64, center: (f64, f64, f64)) {
+    let cos_angle = angle_rad.cos();
+    let sin_angle = angle_rad.sin();
+    
+    // Helper closure to rotate a point around the common center
+    let rotate_point = |x: f64, y: f64| -> (f64, f64) {
+        let translated_x = x - center.0;
+        let translated_y = y - center.1;
+        
+        let rotated_x = translated_x * cos_angle - translated_y * sin_angle;
[…]
