### 2025-11-17 f8a5483
replace pytests with simple test for now
