+++ tests/conftest.py
@@ -1,7 +1,20 @@
 # tests/conftest.py
 import pytest
 import numpy as np
-from multimodars import PyContour, PyContourPoint, PyGeometry, numpy_to_geometry
+from multimodars import PyContour, PyContourPoint, PyGeometry, PyFrame
+from multimodars._converters import numpy_to_geometry
+
+
+def _compute_centroid(points):
+    """Compute centroid from a list of PyContourPoints"""
+    if not points:
+        return (0.0, 0.0, 0.0)
+    
+    sum_x = sum(p.x for p in points)
+    sum_y = sum(p.y for p in points) 
+    sum_z = sum(p.z for p in points)
+    n = len(points)
+    return (sum_x / n, sum_y / n, sum_z / n)
 
 
 def _create_round_contour(
@@ -20,9 +33,19 @@ def _create_round_contour(
     for idx, theta in enumerate(angles):
         x = cx + radius * np.cos(theta)
         y = cy + radius * np.sin(theta)
-        pt = PyContourPoint(0, idx, x, y, cz, aortic)
+        pt = PyContourPoint(frame_index=contour_id, point_index=idx, x=x, y=y, z=cz, aortic=aortic)
         points.append(pt)
-    contour = PyContour(contour_id, points)
+    
+    centroid = _compute_centroid(points)
+    contour = PyContour(
+        id=contour_id,
+        original_frame=contour_id,
+        points=points,
+        centroid=centroid,
+        aortic_thickness=None,
+        pulmonary_thickness=None,
+        kind="Lumen"
+    )
     return contour
 
 
@@ -43,27 +66,44 @@ def _create_elliptic_contour(
     for idx, theta in enumerate(angles):
         x = cx + major_radius * np.cos(theta)
         y = cy + minor_radius * np.sin(theta)
-        pt = PyContourPoint(0, idx, x, y, cz, aortic)
+        pt = PyContourPoint(frame_index=contour_id, point_index=idx, x=x, y=y, z=cz, aortic=aortic)
         points.append(pt)
-    contour = PyContour(contour_id, points)
+    
+    centroid = _compute_centroid(points)
+    contour = PyContour(
+        id=contour_id,
+        original_frame=contour_id,
+        points=points,
+        centroid=centroid,
+        aortic_thickness=None,
+        pulmonary_thickness=None,
+        kind="Lumen"
+    )
     return contour
 
 
 @pytest.fixture
 def sample_contour_points():
     return [
-        PyContourPoint(0, 0, 0.0, 4.0, 3.0, False),
-        PyContourPoint(0, 1, 0.0, 0.0, 3.0, False),
-        PyContourPoint(0, 2, 4.0, 0.0, 3.0, True),
-        PyContourPoint(0, 3, 4.0, 4.0, 3.0, True),
+        PyContourPoint(frame_index=0, point_index=0, x=0.0, y=4.0, z=3.0, aortic=False),
+        PyContourPoint(frame_index=0, point_index=1, x=0.0, y=0.0, z=3.0, aortic=False),
+        PyContourPoint(frame_index=0, point_index=2, x=4.0, y=0.0, z=3.0, aortic=True),
+        PyContourPoint(frame_index=0, point_index=3, x=4.0, y=4.0, z=3.0, aortic=True),
     ]
 
 
 @pytest.fixture
 def sample_contour(sample_contour_points):
-    contour = PyContour(0, sample_contour_points)
-    # Check centroid is as expected
-    assert contour.centroid == (2.0, 2.0, 3.0)
+    centroid = _compute_centroid(sample_contour_points)
+    contour = PyContour(
+        id=0,
+        original_frame=0,
+        points=sample_contour_points,
+        centroid=centroid,
+        aortic_thickness=None,
+        pulmonary_thickness=None,
[…]
