+++ multimodars/__init__.py
@@ -9,66 +9,40 @@ from ._converters import (
 from ._wrappers import (
     from_file,
     from_array,
-    to_centerline,
 )
-from .io.write_geometries import centerline_to_obj
-from .io.read_geometrical import read_mesh
 
 __all__ = [
-    # rust‑backend functions
-    # (fill these in manually, or use `multimodars = dir()` hack to auto‑generate)
-    "align_three_point",
-    "align_manual",
-    "from_file_full",
-    "from_file_doublepair",
-    "from_file_singlepair",
-    "from_file_single",
-    "create_catheter_geometry",
-    "from_array_single",
-    "from_array_full",
-    "from_array_doublepair",
-    "from_array_singlepair",
-    # rust class bound methods
-    ## PyContourPoint
+    # Core classes
     "PyContourPoint",
-    "distance",
-    ## PyContour
-    "PyContour",
-    "compute_centroid",
-    "points_as_tuples",
-    "find_farthest_points",
-    "find_closest_opposite",
-    "get_elliptic_ratio",
-    "get_area",
-    "rotate",
-    "translate",
-    "sort_contour_points",
-    ## PyFrame
+    "PyContour", 
     "PyFrame",
-    "rotate",
-    "translate",
-    "sort_frame_points",
-    ## PyGeometry
     "PyGeometry",
     "PyGeometryPair",
-    "rotate",
-    "translate",
-    "smooth_contours",
-    "reorder",
-    ## PyCenterline
-    "PyCenterlinePoint",
     "PyCenterline",
-    "from_contour_points",
-    # converters
+    "PyCenterlinePoint",
+    "PyInputData",
+    "PyRecord",
+    "PyContourType",
+    
+    # Converter functions
     "to_array",
-    "numpy_to_geometry",
+    "numpy_to_geometry", 
     "numpy_to_centerline",
-    # wrappers
+    
+    # Wrapper functions
     "from_file",
     "from_array",
-    "to_centerline",
+    
+    # Processing functions (from Rust)
+    "from_file_full",
+    "from_file_doublepair", 
+    "from_file_singlepair",
+    "from_file_single",
+    "from_array_full",
+    "from_array_doublepair",
+    "from_array_singlepair", 
+    "from_array_single",
+    "align_three_point",
+    "align_manual",
     "to_obj",
-    # io
-    "centerline_to_obj",
-    "read_mesh",
 ]

[…]
