+++ docs/api/classes/PyInputData.rst
@@ -1,0 +1,12 @@
+PyInputData
+===========
+
+.. currentmodule:: multimodars
+
+
+.. autoclass:: PyInputData
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :inherited-members:
+   :exclude-members: lumen, eem, calcification, sidebranch, record, ref_point, diastole, label
