+++ docs/changelog.md
@@ -1,0 +1,2 @@
+```{include} ../CHANGELOG.md
+    :parser: myst_parser.sphinx_
