+++ multimodars/_converters.py
@@ -37,7 +37,14 @@ def to_array(generic) -> Union[np.ndarray, dict, Tuple[dict, dict]]:
     TypeError
         If the input type is not one of the supported multimodars types.
     """
-    from . import PyContour, PyCenterline, PyGeometry, PyGeometryPair, PyFrame, PyInputData
+    from . import (
+        PyContour,
+        PyCenterline,
+        PyGeometry,
+        PyGeometryPair,
+        PyFrame,
+        PyInputData,
+    )
 
     if isinstance(generic, PyContour):
         pts = [(p.frame_index, p.x, p.y, p.z) for p in generic.points]
@@ -75,23 +82,29 @@ def to_array(generic) -> Union[np.ndarray, dict, Tuple[dict, dict]]:
 def _frame_to_numpy(frame) -> dict[str, np.ndarray]:
     """Convert PyFrame to dictionary of numpy arrays."""
     result = {}
-    
+
     # Add lumen
     lumen_pts = [(p.frame_index, p.x, p.y, p.z) for p in frame.lumen.points]
-    result["lumen"] = np.array(lumen_pts, dtype=float) if lumen_pts else np.zeros((0, 4), dtype=float)
-    
+    result["lumen"] = (
+        np.array(lumen_pts, dtype=float) if lumen_pts else np.zeros((0, 4), dtype=float)
+    )
+
     # Add extras
     for contour_type, contour in frame.extras.items():
         pts = [(p.frame_index, p.x, p.y, p.z) for p in contour.points]
-        result[contour_type.lower()] = np.array(pts, dtype=float) if pts else np.zeros((0, 4), dtype=float)
-    
+        result[contour_type.lower()] = (
+            np.array(pts, dtype=float) if pts else np.zeros((0, 4), dtype=float)
+        )
+
     # Add reference point
     if frame.reference_point:
         ref = frame.reference_point
-        result["reference"] = np.array([[ref.frame_index, ref.x, ref.y, ref.z]], dtype=float)
+        result["reference"] = np.array(
+            [[ref.frame_index, ref.x, ref.y, ref.z]], dtype=float
+        )
     else:
         result["reference"] = np.zeros((0, 4), dtype=float)
-    
+
     return result
 
 
@@ -106,7 +119,7 @@ def _geometry_to_numpy(geom) -> dict[str, np.ndarray]:
         "wall": np.zeros((0, 4), dtype=float),
         "reference": np.zeros((0, 4), dtype=float),
     }
-    
+
     # Collect all points from all frames
     for frame in geom.frames:
         frame_data = _frame_to_numpy(frame)
@@ -116,7 +129,7 @@ def _geometry_to_numpy(geom) -> dict[str, np.ndarray]:
                     result[key] = frame_data[key]
                 else:
                     result[key] = np.vstack([result[key], frame_data[key]])
-    
+
     return result
 
 
@@ -131,7 +144,7 @@ def _input_data_to_numpy(input_data) -> dict[str, Union[np.ndarray, List[str], b
         "diastole": input_data.diastole,
         "label": input_data.label,
     }
-    
+
     # Process lumen (required)
     if input_data.lumen:
         lumen_pts = []
@@ -139,7 +152,7 @@ def _input_data_to_numpy(input_data) -> dict[str, Union[np.ndarray, List[str], b
             lumen_pts.extend([(p.frame_index, p.x, p.y, p.z) for p in contour.points])
         if lumen_pts:
             result["lumen"] = np.array(lumen_pts, dtype=float)
-    
+
     # Process optional contour types
     for contour_type in ["eem", "calcification", "sidebranch"]:
         contours = getattr(input_data, contour_type)
[…]
