+++ multimodars/__init__.py
@@ -1,11 +1,7 @@
 from __future__ import annotations
 
 from .multimodars import *
-from ._converters import (
-    to_array, 
-    numpy_to_geometry, 
-    numpy_to_centerline
-)
+from ._converters import to_array, numpy_to_geometry, numpy_to_centerline
 from ._wrappers import (
     from_file,
     from_array,
@@ -14,7 +10,7 @@ from ._wrappers import (
 __all__ = [
     # Core classes
     "PyContourPoint",
-    "PyContour", 
+    "PyContour",
     "PyFrame",
     "PyGeometry",
     "PyGeometryPair",
@@ -23,24 +19,21 @@ __all__ = [
     "PyInputData",
     "PyRecord",
     "PyContourType",
-    
     # Converter functions
     "to_array",
-    "numpy_to_geometry", 
+    "numpy_to_geometry",
     "numpy_to_centerline",
-    
     # Wrapper functions
     "from_file",
     "from_array",
-    
     # Processing functions (from Rust)
     "from_file_full",
-    "from_file_doublepair", 
+    "from_file_doublepair",
     "from_file_singlepair",
     "from_file_single",
     "from_array_full",
     "from_array_doublepair",
-    "from_array_singlepair", 
+    "from_array_singlepair",
     "from_array_single",
     "align_three_point",
     "align_manual",

[…]
