### 2025-11-18 c99bcb5
remove unnecessary assignment in postprocessing bump version
