+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "multimodars"
-version = "0.1.0"
+version = "0.1.1"
 authors = [
     { name = "Anselm W. Stark"}
 ]
