+++ docs/api/utilities.rst
@@ -1,7 +1,5 @@
 Utility Functions
 =================
 
-.. autofunction:: multimodars.create_catheter_geometry
 .. autofunction:: multimodars.to_obj
 .. autofunction:: multimodars.centerline_to_obj
-.. autofunction:: multimodars.read_mesh
