### 2025-11-19 caf826f
add centerline sphere bounded points identification
