+++ src/lib.rs
@@ -7,6 +7,7 @@ use crate::intravascular::binding::{
     from_array_doublepair, from_array_full, from_array_single, from_array_singlepair,
     from_file_doublepair, from_file_full, from_file_single, from_file_singlepair, to_obj,
 };
+use crate::ccta::binding::label_py::find_centerline_bounded_points_simple;
 
 use pyo3::prelude::*;
 use pyo3::wrap_pyfunction;
@@ -36,6 +37,7 @@ fn multimodars(_py: Python, m: pyo3::prelude::Bound<'_, PyModule>) -> PyResult<(
     m.add_function(wrap_pyfunction!(from_array_singlepair, m.clone())?)?;
     m.add_function(wrap_pyfunction!(from_array_single, m.clone())?)?;
     m.add_function(wrap_pyfunction!(to_obj, m.clone())?)?;
+    m.add_function(wrap_pyfunction!(find_centerline_bounded_points_simple, m.clone())?)?;
 
     // Updated class registration
     m.add_class::<PyInputData>()?;
