+++ src/ccta/label_mesh/mod.rs
@@ -1,0 +1,1 @@
+pub mod label_coronary;
