+++ src/ccta/binding/mod.rs
@@ -1,0 +1,1 @@
+pub mod label_py;
