### 2025-11-21 634dde4
make fill hole functions less restrictive
