+++ docs/api/utilities.rst
@@ -2,4 +2,3 @@ Utility Functions
 =================
 
 .. autofunction:: multimodars.to_obj
-.. autofunction:: multimodars.centerline_to_obj
