+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "multimodars"
-version = "0.1.1"
+version = "0.1.2"
 edition = "2021"
 license = "MIT"
 
@@ -9,14 +9,14 @@ name = "multimodars"
 crate-type = ["cdylib"]
 
 [dependencies]
-pyo3 = "0.25.0"
-csv = "1.1"
-serde = { version = "1.0.219", features = ["derive"]}
-image = "0.25.5"
-nalgebra = "0.34.0"
-toml = "0.9.5"
+pyo3 = "0.27.1"
+csv = "1.4.0"
+serde = { version = "1.0.228", features = ["derive"]}
+image = "0.25.9"
+nalgebra = "0.34.1"
+toml = "0.9.8"
 rayon = "1.7.1"
 crossbeam = "0.8.4"
-anyhow = "1.0.97"
+anyhow = "1.0.100"
 approx = "0.5"
-stl_io = "0.8.5"
+stl_io = "0.10.0"
