+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "multimodars"
-version = "0.1.1"
+version = "0.1.2"
 authors = [
     { name = "Anselm W. Stark"}
 ]
