### 2025-11-27 b4e9a4f
implement ray-triangle intersection to identify aortic faces
