+++ multimodars/io/read_geometrical.py
@@ -65,43 +65,176 @@ def read_mesh(path: Path | str) -> trimesh.base.Trimesh:
     return mesh
 
 
+def find_faces_for_points(mesh: trimesh.Trimesh, points_found, tol: float = 1e-6):
+    """
+    For each point in points_found find nearest vertex on `mesh` (within tol)
+    and return the list of face indices that reference any of those vertices.
+
+    Returns:
+        List[int] -- indices into mesh.faces
+    """
+    points_array = np.asarray(points_found, dtype=np.float64)
+    if points_array.size == 0:
+        return []
+
+    found_vertex_indices = set()
+    verts = mesh.vertices  # (V,3)
+
+    # loop is OK for moderate point counts; if you have many points use a KDTree
+    for p in points_array:
+        distances = np.linalg.norm(verts - p, axis=1)
+        closest_idx = int(np.argmin(distances))
+        if distances[closest_idx] <= tol:
+            found_vertex_indices.add(closest_idx)
+
+    if not found_vertex_indices:
+        return []
+
+    # gather face indices that include any matched vertex
+    face_indices = []
+    for i, face in enumerate(mesh.faces):
+        if (face[0] in found_vertex_indices) or (face[1] in found_vertex_indices) or (face[2] in found_vertex_indices):
+            face_indices.append(i)
+
+    return face_indices
+
+
+def prepare_faces_for_rust(mesh: trimesh.Trimesh, *, points=None, face_indices=None, tol: float = 1e-6):
+    """
+    Convert selected mesh faces to the Rust-friendly format:
+      [ ((v0x,v0y,v0z), (v1x,v1y,v1z), (v2x,v2y,v2z)), ... ]
+
+    Args:
+        mesh: trimesh.Trimesh
+        points: optional list/array of points. If provided, faces are found by calling
+                find_faces_for_points(mesh, points, tol=tol)
+        face_indices: optional explicit list of face indices to convert
+        tol: tolerance for matching points -> vertices
+
+    Returns:
+        List of face tuples (v0, v1, v2) where each v* is a (x,y,z) tuple of floats.
+    """
+    if face_indices is None:
+        if points is not None:
+            face_indices = find_faces_for_points(mesh, points, tol=tol)
+        else:
+            # default to all faces (backwards-compatible)
+            face_indices = list(range(len(mesh.faces)))
+
+    rust_faces = []
+    for fi in face_indices:
+        face = mesh.faces[fi]
+        v0 = tuple(map(float, mesh.vertices[face[0]]))
+        v1 = tuple(map(float, mesh.vertices[face[1]]))
+        v2 = tuple(map(float, mesh.vertices[face[2]]))
+        rust_faces.append((v0, v1, v2))
+    return rust_faces
+
 if __name__ == "__main__":
-    # import pyvista as pv
-    # import numpy as np
-    # vtk_path = 'data/center_smoothed.vtk'
-
-    # mesh = pv.read(vtk_path)
-    # points = mesh.points
-
-    # np.savetxt("data/centerline_narco119.csv", points, delimiter=",", fmt="%.6f")
-
     from trimesh.points import PointCloud
     import multimodars as mm
 
-    mesh = read_mesh('data/output/aligned/mesh_000_None.obj')
-    mesh2 = read_mesh('data/NARCO_119.stl')
-    print(mesh.vertices)
-    print(mesh.faces)
-    points = np.loadtxt("data/centerlines_narco119.csv", delimiter=",")
-    pc = PointCloud(points)
+    # Load ONLY the coronary mesh (mesh2) - ignore the aorta mesh for now
+    mesh2 = read_mesh('data/NARCO_119.stl')  # Coronary mesh
     
[…]
