### 2025-11-28 9e287c1
implement stable labeling of mesh components
