+++ multimodars/ccta/adjust_ccta.py
@@ -1,0 +1,309 @@
+import numpy as np
+import trimesh
+from pathlib import Path
+import warnings
+
+
+def label_geometry(
+    path_ccta_geometry: Path | str,
+    path_centerline_aorta: Path | str,
+    path_centerline_rca: Path | str,
+    path_centerline_lca: Path | str,
+    anomalous_rca: bool = False,
+    anomalous_lca: bool = False,
+    n_points_intramural: int = 120,
+    bounding_sphere_radius_mm: float = 3.0,
+    tolerance_float: float = 1e-6,
+    control_plot: bool = True,
+):
+    import multimodars as mm
+    from multimodars.io.read_geometrical import read_mesh
+    
+    try:
+        mesh = read_mesh(path_ccta_geometry)
+        print(f"Loaded mesh: {len(mesh.vertices)} vertices, {len(mesh.faces)} faces")
+    except Exception as e:
+        print(f"Error reading CCTA mesh from {path_ccta_geometry}: {e}")
+        raise
+
+    try:
+        cl_aorta_raw = np.genfromtxt(path_centerline_aorta, delimiter=',')
+        cl_aorta = mm.numpy_to_centerline(cl_aorta_raw)
+        print(f"Loaded aorta centerline: {len(cl_aorta.points)} points")
+    except Exception as e:
+        print(f"Error reading Aorta centerline from {path_centerline_aorta}: {e}")
+        raise
+
+    try:
+        cl_lca_raw = np.genfromtxt(path_centerline_lca, delimiter=',')
+        cl_lca = mm.numpy_to_centerline(cl_lca_raw)
+        print(f"Loaded LCA centerline: {len(cl_lca.points)} points")
+    except Exception as e:
+        print(f"Error reading LCA centerline from {path_centerline_lca}: {e}")
+        raise
+
+    try:
+        cl_rca_raw = np.genfromtxt(path_centerline_rca, delimiter=',')
+        cl_rca = mm.numpy_to_centerline(cl_rca_raw)
+        print(f"Loaded RCA centerline: {len(cl_rca.points)} points")
+    except Exception as e:
+        print(f"Error reading RCA centerline from {path_centerline_rca}: {e}")
+        raise
+
+    points_list = [tuple(vertex) for vertex in mesh.vertices.tolist()]
+    
+    # Find points for RCA and LCA
+    rca_points_found = mm.find_centerline_bounded_points_simple(
+        cl_rca, 
+        points_list, 
+        bounding_sphere_radius_mm
+    )
+    lca_points_found = mm.find_centerline_bounded_points_simple(
+        cl_lca, 
+        points_list, 
+        bounding_sphere_radius_mm
+    )
+    
+    print(f"RCA points found: {len(rca_points_found)}")
+    print(f"LCA points found: {len(lca_points_found)}")
+    
+    # Apply occlusion removal for anomalous vessels
+    rca_removed_points = []
+    lca_removed_points = []
+    
+    if anomalous_rca:
+        print("Applying occlusion removal for anomalous RCA...")
+        rca_faces_for_rust = prepare_faces_for_rust(mesh, points=rca_points_found, tol=tolerance_float)
+        final_rca_points_found = mm.remove_occluded_points_ray_triangle(
+            centerline_coronary=cl_rca,
+            centerline_aorta=cl_aorta,
+            range_coronary=n_points_intramural,
+            points=rca_points_found,
+            faces=rca_faces_for_rust
+        )
+        rca_removed_points = [p for p in rca_points_found if p not in final_rca_points_found]
+        print(f"RCA: removed {len(rca_removed_points)} occluded points")
+    else:
+        final_rca_points_found = rca_points_found.copy()
+
+    if anomalous_lca:
[…]
