### 2025-11-30 77882ae
add centerline-based scaling of coronary diameter
