+++ src/ccta/label_mesh/mod.rs
@@ -1,1 +1,0 @@
-pub mod label_coronary;
