+++ multimodars/ccta/adjust_ccta.py
@@ -52,7 +52,7 @@ def label_geometry(
 
     points_list = [tuple(vertex) for vertex in mesh.vertices.tolist()]
     
-    # Find points for RCA and LCA
+    # Rust implementation using a rolling sphere with fixed radius
     rca_points_found = mm.find_centerline_bounded_points_simple(
         cl_rca, 
         points_list, 
@@ -67,13 +67,14 @@ def label_geometry(
     print(f"RCA points found: {len(rca_points_found)}")
     print(f"LCA points found: {len(lca_points_found)}")
     
-    # Apply occlusion removal for anomalous vessels
     rca_removed_points = []
     lca_removed_points = []
     
     if anomalous_rca:
         print("Applying occlusion removal for anomalous RCA...")
         rca_faces_for_rust = prepare_faces_for_rust(mesh, points=rca_points_found, tol=tolerance_float)
+        # Rust implementation, that creates ray between aortic and coronary centerline, and
+        # removes faces if 3 consecutive faces are "pierced" by the ray
         final_rca_points_found = mm.remove_occluded_points_ray_triangle(
             centerline_coronary=cl_rca,
             centerline_aorta=cl_aorta,
@@ -101,7 +102,6 @@ def label_geometry(
     else:
         final_lca_points_found = lca_points_found.copy()
 
-    # Create debug plot if requested
     if control_plot:
         debug_plot(
             mesh=mesh,
@@ -114,7 +114,6 @@ def label_geometry(
             cl_aorta=cl_aorta
         )
 
-    # Return results
     results = {
         'mesh': mesh,
         'aorta_points': find_aortic_points(mesh.vertices, final_rca_points_found, final_lca_points_found),
@@ -154,7 +153,6 @@ def debug_plot(
     """
     from trimesh.points import PointCloud
     
-    # Get all mesh vertices and find aortic points
     all_vertices = mesh.vertices
     aortic_points = find_aortic_points(all_vertices, rca_points, lca_points)
     
@@ -166,7 +164,6 @@ def debug_plot(
     print(f"RCA reassigned points (red): {len(rca_removed_points)}")
     print(f"LCA reassigned points (red): {len(lca_removed_points)}")
     
-    # Create point clouds with different colors
     scenes = []
     
     # Scene 1: Mesh with all points
@@ -177,32 +174,27 @@ def debug_plot(
     mesh_visual.visual.face_colors = [200, 200, 200, 100]  # Semi-transparent gray
     scene1_geoms.append(mesh_visual)
     
-    # Add aortic points (yellow)
     if aortic_points:
         aortic_array = np.array(aortic_points, dtype=np.float64)
         aortic_colors = np.tile([255, 255, 0, 255], (len(aortic_points), 1))  # Yellow
         scene1_geoms.append(PointCloud(aortic_array, colors=aortic_colors))
     
-    # Add RCA points (blue)
     if rca_points:
         rca_array = np.array(rca_points, dtype=np.float64)
         rca_colors = np.tile([0, 0, 255, 255], (len(rca_points), 1))  # Blue
         scene1_geoms.append(PointCloud(rca_array, colors=rca_colors))
     
-    # Add LCA points (green)
     if lca_points:
         lca_array = np.array(lca_points, dtype=np.float64)
         lca_colors = np.tile([0, 255, 0, 255], (len(lca_points), 1))  # Green
         scene1_geoms.append(PointCloud(lca_array, colors=lca_colors))
     
-    # Add removed points (red)
     all_removed = rca_removed_points + lca_removed_points
     if all_removed:
         removed_array = np.array(all_removed, dtype=np.float64)
[…]
