+++ src/ccta/mod.rs
@@ -1,3 +1,3 @@
 pub mod binding;
 pub mod io;
-pub mod label_mesh;
+pub mod adjust_mesh;
