+++ src/ccta/binding/mod.rs
@@ -1,1 +1,1 @@
-pub mod label_py;
+pub mod ccta_py;

[…]
