### 2025-12-01 996ffda
add 3 point alignment with hausdorff distance adjustment function
