+++ src/intravascular/binding/align.rs
@@ -1,6 +1,6 @@
 use crate::intravascular::{
     binding::classes::{PyCenterline, PyContourType, PyGeometryPair},
-    centerline_align::{align_manual_rs, align_three_point_rs},
+    centerline_align::{align_manual_rs, align_three_point_rs, align_combined_rs},
 };
 use pyo3::prelude::*;
 
@@ -156,6 +156,101 @@ pub fn align_manual(
         geom_pair_rs,
         rotation_angle,
         ref_point,
+        write,
+        watertight,
+        interpolation_steps,
+        output_dir,
+        rust_contour_types,
+        case_name,
+    )
+    .map_err(|e| pyo3::exceptions::PyRuntimeError::new_err(e.to_string()))?;
+
+    let aligned_py: PyGeometryPair = geom_pair.into();
+    let resampled_cl: PyCenterline = cl.into();
+
+    Ok((aligned_py, resampled_cl))
+}
+
+/// Creates centerline-aligned meshes for diastolic and systolic geometries
+/// based on one reference point (e.g. aorta ostium, upper, lower) and Hausdorff distances.
+///
+/// Args:
+///     centerline: PyCenterline object
+///     geometry_pair: PyGeometryPair object
+///     aortic_ref_pt: Reference point for aortic position
+///     points: List of points to use for Hausdorff distance calculation
+///     angle_step_deg (default 1.0): step size in degrees for rotation search
+///     write (default false): Wether to write aligned meshes to OBJ
+///     watertight (default true): Wether to write shell or watertight mesh to OBJ.
+///     interpolation_steps: Number of interpolation steps
+///     output_dir (default "output/aligned"): Output directory for aligned meshes
+///     case_name (default "None"): Case name for output files
+///
+/// Returns:
+///     PyGeometryPair, PyCenterline (resampled)
+///
+/// Example:
+///     >>> import multimodars as mm
+///     >>> dia, sys = mm.align_three_point(
+///     ...     centerline,
+///     ...     geometry_pair,
+///     ...     (12.2605, -201.3643, 1751.0554),
+///     ...     [(11.7567, -202.1920, 1754.7975), ... ],
+///     ... )
+#[pyfunction]
+#[pyo3(
+    signature = (
+        centerline,
+        geom_pair,
+        aortic_ref_pt,
+        upper_ref_pt,
+        lower_ref_pt,
+        points,
+        angle_step_deg=1.0,
+        angle_range_deg=15.0,
+        index_range=2,
+        write=false,
+        watertight=true,
+        interpolation_steps=28usize,
+        output_dir="output/aligned",
+        contour_types=vec![PyContourType::Lumen, PyContourType::Catheter, PyContourType::Wall],
+        case_name="None",
+    )
+)]
+pub fn align_combined(
+    centerline: PyCenterline,
+    geom_pair: PyGeometryPair,
+    aortic_ref_pt: (f64, f64, f64),
+    upper_ref_pt: (f64, f64, f64),
+    lower_ref_pt: (f64, f64, f64),
+    points: Vec<(f64, f64, f64)>,
+    angle_step_deg: f64,
+    angle_range_deg: f64,  // e.g., 15° in radians
+    index_range: usize, // e.g., 2
+    write: bool,
+    watertight: bool,
+    interpolation_steps: usize,
+    output_dir: &str,
+    contour_types: Vec<PyContourType>,
+    case_name: &str,
+) -> PyResult<(PyGeometryPair, PyCenterline)> {
+    let rust_contour_types: Vec<crate::intravascular::io::geometry::ContourType> =
+        contour_types.iter().map(|ct| ct.into()).collect();
+    let cl_rs = centerline.to_rust_centerline();
[…]
