+++ src/intravascular/centerline_align/align_algorithms.rs
@@ -1,6 +1,8 @@
 use crate::intravascular::io::geometry::{Contour, Geometry};
 use crate::intravascular::io::input::ContourPoint;
 use crate::intravascular::io::input::{Centerline, CenterlinePoint};
+use crate::intravascular::processing::align_between::GeometryPair;
+use crate::intravascular::processing::process_utils::{downsample_contour_points, hausdorff_distance};
 use nalgebra::{Point3, Rotation3, Unit, Vector3};
 
 #[derive(Debug, Clone, Copy)]
@@ -274,6 +276,185 @@ pub fn best_rotation_three_point(
     }
 
     best_angle
+}
+
+/// Refines alignment using Hausdorff distance within a limited search space
+pub fn refine_alignment_hausdorff(
+    geom_pair: &GeometryPair,
+    centerline: &Centerline,
+    initial_cl_ref_idx: usize,
+    initial_rotation: f64,
+    mutated_points: &[ContourPoint],
+    angle_search_range: f64, // e.g., 15° in radians
+    angle_step: f64,
+    index_search_range: usize, // e.g., 2 points
+) -> (f64, usize) {
+    let len_frames = geom_pair.geom_a.frames.len();
+    
+    let mut best_angle = initial_rotation;
+    let mut best_cl_ref_idx = initial_cl_ref_idx;
+    let mut min_hausdorff = f64::MAX;
+    
+    println!("---------------------Refining alignment with Hausdorff---------------------");
+    println!("Initial rotation: {:.4} rad, Initial CL index: {}", initial_rotation, initial_cl_ref_idx);
+    
+    // Search over centerline indices
+    for delta_idx in -(index_search_range as isize)..=(index_search_range as isize) {
+        let current_cl_ref_idx = (initial_cl_ref_idx as isize + delta_idx) as usize;
+        
+        // Ensure we have enough centerline points for all frames
+        if current_cl_ref_idx + len_frames >= centerline.points.len() {
+            continue;
+        }
+        
+        // Create centerline segment for this index
+        let cl_end_idx = current_cl_ref_idx + len_frames;
+        let cl_segment = Centerline {
+            points: centerline.points[current_cl_ref_idx..cl_end_idx].to_vec(),
+        };
+        
+        // Search over rotation angles
+        let mut angle = initial_rotation - angle_search_range;
+        while angle <= initial_rotation + angle_search_range {
+            // Apply rotation and transformation
+            let mut transformed_geompair = rotate_by_best_rotation(geom_pair.clone(), angle);
+            
+            // Use the current centerline point as reference
+            let ref_pt = (
+                centerline.points[current_cl_ref_idx].contour_point.x,
+                centerline.points[current_cl_ref_idx].contour_point.y,
+                centerline.points[current_cl_ref_idx].contour_point.z,
+            );
+            
+            transformed_geompair = apply_transformations(transformed_geompair, &cl_segment, &ref_pt);
+            
+            // Filter mutated points to the current region
+            let filtered_points = filter_points_in_region(
+                mutated_points,
+                &centerline.points[current_cl_ref_idx],
+                &centerline.points[cl_end_idx - 1],
+            );
+            
+            if filtered_points.is_empty() {
+                angle += angle_step;
+                continue;
+            }
+            
+            // Downsample and flatten geometry points
+            let frames = &transformed_geompair.geom_a.frames;
+            let n_points_per_frame = frames[0].lumen.points.len();
+            let mut nested: Vec<Vec<ContourPoint>> = Vec::with_capacity(len_frames);
+            
+            // Calculate downsample ratio
[…]

[…]
