+++ ROADMAP.MD
@@ -3,23 +3,31 @@
 - [x] Bind all Rust structs to python objects
 - [x] Single mesh alignment (ignoring systole and diastole (for OCT))
 - [x] Functions should return obj, as python objects
-- [] Walls not rotated together with contours, and also catheter. 
+- [x] Walls not rotated together with contours, and also catheter. 
     here a helper function that performs operations on all ContourPoints
     inside of a Geometry regardless of type, should solve the problem.
-- [] double check resampling for dia and sys
+- [x] double check resampling for dia and sys
 
 # Centerline alignment module
 - [x] Centerline resampling based on geometry
 - [x] 3-Point alignment versus manual (degree/centerline point)
-- [] versus Hausdorff-alignment (v0.2.0)
-(- [] Create Centerline directly from mesh) (v0.2.0)
+- [x] versus Hausdorff-alignment (v0.2.0)
+(- [] Create Centerline directly from mesh) (v1.0.0)
 
 # Mesh combination (v0.2.0)
 - [x] Read in CCTA mesh (STL/Obj)
 - [x] Create closed mesh for intravascular segment
-- [] Assign regions based on Centerlines (RCA, LCA, Aorta)
-- [] Shrink/Expand CCTA to best match Aortic wall and shrink/expand coronary to match iv images
+- [x] Assign regions based on Centerlines (RCA, LCA, Aorta)
 - [x] ensure works for IVUS and OCT
+- [] Lable points inside of anomalous region, proximal and distal to region
+- [] Shrink/Expand CCTA to best match Aortic wall and shrink/expand coronary to match iv   images
+- [] Remove points inside of anomalous region, connect remaining points to proximal and distal points, without resampling of the anomalous region
+- [] Create ostium frames for smoother connection to aorta (define range cl points, and function of lumen increase)
+- [] Write Pytests for the new functions
+
+# CCTA clinical assessment (v0.3.0)
+- [] automatically assess lumen area, minor-, major axis, mla, intramural length etc. from the CCTA mesh.
+- [] return hausdorff distances per point between aligned ivus mesh and ccta mesh.
 
 # v0.1.0 desired functionality:
 - [x] .from_file(mode = {full, single}, "rest dir", "stress dir", "dia dir", "sys dir")
@@ -35,6 +43,7 @@
 - [x] better summarized print statement
 - [x] for fixture use a idealized geometry to check functionality
 - [x] reference point is not rotated and translated
+- [] better error propagation from Rust to Python (probably in the binding module, handle different python errors)
 
 # Prepublish checklist:
 - [x] tests/ with unit tests checking that importing and running every public function works
@@ -44,6 +53,7 @@
 - [x] Auto-generated API docs
 - [x] High-level "Getting Started" guide
 - [x] CONTRIBUTING.md
-
 - [x] Example scripts Jupyter notebook?
 - [] mypy in pre-commit hooks
+- [] make a dev version with more testing data in data folder and a minimal dataset for the pypi package, that can be loaded instead of read from files.
+- [] make stable version update by reading in a __version__ file
