### 2025-12-02 c2834c9
add (roughly) working labeling of anomalous coronary section
