+++ src/ccta/adjust_mesh/scale_coronary.rs
@@ -1,4 +1,5 @@
 use crate::intravascular::io::input::{CenterlinePoint, Centerline};
+use crate::intravascular::io::geometry::Frame;
 
 pub fn centerline_based_diameter_morphing(
     centerline: &Centerline,
@@ -62,6 +63,150 @@ fn calculate_squared_distance(point: (f64, f64, f64), centerline_point: &Centerl
     let dy = point.1 - centerline_point.contour_point.y;
     let dz = point.2 - centerline_point.contour_point.z;
     dx * dx + dy * dy + dz * dz
+}
+
+pub fn find_points_by_cl_region_rs(
+    centerline: &Centerline,
+    frames: &[Frame],
+    points: &[(f64, f64, f64)],
+) -> (Vec<(f64, f64, f64)>, Vec<(f64, f64, f64)>, Vec<(f64, f64, f64)>) {
+    let mut cumulative_z_dist_frames = 0.0;
+    for i in 1..frames.len() {
+        cumulative_z_dist_frames += (frames[i].centroid.2 - frames[i-1].centroid.2).abs();
+    }
+    cumulative_z_dist_frames /= (frames.len() -1) as f64;
+
+    let centroids_to_match = frames.iter().map(|f| f.centroid).collect::<Vec<(f64, f64, f64)>>();
+    let cl_points_indices: Vec<usize> = find_cl_points_in_range(
+        centerline,
+        &centroids_to_match,
+        cumulative_z_dist_frames,
+    );
+
+    // needed for proximal/distal classification
+    let dist_ref = centroids_to_match[centroids_to_match.len() -1];
+
+    let mut proximal_points: Vec<(f64, f64, f64)> = Vec::new();
+    let mut distal_points: Vec<(f64, f64, f64)> = Vec::new();
+    let mut points_between: Vec<(f64, f64, f64)> = Vec::new();
+
+    let mut remaining_points = points.to_vec();
+    
+    // First pass: find all points between centerline regions
+    remaining_points.retain(|point| {
+        let closest_cl_point = find_closest_centerline_point_optimized(&centerline, *point);
+        let cl_idx = closest_cl_point.contour_point.frame_index as usize;
+        
+        if cl_points_indices.contains(&cl_idx) {
+            points_between.push(*point);
+            false // remove from remaining
+        } else {
+            true // keep in remaining
+        }
+    });
+    
+    // Second pass: classify remaining points as proximal or distal
+    for point in remaining_points.iter() {
+        if point.0 > dist_ref.0 && point.1 > dist_ref.1 && point.2 > dist_ref.2 {
+            proximal_points.push(*point);
+        } else {
+            distal_points.push(*point);
+        }
+    }
+    let (proximal_points, points_between) = clean_up_non_section_points(proximal_points, points_between);
+    let (distal_points, points_between) = clean_up_non_section_points(distal_points, points_between);
+    (proximal_points, distal_points, points_between)
+}
+
+fn find_cl_points_in_range(
+    centerline: &Centerline,
+    points: &[(f64, f64, f64)],
+    search_radius: f64,
+) -> Vec<usize> {
+    let mut selected_points = Vec::new();
+
+    for point in points.iter() {
+        for cl_point in centerline.points.iter() {
+            let dx = point.0 - cl_point.contour_point.x;
+            let dy = point.1 - cl_point.contour_point.y;
+            let dz = point.2 - cl_point.contour_point.z;
+            let distance_squared = dx * dx + dy * dy + dz * dz;
+            if distance_squared <= search_radius * search_radius {
+                selected_points.push(cl_point);
+            }
+        }
+    }
+
+    // remove duplicates
+    selected_points.sort_by_key(|p| p.contour_point.frame_index);
[…]
