### 2025-12-03 98462db
add function to find best scaling factors
