+++ multimodars/ccta/adjust_ccta.py
@@ -213,31 +213,6 @@ def debug_plot(
     print("\nShowing Scene 1: Mesh with colored points")
     print("Colors: Yellow=Aorta, Blue=RCA, Green=LCA, Red=Removed")
     scene1.show()
-    
-    # # Scene 2: Points only (clearer view)
-    # scene2_geoms = []
-    
-    # if aortic_points:
-    #     scene2_geoms.append(PointCloud(aortic_array, colors=aortic_colors))
-    # if rca_points:
-    #     scene2_geoms.append(PointCloud(rca_array, colors=rca_colors))
-    # if lca_points:
-    #     scene2_geoms.append(PointCloud(lca_array, colors=lca_colors))
-    # if all_removed:
-    #     scene2_geoms.append(PointCloud(removed_array, colors=removed_colors))
-    
-    # # Add centerlines
-    # if cl_rca is not None:
-    #     scene2_geoms.append(PointCloud(rca_centerline_points, colors=[0, 100, 200, 255]))
-    # if cl_lca is not None:
-    #     scene2_geoms.append(PointCloud(lca_centerline_points, colors=[0, 150, 0, 255]))
-    # if cl_aorta is not None:
-    #     scene2_geoms.append(PointCloud(aorta_centerline_points, colors=[200, 200, 0, 255]))
-    
-    # scene2 = trimesh.Scene(scene2_geoms)
-    
-    # print("\nShowing Scene 2: Points only (clearer view)")
-    # scene2.show()
 
 
 def prepare_faces_for_rust(mesh: trimesh.Trimesh, *, points=None, face_indices=None, tol: float = 1e-6):
@@ -327,10 +302,8 @@ def plot_anomalous_region(results, centerline):
     
     print(f"\n=== ANOMALOUS REGION VISUALIZATION ===")
     
-    # Create scene
     scene_geoms = []
     
-    # Create point clouds for each group
     if results['proximal_points']:
         proximal_array = np.array(results['proximal_points'])
         proximal_cloud = PointCloud(proximal_array, colors=[0, 0, 255, 255])  # Blue
@@ -446,7 +419,7 @@ def compare_centerline_scaling(
     
     if centerline is not None:
         scene2.add_geometry(PointCloud(centerline_points, colors=centerline_colors))
-    
+
     print("Showing Scene 2: Scaled mesh with RCA region (red) and centerline (cyan)")
     scene2.show()
     
@@ -469,3 +442,286 @@ def compare_centerline_scaling(
     
     print("Showing Scene 3: Side-by-side comparison (Blue=Original, Orange=Scaled)")
     scene3.show()
+
+
+def find_distal_and_proximal_scaling(
+    frames,
+    results: dict,
+    dist_range: int=3,
+    prox_range: int=2,
+    debug_plot: bool=False,
+    percentile_lower=25,
+    percentile_upper=75,
+    anomalous=True,
+) -> Tuple[list, list]:
+    n_anomalous = len(results['anomalous_points'])
+    n_frames = len(frames)
+    dist_ratio = dist_range / n_frames
+    prox_ratio = prox_range / n_frames
+    n_distal_anomalous = int(np.ceil(n_anomalous * dist_ratio))
+    n_proximal_anomalous = int(np.ceil(n_anomalous * prox_ratio))
+
+    frame_points = [(p.x, p.y, p.z) for f in frames for p in f.lumen.points]
+    frame_points_dist = [(p.x, p.y, p.z) for f in frames[-dist_range:] for p in f.lumen.points]
+    frame_points_prox = [(p.x, p.y, p.z) for f in frames[0:prox_range] for p in f.lumen.points]
+
+    # calculate overal centroid of frame_points_dist
[…]
