### 2025-12-17 4d9b743
add cleanup function and converter to inputdata
