+++ src/lib.rs
@@ -1,3 +1,5 @@
+// #![deny(clippy::unwrap_used)]
+
 mod ccta;
 mod intravascular;
 
@@ -12,6 +14,7 @@ use crate::ccta::binding::ccta_py::{
     remove_occluded_points_ray_triangle,
     adjust_diameter_centerline_morphing_simple,
     find_points_by_cl_region,
+    clean_outlier_points,
 };
 
 use pyo3::prelude::*;
@@ -46,6 +49,7 @@ fn multimodars(_py: Python, m: pyo3::prelude::Bound<'_, PyModule>) -> PyResult<(
     m.add_function(wrap_pyfunction!(remove_occluded_points_ray_triangle, m.clone())?)?;
     m.add_function(wrap_pyfunction!(adjust_diameter_centerline_morphing_simple, m.clone())?)?;
     m.add_function(wrap_pyfunction!(find_points_by_cl_region, m.clone())?)?;
+    m.add_function(wrap_pyfunction!(clean_outlier_points, m.clone())?)?;
 
     // Updated class registration
     m.add_class::<PyInputData>()?;
