### 2025-12-18 95cfaad
add stable rust implementation for finding scaling factors
