+++ src/lib.rs
@@ -15,6 +15,7 @@ use crate::ccta::binding::ccta_py::{
     adjust_diameter_centerline_morphing_simple,
     find_points_by_cl_region,
     clean_outlier_points,
+    find_proximal_distal_scaling,
 };
 
 use pyo3::prelude::*;
@@ -50,6 +51,7 @@ fn multimodars(_py: Python, m: pyo3::prelude::Bound<'_, PyModule>) -> PyResult<(
     m.add_function(wrap_pyfunction!(adjust_diameter_centerline_morphing_simple, m.clone())?)?;
     m.add_function(wrap_pyfunction!(find_points_by_cl_region, m.clone())?)?;
     m.add_function(wrap_pyfunction!(clean_outlier_points, m.clone())?)?;
+    m.add_function(wrap_pyfunction!(find_proximal_distal_scaling, m.clone())?)?;
 
     // Updated class registration
     m.add_class::<PyInputData>()?;
