+++ multimodars/ccta/adjust_ccta.py
@@ -175,50 +175,47 @@ def debug_plot(
     print(f"RCA reassigned points (red): {len(rca_removed_points)}")
     print(f"LCA reassigned points (red): {len(lca_removed_points)}")
     
-    scenes = []
-    
-    # Scene 1: Mesh with all points
-    scene1_geoms = []
+    scene_geoms = []
     
     # Add mesh (semi-transparent)
     mesh_visual = mesh.copy()
     mesh_visual.visual.face_colors = [200, 200, 200, 100]  # Semi-transparent gray
-    scene1_geoms.append(mesh_visual)
+    scene_geoms.append(mesh_visual)
     
     if aortic_points:
         aortic_array = np.array(aortic_points, dtype=np.float64)
         aortic_colors = np.tile([255, 255, 0, 255], (len(aortic_points), 1))  # Yellow
-        scene1_geoms.append(PointCloud(aortic_array, colors=aortic_colors))
+        scene_geoms.append(PointCloud(aortic_array, colors=aortic_colors))
     
     if rca_points:
         rca_array = np.array(rca_points, dtype=np.float64)
         rca_colors = np.tile([0, 0, 255, 255], (len(rca_points), 1))  # Blue
-        scene1_geoms.append(PointCloud(rca_array, colors=rca_colors))
+        scene_geoms.append(PointCloud(rca_array, colors=rca_colors))
     
     if lca_points:
         lca_array = np.array(lca_points, dtype=np.float64)
         lca_colors = np.tile([0, 255, 0, 255], (len(lca_points), 1))  # Green
-        scene1_geoms.append(PointCloud(lca_array, colors=lca_colors))
+        scene_geoms.append(PointCloud(lca_array, colors=lca_colors))
     
     all_removed = rca_removed_points + lca_removed_points
     if all_removed:
         removed_array = np.array(all_removed, dtype=np.float64)
         removed_colors = np.tile([255, 0, 0, 255], (len(all_removed), 1))  # Red
-        scene1_geoms.append(PointCloud(removed_array, colors=removed_colors))
+        scene_geoms.append(PointCloud(removed_array, colors=removed_colors))
     
     if cl_rca is not None:
         rca_centerline_points = np.array([(p.contour_point.x, p.contour_point.y, p.contour_point.z) for p in cl_rca.points], dtype=np.float64)
-        scene1_geoms.append(PointCloud(rca_centerline_points, colors=[0, 100, 200, 255]))  # Dark blue
+        scene_geoms.append(PointCloud(rca_centerline_points, colors=[0, 100, 200, 255]))  # Dark blue
     
     if cl_lca is not None:
         lca_centerline_points = np.array([(p.contour_point.x, p.contour_point.y, p.contour_point.z) for p in cl_lca.points], dtype=np.float64)
-        scene1_geoms.append(PointCloud(lca_centerline_points, colors=[0, 150, 0, 255]))  # Dark green
+        scene_geoms.append(PointCloud(lca_centerline_points, colors=[0, 150, 0, 255]))  # Dark green
     
     if cl_aorta is not None:
         aorta_centerline_points = np.array([(p.contour_point.x, p.contour_point.y, p.contour_point.z) for p in cl_aorta.points], dtype=np.float64)
-        scene1_geoms.append(PointCloud(aorta_centerline_points, colors=[200, 200, 0, 255]))  # Dark yellow
+        scene_geoms.append(PointCloud(aorta_centerline_points, colors=[200, 200, 0, 255]))  # Dark yellow
     
-    scene1 = trimesh.Scene(scene1_geoms)
+    scene1 = trimesh.Scene(scene_geoms)
     
[…]
