+++ src/ccta/adjust_mesh/scale_coronary.rs
@@ -1,14 +1,129 @@
+use core::f64;
+use std::collections::HashSet;
 use crate::intravascular::io::input::{CenterlinePoint, Centerline};
 use crate::intravascular::io::geometry::Frame;
 
-#[allow(dead_code)]
 pub fn centerline_based_diameter_optimization(
+    anomalous_points: &[(f64, f64, f64)],
+    n_proximal: usize,
+    n_distal: usize,
     centerline: &Centerline,
-    points: &[(f64, f64, f64)],
+    proximal_reference: &[(f64, f64, f64)],
+    distal_reference: &[(f64, f64, f64)],
+) -> (f64, f64) {
+    let (proximal_points, anomalous_points_new) = find_region_points(&anomalous_points, &proximal_reference, n_proximal);
+    let (distal_points, _) = find_region_points(&anomalous_points_new, &distal_reference, n_distal);
+
+    let start = -2.0f64;
+    let end =  2.0f64;
+    let step = 0.1f64;
+    let steps = ((end - start) / step).round() as i32; // (4.0 / 0.1) => 40
+    
+    let mut min_dist_proximal = f64::MAX;
+    let mut min_dist_distal = f64::MAX;
+    let mut prox_scaling_best = f64::MAX;
+    let mut dist_scaling_best = f64::MAX;
+    
+    for i in 0..=steps {
+        let x = start + i as f64 * step;
+        let temp_points = centerline_based_diameter_morphing(centerline, &proximal_points, x);
+
+        let dist = symmetric_nn_distance(&proximal_reference, &temp_points);
+
+        if dist < min_dist_proximal {
+            min_dist_proximal = dist;
+            prox_scaling_best = x;
+        }
+    }
+    for i in 0..=steps {
+        let x = start + i as f64 * step;
+        let temp_points = centerline_based_diameter_morphing(centerline, &distal_points, x);
+
+        let dist = symmetric_nn_distance(&distal_reference, &temp_points);
+
+        if dist < min_dist_distal {
+            min_dist_distal = dist;
+            dist_scaling_best = x;
+        }
+    }
+    (prox_scaling_best, dist_scaling_best)
+}
+
+fn find_region_points(
+    anomalous_points: &[(f64, f64, f64)],
     reference_points: &[(f64, f64, f64)],
-) -> f64 {
-    let initial_distance = todo!("function that calcualtes distance between the two sets");
-    todo!()
+    n_points: usize,
+) -> (Vec<(f64, f64, f64)>, Vec<(f64, f64, f64)>) {
+    if anomalous_points.is_empty() || reference_points.is_empty() || n_points == 0 {
+        return (Vec::new(), anomalous_points.to_vec());
+    }
+
+    let mut indexed_dists: Vec<(usize, f64)> = anomalous_points
+        .iter()
+        .enumerate()
+        .map(|(i, a_pt)| {
+            let min_sq = reference_points
+                .iter()
+                .map(|r_pt| calculate_squared_distance(a_pt, r_pt))
+                .fold(f64::INFINITY, |m, v| if v < m { v } else { m });
+            (i, min_sq)
+        })
+        .collect();
+
+    indexed_dists.sort_by(|(i1, d1), (i2, d2)| {
+        d1.partial_cmp(d2)
+            .unwrap_or(std::cmp::Ordering::Equal)
+            .then_with(|| i1.cmp(i2))
+    });
+
+    let take = n_points.min(anomalous_points.len());
+    let selected_slice = &indexed_dists[..take];
+
+    let selected_indices: HashSet<usize> = selected_slice.iter().map(|(i, _)| *i).collect();
+
+    let selected_points: Vec<(f64, f64, f64)> = selected_slice
+        .iter()
[…]
