### 2025-12-18 4155dd2
find reference points for aortic wall
