### 2025-12-19 f6faf81
add adjust aortic scaling functionality
