+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "multimodars"
-version = "0.1.1"
+version = "0.1.2"
 edition = "2021"
 license = "MIT"
 
@@ -9,7 +9,7 @@ name = "multimodars"
 crate-type = ["cdylib"]
 
 [dependencies]
-pyo3 = "0.27.1"
+pyo3 = "0.27.2"
 csv = "1.4.0"
 serde = { version = "1.0.228", features = ["derive"]}
 image = "0.25.9"
