### 2026-01-07 47b9c0b
build adjacenecy maps
