+++ test_adjust_ccta.py
@@ -31,110 +31,120 @@ if __name__ == "__main__":
         control_plot=True,
     )
 
-    # Now scale just the RCA region using centerline-based morphing
     mesh = results["mesh"]
-    rca_points = results["rca_points"]
-    rca_cl_raw = np.genfromtxt("data/centerline_rca.csv", delimiter=",")
-    rca_centerline = mm.numpy_to_centerline(rca_cl_raw)
-    aorta_cl_raw = np.genfromtxt("data/centerline_aorta.csv", delimiter=",")
-    ao_centerline = mm.numpy_to_centerline(aorta_cl_raw)
+    print(mesh.faces[0])
+    print(mesh.vertices[0])
 
-    print(f"\n=== CENTERLINE-BASED RCA MORPHING ===")
-    print(f"Original mesh has {len(mesh.vertices)} vertices")
-    print(f"RCA region has {len(rca_points)} vertices")
+    adj_map = mm.build_adjacency_map(mesh.faces.tolist())
+    neighbor_indices = adj_map[0]
+    print(neighbor_indices)
+    neighbor_coords = mesh.vertices[list(neighbor_indices)]
+    print(neighbor_coords)
 
-    # Expand RCA by 2mm using centerline-based morphing
-    scaled_mesh = scale_region_centerline_morphing(
-        mesh=mesh,
-        region_points=rca_points,
-        centerline=rca_centerline,
-        diameter_adjustment_mm=0.5,  # Positive to expand
-    )
+    # # Now scale just the RCA region using centerline-based morphing
+    # mesh = results["mesh"]
+    # rca_points = results["rca_points"]
+    # rca_cl_raw = np.genfromtxt("data/centerline_rca.csv", delimiter=",")
+    # rca_centerline = mm.numpy_to_centerline(rca_cl_raw)
+    # aorta_cl_raw = np.genfromtxt("data/centerline_aorta.csv", delimiter=",")
+    # ao_centerline = mm.numpy_to_centerline(aorta_cl_raw)
 
-    # Create comparison plot
-    compare_centerline_scaling(
-        original_mesh=mesh,
-        scaled_mesh=scaled_mesh,
-        region_points=rca_points,
-        centerline=rca_centerline,
-    )
+    # print(f"\n=== CENTERLINE-BASED RCA MORPHING ===")
+    # print(f"Original mesh has {len(mesh.vertices)} vertices")
+    # print(f"RCA region has {len(rca_points)} vertices")
 
-    # Return both the original and scaled mesh for further use
-    results["scaled_mesh"] = scaled_mesh
+    # # Expand RCA by 2mm using centerline-based morphing
+    # scaled_mesh = scale_region_centerline_morphing(
+    #     mesh=mesh,
+    #     region_points=rca_points,
+    #     centerline=rca_centerline,
+    #     diameter_adjustment_mm=0.5,  # Positive to expand
+    # )
 
-    rest, _ = mm.from_file_singlepair("data/ivus_rest", write_obj=False)
+    # # Create comparison plot
+    # compare_centerline_scaling(
+    #     original_mesh=mesh,
+    #     scaled_mesh=scaled_mesh,
+    #     region_points=rca_points,
+    #     centerline=rca_centerline,
+    # )
 
-    aligned, resampled_cl = mm.align_combined(
-        rca_centerline,
-        rest,
-        (12.2605, -201.3643, 1751.0554),
-        (11.7567, -202.1920, 1754.7975),
-        (15.6605, -202.1920, 1749.9655),
-        results["rca_points"],
-        angle_range_deg=10.0,
-        write=True,
-        watertight=True,
-        output_dir="test",
-    )
+    # # Return both the original and scaled mesh for further use
+    # results["scaled_mesh"] = scaled_mesh
 
[…]
