+++ test_adjust_ccta.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 import sys
 import os
+import timeit
 
 # Add the current directory to Python path so we can import multimodars
 sys.path.insert(0, os.path.dirname(__file__))
@@ -19,7 +20,7 @@ if __name__ == "__main__":
     import trimesh
     import multimodars as mm
 
-    results = label_geometry(
+    results, _ = label_geometry(
         path_ccta_geometry="data/NARCO_119.stl",
         path_centerline_aorta="data/centerline_aorta.csv",
         path_centerline_rca="data/centerline_rca.csv",
@@ -35,116 +36,120 @@ if __name__ == "__main__":
     print(mesh.faces[0])
     print(mesh.vertices[0])
 
+    print(results['rca_points'][:2])
+
+    # exec_time = timeit.timeit("mm.build_adjacency_map(mesh.faces.tolist())", globals=globals(), number=10)
+    # print(f"Adjacency map build time over 10 runs: {exec_time:.4f} seconds")
     adj_map = mm.build_adjacency_map(mesh.faces.tolist())
     neighbor_indices = adj_map[0]
     print(neighbor_indices)
     neighbor_coords = mesh.vertices[list(neighbor_indices)]
     print(neighbor_coords)
 
-    # # Now scale just the RCA region using centerline-based morphing
-    # mesh = results["mesh"]
-    # rca_points = results["rca_points"]
-    # rca_cl_raw = np.genfromtxt("data/centerline_rca.csv", delimiter=",")
-    # rca_centerline = mm.numpy_to_centerline(rca_cl_raw)
-    # aorta_cl_raw = np.genfromtxt("data/centerline_aorta.csv", delimiter=",")
-    # ao_centerline = mm.numpy_to_centerline(aorta_cl_raw)
+    # Now scale just the RCA region using centerline-based morphing
+    mesh = results["mesh"]
+    rca_points = results["rca_points"]
+    rca_cl_raw = np.genfromtxt("data/centerline_rca.csv", delimiter=",")
+    rca_centerline = mm.numpy_to_centerline(rca_cl_raw)
+    aorta_cl_raw = np.genfromtxt("data/centerline_aorta.csv", delimiter=",")
+    ao_centerline = mm.numpy_to_centerline(aorta_cl_raw)
 
-    # print(f"\n=== CENTERLINE-BASED RCA MORPHING ===")
-    # print(f"Original mesh has {len(mesh.vertices)} vertices")
-    # print(f"RCA region has {len(rca_points)} vertices")
+    print(f"\n=== CENTERLINE-BASED RCA MORPHING ===")
+    print(f"Original mesh has {len(mesh.vertices)} vertices")
+    print(f"RCA region has {len(rca_points)} vertices")
 
-    # # Expand RCA by 2mm using centerline-based morphing
-    # scaled_mesh = scale_region_centerline_morphing(
-    #     mesh=mesh,
-    #     region_points=rca_points,
-    #     centerline=rca_centerline,
-    #     diameter_adjustment_mm=0.5,  # Positive to expand
-    # )
+    # Expand RCA by 2mm using centerline-based morphing
+    scaled_mesh = scale_region_centerline_morphing(
+        mesh=mesh,
+        region_points=rca_points,
+        centerline=rca_centerline,
+        diameter_adjustment_mm=0.5,  # Positive to expand
+    )
 
-    # # Create comparison plot
-    # compare_centerline_scaling(
-    #     original_mesh=mesh,
-    #     scaled_mesh=scaled_mesh,
-    #     region_points=rca_points,
-    #     centerline=rca_centerline,
-    # )
+    # Create comparison plot
+    compare_centerline_scaling(
+        original_mesh=mesh,
+        scaled_mesh=scaled_mesh,
[…]
