+++ multimodars/ccta/adjust_ccta.py
@@ -16,7 +16,7 @@ def label_geometry(
     bounding_sphere_radius_mm: float = 3.0,
     tolerance_float: float = 1e-6,
     control_plot: bool = True,
-):
+) -> Tuple[dict, Tuple[any, any, any]]:
     import multimodars as mm
     from multimodars.io.read_geometrical import read_mesh
     
@@ -73,7 +73,7 @@ def label_geometry(
     
     if anomalous_rca:
         print("Applying occlusion removal for anomalous RCA...")
-        rca_faces_for_rust = prepare_faces_for_rust(mesh, points=rca_points_found, tol=tolerance_float)
+        rca_faces_for_rust = _prepare_faces_for_rust(mesh, points=rca_points_found, tol=tolerance_float)
         # Rust implementation, that creates ray between aortic and coronary centerline, and
         # removes faces if 3 consecutive faces are "pierced" by the ray
         final_rca_points_found = mm.remove_occluded_points_ray_triangle(
@@ -90,7 +90,7 @@ def label_geometry(
 
     if anomalous_lca:
         print("Applying occlusion removal for anomalous LCA...")
-        lca_faces_for_rust = prepare_faces_for_rust(mesh, points=lca_points_found, tol=tolerance_float)
+        lca_faces_for_rust = _prepare_faces_for_rust(mesh, points=lca_points_found, tol=tolerance_float)
         final_lca_points_found = mm.remove_occluded_points_ray_triangle(
             centerline_coronary=cl_lca,
             centerline_aorta=cl_aorta,
@@ -104,36 +104,20 @@ def label_geometry(
         final_lca_points_found = lca_points_found.copy()
 
     print(f"Removing LCA and RCA island points...")
-    aortic_points = find_aortic_points(mesh.vertices, final_rca_points_found, final_lca_points_found)
+    aortic_points = _find_aortic_points(mesh.vertices, final_rca_points_found, final_lca_points_found)
     print(f"length before: {len(final_lca_points_found)}")
     final_lca_points, final_aortic_points = mm.clean_outlier_points(
         final_lca_points_found, 
         aortic_points,
         2.0,
-        0.4) # based on patient data, needs more stable option in the future
+        0.4) # based on patient data, only precleaning anyways, rest done by final_reclassification
     final_rca_points, final_aortic_points = mm.clean_outlier_points(
         final_rca_points_found,
         final_aortic_points,
         2.0,
         0.4)
-    aortic_points = find_aortic_points(mesh.vertices, final_rca_points, final_lca_points)
+    aortic_points = _find_aortic_points(mesh.vertices, final_rca_points, final_lca_points)
     print(f"length after: {len(final_lca_points)}")
-
-    # testing cleaning
-    labels = np.zeros(len(mesh.vertices), dtype=np.uint8)
-    labels
-
-    if control_plot:
-        debug_plot(
-            mesh=mesh,
-            rca_points=final_rca_points,
-            lca_points=final_lca_points,
-            rca_removed_points=rca_removed_points,
-            lca_removed_points=lca_removed_points,
-            cl_rca=cl_rca,
-            cl_lca=cl_lca,
-            cl_aorta=cl_aorta
-        )
 
     results = {
         'mesh': mesh,
@@ -143,11 +127,76 @@ def label_geometry(
         'rca_removed_points': rca_removed_points,
         'lca_removed_points': lca_removed_points
     }
+
+    # final reclassification based on adjacency map
+    print("Applying final reclassification based on adjacency map...")
+    new_results = _final_reclassification(results)
+
+    if control_plot:
[…]
