+++ examples/test_adjust_ccta.py
@@ -1,0 +1,157 @@
+#!/usr/bin/env python3
+import sys
+import os
+
+# Add the current directory to Python path so we can import multimodars
+sys.path.insert(0, os.path.dirname(__file__))
+
+from multimodars.ccta.adjust_ccta import (
+    label_geometry,
+    scale_region_centerline_morphing,
+    compare_centerline_scaling,
+    label_anomalous_region,
+    find_distal_and_proximal_scaling,
+    find_aorta_scaling,
+)
+
+if __name__ == "__main__":
+    import numpy as np
+    import trimesh
+    import multimodars as mm
+
+    results, _ = label_geometry(
+        path_ccta_geometry="data/NARCO_119.stl",
+        path_centerline_aorta="data/centerline_aorta.csv",
+        path_centerline_rca="data/centerline_rca.csv",
+        path_centerline_lca="data/centerline_lca.csv",
+        bounding_sphere_radius_mm=3.5,
+        n_points_intramural=200,
+        anomalous_rca=True,
+        anomalous_lca=False,
+        control_plot=True,
+    )
+
+    mesh = results["mesh"]
+    print(mesh.faces[0])
+    print(mesh.vertices[0])
+
+    print(results['rca_points'][:2])
+
+    adj_map = mm.build_adjacency_map(mesh.faces.tolist())
+    neighbor_indices = adj_map[0]
+    print(neighbor_indices)
+    neighbor_coords = mesh.vertices[list(neighbor_indices)]
+    print(neighbor_coords)
+
+    # Now scale just the RCA region using centerline-based morphing
+    mesh = results["mesh"]
+    rca_points = results["rca_points"]
+    rca_cl_raw = np.genfromtxt("data/centerline_rca.csv", delimiter=",")
+    rca_centerline = mm.numpy_to_centerline(rca_cl_raw)
+    aorta_cl_raw = np.genfromtxt("data/centerline_aorta.csv", delimiter=",")
+    ao_centerline = mm.numpy_to_centerline(aorta_cl_raw)
+
+    print(f"\n=== CENTERLINE-BASED RCA MORPHING ===")
+    print(f"Original mesh has {len(mesh.vertices)} vertices")
+    print(f"RCA region has {len(rca_points)} vertices")
+
+    # Expand RCA by 2mm using centerline-based morphing
+    scaled_mesh = scale_region_centerline_morphing(
+        mesh=mesh,
+        region_points=rca_points,
+        centerline=rca_centerline,
+        diameter_adjustment_mm=0.5,  # Positive to expand
+    )
+
+    # Create comparison plot
+    compare_centerline_scaling(
+        original_mesh=mesh,
+        scaled_mesh=scaled_mesh,
+        region_points=rca_points,
+        centerline=rca_centerline,
+    )
+
+    # Return both the original and scaled mesh for further use
+    results["scaled_mesh"] = scaled_mesh
+
+    rest, _ = mm.from_file_singlepair("data/ivus_rest", write_obj=False)
+
+    print("+_+_+_+_+_+_+_+_+_+_+_+_+__+_+_+_+_+_+_+_+_+__")
+    print(rest.geom_a.frames[0].lumen)
+
+    aligned, resampled_cl = mm.align_combined(
+        rca_centerline,
+        rest,
+        (12.2605, -201.3643, 1751.0554),
+        (11.7567, -202.1920, 1754.7975),
+        (15.6605, -202.1920, 1749.9655),
+        results["rca_points"],
+        angle_range_deg=10.0,
+        write=True,
+        watertight=True,
+        output_dir="test",
+    )
+    print("+_+_+_+_+_+_+_+_+_+_+_+_+__+_+_+_+_+_+_+_+_+__")
+    print(aligned.geom_a.frames[0].lumen)
+
[…]
