### 2026-01-10 c371224
prepare examples to be moved to releases
