### 2026-01-10 ead393f
reduce fixture data size ensuring tests still pass
