### 2026-01-10 0c65e74
remove examples from package
