+++ examples/blender_script.py
@@ -1,431 +1,0 @@
-# overhauled_blender_4x_animation.py
-# Compatible with Blender 4.x (4.4)
-import bpy
-import os
-import math
-import bmesh
-from mathutils import Vector, Matrix, Quaternion
-
-# -----------------------------
-# Configuration - update this
-# -----------------------------
-obj_directory = r"D:/00_coding/multimoda-rs/data/output/rest"
-object_end_frame = 60
-z_orbit_frames = 240
-y_orbit_frames = 240
-total_frames = object_end_frame + z_orbit_frames + y_orbit_frames
-frame_rate = 60
-
-# Lighting config
-light_radius = 10.0
-num_lights = 12
-light_energy = 1000.0   # Blender 4's lights are more physically scaled; increase if too dim.
-
-# -----------------------------
-# Utilities
-# -----------------------------
-def safe_print(*args, **kwargs):
-    try:
-        print(*args, **kwargs)
-    except Exception:
-        pass
-
-def clean_scene():
-    """Remove all objects, lights, cameras and collections (fresh start)."""
-    bpy.ops.object.select_all(action='SELECT')
-    bpy.ops.object.delete(use_global=False)
-    # Remove orphan collections (optional, keep the master collection)
-    for coll in list(bpy.data.collections):
-        if coll.name != "Collection":
-            try:
-                bpy.data.collections.remove(coll)
-            except Exception:
-                pass
-
-def get_new_objects(before_set):
-    """Return list of newly-created objects compared to before_set."""
-    return [o for o in bpy.data.objects if o.name not in before_set]
-
-# -----------------------------
-# Normals flipping (BMesh)
-# -----------------------------
-def flip_normals_for_object(obj):
-    """Use BMesh to flip normals robustly for a single mesh object."""
-    if obj.type != 'MESH' or not obj.data.polygons:
-        return
-
-    mesh = obj.data
-    # get a copy of one polygon normal before flipping (world space)
-    try:
-        normal_before = (obj.matrix_world @ mesh.polygons[0].normal).copy()
-    except Exception:
-        normal_before = None
-
-    bm = bmesh.new()
-    try:
-        bm.from_mesh(mesh)
-        # Ensure lookup table and faces exist
-        bm.faces.ensure_lookup_table()
-        if len(bm.faces) > 0:
-            # Reverse all faces (flip winding -> flips normals)
-            bmesh.ops.reverse_faces(bm, faces=bm.faces)
-            # Write back and update
-            bm.to_mesh(mesh)
-            mesh.update()
-    finally:
-        bm.free()
-
-    try:
-        normal_after = (obj.matrix_world @ mesh.polygons[0].normal).copy()
-    except Exception:
-        normal_after = None
-
-    safe_print(f"[Normals] {obj.name} before: {normal_before}, after: {normal_after}")
-
-def flip_normals_for_selected():
-    """Flip normals for all selected mesh objects (safe)."""
-    for obj in list(bpy.context.selected_objects):
-        flip_normals_for_object(obj)
-
-# -----------------------------
-# Import helpers (Blender 4.x)
-# -----------------------------
-def import_obj_file(filepath):
-    """
-    Import a single OBJ using the new Blender 4.x operator.
-    Returns list of objects created by the import.
-    """
-    if not os.path.exists(filepath):
-        safe_print(f"[Import] File not found: {filepath}")
-        return []
-
-    before = {o.name for o in bpy.data.objects}
-    try:
-        # Call the new OBJ importer (Blender 4.x)
-        # The operator is known as bpy.ops.wm.obj_import in 4.x
-        bpy.ops.wm.obj_import(filepath=filepath)
-    except Exception as e:
[…]
