+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "multimodars"
-version = "0.1.2"
+version = "0.2.0"
 authors = [
     { name = "Anselm W. Stark"}
 ]
