### 2026-01-10 f310f60
change version control pyproject.toml and CI.yml
