+++ src/ccta/binding/ccta_py.rs
@@ -1,16 +1,16 @@
 // src/ccta/binding/label_py.rs
 use std::collections::{HashMap, HashSet};
 
-use crate::intravascular::binding::classes::{PyCenterline, PyFrame};
 use crate::ccta::adjust_mesh::label_coronary::find_centerline_bounded_points;
-use crate::ccta::adjust_mesh::label_coronary::{Triangle, remove_occluded_points_ray_triangle_rust};
-use crate::ccta::adjust_mesh::scale_coronary::{
-    centerline_based_diameter_morphing, 
-    find_points_by_cl_region_rs, 
-    clean_up_non_section_points,
-    centerline_based_diameter_optimization,
-    centerline_based_aortic_diameter_optimization,
+use crate::ccta::adjust_mesh::label_coronary::{
+    remove_occluded_points_ray_triangle_rust, Triangle,
 };
+use crate::ccta::adjust_mesh::scale_coronary::{
+    centerline_based_aortic_diameter_optimization, centerline_based_diameter_morphing,
+    centerline_based_diameter_optimization, clean_up_non_section_points,
+    find_points_by_cl_region_rs,
+};
+use crate::intravascular::binding::classes::{PyCenterline, PyFrame};
 use pyo3::prelude::*;
 
 /// Finds points that are bounded by spheres along a coronary vessel centerline.
@@ -26,7 +26,7 @@ use pyo3::prelude::*;
 ///
 /// Example:
 ///     >>> import multimodars as mm
-///     >>> 
+///     >>>
 ///     >>> # Load centerline and point cloud  
 ///     >>> centerline = mm.load_centerline("path/to/centerline.json")
 ///     >>> points = [(0.0, 0.0, 0.0), (1.0, 0.0, 0.0), ...]  # or mesh.vertices.tolist()
@@ -41,19 +41,15 @@ pub fn find_centerline_bounded_points_simple(
     radius: f64,
 ) -> PyResult<Vec<(f64, f64, f64)>> {
     let rust_centerline = centerline.to_rust_centerline();
-    
+
     // Call Rust function directly - no complex conversions needed
-    let result_points = find_centerline_bounded_points(
-        rust_centerline, 
-        &points, 
-        radius
-    );
-    
+    let result_points = find_centerline_bounded_points(rust_centerline, &points, radius);
+
     Ok(result_points)
 }
 
 /// Remove occluded / overlapping points seen from the centerline.
-/// 
+///
 /// For each centerline point, we keep at most one point per angular bin
 /// (quantized theta/phi) — the point with the smallest radius (closest to the center).
 ///
@@ -77,13 +73,13 @@ pub fn remove_occluded_points_ray_triangle(
 ) -> PyResult<Vec<(f64, f64, f64)>> {
     let rust_centerline_coronary = centerline_coronary.to_rust_centerline();
     let rust_centerline_aorta = centerline_aorta.to_rust_centerline();
-    
+
     // Convert faces to Triangle format
     let triangles: Vec<Triangle> = faces
         .into_iter()
         .map(|(v0, v1, v2)| Triangle::new(v0, v1, v2))
         .collect();
-    
+
     let result = remove_occluded_points_ray_triangle_rust(
         &rust_centerline_coronary,
         &rust_centerline_aorta,
@@ -95,12 +91,12 @@ pub fn remove_occluded_points_ray_triangle(
 }
 
 /// Adjust centerline-based diameter by morphing points outward or inward.
-/// 
+///
 /// Args:
 ///     centerline: PyCenterline object representing the vessel centerline
 ///     points: List of (x, y, z) tuples containing point coordinates
 ///     diameter_adjustment_mm: Amount to adjust diameter (positive to expand, negative to contract)
-/// 
+///
 /// Returns:
[…]
