+++ src/ccta/adjust_mesh/scale_coronary.rs
@@ -1,8 +1,8 @@
+use super::calculate_squared_distance;
+use crate::intravascular::io::geometry::Frame;
+use crate::intravascular::io::input::{Centerline, CenterlinePoint};
 use core::f64;
 use std::collections::HashSet;
-use crate::intravascular::io::input::{CenterlinePoint, Centerline};
-use crate::intravascular::io::geometry::Frame;
-use super::calculate_squared_distance;
 
 pub fn centerline_based_aortic_diameter_optimization(
     intramural_points: &[(f64, f64, f64)],
@@ -10,13 +10,13 @@ pub fn centerline_based_aortic_diameter_optimization(
     centerline: &Centerline,
 ) -> f64 {
     let start = -2.0f64;
-    let end =  2.0f64;
+    let end = 2.0f64;
     let step = 0.1f64;
     let steps = ((end - start) / step).round() as i32; // (4.0 / 0.1) => 40
-    
+
     let mut min_dist = f64::MAX;
     let mut scaling_best = f64::MAX;
-    
+
     for i in 0..=steps {
         let x = start + i as f64 * step;
         let temp_points = centerline_based_diameter_morphing(centerline, &intramural_points, x);
@@ -39,19 +39,20 @@ pub fn centerline_based_diameter_optimization(
     proximal_reference: &[(f64, f64, f64)],
     distal_reference: &[(f64, f64, f64)],
 ) -> (f64, f64) {
-    let (proximal_points, anomalous_points_new) = find_region_points(&anomalous_points, &proximal_reference, n_proximal);
+    let (proximal_points, anomalous_points_new) =
+        find_region_points(&anomalous_points, &proximal_reference, n_proximal);
     let (distal_points, _) = find_region_points(&anomalous_points_new, &distal_reference, n_distal);
 
     let start = -2.0f64;
-    let end =  2.0f64;
+    let end = 2.0f64;
     let step = 0.1f64;
     let steps = ((end - start) / step).round() as i32; // (4.0 / 0.1) => 40
-    
+
     let mut min_dist_proximal = f64::MAX;
     let mut min_dist_distal = f64::MAX;
     let mut prox_scaling_best = f64::MAX;
     let mut dist_scaling_best = f64::MAX;
-    
+
     for i in 0..=steps {
         let x = start + i as f64 * step;
         let temp_points = centerline_based_diameter_morphing(centerline, &proximal_points, x);
@@ -117,7 +118,13 @@ fn find_region_points(
     let remaining_points: Vec<(f64, f64, f64)> = anomalous_points
         .iter()
         .enumerate()
-        .filter_map(|(i, pt)| if selected_indices.contains(&i) { None } else { Some(*pt) })
+        .filter_map(|(i, pt)| {
+            if selected_indices.contains(&i) {
+                None
+            } else {
+                Some(*pt)
+            }
+        })
         .collect();
 
     (selected_points, remaining_points)
@@ -131,21 +138,23 @@ fn symmetric_nn_distance(a: &[(f64, f64, f64)], b: &[(f64, f64, f64)]) -> f64 {
         return f64::INFINITY;
     }
 
-    let sum_a_to_b: f64 = a.iter()
+    let sum_a_to_b: f64 = a
+        .iter()
         .map(|pa| {
             b.iter()
-             .map(|pb| calculate_squared_distance(pa, pb))
-             .fold(f64::INFINITY, |m, v| if v < m { v } else { m })
+                .map(|pb| calculate_squared_distance(pa, pb))
+                .fold(f64::INFINITY, |m, v| if v < m { v } else { m })
         })
         .sum();
 
[…]
