+++ src/ccta/mod.rs
@@ -1,3 +1,3 @@
+pub mod adjust_mesh;
 pub mod binding;
 pub mod io;
-pub mod adjust_mesh;
