### 2026-01-12 1961fb6
fix documentation error and add dynamic version
