### 2026-01-14 cd94bf9
fix type annotations versioning conflict
