+++ paper/bibliography.bib
@@ -54,15 +54,12 @@
   doi={10.1038/s41598-020-74264-w}
 }
 
-@article{bourantas2013new,
-  title={A new methodology for accurate 3-dimensional coronary artery reconstruction using routine intravascular ultrasound and angiographic data: implications for widespread assessment of endothelial shear stress in humans},
-  author={Bourantas, Christos V and Papafaklis, Michail I and Athanasiou, Lambros and Kalatzis, Fanis G and Naka, Katerina K and Siogkas, Panagiotis K and Takahashi, Saeko and Saito, Shigeru and Fotiadis, Dimitrios I and Feldman, Charles L and others},
-  journal={EuroIntervention},
-  volume={9},
-  number={5},
-  pages={582--93},
-  year={2013},
-  doi={10.4244/EIJV9I5A94}
+@article{ilic2025comprehensive,
+  title={A comprehensive workflow for CCTA and OCT data fusion with 3D printing validation: advancing patient-specific testing environments for percutaneous coronary intervention devices},
+  author={Ilic, Marc and H{\"a}ner, Jonas and Lehmann, Julian and Stark, Anselm W and Illi, Jo{\"e}l and Gr{\"a}ni, Christoph and Bentele, Marc and Baumann-Zumstein, Philine and Busch, Julia and Haeberlin, Andreas},
+  journal={BioMedical Engineering OnLine},
+  year={2025},
+  publisher={Springer}
 }
 
 @article{stark2025automated,
