+++ paper/bibliography.bib
@@ -10,15 +10,6 @@
   doi={10.1016/j.jaccas.2025.104741}
 }
 
-@article{stark2025ivus,
-  title={IVUS Is Superior to OCT for the Visualization of Dynamic Anatomic Vessel Lumen Changes in Anomalous Coronary Artery},
-  author={Stark, Anselm Walter and Bigler, Marius Reto and Kakizaki, Ryota and R{\"a}ber, Lorenz and Gr{\"a}ni, Christoph},
-  journal={Catheterization and Cardiovascular Interventions},
-  year={2025},
-  publisher={Wiley Online Library},
-  doi={10.1002/ccd.31660}
-}
-
 @article{van20103d,
   title={3D fusion of intravascular ultrasound and coronary computed tomography for in-vivo wall shear stress analysis: a feasibility study},
   author={van der Giessen, Alina G and Schaap, Michiel and Gijsen, Frank JH and Groen, Harald C and van Walsum, Theo and Mollet, Nico R and Dijkstra, Jouke and van de Vosse, Frans N and Niessen, Wiro J and de Feyter, Pim J and others},
@@ -54,6 +45,28 @@
   doi={10.1038/s41598-020-74264-w}
 }
 
+@article{bourantas2013new,
+  title={A new methodology for accurate 3-dimensional coronary artery reconstruction using routine intravascular ultrasound and angiographic data: implications for widespread assessment of endothelial shear stress in humans},
+  author={Bourantas, Christos V and Papafaklis, Michail I and Athanasiou, Lambros and Kalatzis, Fanis G and Naka, Katerina K and Siogkas, Panagiotis K and Takahashi, Saeko and Saito, Shigeru and Fotiadis, Dimitrios I and Feldman, Charles L and others},
+  journal={EuroIntervention},
+  volume={9},
+  number={5},
+  pages={582--93},
+  year={2013},
+  doi={10.4244/EIJV9I5A94}
+}
+
+@article{boogers2012automated,
+  title={Automated quantification of coronary plaque with computed tomography: comparison with intravascular ultrasound using a dedicated registration algorithm for fusion-based quantification},
+  author={Boogers, Mark J and Broersen, Alexander and Van Velzen, Jo{\"e}lla E and De Graaf, Fleur R and El-Naggar, Heba M and Kitslaar, Pieter H and Dijkstra, Jouke and Delgado, Victoria and Boersma, Eric and de Roos, Albert and others},
+  journal={European heart journal},
+  volume={33},
+  number={8},
+  pages={1007--1016},
+  year={2012},
+  publisher={The University of Chicago Press}
+}
+
 @article{ilic2025comprehensive,
   title={A comprehensive workflow for CCTA and OCT data fusion with 3D printing validation: advancing patient-specific testing environments for percutaneous coronary intervention devices},
   author={Ilic, Marc and H{\"a}ner, Jonas and Lehmann, Julian and Stark, Anselm W and Illi, Jo{\"e}l and Gr{\"a}ni, Christoph and Bentele, Marc and Baumann-Zumstein, Philine and Busch, Julia and Haeberlin, Andreas},
@@ -63,11 +76,10 @@
 }
 
 @article{stark2025automated,
-  title={Automated IntraVascular UltraSound Image Processing and Quantification of Coronary Artery Anomalies: The AIVUS-CAA software},
-  author={Stark, Anselm W and Mohammadi Kazaj, Pooya and Balzer, Sebastian and Ilic, Marc and Bergamin, Manuel and Kakizaki, Ryota and Giannopoulos, Andreas and Haeberlin, Andreas and R{\"a}ber, Lorenz and Shiri, Isaac and others},
[…]
