### 2026-02-16 #8938 @yungselm
Hi @arfon and Hi @rkurchin

Thank you very much for the reminder and I'm very sorry for the delay.

I have updated the submission as requested and additionally updated the software:
- Manuscript: The manuscript now includes the sections as provided in the example (Summary, Statement of need, State of the field, Software design, Research impact statement and AI usage disclosure).
- Software & versioning: the project has advanced from v0.0.6 (first submission) to v0.2.1. All substantive changes, fixes and new features are documented in `CHANGELOG.md`.

Thanks again for the review and for the clarification, I appreciate your time.

Please let me know if you need any additional information!
