### 2026-02-23 #8938 @rkurchin
Erick, would you be able to take this one?

(and @jonaspleyer, thanks for volunteering! I'll let the handling editor act on this once there is one)
