### 2026-02-27 #8938 @erickmartins
@richardkoehler @crnh @sstadick would any of you have time and be willing to review this submission for JOSS? As you might know, we carry out our checklist-driven reviews here in GitHub issues and follow these [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). Thanks for your time!
