### 2026-02-27 #8938 @richardkoehler
Sure, I'd be happy to review. I've also reviewed for JOSS before so I am familiar with the guidelines. I can probably get it done by next week.
