### 2026-03-02 #8938 @crnh
> [@richardkoehler](https://github.com/richardkoehler) [@crnh](https://github.com/crnh) [@sstadick](https://github.com/sstadick) would any of you have time and be willing to review this submission for JOSS?

@erickmartins I would be willing to review, but do not specifically have experience with cardiac imaging.
