### 2026-03-02 #10131 @crnh
- Multiple developers contributed, but no issues or pull requests from external users are visible.
  - The project has a license, documentation, tests (both for the Python and Rust side), releases and contribution guidelines. The contribution guidelines could be improved.
  - The Python tests passed on my machine.

  - No example code is provided to verify the performance claims in the 'Performance and parallelisation' section (lines 102-104).

  - The documentation very briefly mentions the aim of the tool
  - All functionality is documented to a satisfactory level. However, core functions like `from_file` and `from_array` are difficult to find, because they are located in a `wrappers` submodule. The discoverability of these functions could be improved.

  - The research impact statement refers to a case report in JACC, but does not cite it.

## AI usage disclosure

AI was used for grammar checking.
