### 2026-03-03 #10131 @crnh
I finished the review, you can find my comments below.
Since I was not able to run the examples, I could not verify the functionality of the software and advise a **major revision**.
The paper is clear and well-written, but the authors should at least rewrite the examples to work with the latest version of the software and include all required example data, so that new users can try the software without running into issues right away.

## Code

### Running the examples

I tested the examples with `multimodars` version 0.2.1, which is the latest version available on PyPI.

#### Quickstart example in Readme

- `from_file` should check if all required parameters are present and raise an exception with a clear message if not (the same applies to other wrapper functions). The example code does not seem to specify all required parameters.
- The file `data/centerline_raw.csv` and other example data is distributed as part of the releases, which is unintuitive. This should at least be mentioned in the Readme, but ideally all example data should just be included in the repository.

#### `ivus_to_centerline.ipynb` from the examples folder

- The data folders in the examples folder (e.g. `examples/data/ivus_rest`) are empty, causing `from_file_full` to raise an unclear exception:
  ```
  RuntimeError: Failed to prepare geometries for full processing
  ```
  The required data should be included in the data folder (I now copied it from the repository) and the exception should specify what exactly went wrong.
- Cell 2 loads the example data from different file names than the names used to save them in cell 1 (e.g. `dia_lumen.obj` vs `rest_dia.obj`), which causes the code to raise an exception.
- Rendering the meshes depends on `nbformat`, but this dependency is not listed. The notebook also depends on `scipy`, which is not listed.
- The notebook imports the same dependencies multiple times.
- I was unable to run the stent example, because I could not find the required data in the repository or the releases.
- The data required for the 'Alignment from array' example is not included in the examples archive and had to be downloaded separately from the repository. The output of the example is not visualized.
- The other examples in the notebook run without errors and the visualizations are clear.

#### Tutorial - Intravascular Module (from the documentation)

- It would be very helpful if the tutorial could be downloaded as a Jupyter notebook right from the documentation. The documentation states the notebook should be in the releases, but I could not find it there.
- The tutorial seems to have been written for a version of the software with a different interface, e.g. arguments like `rest_input_path` and `rest_output_path` are not supported by `from_file`, but it does take two parameters named `input_path_a` and `input_path_b`. However, the docstring still mentions the old parameters. Furthermore, other required parameters (e.g. `label`) seem to be missing from the tutorial.

I hope to evaluate the remaining parts of the tutorials once these issues have been addressed.

### Code style and interface design

- The code uses old-style annotations (e.g. `List` instead of `list`, `Union` instead of `|`). These are required if supporting Python 3.8 and older, but since this Python version is EOL, the authors should consider dropping support for it and using modern type annotations.
- The Python code uses star imports, which makes the interface opaque. For example, `PyContour` is explicitly exported in `__init__.py` but imported through a star import.
- The Python code uses imports outside the top level, which is generally discouraged.
- No type stubs are included for the parts of the interface that are implemented in Rust, which makes the interface less clear and prevents type checking of code that uses the library.
  The authors should consider adding stubs for at least the parts of the interface that they consider to be public.
  This would significantly improve the user experience.
  See e.g. https://www.maturin.rs/project_layout.html#adding-python-type-information and https://pyo3.rs/v0.27.1/type-stub.html for more information on how to do this.
- `multimodars` provides a unified interface to multiple functions through a set of wrapper functions, such as `from_file` and `from_array`. This is good, but any parameters that are required by the underlying functions are passed through `**kwargs`, which requires the user to consult the documentation of these functions to understand the meaning of the parameters. The interface could be improved in a few ways:
  - The wrapper functions could explicitly document the meaning of _all_ supported parameters;
  - The wrapper functions could use a combination of `TypedDict` unpacking (to specify which parameters are supported) combined with overloads to specify which parameters are required for which mode. This would improve the user experience by enabling auto-completions and type checking for the parameters of the wrapper functions. More information: https://typing.python.org/en/latest/spec/callables.html#unpack-kwargs and https://typing.python.org/en/latest/spec/overload.html.

### Contributing

- `CONTRIBUTING.md` lacks the following sections that are mentioned in the TOC: Suggesting Enhancements, Your First Code Contribution, Pull Request Process, Coding Style & Tests, Writing Documentation, Where to Get Help.
  These sections should be added to the guidelines or removed from the TOC.
- `CONTRIBUTING.md` does not mention the required code style (probably Black) and docstring style (looks like NumPyDoc).
  The code style can partially be inferred from the pre-commit configuration, but it should be explicitly stated in the contributing guidelines.
- The bug issue template can be improved. It currently asks which browser is used and asks for information about the smartphone running the software, both of which are not relevant for this project.

### Documentation

- The documentation contains phrases like "There are two ways you can use pyradiomics: 1. Install via pip 2. Install from source", which should be removed or updated to match the actual module name.
- https://multimoda-rs.readthedocs.io/en/latest/installation.html mentions a minimum Python version of 3.12, which does not match the minimum of 3.8 in `pyproject.toml`.
- The project combines multiple docstring styles, e.g. NumPyDoc in `from_file`, Google-style in `centerline_to_obj` and no predefined style in `read_mesh`. The documentation would be improved by using a consistent style across the codebase.
- The repository could link more clearly to the documentation by listing it in the 'About' section.
- API documentation for core functions like `from_file` and `from_array` are difficult to find, because they are located in a `wrappers` submodule. The discoverability of these functions could be improved.

### CI

- The CI pipeline only tests the library against Python 3.12, and could be improved by testing with all supported Python versions.

## Paper

### Authors

- Christoph Gräni does not seem to have contributed to the code and the role of this author is unclear.

### Figures

- The text in Figure 2 is too small.

### Citations

- Citations are not formatted consistently. For example, in one place a paper by Stark is cited as "Stark, Anselm Walter", while in another place it is cited as "Stark, Anselm W.".
- DOIs are not provided for citations 1, 4, and 6, but these are available and should be included.

### Language

- Figure 3 caption: 'inital' should be 'initial'.
- Line 104: 'mutliscale' should be 'multiscale'.
- Line 102: 'documentation docstrings' is a pleonasm. Remove one of the two words, or rephrase to 'documentation and docstrings'.
