### 2026-03-03 7cbb1df
update examples in README, fix documentations and add doi to bibliography
