+++ paper/bibliography.bib
@@ -64,7 +64,8 @@
   number={8},
   pages={1007--1016},
   year={2012},
-  publisher={The University of Chicago Press}
+  publisher={The University of Chicago Press},
+  doi={10.1093/eurheartj/ehr465}
 }
 
 @article{ilic2025comprehensive,
@@ -72,7 +73,8 @@
   author={Ilic, Marc and H{\"a}ner, Jonas and Lehmann, Julian and Stark, Anselm W and Illi, Jo{\"e}l and Gr{\"a}ni, Christoph and Bentele, Marc and Baumann-Zumstein, Philine and Busch, Julia and Haeberlin, Andreas},
   journal={BioMedical Engineering OnLine},
   year={2025},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1186/s12938-025-01501-6 }
 }
 
 @article{stark2025automated,
@@ -81,5 +83,6 @@
   journal={Computer Methods and Programs in Biomedicine},
   pages={109065},
   year={2025},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.cmpb.2025.109065}
 }
