### 2026-03-03 93d245b
update documentation with new example data
