+++ docs/tutorial_intravascular.rst
@@ -27,7 +27,7 @@ After pip installing or locally building the package, install it in the familiar
     import multimodars as mm
 
 To run the whole workflow from .csv files the following requirements have to be met.
-Files should be named ``diastolic_contours.csv``, ``systolic_contours.csv``, 
+Files should be named ``diastolic_contours.csv``, ``systolic_contours.csv``,
 ``diastolic_reference_points.csv`` and ``systolic_reference_points.csv`` depending on the required analysis.
 Every file should be structured in the following way (no headers):
 
@@ -44,8 +44,8 @@ Every file should be structured in the following way (no headers):
 +-------+---------+---------+---------+
 
 To acquire meaningful measurement data, the coordinates should be provided in mm or another SI unit instead of pixel values.
-Optionally a record file can be provided `combined_sorted_manual.csv`, which should have the following structure, here the first column should contain the desired frame order and "measurement_1" 
-represent the thickness of the wall between aorta and coronary and "measurement_2" for the thickness between pulmonary artery and coronary (position just for demonstration). This is based on the 
+Optionally a record file can be provided `combined_sorted_manual.csv`, which should have the following structure, here the first column should contain the desired frame order and "measurement_1"
+represent the thickness of the wall between aorta and coronary and "measurement_2" for the thickness between pulmonary artery and coronary (position just for demonstration). This is based on the
 output of the `AIVUS-CAA software <https://github.com/AI-in-Cardiovascular-Medicine/AIVUS-CAA/>`_:
 
 +-----------------+---------------+---------------+---------------+---------------+
@@ -64,73 +64,98 @@ output of the `AIVUS-CAA software <https://github.com/AI-in-Cardiovascular-Medic
 | 47              |  18.78        |       S       |               |               |
 +-----------------+---------------+---------------+---------------+---------------+
 
-This simplifies the workflow, by just providing a directory to automatically process:
+The full workflow comparing rest vs. stress and diastole vs. systole can be run with:
 
 .. code-block:: python
 
-    rest, stress, dia, sys, (rest_logs, stress_logs, dia_logs, sys_logs) = mm.from_file_full(
-        input_path_a="examples/data/ivus_rest",
-        input_path_b="examples/data/ivus_stress",
-        output_path_a="examples/data/output/rest",
-        output_path_b="examples/data/output/stress",
-        output_path_c="examples/data/output/diastole",
-        output_path_d="examples/data/output/systole"
-    )
-
-However the preferred more flexible way is from numpy arrays.
-
-2. Workflow numpy arrays and Finetuning
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Here the geometry can be directly build from arrays with the same structure as before:
-
-+------------+----------+----------+----------+
-| ...        |   ...    |   ...    |   ...    |
-+------------+----------+----------+----------+
-| 771        | 2.4862   |  6.7096  |  24.5370 |
-+------------+----------+----------+----------+
-| 771        | 2.5118   |  6.7017  |  24.5370 |
-+------------+----------+----------+----------+
-| 771        | 2.5370   |  6.6936  |  24.5370 |
-+------------+----------+----------+----------+
-| ...        |   ...    |   ...    |   ...    |
-+------------+----------+----------+----------+
-
[…]
