+++ README.md
@@ -16,9 +16,9 @@
 
 <figure class="epigraph" style="text-align: center; font-style: italic;">
   <blockquote>
-    “One package to fuse them all.”
+    "One package to fuse them all."
   </blockquote>
-  <figcaption>— <cite>The Lord of the Rings (probably)</cite></figcaption>
+  <figcaption>— <cite>The Lord of the Rings (probably)</cite></figcaption>
 </figure>
 
 ---
@@ -26,7 +26,7 @@
 > A high‑performance, Rust‑accelerated toolkit for multi‑modality cardiac image fusion and registration ﮩ٨ـﮩﮩ٨ـ♡ﮩ٨ـﮩﮩ٨ـ.
 
 ---
-# Overview 
+# Overview
 `multimoda-rs` is a high-performance toolkit developed to enable the study of dynamic vessel deformation in coronary artery anomalies (CAAs), where quantifying lumen changes under stress and rest is critical. It addresses the general challenge of aligning and fusing diverse cardiac imaging modalities, such as CCTA, IVUS, OCT, and MRI—into a unified, high‑resolution 3D model. While CCTA provides comprehensive volumetric context, intravascular modalities (IVUS and OCT) offer sub‑millimeter resolution along the vessel lumen, and MRI (LGE) reveals tissue characteristics like scar and edema. This library leverages Rust for computationally intensive registration steps, delivering faster performance than pure Python implementations.
 
 ## Key Features
@@ -75,40 +75,60 @@ maturin develop
 
 
 ## Quickstart Example
-Download the example data from latest realease and add to root. Run the script with the provided test cases, to ensure sufficient set up.
+Download the example data from the latest release and place it in your working directory. A Jupyter Notebook with step-by-step examples is available at `examples/ivus_to_centerline.ipynb`.
+
 ```python
 import multimodars as mm
 import numpy as np
 
 # IVUS pullbacks: full alignment of rest/stress & diastole/systole
 rest, stress, dia, sys, _ = mm.from_file_full(
-    input_path_a="examples/data/ivus_rest",
-    input_path_b="examples/data/ivus_stress",
-    write_obj=False,
+    input_path_a="ivus_rest",
+    input_path_b="ivus_stress",
+    label="full",
+    step_rotation_deg=0.1,
+    range_rotation_deg=90,
+    image_center=(4.5, 4.5),
+    radius=0.5,
+    n_points=20,
+    write_obj=True,
+    watertight=False,
+    output_path_a="output/rest",
+    output_path_b="output/stress",
+    output_path_c="output/diastole",
+    output_path_d="output/systole",
+    interpolation_steps=28,
+    contour_types=[mm.PyContourType.Lumen, mm.PyContourType.Catheter, mm.PyContourType.Wall]
 )
 
-# Load raw centerline
-cl_raw = np.genfromtxt("examples/data/centerline_raw.csv", delimiter=",")
+# Load raw centerline and align geometry onto it
+cl_raw = np.genfromtxt("centerline_raw.csv", delimiter=",")
 centerline = mm.numpy_to_centerline(cl_raw)
 
-# Align geometry pair onto centerline
 aligned_pair, cl_resampled = mm.align_three_point(
     centerline=centerline,
-    geometry_pair=rest,                # e.g. Rest geometry (dia/sys)
-    aortic_ref_pt=(12.26, -201.36, 1751.06),
-    upper_ref_pt=(11.76, -202.19, 1754.80),
-    lower_ref_pt=(15.66, -202.19, 1749.97)
+    geometry_pair=rest,
+    aortic_ref_pt=(12.2605, -201.3643, 1751.0554),
[…]
