### 2026-03-04 320479a
add docstrings for adjust_ccta, remove non-functional wrappers
