+++ multimodars/ccta/adjust_ccta.py
@@ -17,6 +17,68 @@ def label_geometry(
     tolerance_float: float = 1e-6,
     control_plot: bool = True,
 ) -> Tuple[dict, Tuple[any, any, any]]:
+    """Label CCTA mesh vertices as aorta, RCA, or LCA using centerline-based region detection.
+
+    Loads a 3-D surface mesh and three centerlines (aorta, RCA, LCA), then assigns
+    each mesh vertex to one of the anatomical regions.  For anomalous vessels an
+    additional occlusion-removal step uses ray-triangle intersection to strip
+    intramural segments, followed by adjacency-map reclassification to clean up
+    isolated mis-labelled vertices. Herfore, a ray is cast from every aorta point to the centerline
+    points of the anomalous section and if 3 faces are intersected by the ray the points from
+    the first face must correspond to the intramural section.
+
+    Parameters
+    ----------
+    path_ccta_geometry : Path or str
+        Path to the CCTA surface mesh file (any format supported by
+        :func:`multimodars.io.read_geometrical.read_mesh`).
+    path_centerline_aorta : Path or str
+        Path to a CSV file containing the aortic centerline (comma-delimited,
+        columns: x, y, z, …).
+    path_centerline_rca : Path or str
+        Path to a CSV file containing the RCA centerline.
+    path_centerline_lca : Path or str
+        Path to a CSV file containing the LCA centerline.
+    anomalous_rca : bool, optional
+        When ``True`` applies ray-triangle occlusion removal to the RCA region
+        to handle anomalous (intramural) courses.  Default is ``False``.
+    anomalous_lca : bool, optional
+        When ``True`` applies ray-triangle occlusion removal to the LCA region.
+        Default is ``False``.
+    n_points_intramural : int, optional
+        Number of coronary centerline points examined during occlusion removal
+        (the intramural segment length).  Default is ``120``.
+    bounding_sphere_radius_mm : float, optional
+        Radius in millimetres of the rolling sphere used to collect candidate
+        mesh vertices around each centerline point.  Default is ``3.0``.
+    tolerance_float : float, optional
+        Distance tolerance used when matching mesh vertices to points during
+        face lookup.  Default is ``1e-6``.
+    control_plot : bool, optional
+        When ``True`` opens an interactive 3-D scene showing the labelled mesh
+        after processing.  Default is ``True``.
+
+    Returns
+    -------
+    results : dict
+        Dictionary with keys:
+
+        * ``"mesh"`` - the original :class:`trimesh.Trimesh` object.
+        * ``"aorta_points"`` - list of ``(x, y, z)`` tuples for aortic vertices.
+        * ``"rca_points"`` - list of ``(x, y, z)`` tuples for RCA vertices.
+        * ``"lca_points"`` - list of ``(x, y, z)`` tuples for LCA vertices.
+        * ``"rca_removed_points"`` - RCA vertices removed by occlusion detection.
+        * ``"lca_removed_points"`` - LCA vertices removed by occlusion detection.
+
+    centerlines : tuple
+        A 3-tuple ``(cl_rca, cl_lca, cl_aorta)`` of ``PyCenterline`` objects.
+
+    Raises
+    ------
+    Exception
+        Re-raises any error that occurs while reading the mesh or centerline
+        files, after printing a descriptive message.
+    """
     import multimodars as mm
     from multimodars.io.read_geometrical import read_mesh
 
@@ -154,8 +216,28 @@ def label_geometry(
 def _prepare_faces_for_rust(
     mesh: trimesh.Trimesh, *, points=None, face_indices=None, tol: float = 1e-6
 ):
-    """
-    Convert selected mesh faces to the Rust-friendly format.
+    """Convert selected mesh faces to the Rust-friendly format.
+
+    Parameters
[…]
