+++ multimodars/_wrappers.py
@@ -1,725 +1,0 @@
-from typing import Literal, Any, Dict, Tuple, Union
-import multimodars as mm
-
-Mode = Literal["full", "doublepair", "singlepair", "single"]
-
-
-def from_file(
-    mode: Mode,
-    **kwargs: Any,
-) -> Union[
-    # full returns 4 geometry pairs + 4 log-lists
-    Tuple[
-        mm.PyGeometryPair,
-        mm.PyGeometryPair,
-        mm.PyGeometryPair,
-        mm.PyGeometryPair,
-        list,
-        list,
-        list,
-        list,
-    ],
-    # doublepair returns 2 geom + 4 log-lists
-    Tuple[
-        mm.PyGeometryPair,
-        mm.PyGeometryPair,
-        list,
-        list,
-        list,
-        list,
-    ],
-    # singlepair returns 1 geom‐pair + 2 log-lists
-    Tuple[
-        mm.PyGeometryPair,
-        mm.PyGeometryPair,
-        list,
-        list,
-    ],
-    # single returns 1 geom + 1 log-list
-    Tuple[
-        mm.PyGeometryPair,
-        list,
-    ],
-]:
-    """
-    A unified entrypoint for all `from_file_*` variants.
-
-    Parameters
-    ----------
-    mode : {"full","doublepair","singlepair","single"}
-        Selects which low-level function is called.
-
-          - "full"       → needs rest_input_path, stress_input_path,
-                            rest_output_path, stress_output_path,
-                            diastole_output_path, systole_output_path,
-                            step_rotation_deg, range_rotation_deg,
-                            interpolation_steps, image_center, radius, n_points
-          - "doublepair" → needs rest_input_path, stress_input_path,
-                            rest_output_path, stress_output_path,
-                            step_rotation_deg, range_rotation_deg,
-                            interpolation_steps, image_center, radius, n_points
-          - "singlepair" → needs input_path, output_path,
-                            step_rotation_deg, range_rotation_deg,
-                            interpolation_steps, image_center, radius, n_points
-          - "single"     → needs input_path, output_path,
-                            step_rotation_deg, range_rotation_deg,
-                            diastole, image_center, radius, n_points
-
-    Other Parameters
-    ----------------
-    **kwargs : dict
-        Keyword arguments required depend on `mode` (see above).
-
-    Returns
-    -------
-    Union[
-        # full now returns 4 geometry pairs *and* 4 log-lists
-        Tuple[
-        PyGeometryPair, PyGeometryPair, PyGeometryPair, PyGeometryPair,
-        list, list, list, list,
-        ],
-        # doublepair returns 2 geom + 4 log-lists
-        Tuple[
-        PyGeometryPair, PyGeometryPair,
-        list, list, list, list,
-        ],
-        # singlepair returns 1 geom-pair + 2 log-lists
-        Tuple[
-        PyGeometryPair, PyGeometryPair,
-        list,list,
-        ],
-        # single returns 1 geom + 1 log-list
-        Tuple[
-        PyGeometryPair,
-        list,
-        ],
-    ]:
-        The exact return shape depends on `mode`.
-
-    Raises
-    ------
-    ValueError
-        If an unsupported `mode` is passed.
-    """
-    from . import PyContourType, PyGeometry, PyGeometryPair
-    from . import (
-        from_file_full,
-        from_file_doublepair,
-        from_file_singlepair,
-        from_file_single,
-    )
-
-    defaults = {
-        "diastole": True,
-        "step_rotation_deg": 0.01,
-        "range_rotation_deg": 30,
-        "image_center": (4.5, 4.5),
-        "radius": 0.5,
-        "n_points": 20,
-        "write_obj": True,
[…]
