+++ docs/api/index.rst
@@ -11,3 +11,4 @@ API Reference
    centerline
    utilities
    converters_and_wrappers
+   adjust_ccta
