+++ docs/api/centerline.rst
@@ -3,3 +3,4 @@ Centerline Functions
 
 .. autofunction:: multimodars.align_three_point
 .. autofunction:: multimodars.align_manual
+.. autofunction:: multimodars.align_combined
