+++ paper/bibliography.bib
@@ -1,6 +1,6 @@
 @article{stark2025true,
   title={True Pulsatile Lumen Visualization in Coronary Artery Anomalies Using Controlled Transducer Pullback and Automated IVUS Segmentation},
-  author={Stark, Anselm Walter and Bigler, Marius Reto and R{\"a}ber, Lorenz and Gr{\"a}ni, Christoph},
+  author={Stark, Anselm W and Bigler, Marius R and R{\"a}ber, Lorenz and Gr{\"a}ni, Christoph},
   journal={Case Reports},
   volume={30},
   number={22},
@@ -72,15 +72,18 @@
   title={A comprehensive workflow for CCTA and OCT data fusion with 3D printing validation: advancing patient-specific testing environments for percutaneous coronary intervention devices},
   author={Ilic, Marc and H{\"a}ner, Jonas and Lehmann, Julian and Stark, Anselm W and Illi, Jo{\"e}l and Gr{\"a}ni, Christoph and Bentele, Marc and Baumann-Zumstein, Philine and Busch, Julia and Haeberlin, Andreas},
   journal={BioMedical Engineering OnLine},
-  year={2025},
+  volume={25},
+  number={15},
+  year={2026},
   publisher={Springer},
-  doi={10.1186/s12938-025-01501-6 }
+  doi={10.1186/s12938-025-01501-6}
 }
 
 @article{stark2025automated,
   title={Automated intravascular ultrasound image processing and quantification of coronary artery anomalies: the AIVUS-CAA software},
   author={Stark, Anselm W and Kazaj, Pooya Mohammadi and Balzer, Sebastian and Ilic, Marc and Bergamin, Manuel and Kakizaki, Ryota and Giannopoulos, Andreas and Haeberlin, Andreas and R{\"a}ber, Lorenz and Shiri, Isaac and others},
   journal={Computer Methods and Programs in Biomedicine},
+  volume={272},
   pages={109065},
   year={2025},
   publisher={Elsevier},
