+++ docs/api/adjust_ccta.rst
@@ -1,0 +1,6 @@
+Adjust CCTA
+===========
+
+.. automodule:: multimodars.ccta
+   :members:
+   :no-index:
