+++ docs/api/converters_and_wrappers.rst
@@ -4,10 +4,3 @@ Converters
 .. automodule:: multimodars._converters
    :members:
    :no-index:
-
-Wrappers
-=========
-
-.. automodule:: multimodars._wrappers
-   :members:
-   :no-index:
