### 2026-03-04 2a95755
unify docstring style
