+++ multimodars/_converters.py
@@ -197,19 +197,38 @@ def numpy_to_inputdata(
     label: str = "",
 ) -> mm.PyInputData:
     """
-    Build a PyInputData from numpy arrays, grouping by frame_index into frames.
+    Build a ``PyInputData`` from numpy arrays, grouping by frame index into frames.
 
-    Each row in the ``*_arr`` is [frame_index, x, y, z].
+    Each row in the array arguments must be ``[frame_index, x, y, z]``.
 
-    Returns a PyInputData containing:
-      - lumen: list of PyContour (one per frame found in lumen_arr)
-      - eem: optional list of PyContour (only contours that exist)
-      - calcification: optional list of PyContour (only contours that exist)
-      - sidebranch: optional list of PyContour (only contours that exist)
-      - record: optional list of PyRecord (converted from record array)
-      - ref_point: PyContourPoint (first reference row, or default)
-      - diastole: boolean
-      - label: str
+    Parameters
+    ----------
+    lumen_arr : np.ndarray
+        ``(N, 4)`` array of lumen points ``[frame_index, x, y, z]``.
+    record : np.ndarray, optional
+        ``(M, 4)`` array of records ``[frame, phase, measurement_1, measurement_2]``.
+    ref_point : np.ndarray
+        ``(1, 4)`` or ``(4,)`` array ``[frame_index, x, y, z]`` for the reference point.
+    diastole : bool
+        ``True`` if the data corresponds to the diastolic phase.
+    eem_arr : np.ndarray, optional
+        ``(N, 4)`` array of EEM points ``[frame_index, x, y, z]``.  Default is ``None``.
+    calcification : np.ndarray, optional
+        ``(N, 4)`` array of calcification points.  Default is ``None``.
+    sidebranch : np.ndarray, optional
+        ``(N, 4)`` array of side-branch points.  Default is ``None``.
+    label : str, optional
+        Label string for the input data.  Default is ``""``.
+
+    Returns
+    -------
+    input_data : PyInputData
+        Input data object with contours grouped by frame index.
+
+    Raises
+    ------
+    ValueError
+        If ``lumen_arr`` is empty.
     """
     import numpy as np
     from . import PyContour, PyContourPoint, PyRecord, PyInputData
@@ -414,33 +433,35 @@ def numpy_to_geometry(
     label: str = "",
 ) -> mm.PyGeometry:
     """
-    Build a PyGeometry from numpy arrays, grouping by frame_index into frames.
+    Build a ``PyGeometry`` from numpy arrays, grouping by frame index into frames.
 
-    Each row in the ``*_arr`` is [frame_index, x, y, z].
-
-    Returns a PyGeometry containing frames with:
-      - lumen: PyContour for each frame
-      - extras: dictionary with optional EEM, Catheter, Wall contours
-      - reference_point: PyContourPoint from reference_arr (if provided)
+    Each row in the array arguments must be ``[frame_index, x, y, z]``.
 
     Parameters
     ----------
     lumen_arr : np.ndarray
-        (N,4) array of lumen points [frame_index, x, y, z]
+        ``(N, 4)`` array of lumen points ``[frame_index, x, y, z]``.
     eem_arr : np.ndarray, optional
-        (M,4) array of EEM points [frame_index, x, y, z]
+        ``(M, 4)`` array of EEM points ``[frame_index, x, y, z]``.  Default is ``None``.
     catheter_arr : np.ndarray, optional
-        (K,4) array of catheter points [frame_index, x, y, z]
+        ``(K, 4)`` array of catheter points ``[frame_index, x, y, z]``.  Default is ``None``.
     wall_arr : np.ndarray, optional
-        (L,4) array of wall points [frame_index, x, y, z]
[…]
