+++ src/intravascular/binding/align.rs
@@ -4,35 +4,55 @@ use crate::intravascular::{
 };
 use pyo3::prelude::*;
 
-/// Creates centerline-aligned meshes for diastolic and systolic geometries
-/// based on three reference points (aorta, upper section, lower section).
-/// Only works for elliptic vessels e.g. coronary artery anomalies.
+/// Align diastolic and systolic meshes to the centerline using three reference points.
 ///
-/// Args:
-///     centerline: PyCenterline object
-///     geometry_pair: PyGeometryPair object
-///     aortic_ref_pt: Reference point for aortic position
-///     upper_ref_pt: Upper reference point
-///     lower_ref_pt: Lower reference point
-///     angle_step_deg (default 1.0): step size in degrees for rotation search
-///     write (default false): Wether to write aligned meshes to OBJ
-///     watertight (default true): Wether to write shell or watertight mesh to OBJ.
-///     interpolation_steps: Number of interpolation steps
-///     output_dir (default "output/aligned"): Output directory for aligned meshes
-///     case_name (default "None"): Case name for output files
+/// Creates centerline-aligned meshes for a diastolic/systolic geometry pair
+/// based on three anatomical reference points (aorta, upper section, lower
+/// section).  Only works for elliptic vessels such as coronary artery
+/// anomalies.
 ///
-/// Returns:
-///     PyGeometryPair, PyCenterline (resampled)
+/// Parameters
+/// ----------
+/// centerline : PyCenterline
+///     Centerline of the vessel.
+/// geometry_pair : PyGeometryPair
+///     Diastolic/systolic geometry pair to align.
+/// aortic_ref_pt : tuple of float
+///     ``(x, y, z)`` reference point at the aortic ostium.
+/// upper_ref_pt : tuple of float
+///     ``(x, y, z)`` upper section reference point.
+/// lower_ref_pt : tuple of float
+///     ``(x, y, z)`` lower section reference point.
+/// angle_step_deg : float, optional
+///     Step size in degrees for the rotation search.  Default is ``1.0``.
+/// write : bool, optional
+///     Whether to write the aligned meshes to OBJ files.  Default is ``False``.
+/// watertight : bool, optional
+///     Whether to write a watertight or shell mesh to OBJ.  Default is ``True``.
+/// interpolation_steps : int, optional
+///     Number of interpolation steps between phases.  Default is ``28``.
+/// output_dir : str, optional
+///     Output directory for aligned meshes.  Default is ``"output/aligned"``.
+/// case_name : str, optional
+///     Case name used as a filename prefix.  Default is ``"None"``.
 ///
-/// Example:
-///     >>> import multimodars as mm
-///     >>> dia, sys = mm.align_three_point(
-///     ...     centerline,
-///     ...     geometry_pair,
-///     ...     (12.2605, -201.3643, 1751.0554),
-///     ...     (11.7567, -202.1920, 1754.7975),
-///     ...     (15.6605, -202.1920, 1749.9655),
-///     ... )
+/// Returns
+/// -------
+/// geom_pair : PyGeometryPair
+///     Aligned geometry pair.
+/// centerline : PyCenterline
+///     Resampled centerline.
+///
+/// Examples
+/// --------
+/// >>> import multimodars as mm
+/// >>> dia, sys = mm.align_three_point(
+/// ...     centerline,
+/// ...     geometry_pair,
+/// ...     (12.2605, -201.3643, 1751.0554),
+/// ...     (11.7567, -202.1920, 1754.7975),
+/// ...     (15.6605, -202.1920, 1749.9655),
+/// ... )
 #[pyfunction]
 #[pyo3(
     signature = (
@@ -92,33 +112,51 @@ pub fn align_three_point(
[…]
