+++ docs/api/adjust_ccta.rst
@@ -1,6 +1,6 @@
 Adjust CCTA
 ===========
 
-.. automodule:: multimodars.ccta
+.. automodule:: multimodars.ccta.adjust_ccta
    :members:
    :no-index:

[…]
