### 2026-03-04 257e97d
improve PyFrame representation
