+++ src/intravascular/binding/classes.rs
@@ -254,7 +254,6 @@ impl From<&InputData> for PyInputData {
     }
 }
 
-// Also provide owned conversion if you want to convert InputData (by value)
 impl From<InputData> for PyInputData {
     fn from(input: InputData) -> Self {
         PyInputData::from(&input)
@@ -356,7 +355,6 @@ impl PyContourPoint {
     }
 }
 
-// Conversion from Rust type to Python type
 impl From<&ContourPoint> for PyContourPoint {
     fn from(point: &ContourPoint) -> Self {
         Self {
@@ -763,7 +761,6 @@ impl PyContourType {
     }
 }
 
-// Conversion between Rust ContourType and PyContourType
 impl From<ContourType> for PyContourType {
     fn from(contour_type: ContourType) -> Self {
         match contour_type {
@@ -790,26 +787,10 @@ impl From<PyContourType> for ContourType {
     }
 }
 
-// Also implement for references
 impl From<&PyContourType> for ContourType {
     fn from(py_contour_type: &PyContourType) -> Self {
         (*py_contour_type).into()
     }
-}
-
-#[pyclass]
-#[derive(Debug, Clone)]
-pub struct PyFrame {
-    #[pyo3(get, set)]
-    pub id: u32,
-    #[pyo3(get, set)]
-    pub centroid: (f64, f64, f64),
-    #[pyo3(get, set)]
-    pub lumen: PyContour,
-    #[pyo3(get, set)]
-    pub extras: HashMap<String, PyContour>, // String keys for ContourType
-    #[pyo3(get, set)]
-    pub reference_point: Option<PyContourPoint>,
 }
 
 /// Python representation of a single intravascular imaging frame.
@@ -837,6 +818,21 @@ pub struct PyFrame {
 /// ...     extras={"Eem": eem_contour},
 /// ...     reference_point=ref_point
 /// ... )
+#[pyclass]
+#[derive(Debug, Clone)]
+pub struct PyFrame {
+    #[pyo3(get, set)]
+    pub id: u32,
+    #[pyo3(get, set)]
+    pub centroid: (f64, f64, f64),
+    #[pyo3(get, set)]
+    pub lumen: PyContour,
+    #[pyo3(get, set)]
+    pub extras: HashMap<String, PyContour>, // String keys for ContourType
+    #[pyo3(get, set)]
+    pub reference_point: Option<PyContourPoint>,
+}
+
 #[pymethods]
 impl PyFrame {
     #[new]
@@ -868,6 +864,17 @@ impl PyFrame {
         )
     }
 
+    /// Rotate all contour points in the frame by the given angle.
+    ///
+    /// Parameters
+    /// ----------
+    /// angle_deg : float
+    ///     Rotation angle in degrees.
+    ///
+    /// Returns
+    /// -------
+    /// frame : PyFrame
+    ///     New frame with all contours rotated.
     #[pyo3(signature = (angle_deg))]
     pub fn rotate(&self, angle_deg: f64) -> PyResult<PyFrame> {
         let mut rust_frame = self.to_rust_frame()?;
@@ -875,6 +882,21 @@ impl PyFrame {
         Ok(PyFrame::from(&rust_frame))
     }
 
+    /// Translate all contour points in the frame by the given offsets.
+    ///
+    /// Parameters
+    /// ----------
+    /// dx : float
+    ///     Translation along the x-axis in mm.
+    /// dy : float
+    ///     Translation along the y-axis in mm.
+    /// dz : float
+    ///     Translation along the z-axis in mm.
+    ///
+    /// Returns
+    /// -------
+    /// frame : PyFrame
+    ///     New frame with all contours translated.
[…]
