+++ docs/tutorial_ccta.rst
@@ -3,7 +3,8 @@
 
     Also download the example data and jupyter notebooks from the current release from the
     `Github Releases page <https://github.com/yungselm/multimoda-rs/releases>`_ to follow
-    this tutorial step-by-step.
+    this tutorial step-by-step. Specifically, the test_adjust_ccta.py allows you to also 
+    display the trimesh debug plots.
 
 Tutorial - CCTA Module
 =======================
@@ -25,10 +26,10 @@ that combines both steps, and performs several checks. Additionally is the geome
     import multimodars as mm
 
     results, (rca_cl, lca_cl, ao_cl) = mm.label_geometry(
-        path_ccta_geometry="../data/NARCO_119.stl",
-        path_centerline_aorta="../data/centerline_aorta.csv",
-        path_centerline_rca="../data/centerline_rca.csv",
-        path_centerline_lca="../data/centerline_lca.csv",
+        path_ccta_geometry="data/NARCO_119.stl",
+        path_centerline_aorta="data/centerline_aorta.csv",
+        path_centerline_rca="data/centerline_rca.csv",
+        path_centerline_lca="data/centerline_lca.csv",
         bounding_sphere_radius_mm=3.5, # rolling sphere radius for initial vessel labeling
         n_points_intramural=200, # points centerline to define end of intramural section, if unsure keep large
         anomalous_rca=True, # cleans wrongly labeled points intramural part
