### 2026-03-04 27d5421
numpy_to_inputdata record should default to None
