### 2026-03-04 51187fc
update templates and add __future__ annotation
