+++ multimodars/ccta/adjust_ccta.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import numpy as np
 import trimesh
 from pathlib import Path
