+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -12,27 +12,20 @@ A clear and concise description of what the bug is.
 
 **To Reproduce**
 Steps to reproduce the behavior:
-1. Go to '...'
-2. Click on '....'
-3. Scroll down to '....'
-4. See error
+1. Load data with '...'
+2. Call function '...'
+3. See error
 
 **Expected behavior**
 A clear and concise description of what you expected to happen.
 
+**Environment (please complete the following information):**
+ - OS: [e.g. Windows, Linux, macOS]
+ - Python version: [e.g. 3.10]
+ - Package version: [e.g. 0.1.0]
+
 **Screenshots**
 If applicable, add screenshots to help explain your problem.
-
-**Desktop (please complete the following information):**
- - OS: [e.g. iOS]
- - Browser [e.g. chrome, safari]
- - Version [e.g. 22]
-
-**Smartphone (please complete the following information):**
- - Device: [e.g. iPhone6]
- - OS: [e.g. iOS8.1]
- - Browser [e.g. stock browser, safari]
- - Version [e.g. 22]
 
 **Additional context**
 Add any other context about the problem here.
