### 2026-03-04 #10131 @yungselm
Thank you @crnh for the detailed report, @jonaspleyer for confirming this and @erickmartins for overseeing the review.

The root cause was a refactor where example data was moved out of the package (to keep it lightweight) into a separate `examples/` folder, but the README, documentation, and example data were not updated accordingly.

The following has been addressed in v0.2.2:
- The correct `example.zip` file has been attached to the v0.2.2 release, now including the complete example data
- The Jupyter notebook has been adjusted to match the correct folder structure when extracting `example.zip` into the project
- The generic `from_file` wrapper has been removed in favour of more explicit functions (`from_file_full`, `from_file_singlepair`, etc.), each with specific parameter checking and clear error messages
- The README quickstart example has been updated to reflect the new API and datastructure
- All docstrings have been updated and standardized to improve the ReadTheDocs output

I verified everything end-to-end by building the package from scratch in a fresh WSL2 environment. And one of my collaborators ran the project on native linux. So far I could not test it on Windows.

Apologies for this mistake from my side and thank you all for your time and effort on reviewing this project.
