+++ paper/REVIEWER1.md
@@ -1,0 +1,117 @@
+We thank @crnh for the careful evaluation of our manuscript and codebase and the appreciation of our work.
+
+## Code
+### Running the examples
+I tested the examples with multimodars version 0.2.1, which is the latest version available on PyPI.
+
+### Quickstart example in Readme
+- [x] from_file should check if all required parameters are present and raise an exception with a clear message if not (the same applies to other wrapper functions). The example code does not seem to specify all required parameters.
+
+**Answer:** I agree that the current from_file wrapper had insufficient error handling. However, for the current version I decided to remove the wrapper functions entirely from the project and focus on the more explicit entry points (e.g. from_file_full), which have  error handling already implemented.
+
+*Removed:* multimodars/io/_wrappers.py
+
+- [x] The file data/centerline_raw.csv and other example data is distributed as part of the releases, which is unintuitive. This should at least be mentioned in the Readme, but ideally all example data should just be included in the repository.
+
+**Answer:** The reviewer is correct that this is less intuitive, and I originally had all the example data in the repository. Since the datavolume kept increasing, with added 3D functionality, I decided on attaching data to specific relases, this ensures the same set up even if more 3D model data will be provided in the future. However, I made it clearer how to access the data by linking in the README.md
+
+*Changed:* README.md "Download the [example.zip](https://github.com/yungselm/multimoda-rs/releases/tag/v0.2.2) from the latest release..."
+
+### ``ivus_to_centerline.ipynb`` from the examples folder
+
+- [x] The data folders in the examples folder (e.g. examples/data/ivus_rest) are empty, causing from_file_full to raise an unclear exception:
+
+**Answer:** I thank the reviewer for spotting this error early. As I previously explained, this error occured due to a refactor where example data was moved out of the package. I now added the correct ``examples.zip`` to the github repository, and updated the Jupyter Notebook. Additionally as stated above removed the faulty _wrappers.
+
+*Added:* Complete example data added to v0.2.2
+
+*Removed:* multimodars/io/_wrappers.py
+
+- [x] RuntimeError: Failed to prepare geometries for full processing
+
+**Answer:** I thank the reviewer for their valuable feedback, as discussed on the JOSS review page, this should now be resolved.
+
+- [x] The required data should be included in the data folder (I now copied it from the repository) and the exception should specify what exactly went wrong.
+
+**Answer:** As stated above, the reasoning for this was to keep the repository lean for future additional data. If the reviewer does not agree with this design choice, I will adjust the repository accordingly.
+
+- [x] Cell 2 loads the example data from different file names than the names used to save them in cell 1 (e.g. dia_lumen.obj vs rest_dia.obj), which causes the code to raise an exception.
+- [x] The notebook imports the same dependencies multiple times.
+- [x] I was unable to run the stent example, because I could not find the required data in the repository or the releases.
+- [x] The data required for the 'Alignment from array' example is not included in the examples archive and had to be downloaded separately from the repository. The output of the example is not visualized.
+- [x] The other examples in the notebook run without errors and the visualizations are clear.
+
+**Answer:** I thank the reviewr for his patience, I updated the whole Jupyter notebook, to match the new version v0.2.2.
+
+*Changed:* Completely overworked examples Jupyter Notebook.
+
+- [x] Rendering the meshes depends on nbformat, but this dependency is not listed. The notebook also depends on scipy, which is not listed.
+
+**Answer:** I agree with the reviewer, that this can be confusing to the user. My idea was that the Jupyter notebooks are something seperate from the core repository, therefore the dependencies are loaded only if a user decides to download the examples folder. If the reviewer wants this changed I will add dev dependencies and add the full files to the main directory.
+
[…]
