+++ multimodars/ccta/adjust_ccta.py
@@ -4,7 +4,19 @@ import numpy as np
 import trimesh
 from pathlib import Path
 import warnings
-from typing import Optional, Tuple, List
+from trimesh.points import PointCloud
+from ..multimodars import (
+    find_centerline_bounded_points_simple,
+    remove_occluded_points_ray_triangle,
+    clean_outlier_points,
+    build_adjacency_map,
+    find_points_by_cl_region,
+    adjust_diameter_centerline_morphing_simple,
+    find_proximal_distal_scaling,
+    find_aortic_scaling,
+)
+from .._converters import numpy_to_centerline
+from ..io.read_geometrical import read_mesh
 
 
 def label_geometry(
@@ -18,7 +30,7 @@ def label_geometry(
     bounding_sphere_radius_mm: float = 3.0,
     tolerance_float: float = 1e-6,
     control_plot: bool = True,
-) -> Tuple[dict, Tuple[any, any, any]]:
+) -> tuple[dict, tuple[any, any, any]]:
     """Label CCTA mesh vertices as aorta, RCA, or LCA using centerline-based region detection.
 
     Loads a 3-D surface mesh and three centerlines (aorta, RCA, LCA), then assigns
@@ -81,9 +93,6 @@ def label_geometry(
         Re-raises any error that occurs while reading the mesh or centerline
         files, after printing a descriptive message.
     """
-    import multimodars as mm
-    from multimodars.io.read_geometrical import read_mesh
-
     try:
         mesh = read_mesh(path_ccta_geometry)
         print(f"Loaded mesh: {len(mesh.vertices)} vertices, {len(mesh.faces)} faces")
@@ -93,7 +102,7 @@ def label_geometry(
 
     try:
         cl_aorta_raw = np.genfromtxt(path_centerline_aorta, delimiter=",")
-        cl_aorta = mm.numpy_to_centerline(cl_aorta_raw)
+        cl_aorta = numpy_to_centerline(cl_aorta_raw)
         print(f"Loaded aorta centerline: {len(cl_aorta.points)} points")
     except Exception as e:
         print(f"Error reading Aorta centerline from {path_centerline_aorta}: {e}")
@@ -101,7 +110,7 @@ def label_geometry(
 
     try:
         cl_lca_raw = np.genfromtxt(path_centerline_lca, delimiter=",")
-        cl_lca = mm.numpy_to_centerline(cl_lca_raw)
+        cl_lca = numpy_to_centerline(cl_lca_raw)
         print(f"Loaded LCA centerline: {len(cl_lca.points)} points")
     except Exception as e:
         print(f"Error reading LCA centerline from {path_centerline_lca}: {e}")
@@ -109,7 +118,7 @@ def label_geometry(
 
     try:
         cl_rca_raw = np.genfromtxt(path_centerline_rca, delimiter=",")
-        cl_rca = mm.numpy_to_centerline(cl_rca_raw)
+        cl_rca = numpy_to_centerline(cl_rca_raw)
         print(f"Loaded RCA centerline: {len(cl_rca.points)} points")
     except Exception as e:
         print(f"Error reading RCA centerline from {path_centerline_rca}: {e}")
@@ -118,10 +127,10 @@ def label_geometry(
     points_list = [tuple(vertex) for vertex in mesh.vertices.tolist()]
 
     # Rust implementation using a rolling sphere with fixed radius
-    rca_points_found = mm.find_centerline_bounded_points_simple(
+    rca_points_found = find_centerline_bounded_points_simple(
         cl_rca, points_list, bounding_sphere_radius_mm
     )
-    lca_points_found = mm.find_centerline_bounded_points_simple(
[…]
