### 2026-03-06 7c7ceb1
add type stubs
