+++ paper/REVIEWER1.md
@@ -67,7 +67,10 @@ I tested the examples with multimodars version 0.2.1, which is the latest versio
 
 **Answer:** Thank you very much for this careful evaulation, I adusted all the modules to use only top level imports and adjusted the type annotation.
 
-- [] No type stubs are included for the parts of the interface that are implemented in Rust, which makes the interface less clear and prevents type checking of code that uses the library. The authors should consider adding stubs for at least the parts of the interface that they consider to be public. This would significantly improve the user experience. See e.g. https://www.maturin.rs/project_layout.html#adding-python-type-information and https://pyo3.rs/v0.27.1/type-stub.html for more information on how to do this.
+- [x] No type stubs are included for the parts of the interface that are implemented in Rust, which makes the interface less clear and prevents type checking of code that uses the library. The authors should consider adding stubs for at least the parts of the interface that they consider to be public. This would significantly improve the user experience. See e.g. https://www.maturin.rs/project_layout.html#adding-python-type-information and https://pyo3.rs/v0.27.1/type-stub.html for more information on how to do this.
+
+**Answer:** I did not know this feature! The automatic generation sounds very interesting, however since it was tagged as still experimental I decided for a manual version for now, since it also does not seem to support the function-based declaration.
+
 - [] ``multimodars`` provides a unified interface to multiple functions through a set of wrapper functions, such as from_file and from_array. This is good, but any parameters that are required by the underlying functions are passed through **kwargs, which requires the user to consult the documentation of these functions to understand the meaning of the parameters. The interface could be improved in a few ways:
     - [] The wrapper functions could explicitly document the meaning of all supported parameters;
     - [] The wrapper functions could use a combination of TypedDict unpacking (to specify which parameters are supported) combined with overloads to specify which parameters are required for which mode. This would improve the user experience by enabling auto-completions and type checking for the parameters of the wrapper functions. More information: https://typing.python.org/en/latest/spec/callables.html#unpack-kwargs and https://typing.python.org/en/latest/spec/overload.html.
