+++ CONTRIBUTING.md
@@ -56,6 +56,109 @@ If you find a bug, please help us fix it by opening a [a new issue](https://gith
     - Python version
     - OS and architecture
 
+## Suggesting Enhancements
+Have an idea for a new feature or an improvement to existing behaviour? Open a [feature request issue](https://github.com/yungselm/multimoda-rs/issues/new/choose) and include:
+- **Summary**: A concise description of the enhancement
+- **Motivation**: Why would this be useful? What problem does it solve?
+- **Proposed solution**: How you envision it working (API sketch, example usage, etc.)
+- **Alternatives considered**: Other approaches you've thought about
+
+We appreciate well-scoped proposals with clear use cases.
+
+## Your First Code Contribution
+Not sure where to start? Look for issues labelled [`good first issue`](https://github.com/yungselm/multimoda-rs/issues?q=is%3Aissue+label%3A%22good+first+issue%22) or [`help wanted`](https://github.com/yungselm/multimoda-rs/issues?q=is%3Aissue+label%3A%22help+wanted%22).
+
+**Setup** (Python + Rust hybrid via [maturin](https://github.com/PyO3/maturin)):
+```bash
+# 1. Fork the repo on GitHub, then clone your fork
+git clone https://github.com/<your-username>/multimoda-rs.git
+cd multimoda-rs
+
+# 2. Create and activate a virtual environment
+python -m venv .venv
+source .venv/bin/activate   # Windows: .venv\Scripts\activate
+
+# 3. Install the package in development mode (compiles the Rust extension)
+pip install -e ".[dev]"
+
+# 4. Install pre-commit hooks
+pre-commit install
+
+# 5. Verify everything works
+pytest
+cargo test --lib
+```
+
+## Pull Request Process
+1. **Branch** – create a descriptive branch (`feature/my-feature`, `fix/issue-123`).
+2. **Commit** – write clear, atomic commit messages.
+3. **Tests** – add or update tests to cover your changes; all existing tests must pass.
+4. **Docs** – update docstrings and, if applicable, the RST docs under `docs/`.
+5. **Open the PR** – target the `main` branch. Fill in the PR template with a summary, motivation, and testing notes.
+6. **Review** – address reviewer comments; once approved and CI is green, a maintainer will merge.
+
+PRs that break existing tests or skip the pre-commit hooks will not be merged.
+
+## Coding Style & Tests
+
+### Python
+- **Formatter**: [Black](https://black.readthedocs.io/) (enforced automatically via pre-commit).
+- **Linter**: [Ruff](https://docs.astral.sh/ruff/) with `--fix` (enforced automatically via pre-commit).
+- **Docstrings**: [NumPy docstring style](https://numpydoc.readthedocs.io/en/latest/format.html).
+  Example:
+  ```python
+  def my_function(x: float, y: float) -> float:
+      """Short one-line summary.
+
+      Longer description if needed.
+
+      Parameters
+      ----------
+      x : float
+          Description of x.
+      y : float
+          Description of y.
+
+      Returns
+      -------
+      float
+          Description of the return value.
+      """
+  ```
+- **Type annotations**: use `from __future__ import annotations` at the top of every module and annotate all public functions and methods.
+- **Tests**: place tests under `tests/` and use [pytest](https://docs.pytest.org/). Run with `pytest`.
+
+### Rust
+- **Formatter**: `cargo fmt --all` (enforced automatically via pre-commit).
+- **Tests**: unit tests live alongside the source (`#[cfg(test)]` blocks). Run with `cargo test --lib`.
+
+All checks are wired up via pre-commit and run automatically on every commit. You can also run them manually:
+```bash
+pre-commit run --all-files
+```
+
+## Writing Documentation
[…]
