### 2026-03-06 943f921
fix documentation for python version and make cleaner
