+++ docs/api/adjust_ccta.rst
@@ -1,6 +1,9 @@
 Adjust CCTA
 ===========
 
-.. automodule:: multimodars.ccta.adjust_ccta
-   :members:
-   :no-index:
+.. autofunction:: multimodars.ccta.adjust_ccta.label_geometry
+.. autofunction:: multimodars.ccta.adjust_ccta.label_anomalous_region
+.. autofunction:: multimodars.ccta.adjust_ccta.scale_region_centerline_morphing
+.. autofunction:: multimodars.ccta.adjust_ccta.compare_centerline_scaling
+.. autofunction:: multimodars.ccta.adjust_ccta.find_distal_and_proximal_scaling
+.. autofunction:: multimodars.ccta.adjust_ccta.find_aorta_scaling
