+++ docs/api/converters_and_wrappers.rst
@@ -1,6 +1,9 @@
 Converters
 ==========
 
-.. automodule:: multimodars._converters
-   :members:
-   :no-index:
+.. autofunction:: multimodars._converters.to_array
+.. autofunction:: multimodars._converters.numpy_to_inputdata
+.. autofunction:: multimodars._converters.numpy_to_geometry
+.. autofunction:: multimodars._converters.numpy_to_centerline
+.. autofunction:: multimodars._converters.array_to_pyinputdata
+.. autofunction:: multimodars._converters.geometry_to_frames_array
