+++ .github/workflows/CI.yml
@@ -19,6 +19,27 @@ permissions:
   contents: read
 
 jobs:
+  python-compat:
+    name: Python ${{ matrix.python-version }} compatibility
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ['3.9', '3.10', '3.11', '3.12']
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install Rust
+        uses: dtolnay/rust-toolchain@stable
+      - name: Install Python deps
+        run: |
+          python -m pip install --upgrade pip
+          pip install -e .[test]
+      - name: Run Python tests
+        run: pytest
+
   linux:
     runs-on: ${{ matrix.platform.runner }}
     strategy:
