### 2026-03-06 f326d64
fix missing function in docs and remove non working python version 3.9
