+++ multimodars/__init__.py
@@ -21,6 +21,7 @@ from .multimodars import (
     from_array_single,
     align_three_point,
     align_manual,
+    align_combined,
     to_obj,
     find_centerline_bounded_points_simple,
     find_proximal_distal_scaling,
@@ -68,6 +69,7 @@ __all__ = [
     "from_array_single",
     "align_three_point",
     "align_manual",
+    "align_combined",
     "to_obj",
     "find_centerline_bounded_points_simple",
     "find_proximal_distal_scaling",
