### 2026-03-06 1bc5055
add new methods for PyFrame replacement, finalize review
