+++ paper/REVIEWER1.md
@@ -1,41 +1,41 @@
-We thank @crnh for the careful evaluation of our manuscript and codebase and the appreciation of our work.
+We thank @crnh for the thorough evaluation of our manuscript and codebase and the appreciation of our work. The detailed feedback was genuinely helpful and led to several meaningful improvements.
 
 ## Code
 ### Running the examples
 I tested the examples with multimodars version 0.2.1, which is the latest version available on PyPI.
 
-**Answer:** Please not that the review is now performed on the latest version v0.2.2
+**Answer:** Just a heads-up, the review should now be performed on the latest version v0.2.2, which addresses many of the points raised below.
 
 ### Quickstart example in Readme
 - [x] from_file should check if all required parameters are present and raise an exception with a clear message if not (the same applies to other wrapper functions). The example code does not seem to specify all required parameters.
 
-**Answer:** I agree that the current from_file wrapper had insufficient error handling. However, for the current version I decided to remove the wrapper functions entirely from the project and focus on the more explicit entry points (e.g. from_file_full), which have  error handling already implemented.
+**Answer:** I agree, the old ``from_file`` wrapper had insufficient error handling. Rather than patching it, we decided to remove the wrapper functions entirely and focus on the more explicit entry points (e.g. ``from_file_full``), which already have proper error handling in place.
 
-*Removed:* multimodars/io/_wrappers.py
+*Removed:* ``multimodars/io/_wrappers.py``
 
 - [x] The file data/centerline_raw.csv and other example data is distributed as part of the releases, which is unintuitive. This should at least be mentioned in the Readme, but ideally all example data should just be included in the repository.
 
-**Answer:** The reviewer is correct that this is less intuitive, and I originally had all the example data in the repository. Since the datavolume kept increasing, with added 3D functionality, I decided on attaching data to specific relases, this ensures the same set up even if more 3D model data will be provided in the future. However, I made it clearer how to access the data by linking in the README.md
+**Answer:** We completely agree with you! We originally kept all example data in the repository, but as the data volume grew with added 3D functionality, attaching data to specific releases felt like the cleaner approach, it also ensures a consistent setup going forward. That said, we've made it much easier to find by linking directly to the download in the README.
 
-*Changed:* README.md "Download the [example.zip](https://github.com/yungselm/multimoda-rs/releases/tag/v0.2.2) from the latest release..."
+*Changed:* README.md "Download [examples.zip](https://github.com/yungselm/multimoda-rs/releases/latest/download/examples.zip) (SHA256: `d11ebc7607f43ab4571fb51c9ac9178caac57774cf5d97f4f068ace4eb070fee`) from..."
 
 ### ``ivus_to_centerline.ipynb`` from the examples folder
 
 - [x] The data folders in the examples folder (e.g. examples/data/ivus_rest) are empty, causing from_file_full to raise an unclear exception:
 
-**Answer:** I thank the reviewer for spotting this error early. As I previously explained, this error occured due to a refactor where example data was moved out of the package. I now added the correct ``examples.zip`` to the github repository, and updated the Jupyter Notebook. Additionally as stated above removed the faulty _wrappers.
+**Answer:** Thanks for catching this early, it was a leftover issue from a refactor where example data was moved out of the package. The correct ``examples.zip`` has now been added to the repository, and the Jupyter Notebook has been updated accordingly. The faulty _wrappers have also been removed as mentioned above.
 
 *Added:* Complete example data added to v0.2.2
 
-*Removed:* multimodars/io/_wrappers.py
+*Removed:* ``multimodars/io/_wrappers.py``
 
 - [x] RuntimeError: Failed to prepare geometries for full processing
 
-**Answer:** I thank the reviewer for their valuable feedback, as discussed on the JOSS review page, this should now be resolved.
[…]
