### 2026-03-09 bcb9cd6
improve stubs
