### 2026-03-09 dd79169
improve stubs add processing in polars style
