+++ multimodars/_processing.py
@@ -1,0 +1,1382 @@
+from __future__ import annotations
+
+from .multimodars import (
+    PyContourType,
+    PyGeometry,
+    PyGeometryPair,
+    PyInputData,
+    PyCenterline,
+    from_file_full as _from_file_full,
+    from_file_doublepair as _from_file_doublepair,
+    from_file_singlepair as _from_file_singlepair,
+    from_file_single as _from_file_single,
+    from_array_full as _from_array_full,
+    from_array_doublepair as _from_array_doublepair,
+    from_array_singlepair as _from_array_singlepair,
+    from_array_single as _from_array_single,
+    align_three_point as _align_three_point,
+    align_manual as _align_manual,
+    align_combined as _align_combined,
+    to_obj as _to_obj,
+    find_centerline_bounded_points_simple as _find_centerline_bounded_points_simple,
+    find_proximal_distal_scaling as _find_proximal_distal_scaling
+)
+
+_AlignLog = list[tuple[int, int, float, float, float, float, float]]
+
+
+def _default_contour_types() -> list[PyContourType]:
+    return [PyContourType.Lumen, PyContourType.Catheter, PyContourType.Wall]
+
+
+# ---------------------------------------------------------------------------
+# Processing functions — from CSV files
+# ---------------------------------------------------------------------------
+
+
+def from_file_full(
+    input_path_a: str,
+    input_path_b: str,
+    label: str = "full",
+    diastole: bool = True,
+    step_rotation_deg: float = 0.5,
+    range_rotation_deg: float = 90.0,
+    image_center: tuple[float, float] = (4.5, 4.5),
+    radius: float = 0.5,
+    n_points: int = 20,
+    write_obj: bool = True,
+    watertight: bool = True,
+    output_path_a: str = "output/rest",
+    output_path_b: str = "output/stress",
+    output_path_c: str = "output/diastole",
+    output_path_d: str = "output/systole",
+    interpolation_steps: int = 28,
+    bruteforce: bool = False,
+    sample_size: int = 500,
+    contour_types: list[PyContourType] | None = None,
+    smooth: bool = True,
+    postprocessing: bool = True,
+) -> tuple[
+    PyGeometryPair,
+    PyGeometryPair,
+    PyGeometryPair,
+    PyGeometryPair,
+    tuple[_AlignLog, _AlignLog, _AlignLog, _AlignLog],
+]:
+    """Process four intravascular imaging geometries in parallel from CSV folders.
+
+    Reads REST and STRESS acquisitions from two input folders, aligns frames
+    within and between each cardiac phase in parallel, and writes interpolated
+    OBJ meshes.
+
+    .. code-block:: text
+
+       Rest           Stress
+       diastole ──▶ diastole
+          │            │
+          ▼            ▼
+       systole ──▶ systole
+
+    .. warning::
+
+       The CSV must have **no header**. Each row is
+       ``(frame index, x-coord (mm), y-coord (mm), z-coord (mm))``:
+
+    .. code-block:: text
+
+       185, 5.32, 2.37, 0.0
+       ...
+
+    Parameters
+    ----------
+    input_path_a : str
+        Path to the REST input folder.
+    input_path_b : str
+        Path to the STRESS input folder.
+    label : str, optional
+        Label used for output filenames.  Default is ``"full"``.
+    diastole : bool, optional
+        Whether to process the diastolic phase.  Default is ``True``.
+    step_rotation_deg : float, optional
+        Rotation step in degrees.  Default is ``0.5``.
+    range_rotation_deg : float, optional
+        Rotation search range (±) in degrees; a range of 90° gives 180°
+        total.  Default is ``90.0``.
+    image_center : tuple of float, optional
+        Image center ``(x, y)`` in mm.  Default is ``(4.5, 4.5)``.
+    radius : float, optional
[…]
