### 2026-03-09 75c6fbd
move debug plots to own module
