+++ multimodars/ccta/__init__.py
@@ -1,1 +1,2 @@
 from . import adjust_ccta
+from . import debug_plots
