+++ multimodars/ccta/debug_plots.py
@@ -1,0 +1,272 @@
+from __future__ import annotations
+
+import numpy as np
+import trimesh
+from trimesh.points import PointCloud
+
+
+def labeled_geometry_plot(
+    mesh: trimesh.Trimesh,
+    rca_points: list,
+    lca_points: list,
+    rca_removed_points: list,
+    lca_removed_points: list,
+    cl_rca=None,
+    cl_lca=None,
+    cl_aorta=None,
+):
+    """Open an interactive 3-D scene visualising the labelled mesh regions.
+
+    Colour coding:
+
+    * **Yellow** - aortic vertices (not in RCA or LCA).
+    * **Blue** - RCA coronary vertices.
+    * **Green** - LCA coronary vertices.
+    * **Red** - vertices removed by occlusion detection.
+    * **Dark blue / dark green / dark yellow** - RCA / LCA / aorta centerline
+      points.
+
+    Parameters
+    ----------
+    mesh : trimesh.Trimesh
+        The surface mesh displayed semi-transparently in the background.
+    rca_points : list of tuple
+        RCA vertex coordinates.
+    lca_points : list of tuple
+        LCA vertex coordinates.
+    rca_removed_points : list of tuple
+        RCA vertices that were flagged by occlusion removal.
+    lca_removed_points : list of tuple
+        LCA vertices that were flagged by occlusion removal.
+    cl_rca : PyCenterline, optional
+        RCA centerline object; plotted when provided.
+    cl_lca : PyCenterline, optional
+        LCA centerline object; plotted when provided.
+    cl_aorta : PyCenterline, optional
+        Aorta centerline object; plotted when provided.
+    """
+    rca_set = set(rca_points)
+    lca_set = set(lca_points)
+    aortic_points = [
+        tuple(vertex)
+        for vertex in mesh.vertices
+        if tuple(vertex) not in rca_set and tuple(vertex) not in lca_set
+    ]
+
+    print(f"\n=== DEBUG PLOT STATISTICS ===")
+    print(f"Total mesh vertices: {len(mesh.vertices)}")
+    print(f"Aortic points (yellow): {len(aortic_points)}")
+    print(f"RCA coronary points (blue): {len(rca_points)}")
+    print(f"LCA coronary points (green): {len(lca_points)}")
+    print(f"RCA reassigned points (red): {len(rca_removed_points)}")
+    print(f"LCA reassigned points (red): {len(lca_removed_points)}")
+
+    scene_geoms = []
+
+    # Add mesh (semi-transparent)
+    mesh_visual = mesh.copy()
+    mesh_visual.visual.face_colors = [200, 200, 200, 100]  # Semi-transparent gray
+    scene_geoms.append(mesh_visual)
+
+    if aortic_points:
+        aortic_array = np.array(aortic_points, dtype=np.float64)
+        aortic_colors = np.tile([255, 255, 0, 255], (len(aortic_points), 1))  # Yellow
+        scene_geoms.append(PointCloud(aortic_array, colors=aortic_colors))
+
+    if rca_points:
+        rca_array = np.array(rca_points, dtype=np.float64)
+        rca_colors = np.tile([0, 0, 255, 255], (len(rca_points), 1))  # Blue
+        scene_geoms.append(PointCloud(rca_array, colors=rca_colors))
+
+    if lca_points:
+        lca_array = np.array(lca_points, dtype=np.float64)
+        lca_colors = np.tile([0, 255, 0, 255], (len(lca_points), 1))  # Green
+        scene_geoms.append(PointCloud(lca_array, colors=lca_colors))
+
+    all_removed = rca_removed_points + lca_removed_points
+    if all_removed:
+        removed_array = np.array(all_removed, dtype=np.float64)
[…]
