### 2026-03-09 f3100f1
add a point deleting function to adjust_ccta and set up stitching
