+++ multimodars/ccta/adjust_ccta.py
@@ -4,8 +4,8 @@ import numpy as np
 import trimesh
 from pathlib import Path
 import warnings
-from trimesh.points import PointCloud
 from ..multimodars import (
+    PyGeometry,
     find_centerline_bounded_points_simple,
     remove_occluded_points_ray_triangle,
     clean_outlier_points,
@@ -17,6 +17,7 @@ from ..multimodars import (
 )
 from .._converters import numpy_to_centerline
 from ..io.read_geometrical import read_mesh
+from .debug_plots import labeled_geometry_plot, plot_anomalous_region, compare_centerline_scaling
 
 
 def label_geometry(
@@ -34,7 +35,7 @@ def label_geometry(
     """Label CCTA mesh vertices as aorta, RCA, or LCA using centerline-based region detection.
 
     Loads a 3-D surface mesh and three centerlines (aorta, RCA, LCA), then assigns
-    each mesh vertex to one of the anatomical regions.  For anomalous vessels an
+    each mesh vertex to one of the anatomical regions. For anomalous vessels an
     additional occlusion-removal step uses ray-triangle intersection to strip
     intramural segments, followed by adjacency-map reclassification to clean up
     isolated mis-labelled vertices. Herfore, a ray is cast from every aorta point to the centerline
@@ -188,11 +189,15 @@ def label_geometry(
     final_lca_points, final_aortic_points = clean_outlier_points(
         final_lca_points_found, aortic_points, 2.0, 0.4
     )  # based on patient data, only precleaning anyways, rest done by final_reclassification
-    final_rca_points, final_aortic_points = clean_outlier_points(
+    final_rca_points, _aortic_points = clean_outlier_points(
         final_rca_points_found, final_aortic_points, 2.0, 0.4
     )
-    aortic_points = _find_aortic_points(
+    final_aortic_points = _find_aortic_points(
         mesh.vertices, final_rca_points, final_lca_points
+    )
+    # add also the rca_removed points and lca_removed points to aortic points
+    final_aortic_points = list(
+        set(final_aortic_points) | set(rca_removed_points) | set(lca_removed_points)
     )
     print(f"length after: {len(final_lca_points)}")
 
@@ -210,7 +215,7 @@ def label_geometry(
     new_results = _final_reclassification(results)
 
     if control_plot:
-        _labeled_geometry_plot(
+        labeled_geometry_plot(
             mesh=new_results["mesh"],
             rca_points=new_results["rca_points"],
             lca_points=new_results["lca_points"],
@@ -449,128 +454,6 @@ def _final_reclassification(results: dict) -> dict:
     return updated_results
 
 
-def _labeled_geometry_plot(
-    mesh: trimesh.Trimesh,
-    rca_points: list,
-    lca_points: list,
-    rca_removed_points: list,
-    lca_removed_points: list,
-    cl_rca=None,
-    cl_lca=None,
-    cl_aorta=None,
-):
-    """Open an interactive 3-D scene visualising the labelled mesh regions.
-
-    Colour coding:
-
-    * **Yellow** - aortic vertices (not in RCA or LCA).
-    * **Blue** - RCA coronary vertices.
-    * **Green** - LCA coronary vertices.
-    * **Red** - vertices removed by occlusion detection.
-    * **Dark blue / dark green / dark yellow** - RCA / LCA / aorta centerline
-      points.
-
-    Parameters
-    ----------
-    mesh : trimesh.Trimesh
-        The surface mesh displayed semi-transparently in the background.
-    rca_points : list of tuple
-        RCA vertex coordinates.
-    lca_points : list of tuple
-        LCA vertex coordinates.
[…]
