### 2026-03-10 2e0cb62
add prototype stitching
