+++ multimodars/ccta/adjust_ccta.py
@@ -15,7 +15,7 @@ from ..multimodars import (
     find_proximal_distal_scaling,
     find_aortic_scaling,
 )
-from .._converters import numpy_to_centerline
+from .._converters import numpy_to_centerline, geometry_to_trimesh
 from ..io.read_geometrical import read_mesh
 from .debug_plots import labeled_geometry_plot, plot_anomalous_region, compare_centerline_scaling
 
@@ -604,27 +604,85 @@ def remove_anomalous_points_from_mesh(results: dict) -> dict:
     return updated
 
 
+
 def stitch_ccta_to_intravascular(
     iv_mesh: PyGeometry,
     mesh: trimesh.Trimesh,
     results: dict,
-) -> trimesh.Trimesh:
+) -> dict:
+    """
+    Uses an aligned intravascular mesh and stitches it to a CCTA mesh using the following steps.
+    """
     proximal_centroid = iv_mesh.frames[0].centroid
     distal_centroid = iv_mesh.frames[-1].centroid
+    proximal_points = iv_mesh.frames[0].lumen.points
+    distal_points = iv_mesh.frames[-1].lumen.points
 
+    prox_boundary_pts, dist_boundary_pts = _prepare_prox_dist_boundary_pts(
+        mesh,
+        results,
+        proximal_centroid,
+        distal_centroid,
+    )
+    prox_point_step = len(proximal_points) // len(prox_boundary_pts)
+    dist_point_step = len(distal_points) // len(dist_boundary_pts)
+
+    # Adjust start point
+    prox_boundary_pts, dist_boundary_pts = _adjust_start_point(
+        prox_boundary_pts,
+        dist_boundary_pts,
+        proximal_points,
+        distal_points,
+    )
+
+    # Check direction of boundary points by keeping only
+    prox_boundary_pts, dist_boundary_pts = _check_ring_direction(
+        prox_boundary_pts,
+        dist_boundary_pts,
+        proximal_points,
+        distal_points,
+        prox_point_step,
+        dist_point_step,
+    )
+
+    # Step 3: stitch each boundary ring to its IV ring
+    prox_patch = _stitch_boundary_ring(prox_boundary_pts, proximal_points, prox_point_step)
+    dist_patch = _stitch_boundary_ring(dist_boundary_pts, distal_points, dist_point_step)
+    test_mesh = geometry_to_trimesh(iv_mesh)
+    test_mesh.update_faces(test_mesh.unique_faces())
+    test_mesh.update_faces(test_mesh.nondegenerate_faces())
+    test_mesh.fix_normals()
+    mesh = trimesh.util.concatenate([mesh, prox_patch, dist_patch, test_mesh])
+    mesh.merge_vertices()
+    mesh.export("test_mesh.stl")
+    results["prox_boundary_points"] = prox_boundary_pts
+    results["dist_boundary_points"] = dist_boundary_pts
+    results["mesh"] = mesh
+    
+    return results 
+
+
+def _prepare_prox_dist_boundary_pts(
+        mesh: trimesh.Trimesh, 
+        results: dict, 
+        prox_centroid: tuple[float, float, float],
+        dist_centroid: tuple[float, float, float],
+) -> tuple[list, list]:
     proximal_boundary_pts = []
     distal_boundary_pts = []
-
+    print(f"Number of boundary points: {len(results["boundary_points"])}")
     for pt in results["boundary_points"]:
-        distance_prox = np.linalg.norm(np.array(proximal_centroid) - np.array(pt))
-        distance_dist = np.linalg.norm(np.array(distal_centroid) - np.array(pt))
+        distance_prox = np.linalg.norm(np.array(prox_centroid) - np.array(pt))
[…]
