### 2026-03-10 4420c16
fix stitching by using an outward_reference
