### 2026-03-10 3fab54f
fix normals after conversion and stitching
