+++ pyproject.toml
@@ -11,6 +11,7 @@ dependencies = [
     "numpy>=2.0.0,<3.0.0",
     "trimesh>=4.8.0",
     "pyglet<2",
+    "pymeshlab",
 ]
 classifiers = [
     "Programming Language :: Rust",
