### 2026-03-11 04c055a
small refactor of ccta modules
