+++ multimodars/ccta/__init__.py
@@ -1,3 +1,4 @@
-from . import adjust_ccta
+from . import manipulating
+from . import labeling
 from . import debug_plots
 from . import fixing_functions
