+++ multimodars/ccta/adjust_ccta.py
@@ -1,1178 +1,0 @@
-from __future__ import annotations
-
-import numpy as np
-import trimesh
-from pathlib import Path
-import warnings
-from ..multimodars import (
-    PyGeometry,
-    find_centerline_bounded_points_simple,
-    remove_occluded_points_ray_triangle,
-    clean_outlier_points,
-    build_adjacency_map,
-    find_points_by_cl_region,
-    adjust_diameter_centerline_morphing_simple,
-    find_proximal_distal_scaling,
-    find_aortic_scaling,
-)
-from .._converters import numpy_to_centerline, geometry_to_trimesh
-from ..io.read_geometrical import read_mesh
-from .debug_plots import labeled_geometry_plot, plot_anomalous_region, compare_centerline_scaling
-
-
-def label_geometry(
-    path_ccta_geometry: Path | str,
-    path_centerline_aorta: Path | str,
-    path_centerline_rca: Path | str,
-    path_centerline_lca: Path | str,
-    anomalous_rca: bool = False,
-    anomalous_lca: bool = False,
-    n_points_intramural: int = 120,
-    bounding_sphere_radius_mm: float = 3.0,
-    tolerance_float: float = 1e-6,
-    control_plot: bool = True,
-) -> tuple[dict, tuple[any, any, any]]:
-    """Label CCTA mesh vertices as aorta, RCA, or LCA using centerline-based region detection.
-
-    Loads a 3-D surface mesh and three centerlines (aorta, RCA, LCA), then assigns
-    each mesh vertex to one of the anatomical regions. For anomalous vessels an
-    additional occlusion-removal step uses ray-triangle intersection to strip
-    intramural segments, followed by adjacency-map reclassification to clean up
-    isolated mis-labelled vertices. Herfore, a ray is cast from every aorta point to the centerline
-    points of the anomalous section and if 3 faces are intersected by the ray the points from
-    the first face must correspond to the intramural section.
-
-    Parameters
-    ----------
-    path_ccta_geometry : Path or str
-        Path to the CCTA surface mesh file (any format supported by
-        :func:`multimodars.io.read_geometrical.read_mesh`).
-    path_centerline_aorta : Path or str
-        Path to a CSV file containing the aortic centerline (comma-delimited,
-        columns: x, y, z, …).
-    path_centerline_rca : Path or str
-        Path to a CSV file containing the RCA centerline.
-    path_centerline_lca : Path or str
-        Path to a CSV file containing the LCA centerline.
-    anomalous_rca : bool, optional
-        When ``True`` applies ray-triangle occlusion removal to the RCA region
-        to handle anomalous (intramural) courses.  Default is ``False``.
-    anomalous_lca : bool, optional
-        When ``True`` applies ray-triangle occlusion removal to the LCA region.
-        Default is ``False``.
-    n_points_intramural : int, optional
-        Number of coronary centerline points examined during occlusion removal
-        (the intramural segment length).  Default is ``120``.
-    bounding_sphere_radius_mm : float, optional
-        Radius in millimetres of the rolling sphere used to collect candidate
-        mesh vertices around each centerline point.  Default is ``3.0``.
-    tolerance_float : float, optional
-        Distance tolerance used when matching mesh vertices to points during
-        face lookup.  Default is ``1e-6``.
-    control_plot : bool, optional
-        When ``True`` opens an interactive 3-D scene showing the labelled mesh
-        after processing.  Default is ``True``.
-
-    Returns
-    -------
-    results : dict
-        Dictionary with keys:
-
[…]

[…]
