### 2026-03-12 5c2fb64
simplify plotting
