### 2026-03-12 5d27755
change pymeshlab and pyglet  as optional dependencies
