### 2026-03-12 635495c
add try except for pymeshlab
