+++ multimodars/ccta/manipulating.py
@@ -93,7 +93,6 @@ def find_distal_and_proximal_scaling(
     results: dict,
     dist_range: int = 3,
     prox_range: int = 2,
-    debug_plot: bool = True,
 ) -> tuple[float, float]:
     """Compute the optimal radial scaling factors for the proximal and distal segments.
 
@@ -135,7 +134,7 @@ def find_distal_and_proximal_scaling(
     n_anomalous_points = len(results["anomalous_points"])
     n_section: int = int(np.ceil(0.25 * n_anomalous_points))
 
-    print("=== Finding best scaling factors ===")
+    print("\n=== Finding best scaling factors ===")
     prox_scaling, dist_scaling = find_proximal_distal_scaling(
         results["anomalous_points"],
         n_section,
@@ -144,8 +143,9 @@ def find_distal_and_proximal_scaling(
         frame_points_prox,
         frame_points_dist,
     )
-    print(f"Best proximal scaling: {prox_scaling}")
-    print(f"Best distal scaling: {dist_scaling}")
+    print(f"Best proximal scaling: {np.round(prox_scaling, 2)} mm")
+    print(f"Best distal scaling: {np.round(dist_scaling, 2)} mm")
+    print("====================================\n")
 
     return prox_scaling, dist_scaling
 
@@ -154,7 +154,6 @@ def find_aorta_scaling(
     frames,
     centerline,
     results: dict,
-    debug_plot: bool = True,
 ) -> float:
     """Compute the optimal radial scaling factor for the aortic wall region.
 
@@ -182,13 +181,14 @@ def find_aorta_scaling(
     """
     reference_points = _extract_wall_from_frames(frames)
 
-    print("=== Finding best scaling factor ===")
+    print("\n=== Finding best scaling factor ===")
     scaling = find_aortic_scaling(
         results["rca_removed_points"],  # For now work with removed points
         reference_points,
         centerline,
     )
-    print(f"Best aortic scaling: {scaling}")
+    print(f"Best aortic scaling: {np.round(scaling, 2)} mm")
+    print("====================================\n")
 
     return scaling
 
@@ -255,34 +255,50 @@ def _extract_wall_from_frames(frames) -> list[tuple[float, float, float]]:
     return reference_points
 
 
-def remove_anomalous_points_from_mesh(results: dict) -> dict:
-    """Remove anomalous vertices from the mesh and update all point lists.
+def remove_labeled_points_from_mesh(
+    results: dict,
+    region_keys: list[str] | str = "anomalous_points",
+) -> dict:
+    """Remove one or more labeled regions of vertices from the mesh.
 
-    Deletes the vertices listed under ``"anomalous_points"`` (and any faces
-    referencing them) from the mesh, remaps the remaining faces, and rebuilds
-    every coordinate list in *results* to reflect the new vertex indices.
+    Collects all points stored under *region_keys*, deletes the corresponding
+    vertices (and any faces referencing them) from the mesh, remaps the
+    remaining faces, and rebuilds every coordinate list in *results* to
+    reflect the new vertex indices.
 
     Parameters
     ----------
     results : dict
-        Dictionary produced by :func:`label_anomalous_region` containing at
-        minimum the keys ``"mesh"`` and ``"anomalous_points"``.  Any of
+        Dictionary containing at minimum the key ``"mesh"``.  Any of
         ``"aorta_points"``, ``"rca_points"``, ``"lca_points"``,
         ``"rca_removed_points"``, ``"lca_removed_points"``,
         ``"proximal_points"``, and ``"distal_points"`` are also updated if
         present.
+    region_keys : str or list of str
+        Key(s) in *results* whose point lists should be removed from the mesh.
[…]
