### 2026-03-12 568e95a
add manual_hole filling function
