### 2026-03-12 766a558
replace anomalous points by pygeometry in stitch_ccta_to_intravascular
