+++ ROADMAP.MD
@@ -33,6 +33,9 @@
 - [] automatically find the required reference points for alignment
 - [] automatically assess lumen area, minor-, major axis, mla, intramural length etc. from the CCTA mesh.
 - [] return hausdorff distances per point between aligned ivus mesh and ccta mesh.
+- [] function to automatically assign wall thickness to ccta and combined ivus/ccta mesh
+- [] helper function to automatically export only certain points on mesh (e.g. RCA or LCA)
+- [] function to combine anomalous section only with the distal part
 
 # v0.1.0 desired functionality:
 - [x] .from_file(mode = {full, single}, "rest dir", "stress dir", "dia dir", "sys dir")
@@ -56,8 +59,8 @@
 - [x] High-level "Getting Started" guide
 - [x] CONTRIBUTING.md
 - [x] Example scripts Jupyter notebook?
-- [] make stable version update by reading in a __version__ file
+- [x] make stable version update by reading in a version directly from ``Cargo.toml``
+- [x] move example data to a downloadable folder in release and adjust tests to only use fixtures
 - [] Integrate coverage measurement (e.g. pytest-cov) and upload reports to Coverall or Codecov
 - [] Status badge for coverage to README
 - [] mypy in pre-commit hooks
-- [] move example data to a downloadable folder in release and adjust tests to only use fixtures
