### 2026-03-12 4760457
add wrapper functions to ccta module
