+++ multimodars/ccta/labeling.py
@@ -1,3 +1,6 @@
+from __future__ import annotations
+
+
 from pathlib import Path
 import trimesh
 import numpy as np
@@ -8,23 +11,25 @@ from ..multimodars import (
     clean_outlier_points,
     build_adjacency_map,
     find_points_by_cl_region,
+    PyCenterline,
 )
 from .._converters import numpy_to_centerline
 from ..io.read_geometrical import read_mesh
 from .debug_plots import plot_results_key, compare_centerline_scaling
 
+
 def label_geometry(
-    path_ccta_geometry: Path | str,
-    path_centerline_aorta: Path | str,
-    path_centerline_rca: Path | str,
-    path_centerline_lca: Path | str,
+    path_ccta_geometry: Path | str | trimesh.Trimesh,
+    path_centerline_aorta: Path | str | PyCenterline,
+    path_centerline_rca: Path | str | PyCenterline,
+    path_centerline_lca: Path | str | PyCenterline,
     anomalous_rca: bool = False,
     anomalous_lca: bool = False,
     n_points_intramural: int = 120,
     bounding_sphere_radius_mm: float = 3.0,
     tolerance_float: float = 1e-6,
     control_plot: bool = True,
-) -> tuple[dict, tuple[any, any, any]]:
+) -> tuple[dict, tuple[PyCenterline, PyCenterline, PyCenterline]]:
     """Label CCTA mesh vertices as aorta, RCA, or LCA using centerline-based region detection.
 
     Loads a 3-D surface mesh and three centerlines (aorta, RCA, LCA), then assigns
@@ -87,36 +92,61 @@ def label_geometry(
         Re-raises any error that occurs while reading the mesh or centerline
         files, after printing a descriptive message.
     """
-    try:
-        mesh = read_mesh(path_ccta_geometry)
-        print(f"Loaded mesh: {len(mesh.vertices)} vertices, {len(mesh.faces)} faces")
-    except Exception as e:
-        print(f"Error reading CCTA mesh from {path_ccta_geometry}: {e}")
-        raise
+    if isinstance(path_ccta_geometry, trimesh.Trimesh):
+        mesh = path_ccta_geometry
+        print(f"Using provided mesh: {len(mesh.vertices)} vertices, {len(mesh.faces)} faces")
+    else:
+        try:
+            mesh = read_mesh(path_ccta_geometry)
+            print(f"Loaded mesh: {len(mesh.vertices)} vertices, {len(mesh.faces)} faces")
+        except Exception as e:
+            print(f"Error reading CCTA mesh from {path_ccta_geometry}: {e}")
+            raise
 
-    try:
-        cl_aorta_raw = np.genfromtxt(path_centerline_aorta, delimiter=",")
-        cl_aorta = numpy_to_centerline(cl_aorta_raw)
-        print(f"Loaded aorta centerline: {len(cl_aorta.points)} points")
-    except Exception as e:
-        print(f"Error reading Aorta centerline from {path_centerline_aorta}: {e}")
-        raise
+    if isinstance(path_centerline_aorta, PyCenterline):
+        cl_aorta = path_centerline_aorta
+        print(f"Using provided aorta centerline: {len(cl_aorta.points)} points")
+    elif isinstance(path_centerline_aorta, np.ndarray):
+        cl_aorta = numpy_to_centerline(path_centerline_aorta)
+        print(f"Using provided aorta centerline: {len(cl_aorta.points)} points")
+    else:
+        try:
+            cl_aorta_raw = np.genfromtxt(path_centerline_aorta, delimiter=",")
+            cl_aorta = numpy_to_centerline(cl_aorta_raw)
+            print(f"Loaded aorta centerline: {len(cl_aorta.points)} points")
+        except Exception as e:
[…]
