+++ multimodars/ccta/__init__.py
@@ -1,4 +1,278 @@
+from __future__ import annotations
+
 from . import manipulating
 from . import labeling
 from . import debug_plots
 from . import fixing_functions
+
+from pathlib import Path
+from typing import TYPE_CHECKING
+import trimesh
+
+if TYPE_CHECKING:
+    from ..multimodars import PyCenterline, PyFrame, PyGeometry
+
+# -------------------------------
+# Convenience Functions
+# -------------------------------
+def label(
+    path_ccta_geometry: Path | str | trimesh.Trimesh,
+    path_centerline_aorta: Path | str | PyCenterline,
+    path_centerline_rca: Path | str | PyCenterline,
+    path_centerline_lca: Path | str | PyCenterline,
+    aligned_frames: list[PyFrame],
+    anomalous_rca: bool = False,
+    anomalous_lca: bool = False,
+    n_points_intramural: int = 120,
+    bounding_sphere_radius_mm: float = 3.0,
+    tolerance_float: float = 1e-6,
+    control_plot: bool = True,
+) -> tuple[dict, tuple[PyCenterline, PyCenterline, PyCenterline]]:
+    """Label CCTA mesh vertices as aorta, RCA, or LCA using centerline-based region detection.
+
+    Loads a 3-D surface mesh and three centerlines (aorta, RCA, LCA), then assigns
+    each mesh vertex to one of the anatomical regions. For anomalous vessels an
+    additional occlusion-removal step uses ray-triangle intersection to strip
+    intramural segments, followed by adjacency-map reclassification to clean up
+    isolated mis-labelled vertices. Herfore, a ray is cast from every aorta point to the centerline
+    points of the anomalous section and if 3 faces are intersected by the ray the points from
+    the first face must correspond to the intramural section.
+
+    Additionally partition a coronary region into proximal, anomalous, and distal sub-regions.
+
+    Uses the intravascular imaging frames to determine where along the centerline
+    the anomalous (intramural) segment begins and ends, then tags each mesh
+    vertex accordingly.
+
+    Parameters
+    ----------
+    path_ccta_geometry : Path or str
+        Path to the CCTA surface mesh file (any format supported by
+        :func:`multimodars.io.read_geometrical.read_mesh`).
+    path_centerline_aorta : Path or str
+        Path to a CSV file containing the aortic centerline (comma-delimited,
+        columns: x, y, z, …).
+    path_centerline_rca : Path or str
+        Path to a CSV file containing the RCA centerline.
+    path_centerline_lca : Path or str
+        Path to a CSV file containing the LCA centerline.
+    aligned_frames : list of PyFrame
+        Ordered list of intravascular imaging frames for the vessel.
+    anomalous_rca : bool, optional
+        When ``True`` applies ray-triangle occlusion removal to the RCA region
+        to handle anomalous (intramural) courses.  Default is ``False``.
+    anomalous_lca : bool, optional
+        When ``True`` applies ray-triangle occlusion removal to the LCA region.
+        Default is ``False``.
+    n_points_intramural : int, optional
+        Number of coronary centerline points examined during occlusion removal
+        (the intramural segment length).  Default is ``120``.
+    bounding_sphere_radius_mm : float, optional
+        Radius in millimetres of the rolling sphere used to collect candidate
+        mesh vertices around each centerline point.  Default is ``3.0``.
+    tolerance_float : float, optional
+        Distance tolerance used when matching mesh vertices to points during
+        face lookup.  Default is ``1e-6``.
+    control_plot : bool, optional
[…]

[…]
