### 2026-03-12 379810d
add tests for ccta functions
