+++ tests/test_ccta.py
@@ -1,0 +1,672 @@
+"""Tests for the multimodars.ccta module.
+
+Covers:
+  - labeling:   _find_aortic_points, _find_faces_for_points,
+                _prepare_faces_for_rust, _final_reclassification
+  - fixing_functions: manual_hole_fill, postprocess_stitched_mesh
+  - manipulating: remove_labeled_points_from_mesh,
+                  keep_labeled_points_from_mesh, sync_results_to_mesh,
+                  order_points_list, scale_region_centerline_morphing,
+                  _adjust_start_point, _check_ring_direction,
+                  _stitch_boundary_ring
+"""
+from __future__ import annotations
+
+import importlib
+
+import numpy as np
+import pytest
+import trimesh
+
+from multimodars import PyContourPoint
+from multimodars.ccta.fixing_functions import manual_hole_fill, postprocess_stitched_mesh
+from multimodars.ccta.labeling import (
+    _find_aortic_points,
+    _find_faces_for_points,
+    _final_reclassification,
+    _prepare_faces_for_rust,
+)
+from multimodars.ccta.manipulating import (
+    _adjust_start_point,
+    _check_ring_direction,
+    _stitch_boundary_ring,
+    keep_labeled_points_from_mesh,
+    order_points_list,
+    remove_labeled_points_from_mesh,
+    scale_region_centerline_morphing,
+    sync_results_to_mesh,
+)
+
+
+# ---------------------------------------------------------------------------
+# Shared mesh factories
+# ---------------------------------------------------------------------------
+
+def _make_grid_mesh() -> trimesh.Trimesh:
+    """3x3 grid (9 vertices, 8 triangular faces, z=0 plane).
+
+    Vertex layout:
+        6--7--8
+        |/|/|
+        3--4--5
+        |/|/|
+        0--1--2
+    """
+    verts = np.array(
+        [
+            [0.0, 0.0, 0.0],  # 0  corner
+            [1.0, 0.0, 0.0],  # 1
+            [2.0, 0.0, 0.0],  # 2  corner
+            [0.0, 1.0, 0.0],  # 3
+            [1.0, 1.0, 0.0],  # 4  centre – adjacent to {1,2,3,5,6,7}
+            [2.0, 1.0, 0.0],  # 5
+            [0.0, 2.0, 0.0],  # 6  corner
+            [1.0, 2.0, 0.0],  # 7
+            [2.0, 2.0, 0.0],  # 8  corner
+        ],
+        dtype=float,
+    )
+    faces = np.array(
+        [
+            [0, 1, 3],
+            [1, 4, 3],
+            [1, 2, 4],
+            [2, 5, 4],
+            [3, 4, 6],
+            [4, 7, 6],
+            [4, 5, 7],
+            [5, 8, 7],
+        ]
+    )
+    return trimesh.Trimesh(vertices=verts, faces=faces, process=False)
+
+
+def _make_hex_fan_mesh() -> trimesh.Trimesh:
+    """6 outer vertices (0-5) + centre (6), 6 triangular faces.
+
+    Boundary ring: vertices 0-5 in angular order.
+    """
+    angles = np.linspace(0, 2 * np.pi, 6, endpoint=False)
+    outer = np.column_stack([np.cos(angles), np.sin(angles), np.zeros(6)])
+    centre = np.array([[0.0, 0.0, 0.0]])
+    verts = np.vstack([outer, centre])
+    faces = np.array([[i, (i + 1) % 6, 6] for i in range(6)])
+    return trimesh.Trimesh(vertices=verts, faces=faces, process=False)
+
+
+def _make_iv_pts(coords) -> list[PyContourPoint]:
+    return [
[…]
