### 2026-03-13 aea3907
add region export (connecting)
