### 2026-03-13 0685a6a
add algorithm for stitching vessel either to proximal section or directly to aorta for anomalies
