+++ multimodars/ccta/manipulating.py
@@ -513,12 +513,35 @@ def sync_results_to_mesh(
     return updated
 
 
+def _rotate_to_nearest_iv(boundary_pts: list, iv_pt) -> list:
+    """Rotate a boundary ring so the point nearest to *iv_pt* is first."""
+    iv_arr = np.array([iv_pt.x, iv_pt.y, iv_pt.z])
+    dists = [np.linalg.norm(np.array(pt) - iv_arr) for pt in boundary_pts]
+    start_idx = int(np.argmin(dists))
+    return boundary_pts[start_idx:] + boundary_pts[:start_idx]
+
+
+def _adjust_start_point_by_z(boundary_pts: list) -> list:
+    """Rotate a boundary ring so the point with the highest z-value is first."""
+    start_idx = int(np.argmax([pt[2] for pt in boundary_pts]))
+    return boundary_pts[start_idx:] + boundary_pts[:start_idx]
+
+
 def stitch_ccta_to_intravascular(
     iv_mesh: PyGeometry,
     mesh: trimesh.Trimesh,
     results: dict,
     n_points_iv_cont: int = 100,
+    prox_start_mode: str = "nearest_iv",
+    dist_start_mode: str = "nearest_iv",
 ) -> dict:
+    """
+    ``prox_start_mode`` / ``dist_start_mode`` control how index 0 of each
+    boundary ring is chosen before stitching:
+
+    * ``"nearest_iv"`` (default) - rotate to the point closest to IV point 0.
+    * ``"highest_z"`` - rotate to the point with the largest z-coordinate.
+    """
     """
     Uses an aligned intravascular mesh and stitches it to a CCTA mesh using the following steps.
     """
@@ -539,24 +562,21 @@ def stitch_ccta_to_intravascular(
     dist_point_step = len(distal_points) // len(dist_boundary_pts)
 
     # Adjust start point
-    prox_boundary_pts, dist_boundary_pts = _adjust_start_point(
-        prox_boundary_pts,
-        dist_boundary_pts,
-        proximal_points,
-        distal_points,
-    )
+    if prox_start_mode == "highest_z" or dist_start_mode == "highest_z":
+        iv_mesh = iv_mesh.sort_frame_points()
+        proximal_points = iv_mesh.frames[0].lumen.points
+        distal_points = iv_mesh.frames[-1].lumen.points
+    if prox_start_mode == "highest_z":
+        prox_boundary_pts = _adjust_start_point_by_z(prox_boundary_pts)
+    else:
+        prox_boundary_pts = _rotate_to_nearest_iv(prox_boundary_pts, proximal_points[0])
+    if dist_start_mode == "highest_z":
+        dist_boundary_pts = _adjust_start_point_by_z(dist_boundary_pts)
+    else:
+        dist_boundary_pts = _rotate_to_nearest_iv(dist_boundary_pts, distal_points[0])
 
-    # Check direction of boundary points by keeping only
-    prox_boundary_pts, dist_boundary_pts = _check_ring_direction(
-        prox_boundary_pts,
-        dist_boundary_pts,
-        proximal_points,
-        distal_points,
-        prox_point_step,
-        dist_point_step,
-    )
-
-    # Compute the vessel axis so each patch can be flipped to face outward.
+    # Compute the vessel axis so each patch can be flipped to face outward,
+    # and also used as the consistent reference normal for direction checking.
     # frames[0] is the "proximal" end, frames[-1] the "distal" end.
     # The outward direction for the proximal patch points away from the mesh
     # interior (toward frames[0]), and vice-versa for the distal patch.
@@ -564,6 +584,26 @@ def stitch_ccta_to_intravascular(
     dist_c = np.array(iv_mesh.frames[-1].centroid)
[…]
