### 2026-03-16 #10131 @crnh
I verified the functionality of `multimodars` with v0.2.3 and Python 3.12, and completed the review checklist.
Additionally, I have the following minor comments and suggestions for improvement.

> I agree with the reviewer, that this can be confusing to the user. Our thinking was that the notebooks are somewhat separate from the core package, so their dependencies are only loaded if a user chooses to download the examples folder. We're happy to add dev dependencies and move the full files to the main directory if that's preferred, just let us know.

I agree with the authors that the example dependencies should not be included in the dependencies of the `multimodars`. However, the example notebook could contain a requirements section listing additional dependencies needed to run the example. I noticed the example notebook now contains `%pip install` statements, which is also a good solution. However, when the notebook is executed in a virtualenv created with a popular project manager like `uv`, these statements do not work.

> We weren't aware of this feature, this was a really useful suggestion! The automatic generation approach looked promising but seemed a bit experimental and didn't appear to support function-based declarations, so we've gone with a manually written stub for now.

The written stubs work well, and my editor does not complain about missing imports or non-existent attributes anymore.

## Examples

### `ivus_to_centerline.ipynb`

- The example notebook is not entirely self-contained: it requires images that are included in the repository but not in the examples folder. These images could be embedded in the notebook instead.
- The example now runs successfully.

### Intravascular tutorial

- Most code runs, as far as example data is available.
- Some cells depends on a `results` variable that is not defined nor explained in the notebook. For most cells it is clear that the user should still create this variable, but not for this cell:
  ```python
      aligned, resampled_cl = mm.align_combined(
      centerline,
      rest,
      (12.2605, -201.3643, 1751.0554),
      (11.7567, -202.1920, 1754.7975),
      (15.6605, -202.1920, 1749.9655),
      results["rca_points"],  # [(x, y, z), ...]
      angle_range_deg=10.0,
      write=True,
      watertight=True,
      output_dir="test",
  )
  ```

### CCTA tutorial

- I was able to complete this tutorial and did not encounter any issues.

### General advice on the tutorials

I do not consider this a review comment and this is only a suggestion, not a request for change.
The documentation would benefit from writing the tutorials as Jupyter notebooks using e.g. [`nbsphinx`](https://nbsphinx.readthedocs.io/) or [`myst-nb`](https://myst-nb.readthedocs.io/en/v0.13.2/index.html).
This would allow including outputs in the tutorials, so the user can verify the results.
Furthermore, the notebooks could be run when building the documentation, so the outputs are always up-to-date and errors in the tutorials are detected early.

## Paper

The JACC case report mentioned in the Research impact statement is cited elsewhere, but for clarity it could be cited in the Research impact statement as well.
