### 2026-03-18 8957eac
fix missing stub types
