### 2026-03-18 a5d6f6f
remove unnecessary stl read in rust
