+++ src/ccta/io/input.rs
@@ -1,220 +1,0 @@
-use anyhow::{anyhow, Context, Result};
-use std::collections::{HashMap, HashSet};
-use std::fs::OpenOptions;
-use std::path::Path;
-use stl_io::IndexedMesh;
-
-#[allow(dead_code)]
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub enum Label {
-    LCA,
-    RCA,
-    Intravascular,
-    Aorta,
-    Unknown,
-}
-
-#[allow(dead_code)]
-#[derive(Debug, Clone)]
-pub struct LabeledMesh {
-    pub mesh: IndexedMesh,
-    pub labels: Vec<Option<Label>>,
-    pub label_index: HashMap<Label, HashSet<usize>>,
-}
-
-#[allow(dead_code)]
-impl LabeledMesh {
-    /// Create a new LabeledMesh with all vertices unlabeled
-    pub fn from_mesh(mesh: IndexedMesh) -> Self {
-        let labels = vec![None; mesh.vertices.len()];
-        Self {
-            mesh,
-            labels,
-            label_index: HashMap::new(),
-        }
-    }
-
-    /// Assign a label to a vertex
-    pub fn set_label(&mut self, vertex_idx: usize, label: Label) {
-        if vertex_idx < self.labels.len() {
-            self.labels[vertex_idx] = Some(label);
-        } else {
-            eprintln!("Warning: vertex index {} out of bounds", vertex_idx);
-        }
-    }
-
-    /// Retrieve the label (if any) for a vertex
-    pub fn get_label(&self, vertex_idx: usize) -> Option<&Label> {
-        self.labels.get(vertex_idx).and_then(|o| o.as_ref())
-    }
-
-    /// Count how many vertices have a given label
-    pub fn count_label(&self, label: &Label) -> usize {
-        self.labels
-            .iter()
-            .filter(|l| l.as_ref() == Some(label))
-            .count()
-    }
-
-    /// Retrieve all vertices for a given label
-    pub fn vertices_with_label(&self, label: &Label) -> Vec<usize> {
-        self.labels
-            .iter()
-            .enumerate()
-            .filter_map(|(i, l)| {
-                if l.as_ref() == Some(label) {
-                    Some(i)
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-}
-
-#[allow(dead_code)]
-pub fn read_stl_ccta<P: AsRef<Path>>(path: P) -> Result<LabeledMesh> {
-    let path_ref = path.as_ref();
-
-    if !path_ref.exists() {
-        return Err(anyhow!("File does not exist: {:?}", path_ref.display()));
-    }
-
-    if !path_ref.is_file() {
-        return Err(anyhow!("Path is not a file: {:?}", path_ref.display()));
-    }
-
-    let mut file = OpenOptions::new()
-        .read(true)
-        .open(path_ref)
-        .with_context(|| format!("Failed to open file: {:?}", path_ref.display()))?;
-
-    let stl = stl_io::read_stl(&mut file)
-        .with_context(|| format!("Failed to parse STL file: {:?}", path_ref.display()))?;
-
-    if stl.validate().is_err() {
-        return Err(anyhow!(
-            "STL file is not a valid mesh: {:?}",
-            path_ref.display()
-        ));
-    }
-
-    Ok(LabeledMesh::from_mesh(stl))
-}
-
-#[allow(dead_code)]
-fn read_obj_ccta() -> () {
-    todo!()
-}
-
-// #[cfg(test)]
-// mod input_ccta_tests {
-//     use super::*;
-//     use std::fs;
-//     use std::path::PathBuf;
[…]
