+++ Cargo.toml
@@ -19,4 +19,3 @@ rayon = "1.7.1"
 crossbeam = "0.8.4"
 anyhow = "1.0.100"
 approx = "0.5"
-stl_io = "0.10.0"
