### 2026-03-18 47b168a
add mypy to pre-commit hook and fix type errors
