+++ multimodars/ccta/manipulating.py
@@ -50,14 +50,14 @@ def scale_region_centerline_morphing(
     """
     scaled_mesh = mesh.copy()
 
-    region_vertex_indices = []
+    region_vertex_indices_list: list[int] = []
     region_set = set(region_points)
 
     for idx, vertex in enumerate(scaled_mesh.vertices):
         if tuple(vertex) in region_set:
-            region_vertex_indices.append(idx)
+            region_vertex_indices_list.append(idx)
 
-    region_vertex_indices = np.array(region_vertex_indices)
+    region_vertex_indices = np.array(region_vertex_indices_list)
 
     if len(region_vertex_indices) == 0:
         print("Warning: No vertices found for scaling region")
@@ -82,7 +82,7 @@ def scale_region_centerline_morphing(
     )
 
     # Clear mesh cache since we modified vertices directly
-    scaled_mesh.vertices.flags['WRITEABLE'] = False
+    scaled_mesh.vertices.flags["WRITEABLE"] = False
 
     return scaled_mesh
 
@@ -131,12 +131,12 @@ def find_distal_and_proximal_scaling(
     frame_points_prox = [
         (p.x, p.y, p.z) for f in frames[0:prox_range] for p in f.lumen.points
     ]
-    n_anomalous_points = len(results['anomalous_points'])
+    n_anomalous_points = len(results["anomalous_points"])
     n_section: int = int(np.ceil(0.25 * n_anomalous_points))
 
     print("\n=== Finding best scaling factors ===")
     prox_scaling, dist_scaling = find_proximal_distal_scaling(
-        results['anomalous_points'],
+        results["anomalous_points"],
         n_section,
         n_section,
         centerline,
@@ -180,10 +180,12 @@ def find_aorta_scaling(
         Optimal radial scaling factor for the aortic segment.
     """
     reference_points = _extract_wall_from_frames(frames)
+    if reference_points is None:
+        raise ValueError("No aortic wall points found in frames for scaling reference")
 
     print("\n=== Finding best scaling factor ===")
     scaling = find_aortic_scaling(
-        results['rca_removed_points'],  # For now work with removed points
+        results["rca_removed_points"],  # For now work with removed points
         reference_points,
         centerline,
     )
@@ -193,7 +195,7 @@ def find_aorta_scaling(
     return scaling
 
 
-def _extract_wall_from_frames(frames) -> list[tuple[float, float, float]]:
+def _extract_wall_from_frames(frames) -> list[tuple[float, float, float]] | None:
     """Extract reconstructed aortic wall points from intravascular imaging frames.
 
     Iterates over *frames* looking for those that carry ``aortic_thickness``
@@ -231,7 +233,7 @@ def _extract_wall_from_frames(frames) -> list[tuple[float, float, float]]:
     for frame in frames:
         if frame.lumen.aortic_thickness is None:
             continue
-        if "Wall" not in frame.extras or frame.extras['Wall'] is None:
+        if "Wall" not in frame.extras or frame.extras["Wall"] is None:
             raise ValueError(
                 f"No Wall extras found for frame {getattr(frame, 'frame', '?')}"
             )
@@ -290,13 +292,9 @@ def remove_labeled_points_from_mesh(
     if isinstance(region_keys, str):
         region_keys = [region_keys]
 
-    mesh: trimesh.Trimesh = results['mesh']
+    mesh: trimesh.Trimesh = results["mesh"]
 
-    points_to_remove = [
-        pt
-        for key in region_keys
[…]
