+++ .pre-commit-config.yaml
@@ -28,11 +28,18 @@ repos:
         language: system
         types: [python]
         require_serial: true
-        args: [--quiet]
+        args: [--quiet, --target-version, py312]
 
       - id: python-ruff
         name: Ruff linting
-        entry: ruff
+        entry: ruff check
         language: system
         types: [python]
-        args: [--fix, --exit-non-zero-on-fix]
+        args: [--fix]
+
+      - id: python-mypy
+        name: Mypy type checking
+        entry: mypy multimodars --ignore-missing-imports
+        language: system
+        types: [python]
+        pass_filenames: false
