### 2026-03-19 2939129
fix 3d area calculation and elliptic ratio caculation add aortic wall optimization
