### 2026-03-19 4f28f8f
add wall scaling to python side
