### 2026-03-23 #10131 @richardkoehler
Hi @erickmartins and sorry for the minor delay, I had some urgent items come up. But as you have seen, I've already started work on the checklist and verified the repository, documentation and paper and I will finish up testing the software itself by the end of the week. Since the other two reviewers have already done a thorough check, I do not expect any major surprises here, so I thought I would bring up one topic to discussion, since I am not yet too familiar with the nuances of the JOSS editorial guidelines.

When looking at the repository, I noticed that the repository was created around 2 months (June 6th) before JOSS submission (Sep 9th). A large proportion of the contributions are then clustered in these 2 months (please let me know if I am mistaken, I am not familiar with all the ins-and-outs of GitHub, I find it confusing sometimes!), with likely the clear goal of submission to JOSS.

However the [reviewing guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#development-history-and-open-source-practice) state:

> Development timeline
The project should show evidence of sustained development over time (preferably months or years) rather than rapid, recent code generation. Look for:
>
> - Commit history spanning an extended period
> - Iterative improvements and refinements
> - Evolution of features and capabilities
>
> Good: Commits distributed over 6+ months showing gradual feature development
> OK: Development spanning 6+ months but with sporadic or bursty activity patterns (e.g., concentrated bursts, rather than steady continuous development)
> Not acceptable: All or most commits concentrated in the last few weeks before submission

Then, secondly the [guidelines ](https://joss.readthedocs.io/en/latest/review_criteria.html#collaborative-effort) state that collaborative effort is required. There are 4 contributors, but only the main author has contributed significant amounts of code, while the other three have 1-2 commits each with a dozen lines of code each. Also, apart from the issues of the other JOSS reviewers, there is not a single GitHub issue.

>Collaborative effort
> The commit history should show contributions from multiple developers and evidence of iterative refinement through community feedback:
>
> - Multiple contributors to the codebase
> - Code review through pull requests
> - Responses to issues and feature requests
> - Evidence of community-driven improvements
>
> Good: Multiple developers contributing to the codebase with evidence of iterative refinement through community feedback (issues, pull requests, code review)
> OK: Single author but shows other evidence of community engagement, either in the repository or evidenced in the paper
> Not acceptable: Single author with no evidence of community engagement, external use, or collaborative input

Another item in the checklist I struggled with is the following:

> Research Impact Statement (required section): Does the paper include a "Research Impact Statement" section providing evidence of realized impact (publications, external use, integrations) or credible near-term significance (benchmarks, reproducible materials, community-readiness signals)? Is the evidence presented compelling?

Here, only a case report is mentioned in which this software was used (although in an admittedly very brief search I could not find the case report, since it is not cited in the section). So there are no past major publications or community use (otherwise we might expect issues and PRs), and to me there is no evidence of likely near-term significance as of now.
This is important since the [editorial guide](https://joss.readthedocs.io/en/latest/editing.html#overview-of-editorial-process
) states the following:

> Before searching for reviewers, check the submission against the pre-review screening gates. A submission failing any of these should receive a desk rejection rather than proceeding to review.
>
> Hard gates (reject if not met):
> Public development history ≥ 6 months: Is the repository public for more than six months, with development activity distributed across that period? A recent repo dump or commits concentrated into a short window fails this gate.
>
> Demonstrated research impact: Is there evidence the software is actually being used for research? Look for paper/preprint references, DOIs, cited use by others, or documented integration into research workflows. Potential future use is not sufficient.
>
> Open source practices: Does the repository show the expected signals of an open project? For multi-author projects: issues, PRs, public discussion. For single-author projects: at least several of — tagged releases/changelog, tests/CI, clear documentation, CONTRIBUTING file, stated support expectations. Multiple indicators must be present. Single-author projects should not be rejected solely for lacking a PR workflow if other signals are strong.
>
> Iterative development: Does the commit history show ongoing refinement, or does it look like a one-time code dump? Look at commit distribution over time, not just total commit count.

It seems to me that several of these criteria are not met. @erickmartins could you confirm that these were indeed checked and validated before sending the paper out to review, and that it is a misinterpretation on my side? Note that I am not at all questioning the effort of the author(s) and the scientific validity of the software!
