### 2026-03-23 #10131 @erickmartins
thanks @richardkoehler and no worries about the delay - we know life gets in the way. I will phone-a-friend on this one and tag @rkurchin since if I am reading the editorial guide correctly at least a few of those would be desk-rejections at a stage before I was involved in the process.

re. research impact: I think this is also a valid concern that @yungselm might have a more substantial answer to - is a direct citation to the case report feasible? are there more concrete avenues towards creating a community leveraging this software?
