### 2026-04-01 #10131 @yungselm
Thank you @richardkoehler for the thorough and constructive review!

I'd like to address some points directly already:

**1. Development timeline**
The current repository does not reflect the full development history of the project. `multimodars` is built upon the [`aaoca_compression_simulation`](https://github.com/yungselm/aaoca_compression_simulation) repository, and a large portion of the codebase was migrated when I decided to transition to a `maturin`-based Python/Rust project, rather than a pure Rust implementation. The reason for this was that I expected more researchers to be familiar with Python. I had informed @rkurchin of this when submitting.

**2. Research impact**
The case report referenced in the paper was in press at the time of submission but has since been officially published (25 March 2026): https://doi.org/10.1016/j.jaccas.2026.107028. Additionally, further publications making use of `multimodars` are currently under review, though I cannot cite these yet.

**3. Community engagement**
This is a fair point and one I take seriously. I will encourage my collaborators to contribute more actively to the repository. I also intend to promote the project more broadly (e.g., via LinkedIn) upon returning from my holiday leave. We also are already actively searching for grants around this project. I recognize that building genuine community engagement is an inherently gradual process, but I am committed to working towards it.

I appreciate your time and look forward to your feedback.
