### 2026-04-01 #10131 @richardkoehler
@yungselm Thanks for the clarification and congratulations on the publication of the case report!

I think your answer addresses most of my concerns, and I think the community aspect is sometimes the most difficult part. But since the JOSS specifically also allows single-author software projects, I don't expect your paper to be rejected solely on this ground. And I appreciate your effort to build up an active community around your software.

I will then continue with my review since I am almost done there too and hope that I will have finished next week.
