### 2026-04-04 yungselm/multimoda-rs#11 @jonaspleyer
[JOSS-review] upload data to persistent archive such as zenodo
The data is available via `examples.zip` but I would like to see this data uploaded to a persistent file storage such as Zenodo as well. The source code of the repository must be uploaded there as well as part of the submission so I suggest to upload the data either together with the source code or separately.

Leave this issue open until resolved.

https://github.com/openjournals/joss-reviews/issues/10131#issuecomment-3982924781
