### 2026-04-04 yungselm/multimoda-rs#12 @jonaspleyer
[JOSS-review] Enhance documentation of Tutorial - Intravascular Module
These are some minor things that I noticed while reading the documentation:
## Tutorial Intravascular
### 1. Workflow csv files
- [ ] After "Every file should be structured in the following way (no headers):" add a description of what the columns mean, possibly with the exact units to be used. For researchers not too familiar with the topic this would be extremely helpful.
- [ ] Also provide header names for first table
- [ ] Link to api calls where possible. Using sphinx you should be able to do `:function:'some_func'` or `:class'SomeClass'` and get the correct link (hopefully).

### 2. Workflow numpy arrays and Finetuning
- [ ] Describe how this workflow differs from the previous one. Alternatively this description could be given at the heading of the page. Otherwise users might be confused: is one approach better? Do they serve the same purpose? How do they differ?

### 2. Alignment with a centerline
- [ ] same number as previous one; also adjust all sections after that (increase numbering)

### 4. Utility functions to link to numpy
- [ ] Everything after "Returns" is somewhat unclear to me? What returns that? The preceding code-snippet has no return statement but only variable assignments. Furthermore the code block after "Returns" looks like the function signature of the `to_array` function. Linking there would make sense. Describe the actions in text.

https://github.com/openjournals/joss-reviews/issues/10131#issuecomment-3982924781
