### 2026-04-04 yungselm/multimoda-rs#13 @jonaspleyer
[JOSS-review] Paper suggestions
Here are some suggestions regarding the submitted paper:

- [ ] Figure 1: consider adding link to script/example where images can be found.
- [ ] Figure 2: The caption seems to not describe the lower 2/3 of the figure.
- [ ] section "algorithms" One thing that was slightly unclear to me: As far as I can tell, all algorithms are used for alignment purposes. Therefore it may increase the clarity of the section if you change "Algorithms" to something like "Alignment Algorithms"
- [ ] Heading "Performance and parallelisation" is written in british english. The rest is in american english (compare with "visualization"). Therefore it should be "parallelization"
- [ ] Research impact statement: I am no expert in the topic of ardiac imaging but I feel that this section could be improvided by a few additional references.

https://github.com/openjournals/joss-reviews/issues/10131#issuecomment-3982924781
