### 2026-04-04 #10131 @jonaspleyer
I have completed my review and can report that I am overall very satisfied with the quality of the work. I have had some minor comments and suggestions which I noted in a few issues and one PR:

https://github.com/yungselm/multimoda-rs/pull/10
https://github.com/yungselm/multimoda-rs/issues/11
https://github.com/yungselm/multimoda-rs/issues/12
https://github.com/yungselm/multimoda-rs/issues/13

The preceding discussion spearheaded by @richardkoehler (thanks for your in-depth look) was very helpful in already addressing some of my initial thoughts.
