### 2026-04-21 #10131 @richardkoehler
So I have finally finished the rest of the review, and overall I am very positive, there are very few minor comments from my side, that have also actually been raised in parts by the other reviewers before me, as I have noticed:

## Performance
- [ ] There is a performance claim in the paper on lines 102-104. There should be code, and ideally sample data to reproduce this claim:
> Empirical performance on a 16-core CPU: an OCT pullback with 280 frames and a rotation search range of ±3° (final accuracy 0.01°) saw alignment time reduced from 150𝑠 to 18𝑠 with the optimized multiscale search.

- [ ] Scipy is missing as dependency for mm.label_geometry(), I think it needs to be added to ``pyproject.toml``
- [ ] I would suggest to use ``dependency-groups`` instead of ``optional-dependencies`` in ``pyproject.toml``, since ``optional-dependencies`` are shipped with the package when you publish it, but ``dependency-groups`` are not. In your case, you have additional dependencies that do not need to be shipped, like ``docs`` and ``test``. An example for the use of  ``optional-dependencies`` would for example be if you wanted to install support for excel with ``pandas``, in which case you would do: ``pip install "pandas[excel]"``. See here: https://packaging.python.org/en/latest/specifications/dependency-groups/
- [ ] Using ``dependency-groups`` you can also solve the problem of listing the dependencies that are necessary to run the example notebook, since common package and environment managers like ``uv`` use the dependency-group ``dev`` for these types of dependencies. The ``dev`` dependencies are installed by default when developing, but are like all dependency-groups not distributed with the package when publishing. So I would suggest listing the dependencies necessary for running the notebook (trimesh plotly scipy nbformat and ipykernel) in the group ``dev``.

## Tutorials
In general, it is not possible to follow the tutorials step-by-step, since for example not all the code necessary is included and there are sometimes changes in variable names. I therefore entirely agree with this comment by crnh:

> ### General advice on the tutorials
> I do not consider this a review comment and this is only a suggestion, not a request for change. The documentation would benefit from writing the tutorials as Jupyter notebooks using e.g. [`nbsphinx`](https://nbsphinx.readthedocs.io/) or [`myst-nb`](https://myst-nb.readthedocs.io/en/v0.13.2/index.html). This would allow including outputs in the tutorials, so the user can verify the results. Furthermore, the notebooks could be run when building the documentation, so the outputs are always up-to-date and errors in the tutorials are detected early.

### Intravascular Module
- [ ] Remove tuple brackets from ``(0.0, 1.0, 2.0)`` here since this results in an error: contour_trsl = contour_rot.translate((0.0, 1.0, 2.0))

## The examples zip folder

- [ ] Is there a reason the examples folder has to be downloaded separately and is not included as is in the repository? That would be very helpful.
- [ ] By including the examples folder in the repository you can also write a unit test for the example notebook and make sure that it still runs through when you make any code changes to your library.
- [ ] The notebook ``ivus_to_centerline.ipynb`` yields the following error, since the folders output and unprocessed do not exist. I would suggest including the empty folders in the examples folder, or creating them on-the-fly with os.makedirs or similar: ``ValueError: string is not a file: 'output/unprocessed/dia_lumen.obj'``

## Human and animal research

- [ ] I think I didn't find any information on this but is the sample data (in the repo and the zip folder) real human data or synthetic data? Maybe I also overlooked that you elaborated on this somewhere, but of course data should follow the JOSS policies, so if it is not synthetic data I would recommend adding a small statement on the conformity of the data to your repository (and potentially the paper): https://joss.readthedocs.io/en/latest/policies.html#joss-policies
