### 2026-04-21 #10131 @crnh
Overall I am positive, but two items on [my checklist](https://github.com/openjournals/joss-reviews/issues/10131#issuecomment-3984226026) are still open:

- [ ] As already indicated above by richardkoehler, the performance claims in the paper (lines 102 - 104) cannot be reproduced without additional information. If JOSS only accepts performance claims that are sufficiently substantiated, this point should be addressed.
- [ ] The case report mentioned in the Research impact statement is already cited elsewhere, but for clarity I recommend to cite it in the Research impact statement as well.

Furthermore, as mentioned by the other reviewers, the intravascular tutorial in the documentation cannot be run due to undefined variables.
