+++ src/intravascular/binding/mod.rs
@@ -55,8 +55,11 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
 ///     Path to the REST input folder.
 /// input_path_b : str
 ///     Path to the STRESS input folder.
-/// label : str, optional
-///     Label used for output filenames.  Default is ``"full"``.
+/// labels : list of str, optional
+///     Labels for the four geometries ``[rest_dia, rest_sys, stress_dia,
+///     stress_sys]``.  Must be exactly 4 strings; if a different number is
+///     supplied the last component of each input path is used instead.
+///     Default is ``[]``.
 /// diastole : bool, optional
 ///     Whether to process the diastolic phase.  Default is ``True``.
 /// step_rotation_deg : float, optional
@@ -129,7 +132,7 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
     signature = (
         input_path_a,
         input_path_b,
-        label = "full",
+        labels = vec![],
         diastole = true,
         step_rotation_deg = 0.5f64,
         range_rotation_deg = 90.0f64,
@@ -153,7 +156,7 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
 pub fn from_file_full(
     input_path_a: &str,
     input_path_b: &str,
-    label: &str,
+    labels: Vec<String>,
     diastole: bool,
     step_rotation_deg: f64,
     range_rotation_deg: f64,
@@ -197,7 +200,7 @@ pub fn from_file_full(
         logs_c,
         logs_d,
     ) = full_processing_rs(
-        label.to_string(),
+        labels,
         image_center,
         radius,
         n_points,
@@ -210,7 +213,7 @@ pub fn from_file_full(
         diastole,
         write_obj,
         interpolation_steps,
-        rust_contour_types, // Use converted types
+        rust_contour_types,
         watertight,
         output_path_a,
         output_path_b,
@@ -271,8 +274,11 @@ pub fn from_file_full(
 ///     Path to the REST input folder.
 /// input_path_b : str
 ///     Path to the STRESS input folder.
-/// label : str, optional
-///     Label used for output filenames.  Default is ``"double_pair"``.
+/// labels : list of str, optional
+///     Labels for the four geometries ``[rest_dia, rest_sys, stress_dia,
+///     stress_sys]``.  Must be exactly 4 strings; if a different number is
+///     supplied the last component of each input path is used instead.
+///     Default is ``[]``.
 /// diastole : bool, optional
 ///     Whether to process the diastolic phase.  Default is ``True``.
 /// step_rotation_deg : float, optional
@@ -329,7 +335,7 @@ pub fn from_file_full(
 #[pyo3(signature = (
     input_path_a,
     input_path_b,
-    label = "double_pair",
+    labels = vec![],
     diastole = true,
     step_rotation_deg = 0.5f64,
     range_rotation_deg = 90.0f64,
@@ -350,7 +356,7 @@ pub fn from_file_full(
 pub fn from_file_doublepair(
     input_path_a: &str,
     input_path_b: &str,
-    label: &str,
+    labels: Vec<String>,
     diastole: bool,
     step_rotation_deg: f64,
     range_rotation_deg: f64,
@@ -381,7 +387,7 @@ pub fn from_file_doublepair(
         contour_types.iter().map(|ct| ct.into()).collect();
 
     let (geom_ab_final, geom_cd_final, logs_a, logs_b, logs_c, logs_d) = double_pair_processing_rs(
-        label.to_string(),
[…]
