+++ multimodars/_processing.py
@@ -20,7 +20,7 @@ from .multimodars import (
     to_obj as _to_obj,
     find_centerline_bounded_points_simple as _find_centerline_bounded_points_simple,
     find_proximal_distal_scaling as _find_proximal_distal_scaling,
-    build_adjacency_map as _build_adjacency_map
+    build_adjacency_map as _build_adjacency_map,
 )
 
 _AlignLog = list[tuple[int, int, float, float, float, float, float]]
@@ -38,7 +38,7 @@ def _default_contour_types() -> list[PyContourType]:
 def from_file_full(
     input_path_a: str,
     input_path_b: str,
-    label: str = "full",
+    labels: list[str] | None = None,
     diastole: bool = True,
     step_rotation_deg: float = 0.5,
     range_rotation_deg: float = 90.0,
@@ -94,8 +94,11 @@ def from_file_full(
         Path to the REST input folder.
     input_path_b : str
         Path to the STRESS input folder.
-    label : str, optional
-        Label used for output filenames.  Default is ``"full"``.
+    labels : list of str, optional
+        Labels for the four geometries ``[rest_dia, rest_sys, stress_dia,
+        stress_sys]``.  Must be exactly 4 strings; if a different number is
+        supplied the last component of each input path is used instead.
+        Default is ``[]``.
     diastole : bool, optional
         Whether to process the diastolic phase.  Default is ``True``.
     step_rotation_deg : float, optional
@@ -169,7 +172,7 @@ def from_file_full(
     return _from_file_full(
         input_path_a,
         input_path_b,
-        label,
+        labels or [],
         diastole,
         step_rotation_deg,
         range_rotation_deg,
@@ -194,7 +197,7 @@ def from_file_full(
 def from_file_doublepair(
     input_path_a: str,
     input_path_b: str,
-    label: str = "double_pair",
+    labels: list[str] | None = None,
     diastole: bool = True,
     step_rotation_deg: float = 0.5,
     range_rotation_deg: float = 90.0,
@@ -245,8 +248,11 @@ def from_file_doublepair(
         Path to the REST input folder.
     input_path_b : str
         Path to the STRESS input folder.
-    label : str, optional
-        Label used for output filenames.  Default is ``"double_pair"``.
+    labels : list of str, optional
+        Labels for the four geometries ``[rest_dia, rest_sys, stress_dia,
+        stress_sys]``.  Must be exactly 4 strings; if a different number is
+        supplied the last component of each input path is used instead.
+        Default is ``[]``.
     diastole : bool, optional
         Whether to process the diastolic phase.  Default is ``True``.
     step_rotation_deg : float, optional
@@ -305,7 +311,7 @@ def from_file_doublepair(
     return _from_file_doublepair(
         input_path_a,
         input_path_b,
-        label,
+        labels or [],
         diastole,
         step_rotation_deg,
         range_rotation_deg,
@@ -327,7 +333,7 @@ def from_file_doublepair(
 
 def from_file_singlepair(
     input_path: str,
-    label: str = "single_pair",
+    labels: list[str] | None = None,
     diastole: bool = True,
     step_rotation_deg: float = 0.5,
     range_rotation_deg: float = 90.0,
@@ -371,8 +377,10 @@ def from_file_singlepair(
     ----------
     input_path : str
         Path to the input CSV folder.
-    label : str, optional
-        Label used for output filenames.  Default is ``"single_pair"``.
+    labels : list of str, optional
+        Labels for the two geometries ``[diastole, systole]``.  Must be
+        exactly 2 strings; if a different number is supplied the last
+        component of the input path is used instead.  Default is ``[]``.
     diastole : bool, optional
[…]
