### 2026-04-29 c01cb6c
remove unnecessary diastole arguments
