+++ src/intravascular/binding/mod.rs
@@ -60,8 +60,6 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
 ///     stress_sys]``.  Must be exactly 4 strings; if a different number is
 ///     supplied the last component of each input path is used instead.
 ///     Default is ``[]``.
-/// diastole : bool, optional
-///     Whether to process the diastolic phase.  Default is ``True``.
 /// step_rotation_deg : float, optional
 ///     Rotation step in degrees.  Default is ``0.5``.
 /// range_rotation_deg : float, optional
@@ -133,7 +131,6 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
         input_path_a,
         input_path_b,
         labels = vec![],
-        diastole = true,
         step_rotation_deg = 0.5f64,
         range_rotation_deg = 90.0f64,
         image_center = (4.5f64, 4.5f64),
@@ -157,7 +154,6 @@ pub fn from_file_full(
     input_path_a: &str,
     input_path_b: &str,
     labels: Vec<String>,
-    diastole: bool,
     step_rotation_deg: f64,
     range_rotation_deg: f64,
     image_center: (f64, f64),
@@ -210,7 +206,6 @@ pub fn from_file_full(
         None,
         None,
         None,
-        diastole,
         write_obj,
         interpolation_steps,
         rust_contour_types,
@@ -279,8 +274,6 @@ pub fn from_file_full(
 ///     stress_sys]``.  Must be exactly 4 strings; if a different number is
 ///     supplied the last component of each input path is used instead.
 ///     Default is ``[]``.
-/// diastole : bool, optional
-///     Whether to process the diastolic phase.  Default is ``True``.
 /// step_rotation_deg : float, optional
 ///     Rotation step in degrees.  Default is ``0.5``.
 /// range_rotation_deg : float, optional
@@ -336,7 +329,6 @@ pub fn from_file_full(
     input_path_a,
     input_path_b,
     labels = vec![],
-    diastole = true,
     step_rotation_deg = 0.5f64,
     range_rotation_deg = 90.0f64,
     image_center = (4.5f64, 4.5f64),
@@ -357,7 +349,6 @@ pub fn from_file_doublepair(
     input_path_a: &str,
     input_path_b: &str,
     labels: Vec<String>,
-    diastole: bool,
     step_rotation_deg: f64,
     range_rotation_deg: f64,
     image_center: (f64, f64),
@@ -397,7 +388,6 @@ pub fn from_file_doublepair(
         None,
         None,
         None,
-        diastole,
         write_obj,
         interpolation_steps,
         rust_contour_types,
@@ -457,8 +447,6 @@ pub fn from_file_doublepair(
 ///     Labels for the two geometries ``[diastole, systole]``.
 ///     Must be exactly 2 strings; if a different number is supplied the last
 ///     component of the input path is used instead.  Default is ``[]``.
-/// diastole : bool, optional
-///     Whether to process the diastolic phase.  Default is ``True``.
 /// step_rotation_deg : float, optional
 ///     Rotation step in degrees.  Default is ``0.5``.
 /// range_rotation_deg : float, optional
@@ -515,7 +503,6 @@ pub fn from_file_doublepair(
 #[pyo3(signature = (
     input_path,
     labels = vec![],
-    diastole = true,
     step_rotation_deg = 0.5f64,
     range_rotation_deg = 90.0f64,
     image_center = (4.5f64, 4.5f64),
@@ -534,7 +521,6 @@ pub fn from_file_doublepair(
 pub fn from_file_singlepair(
     input_path: &str,
[…]

[…]
