+++ multimodars/_processing.py
@@ -41,19 +41,19 @@ def from_file_full(
     labels: list[str] | None = None,
     step_rotation_deg: float = 0.5,
     range_rotation_deg: float = 90.0,
+    sample_size: int = 500,
     image_center: tuple[float, float] = (4.5, 4.5),
     radius: float = 0.5,
     n_points: int = 20,
     write_obj: bool = True,
     watertight: bool = True,
+    contour_types: list[PyContourType] | None = None,
     output_path_a: str = "output/rest",
     output_path_b: str = "output/stress",
     output_path_c: str = "output/diastole",
     output_path_d: str = "output/systole",
-    interpolation_steps: int = 28,
+    interpolation_steps: int = 0,
     bruteforce: bool = False,
-    sample_size: int = 500,
-    contour_types: list[PyContourType] | None = None,
     smooth: bool = True,
     postprocessing: bool = True,
 ) -> tuple[
@@ -103,6 +103,9 @@ def from_file_full(
     range_rotation_deg : float, optional
         Rotation search range (±) in degrees; a range of 90° gives 180°
         total.  Default is ``90.0``.
+    sample_size : int, optional
+        Number of points to downsample to during alignment.  Default is
+        ``500``.
     image_center : tuple of float, optional
         Image center ``(x, y)`` in mm.  Default is ``(4.5, 4.5)``.
     radius : float, optional
@@ -114,6 +117,9 @@ def from_file_full(
         Whether to write OBJ files to disk.  Default is ``True``.
     watertight : bool, optional
         Whether to write a watertight or shell mesh.  Default is ``True``.
+    contour_types : list of PyContourType, optional
+        Contour types to export.  Default is
+        ``[PyContourType.Lumen, PyContourType.Catheter, PyContourType.Wall]``.
     output_path_a : str, optional
         Output directory for REST results.  Default is ``"output/rest"``.
     output_path_b : str, optional
@@ -129,12 +135,6 @@ def from_file_full(
     bruteforce : bool, optional
         Whether to use brute-force alignment (one comparison per step).
         Default is ``False``.
-    sample_size : int, optional
-        Number of points to downsample to during alignment.  Default is
-        ``500``.
-    contour_types : list of PyContourType, optional
-        Contour types to export.  Default is
-        ``[PyContourType.Lumen, PyContourType.Catheter, PyContourType.Wall]``.
     smooth : bool, optional
         Whether to smooth frames after alignment using a 3-point moving
         average.  Default is ``True``.
@@ -172,19 +172,19 @@ def from_file_full(
         labels or [],
         step_rotation_deg,
         range_rotation_deg,
+        sample_size,
         image_center,
         radius,
         n_points,
         write_obj,
         watertight,
+        contour_types,
         output_path_a,
         output_path_b,
         output_path_c,
         output_path_d,
         interpolation_steps,
         bruteforce,
-        sample_size,
-        contour_types,
         smooth,
         postprocessing,
     )
@@ -196,17 +196,17 @@ def from_file_doublepair(
     labels: list[str] | None = None,
     step_rotation_deg: float = 0.5,
     range_rotation_deg: float = 90.0,
+    sample_size: int = 500,
     image_center: tuple[float, float] = (4.5, 4.5),
     radius: float = 0.5,
     n_points: int = 20,
     write_obj: bool = True,
     watertight: bool = True,
+    contour_types: list[PyContourType] | None = None,
     output_path_a: str = "output/rest",
     output_path_b: str = "output/stress",
-    interpolation_steps: int = 28,
+    interpolation_steps: int = 0,
     bruteforce: bool = False,
-    sample_size: int = 500,
-    contour_types: list[PyContourType] | None = None,
[…]
