+++ src/intravascular/binding/mod.rs
@@ -31,13 +31,14 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
 /// within and between each cardiac phase in parallel, and writes interpolated
 /// OBJ meshes.
 ///
-/// .. code-block:: text
+/// .. parsed-literal::
 ///
-///    Rest           Stress
-///    diastole ──▶ diastole
-///       │            │
-///       ▼            ▼
-///    systole ──▶ systole
+///    ``input_path_ab`     ``input_path_cd``
+///    **Rest:**              **Stress:**
+///    a) diastole    ──▶    c) diastole
+///        │                      │
+///        ▼                      ▼
+///    b) systole     ──▶    d) systole
 ///
 /// .. warning::
 ///
@@ -51,10 +52,10 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
 ///
 /// Parameters
 /// ----------
-/// input_path_a : str
-///     Path to the REST input folder.
-/// input_path_b : str
-///     Path to the STRESS input folder.
+/// input_path_ab : str
+///     Path to the e.g. REST input folder (``a`` rest diastole and ``b`` rest systole).
+/// input_path_cd : str
+///     Path to the e.g. STRESS input folder (``c`` stress diastole and ``d`` stress systole).
 /// labels : list of str, optional
 ///     Labels for the four geometries ``[rest_dia, rest_sys, stress_dia,
 ///     stress_sys]``.  Must be exactly 4 strings; if a different number is
@@ -65,6 +66,9 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
 /// range_rotation_deg : float, optional
 ///     Rotation search range (±) in degrees; a range of 90° gives 180°
 ///     total.  Default is ``90.0``.
+/// sample_size : int, optional
+///     Number of points to downsample to during alignment.  Default is
+///     ``500``.
 /// image_center : tuple of float, optional
 ///     Image center ``(x, y)`` in mm.  Default is ``(4.5, 4.5)``.
 /// radius : float, optional
@@ -76,27 +80,24 @@ fn logs_to_tuples(logs: Vec<AlignLog>) -> Vec<(u32, u32, f64, f64, f64, f64, f64
 ///     Whether to write OBJ files to disk.  Default is ``True``.
 /// watertight : bool, optional
 ///     Whether to write a watertight or shell mesh.  Default is ``True``.
-/// output_path_a : str, optional
-///     Output directory for REST results.  Default is ``"output/rest"``.
-/// output_path_b : str, optional
-///     Output directory for STRESS results.  Default is ``"output/stress"``.
-/// output_path_c : str, optional
-///     Output directory for DIASTOLE results.  Default is
-///     ``"output/diastole"``.
-/// output_path_d : str, optional
-///     Output directory for SYSTOLE results.  Default is
-///     ``"output/systole"``.
-/// interpolation_steps : int, optional
-///     Number of interpolated meshes between phases.  Default is ``28``.
-/// bruteforce : bool, optional
-///     Whether to use brute-force alignment (one comparison per step).
-///     Default is ``False``.
-/// sample_size : int, optional
-///     Number of points to downsample to during alignment.  Default is
-///     ``500``.
 /// contour_types : list of PyContourType, optional
 ///     Contour types to export.  Default is
 ///     ``[PyContourType.Lumen, PyContourType.Catheter, PyContourType.Wall]``.
+/// output_path_ab : str, optional
+///     Output directory for REST results.  Default is ``"output/rest"`` (with diastole versus systole).
[…]

[…]
