+++ multimodars/_processing.py
@@ -36,8 +36,8 @@ def _default_contour_types() -> list[PyContourType]:
 
 
 def from_file_full(
-    input_path_a: str,
-    input_path_b: str,
+    input_path_ab: str,
+    input_path_cd: str,
     labels: list[str] | None = None,
     step_rotation_deg: float = 0.5,
     range_rotation_deg: float = 90.0,
@@ -48,10 +48,10 @@ def from_file_full(
     write_obj: bool = True,
     watertight: bool = True,
     contour_types: list[PyContourType] | None = None,
-    output_path_a: str = "output/rest",
-    output_path_b: str = "output/stress",
-    output_path_c: str = "output/diastole",
-    output_path_d: str = "output/systole",
+    output_path_ab: str = "output/rest",
+    output_path_cd: str = "output/stress",
+    output_path_ac: str = "output/diastole",
+    output_path_bd: str = "output/systole",
     interpolation_steps: int = 0,
     bruteforce: bool = False,
     smooth: bool = True,
@@ -69,13 +69,17 @@ def from_file_full(
     within and between each cardiac phase in parallel, and writes interpolated
     OBJ meshes.
 
-    .. code-block:: text
+    .. parsed-literal::
 
-       Rest           Stress
-       diastole ──▶ diastole
-          │            │
-          ▼            ▼
-       systole ──▶ systole
+                            ``output_path_ac`` (Diastole: a vs. c)
+                    ┌──────────────────────────────────────────┐
+                    ▼                                          ▼
+        **a** REST diastole                      **c** STRESS diastole
+                │  ``output_path_ab`` (Rest: a+b)        │  ``output_path_cd`` (Stress: c+d)
+                ▼                                        ▼
+        **b** REST systole                       **d** STRESS systole
+                    └──────────────────────────────────────────┘
+                            ``output_path_bd`` (Systole: b vs. d)
 
     .. warning::
 
@@ -89,10 +93,10 @@ def from_file_full(
 
     Parameters
     ----------
-    input_path_a : str
-        Path to the REST input folder.
-    input_path_b : str
-        Path to the STRESS input folder.
+    input_path_ab : str
+        Path to the REST input folder (contains diastolic ``a`` and systolic ``b`` CSVs).
+    input_path_cd : str
+        Path to the STRESS input folder (contains diastolic ``c`` and systolic ``d`` CSVs).
     labels : list of str, optional
         Labels for the four geometries ``[rest_dia, rest_sys, stress_dia,
         stress_sys]``.  Must be exactly 4 strings; if a different number is
@@ -120,15 +124,15 @@ def from_file_full(
     contour_types : list of PyContourType, optional
         Contour types to export.  Default is
         ``[PyContourType.Lumen, PyContourType.Catheter, PyContourType.Wall]``.
-    output_path_a : str, optional
-        Output directory for REST results.  Default is ``"output/rest"``.
-    output_path_b : str, optional
-        Output directory for STRESS results.  Default is ``"output/stress"``.
-    output_path_c : str, optional
-        Output directory for DIASTOLE results.  Default is
+    output_path_ab : str, optional
+        Output directory for REST results (pair a+b).  Default is ``"output/rest"``.
+    output_path_cd : str, optional
+        Output directory for STRESS results (pair c+d).  Default is ``"output/stress"``.
+    output_path_ac : str, optional
+        Output directory for DIASTOLE results (pair a+c).  Default is
         ``"output/diastole"``.
-    output_path_d : str, optional
-        Output directory for SYSTOLE results.  Default is
+    output_path_bd : str, optional
+        Output directory for SYSTOLE results (pair b+d).  Default is
         ``"output/systole"``.
     interpolation_steps : int, optional
[…]
