+++ docs/tutorial_intravascular.rst
@@ -123,8 +123,8 @@ The full workflow comparing rest vs. stress and diastole vs. systole can be run 
 .. code-block:: python
 
     rest, stress, dia, sys, _ = mm.from_file_full(
-        input_path_a="ivus_rest",
-        input_path_b="ivus_stress",
+        input_path_ab="ivus_rest",
+        input_path_cd="ivus_stress",
         step_rotation_deg=0.1,
         range_rotation_deg=90,
         image_center=(4.5, 4.5),
@@ -132,11 +132,11 @@ The full workflow comparing rest vs. stress and diastole vs. systole can be run 
         n_points=20,
         write_obj=True,
         watertight=False,
-        contour_types=[mm.PyContourType.Lumen, mm.PyContourType.Catheter, mm.PyContourType.Wall]
-        output_path_a="output/rest",
-        output_path_b="output/stress",
-        output_path_c="output/diastole",
-        output_path_d="output/systole",
+        contour_types=[mm.PyContourType.Lumen, mm.PyContourType.Catheter, mm.PyContourType.Wall],
+        output_path_ab="output/rest",
+        output_path_cd="output/stress",
+        output_path_ac="output/diastole",
+        output_path_bd="output/systole",
         interpolation_steps=28,
     )
 
@@ -185,8 +185,7 @@ are all Numpy arrays with N(4, ) shape:
 
 4. Finetuning of alignment algorithms (in-depth parameter explanation)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-The core idea of the underlying algorithm is to translate and rotate adjacent intravascular imaging frames to minimize the Hausdorff distances between
-the contours. Efficiency is provided by a coarse angle search and multithreaded preparation of different states.
+The alignment algorithm co-registers adjacent intravascular frames by translating each frame to a common centroid and subsequently identifying the optimal rotation angle that minimises the Hausdorff distance between consecutive contours. Computational efficiency is achieved through a hierarchical coarse-to-fine angular search combined with multi-threaded processing of the different geometric states (see :numref:`fig-align`).
 
 .. figure:: ../paper/figures/Figure4.jpg
    :name: fig-align
@@ -194,20 +193,19 @@ the contours. Efficiency is provided by a coarse angle search and multithreaded 
    :align: center
    :width: 400px
 
-   Schematic of the translation and rotation alignment algorithm, illustrating the rotation range, step size, image centre, and catheter radius parameters.
+   Schematic of the translation and rotation alignment algorithm, illustrating the rotation search range, step size, image centre, and catheter radius parameters.
 
-In this section we are going to discuss the parameters of the alignment functions and how they can influence the final results. Let's take as an
-example the :func:`multimodars.from_array_full` function:
+This section provides an in-depth discussion of each parameter, using :func:`multimodars.from_array_full` as a reference. The only required inputs are the four ``PyInputData`` objects (``input_data_a`` through ``input_data_d``); all remaining parameters have sensible defaults.
 
 .. code-block:: python
 
     import multimodars as mm
 
     rest, stress, diastole, systole, (log_rest, log_stress, log_diastole, log_systole) = mm.from_array_full(
-        input_data_a=diastole_rest_input,
-        input_data_b=systole_rest_input,
-        input_data_c=diastole_stress_input,
-        input_data_d=diastole_stress_input,
+        input_data_a=diastole_rest_input,    # REST diastole
+        input_data_b=systole_rest_input,     # REST systole
+        input_data_c=diastole_stress_input,  # STRESS diastole
[…]
