+++ docs/tutorial_intravascular.rst
@@ -14,14 +14,13 @@ Tutorial - Intravascular Module
 
 This step-by-step tutorial demonstrates how to:
 
-- How to prepare the segmentation data for best results
-- Run the workflow from csv files (AIVUS-CAA outpu)
-- Run the workflow by building input data from numpy arrays
-- Finetuning of alignment algorithms (in-depth parameter explanation)
-- Alignment with a centerline
-- Saving everything as .obj files
-- Utility functions to link to numpy
-- Class methods
+1. Prepare segmentation data for best results
+2. Run the workflow from CSV files (AIVUS-CAA output)
+3. Run the workflow from numpy arrays — including in-depth parameter finetuning
+4. Align a geometry with a CCTA centerline
+5. Save geometries as ``.obj`` files
+6. Convert between ``Py*`` objects and numpy arrays
+7. Use class-level methods for inspection and manipulation
 
 1. General note on segmentation preparation
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -50,6 +49,14 @@ When pullbacks are acquired under both rest and stress conditions, each position
 rest-diastole, rest-systole, stress-diastole, and stress-systole.
 This is particularly relevant in AAOCA, where the intramural segment undergoes a complex, phase- and load-dependent deformation that can only be fully characterised by comparing all four states simultaneously (*full* mode).
 
+.. figure:: ../paper/figures/Figure1.jpg
+   :name: fig-processingmodes
+   :alt: Overview of the four processing modes: full, double-pair, single-pair, and single
+   :align: center
+   :width: 400px
+
+   Overview of the four processing modes supported by the package.
+
 An important practical consequence is that heart rate differs between rest and stress.
 Because the catheter is withdrawn at a constant speed, a higher heart rate during stress results in a greater number of cardiac cycles — and thus a greater number of diastolic-systolic pairs — per unit of pullback length.
 Consequently, the axial inter-frame spacing is effectively compressed at stress relative to rest, and the two pullbacks must be resampled to a common z-grid before any cross-condition comparison can be made.
@@ -76,27 +83,33 @@ After pip installing or locally building the package, install it in the familiar
 
 To run the whole workflow from .csv files with either :func:`multimodars.from_file_full`, :func:`multimodars.from_file_doublepair`, 
 :func:`multimodars.from_file_singlepair` or :func:`multimodars.from_file_single` the following requirements have to be met.
-Files should be named ``diastolic_contours.csv``, ``systolic_contours.csv``,
+Files should be named after `AIVUS <https://github.com/AI-in-Cardiovascular-Medicine/AIVUS-CAA>`_ convention: ``diastolic_contours.csv``, ``systolic_contours.csv``,
 ``diastolic_reference_points.csv`` and ``systolic_reference_points.csv`` depending on the required analysis.
-Every file should be structured in the following way (no headers):
+Every file should be structured in the following way (the files themselves have **no headers**):
 
-+------------+----------+----------+----------+
-| (frame id) | (x (mm)) | (y (mm)) | (z (mm)) |
-+------------+----------+----------+----------+
-| ...        | ...      | ...      | ...      |
-+------------+----------+----------+----------+
-| 771        | 2.4862   | 6.7096   | 24.5370  |
-+------------+----------+----------+----------+
-| 771        | 2.5118   | 6.7017   | 24.5370  |
-+------------+----------+----------+----------+
-| 771        | 2.5370   | 6.6936   | 24.5370  |
-+------------+----------+----------+----------+
-| ...        | ...      | ...      | ...      |
-+------------+----------+----------+----------+
[…]
