### 2026-04-30 f5ebe7d
add scipy as dependency
