+++ pyproject.toml
@@ -10,6 +10,7 @@ requires-python = ">=3.10"
 dependencies = [
     "numpy>=2.0.0,<3.0.0",
     "trimesh>=4.8.0",
+    "scipy>=1.16.2"
 ]
 classifiers = [
     "Programming Language :: Rust",
