### 2026-04-30 316f2b9
downgrade scipy version for python 3.10
