### 2026-04-30 7150d49
clippy changes and add clippy to pre-commit while leaving pytest and rust-test only to CI
