+++ src/ccta/adjust_mesh/scale_coronary.rs
@@ -3,12 +3,14 @@ use crate::intravascular::io::geometry::Frame;
 use crate::intravascular::io::input::{Centerline, CenterlinePoint};
 use std::collections::HashSet;
 
+type Point3D = (f64, f64, f64);
+
 pub fn centerline_based_wall_diameter_optimization(
     centerline: &Centerline,
-    ref_point_coronary: &(f64, f64, f64),
-    aortic_points: &[(f64, f64, f64)],
+    ref_point_coronary: &Point3D,
+    aortic_points: &[Point3D],
 ) -> f64 {
-    if centerline.points.len() == 0 || aortic_points.is_empty() {
+    if centerline.points.is_empty() || aortic_points.is_empty() {
         return 0.0;
     }
 
@@ -24,7 +26,7 @@ pub fn centerline_based_wall_diameter_optimization(
         None => return 0.0,
     };
 
-    let closest_aortic: &(f64, f64, f64) = match aortic_points.iter().min_by(|a, b| {
+    let closest_aortic: &Point3D = match aortic_points.iter().min_by(|a, b| {
         let dist_a = calculate_squared_distance(*a, ref_point_coronary);
         let dist_b = calculate_squared_distance(*b, ref_point_coronary);
         dist_a
@@ -65,8 +67,8 @@ pub fn centerline_based_wall_diameter_optimization(
 }
 
 pub fn centerline_based_aortic_diameter_optimization(
-    intramural_points: &[(f64, f64, f64)],
-    reference_points: &[(f64, f64, f64)],
+    intramural_points: &[Point3D],
+    reference_points: &[Point3D],
     centerline: &Centerline,
 ) -> f64 {
     let start = -2.0f64;
@@ -92,12 +94,12 @@ pub fn centerline_based_aortic_diameter_optimization(
 }
 
 pub fn centerline_based_diameter_optimization(
-    anomalous_points: &[(f64, f64, f64)],
+    anomalous_points: &[Point3D],
     n_proximal: usize,
     n_distal: usize,
     centerline: &Centerline,
-    proximal_reference: &[(f64, f64, f64)],
-    distal_reference: &[(f64, f64, f64)],
+    proximal_reference: &[Point3D],
+    distal_reference: &[Point3D],
 ) -> (f64, f64) {
     let (proximal_points, anomalous_points_new) =
         find_region_points(anomalous_points, proximal_reference, n_proximal);
@@ -139,10 +141,10 @@ pub fn centerline_based_diameter_optimization(
 }
 
 fn find_region_points(
-    anomalous_points: &[(f64, f64, f64)],
-    reference_points: &[(f64, f64, f64)],
+    anomalous_points: &[Point3D],
+    reference_points: &[Point3D],
     n_points: usize,
-) -> (Vec<(f64, f64, f64)>, Vec<(f64, f64, f64)>) {
+) -> (Vec<Point3D>, Vec<Point3D>) {
     if anomalous_points.is_empty() || reference_points.is_empty() || n_points == 0 {
         return (Vec::new(), anomalous_points.to_vec());
     }
@@ -170,12 +172,12 @@ fn find_region_points(
 
     let selected_indices: HashSet<usize> = selected_slice.iter().map(|(i, _)| *i).collect();
 
-    let selected_points: Vec<(f64, f64, f64)> = selected_slice
+    let selected_points: Vec<Point3D> = selected_slice
         .iter()
         .map(|(i, _)| anomalous_points[*i])
         .collect();
 
-    let remaining_points: Vec<(f64, f64, f64)> = anomalous_points
[…]
