+++ src/intravascular/utils/mod.rs
@@ -1,4 +1,4 @@
-pub mod utils;
+pub mod general_utils;
 
 #[cfg(test)]
 pub mod test_utils;
