### 2026-04-30 b738680
bump version to 0.3.0
