### 2026-04-30 79475ea
fix dependencies as suggested by reviewer
