### 2026-05-01 e248aeb
add examples data, provide notebooks in documentation
