+++ examples/stitching.py
@@ -1,0 +1,189 @@
+import os
+from pathlib import Path
+import numpy as np
+import multimodars as mm
+import trimesh
+
+# load the provided example data
+cwd = Path.cwd()
+if (cwd / "examples" / "data").exists():
+    os.chdir(cwd / "examples" / "data")
+elif (cwd / "data").exists():
+    os.chdir(cwd / "data")
+print(f"Working directory: {os.getcwd()}")
+
+rest, (dia_logs, sys_logs) = mm.from_file_singlepair(
+    input_path="ivus_rest",
+    labels=["aligned_dia", "aligned_sys"],
+    output_path="output/rest",
+)
+
+rca_cl_raw = np.genfromtxt("../data/centerline_rca_short.csv", delimiter=",")
+lca_cl_raw = np.genfromtxt("../data/centerline_lca.csv", delimiter=",")
+aorta_cl_raw = np.genfromtxt("../data/centerline_aorta.csv", delimiter=",")
+rca_cl = mm.numpy_to_centerline(rca_cl_raw)
+lca_cl = mm.numpy_to_centerline(lca_cl_raw)
+aorta_cl = mm.numpy_to_centerline(aorta_cl_raw)
+
+results, (rca_cl, lca_cl, ao_cl) = mm.label_geometry(
+    path_ccta_geometry="../data/NARCO_119_noside.stl",
+    path_centerline_aorta="../data/centerline_aorta.csv",
+    path_centerline_rca="../data/centerline_rca_short.csv",
+    path_centerline_lca="../data/centerline_lca.csv",
+    bounding_sphere_radius_mm=3.0,
+    n_points_intramural=100,
+    anomalous_rca=True,
+    anomalous_lca=False,
+    control_plot=False,
+)
+
+aligned, resampled_cl = mm.align_combined(
+    rca_cl,
+    rest,
+    (12.2605, -201.3643, 1751.0554),
+    (11.7567, -202.1920, 1754.7975),
+    (15.6605, -202.1920, 1749.9655),
+    results["rca_points"],
+    angle_range_deg=10.0,
+    write=True,
+    watertight=False,
+    output_dir="test",
+)
+
+results = mm.label_anomalous_region(
+    centerline=rca_cl,
+    frames=aligned.geom_a.frames,
+    results=results,
+    results_key="rca_points",
+    debug_plot=False,
+)
+
+prox_scaling, distal_scaling = mm.find_distal_and_proximal_scaling(
+    frames=aligned.geom_a.frames,
+    centerline=rca_cl,
+    results=results,
+)
+
+aortic_scaling = mm.find_aorta_scaling(
+    frames=aligned.geom_a.frames,
+    cl_aorta=ao_cl,
+    results=results,
+)
+
+aortic_wall_scaling = mm.find_aortic_wall_scaling(
+    frames=aligned.geom_a.frames,
+    cl_aorta=ao_cl,
+    results=results,
+)
+
+# no just scale any desired part of the geometry
+scaled_distal = mm.scale_region_centerline_morphing(
+    mesh=results["mesh"],
+    region_points=results["distal_points"],
+    centerline=rca_cl,
+    diameter_adjustment_mm=distal_scaling,
+)
+results = mm.sync_results_to_mesh(results, results["mesh"], scaled_distal)
+
+scaled_distal_aortic = mm.scale_region_centerline_morphing(
+    mesh=results["mesh"],
+    region_points=results["aorta_points"] + results["rca_removed_points"],
+    centerline=aorta_cl,
+    diameter_adjustment_mm=aortic_scaling,
+)
[…]
