+++ docs/conf.py
@@ -38,8 +38,11 @@ extensions = [
     "sphinx.ext.napoleon",
     "sphinx.ext.viewcode",
     "sphinx_autodoc_typehints",
-    "myst_parser",
+    "myst_nb",
 ]
+
+# myst-nb: render pre-executed notebooks without re-running them
+nb_execution_mode = "off"
 
 # Templates and static files
 _here = os.path.abspath(os.path.dirname(__file__))
