+++ .github/workflows/CI.yml
@@ -19,6 +19,23 @@ permissions:
   contents: read
 
 jobs:
+  notebook-tests:
+    name: Notebook tests
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+      - name: Install Rust
+        uses: dtolnay/rust-toolchain@stable
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -e ".[dev]"
+      - name: Run notebook tests
+        run: pytest --nbmake docs/notebooks/
+
   python-compat:
     name: Python ${{ matrix.python-version }} compatibility
     runs-on: ubuntu-latest
