### 2026-05-01 70e007a
add plotly as docs dependency
