+++ pyproject.toml
@@ -42,6 +42,7 @@ docs = [
     "sphinx-rtd-theme",
     "sphinx-autodoc-typehints",
     "myst-nb",
+    "plotly",
     "maturin>=1.7,<2.0",
     "furo",
 ]
