### 2026-05-01 9f7f316
ensure CI pipeline also installs group dependencies
