### 2026-05-01 45993d4
improve intravascular print statements
