+++ Cargo.toml
@@ -19,3 +19,4 @@ rayon = "1.7.1"
 crossbeam = "0.8.4"
 anyhow = "1.0.100"
 approx = "0.5"
+indicatif = "0.17"
