### 2026-05-01 6fd482d
finish ccta_notebook and improve print statements
