### 2026-05-01 83c3d3a
include developer setup and condense README
