+++ README.md
@@ -26,214 +26,77 @@
 > A high‑performance, Rust‑accelerated toolkit for multi‑modality cardiac image fusion and registration ﮩ٨ـﮩﮩ٨ـ♡ﮩ٨ـﮩﮩ٨ـ.
 
 ---
-# Overview
-`multimoda-rs` is a high-performance toolkit developed to enable the study of dynamic vessel deformation in coronary artery anomalies (CAAs), where quantifying lumen changes under stress and rest is critical. It addresses the general challenge of aligning and fusing diverse cardiac imaging modalities, such as CCTA, IVUS, OCT, and MRI—into a unified, high‑resolution 3D model. While CCTA provides comprehensive volumetric context, intravascular modalities (IVUS and OCT) offer sub‑millimeter resolution along the vessel lumen, and MRI (LGE) reveals tissue characteristics like scar and edema. This library leverages Rust for computationally intensive registration steps, delivering faster performance than pure Python implementations.
+
+## Overview
+
+`multimoda-rs` aligns and fuses diverse cardiac imaging modalities — IVUS, OCT and CCTA — into unified high-resolution 3D models. Originally developed to quantify dynamic lumen deformation in coronary artery anomalies (CAAs), it is equally applicable to longitudinal studies (e.g., pre/post-stenting) and general coronary artery disease workflows. The Rust backend parallelizes computationally intensive registration steps for speeds well beyond pure Python.
 
 ## Key Features
-- IVUS/OCT Contours Registration
-  - Aligns pullback sequences (rest vs. stress, diastole vs. systole) using Hausdorff distance on vessel contours and catheter centroids.
-  - Supports four alignment modes:
-    - *Full*: register all four phases (rest‑dia, rest‑sys, stress‑dia, stress‑sys)
-    - *Double-pair*: two pairs (rest vs. stress).
-    - *Single-pair*: diastole vs. systole.
-    - *Single*: one phase only.
-- Centerline Alignment
-  - Align registered geometries onto a vessel centerline using three‑point or manual rotation methods.
-- Geometry Utilities
-  - Smooth contours, compute areas and elliptic ratios, find farthest/closest point pairs, and more.
+
+- **Intravascular Registration**: align pullback sequences (rest/stress, diastole/systole) using Hausdorff distance on vessel contours and catheter centroids; four modes: *full*, *double-pair*, *single-pair*, *single*.
+- **Centerline Alignment**: register intravascular geometries onto a CCTA-derived centerline via three-point landmark or manual rotation.
+- **CCTA Fusion**: automatically label CCTA geometries by vessel region and morph them to match intravascular measurements.
+- **Flexible Input**: accepts CSV files ([AIVUS](https://github.com/AI-in-Cardiovascular-Medicine/AIVUS-CAA) format) or raw NumPy arrays.
 
 ## Installation
 
-Either directly from PyPI (recommended):
 ```bash
 pip install multimodars
 ```
 
-or by cloning the repo and building the project yourself:
+Optional extras:
+
 ```bash
-# Install rust in case you don't have it on your system
-curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
-
-git clone https://github.com/yungselm/multimoda-rs.git
-python -m venv .venv
-source .venv/bin/activate
-pip install maturin
-. "$HOME/.cargo/env" # Set rust env
-maturin develop
+pip install "multimodars[viz]"      # mesh visualisation (pyglet)
+pip install "multimodars[meshlab]"  # MeshLab integration
 ```
 
-**Note:** In case you get the following error:
-```
-💥 maturin failed
-  Caused by: rustc, the rust compiler, is not installed or not in PATH. This package requires Rust and Cargo to compile extensions. Install it through the system's package manager or via https://rustup.rs/.
-```
[…]
