### 2026-05-01 80cde99
include most important workflow images in README
