### 2026-05-04 e523894
add explanation of labeling algorithms + small implementation par_iter()
