### 2026-05-04 1f98bed
add soft dependency networkx
