### 2026-05-04 80112b3
include meshlab in CI pipeline for notebook
