+++ .github/workflows/CI.yml
@@ -32,7 +32,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -e .
+          pip install -e '.[meshlab]'
           pip install --group dev
       - name: Run notebook tests
         run: pytest --nbmake docs/notebooks/
