+++ .github/workflows/CI.yml
@@ -26,7 +26,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
-          python-version: '3.11'
+          python-version: '3.12.x'
       - name: Install Rust
         uses: dtolnay/rust-toolchain@stable
       - name: Install dependencies
