### 2026-05-04 b5cd4f4
fails only on CI not locally because of libopengl0 dependency
