+++ .github/workflows/CI.yml
@@ -26,9 +26,11 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
-          python-version: '3.12.x'
+          python-version: 3.12.x
       - name: Install Rust
         uses: dtolnay/rust-toolchain@stable
+      - name: Install system dependencies
+        run: sudo apt-get install -y libopengl0
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
