### 2026-05-04 b48b372
document libopengl0 dpendency issue and update versions in CI.yml pipeline
