### 2026-05-04 329c0ae
fix readthedocs pipeline to work with new dependecies structure
