### 2026-05-04 d07995b
simplify readthedocs pipeline
