+++ .readthedocs.yaml
@@ -7,19 +7,7 @@ build:
     rust: "latest"
   jobs:
     post_create_environment:
-      - "curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y --default-toolchain 1.87.0"
-      - ". \"$HOME/.cargo/env\""
-      - "rustup default 1.87.0 || true"
-      - "echo \"Rust installed: $(rustc --version)\""
-      - ". \"$READTHEDOCS_VIRTUALENV_PATH/bin/activate\""
-      - "echo \"Activated venv: $VIRTUAL_ENV\""
-      - "ln -s \"$READTHEDOCS_VIRTUALENV_PATH\" .venv"
-      - "echo \"Symlinked .venv → $READTHEDOCS_VIRTUALENV_PATH\""
-      - "pip install 'maturin>=1.7,<2.0' 'sphinx>=8.0,<9.0' sphinx-rtd-theme sphinx-autodoc-typehints myst-nb plotly furo"
-
-    pre_build:
-      - "maturin develop --release"
-      - 'python -c "import sys; print(\"Python path:\", sys.path)"'
+      - "$READTHEDOCS_VIRTUALENV_PATH/bin/pip install maturin sphinx sphinx-rtd-theme sphinx-autodoc-typehints myst-nb plotly furo"
 
 sphinx:
   configuration: docs/conf.py
@@ -29,4 +17,3 @@ python:
   install:
     - method: pip
       path: .
-
