### 2026-05-04 403ab4c
interactive plot for documentation
