### 2026-05-05 a9fafca
add benchmark suite
