+++ benchmarks/benchmark_cpu_scaling.py
@@ -1,0 +1,258 @@
+"""Benchmark 2: CPU-core scaling for optimised vs. bruteforce alignment.
+
+Tests from_array_single on the bundled OCT example dataset (280 frames) at
+varying RAYON_NUM_THREADS values (1 → 16 cores).
+
+Parameters follow the docs/notebooks:
+  step_rotation_deg = 0.01, range_rotation_deg = 6.0
+
+Because rayon's global thread pool is initialised once per process, every
+(n_cores x bruteforce x repeat) combination is run in a fresh subprocess
+so that RAYON_NUM_THREADS takes effect before the first rayon call.
+
+Output: benchmarks/results/cpu_scaling.png
+"""
+
+import json
+import statistics
+import subprocess
+import sys
+from pathlib import Path
+
+import matplotlib.pyplot as plt
+
+# ---------------------------------------------------------------------------
+# Configuration
+# ---------------------------------------------------------------------------
+
+_ROOT = Path(__file__).parent.parent
+OCT_DATA_PATH = _ROOT / "examples" / "data" / "oct_single"
+RESULTS_DIR = Path(__file__).parent / "results"
+
+CORE_COUNTS: list[int] = [2, 4, 8, 12, 16]
+REPEATS: int = 3
+
+# Parameters from docs/notebooks (OCT single reconstruction)
+OCT_STEP = 0.01
+OCT_RANGE = 6.0
+OCT_CENTER = (5.0, 5.0)
+OCT_N_POINTS = 40
+
+# ---------------------------------------------------------------------------
+# Worker script template (run inside each subprocess)
+# ---------------------------------------------------------------------------
+
+_WORKER_TEMPLATE = """\
+import os, sys, json, time
+os.environ["RAYON_NUM_THREADS"] = "{n_cores}"
+
+import numpy as np
+sys.path.insert(0, r"{root}")
+import multimodars as mm
+
+data_dir = r"{data_path}"
+oct_raw = np.genfromtxt(data_dir + "/oct_contours_raw.csv", delimiter=",")
+oct_ref = np.genfromtxt(data_dir + "/oct_ref.csv",          delimiter=",")
+
+oct_input = mm.numpy_to_inputdata(
+    lumen_arr=oct_raw,
+    ref_point=oct_ref,
+    record=None,
+    diastole=True,
+    label="oct",
+)
+
+times = []
+for _ in range({repeats}):
+    t0 = time.perf_counter()
+    mm.from_array_single(
+        input_data=oct_input,
+        step_rotation_deg={step},
+        range_rotation_deg={rng},
+        sample_size=200,
+        image_center={center},
+        n_points={n_points},
+        write_obj=False,
+        smooth=False,
+        bruteforce={bruteforce},
+    )
+    times.append(time.perf_counter() - t0)
+
+print(json.dumps(times))
+"""
+
+
+# ---------------------------------------------------------------------------
+# Subprocess runner
+# ---------------------------------------------------------------------------
+
+
+def _run_worker(n_cores: int, bruteforce: bool) -> list[float]:
+    script = _WORKER_TEMPLATE.format(
+        n_cores=n_cores,
+        root=str(_ROOT),
+        data_path=str(OCT_DATA_PATH),
+        repeats=REPEATS,
+        step=OCT_STEP,
+        rng=OCT_RANGE,
+        center=OCT_CENTER,
+        n_points=OCT_N_POINTS,
+        bruteforce=str(bruteforce),
+    )
+    proc = subprocess.run(
+        [sys.executable, "-c", script],
+        capture_output=True,
+        text=True,
+    )
+    if proc.returncode != 0 or not proc.stdout.strip():
[…]
