+++ multimodars/__init__.py
@@ -118,7 +118,7 @@ __all__ = [
 # def _print_banner():
 #     v = version("multimodars")
 #     print(r"""
-#   .__   __  .__                   .___
+#               .__   __  .__                   .___
 #   _____  __ __|  |_/  |_|__| _____   ____   __| _/____ _______  ______
 #  /     \|  |  \  |\   __\  |/     \ /  _ \ / __ |\__  \\_  __ \/  ___/
 # |  Y Y  \  |  /  |_|  | |  |  Y Y  (  <_> ) /_/ | / __ \|  | \/\___ \
