+++ benchmarks/benchmark_bruteforce_stepsize.py
@@ -1,0 +1,196 @@
+"""Benchmark 1: bruteforce vs. optimised alignment at varying step sizes.
+
+Tests from_file_full on the bundled IVUS rest/stress example data with
+step_rotation_deg in {5.0, 2.5, 1.0, 0.5, 0.1} degrees.
+
+write_obj / smooth / postprocessing are all disabled so only the alignment
+algorithm itself is timed.
+
+Output: benchmarks/results/bruteforce_stepsize.png
+"""
+
+import statistics
+import time
+from pathlib import Path
+
+import matplotlib.pyplot as plt
+
+import multimodars as mm
+
+# ---------------------------------------------------------------------------
+# Configuration
+# ---------------------------------------------------------------------------
+
+_ROOT = Path(__file__).parent.parent
+DATA_ROOT = _ROOT / "examples" / "data"
+INPUT_AB = str(DATA_ROOT / "ivus_rest")
+INPUT_CD = str(DATA_ROOT / "ivus_stress")
+RESULTS_DIR = Path(__file__).parent / "results"
+
+STEP_SIZES: list[float] = [5.0, 2.5, 1.0, 0.5, 0.1, 0.05]
+RANGE_DEG: float = 90.0
+REPEATS: int = 3
+
+
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+
+def _n_steps(step: float) -> float:
+    return (RANGE_DEG * 2) / step
+
+
+def _run_once(step: float, bruteforce: bool) -> float:
+    t0 = time.perf_counter()
+    mm.from_file_full(
+        input_path_ab=INPUT_AB,
+        input_path_cd=INPUT_CD,
+        step_rotation_deg=step,
+        range_rotation_deg=RANGE_DEG,
+        write_obj=False,
+        smooth=False,
+        postprocessing=False,
+        bruteforce=bruteforce,
+        interpolation_steps=0,
+    )
+    return time.perf_counter() - t0
+
+
+# ---------------------------------------------------------------------------
+# Benchmark
+# ---------------------------------------------------------------------------
+
+
+def run_benchmark() -> dict[str, list[float]]:
+    results: dict[str, list[float]] = {"bruteforce": [], "optimised": []}
+
+    for step in STEP_SIZES:
+        print(f"step={step:4.1f}°  bruteforce=True  ", end="", flush=True)
+        times_bf = [_run_once(step, bruteforce=True) for _ in range(REPEATS)]
+        med_bf = statistics.median(times_bf)
+        results["bruteforce"].append(med_bf)
+        print(f"{med_bf:.2f} s")
+
+        print(f"step={step:4.1f}°  bruteforce=False ", end="", flush=True)
+        times_opt = [_run_once(step, bruteforce=False) for _ in range(REPEATS)]
+        med_opt = statistics.median(times_opt)
+        results["optimised"].append(med_opt)
+        speedup = med_bf / med_opt if med_opt > 0 else float("inf")
+        print(f"{med_opt:.2f} s  (speedup {speedup:.1f}x)")
+
+    return results
+
+
+# ---------------------------------------------------------------------------
+# Plotting
+# ---------------------------------------------------------------------------
+
+
+def plot_results(results: dict[str, list[float]]) -> None:
+    RESULTS_DIR.mkdir(parents=True, exist_ok=True)
+
+    n_steps_list = [_n_steps(s) for s in STEP_SIZES]
+    bf_times = results["bruteforce"]
+    opt_times = results["optimised"]
[…]
