### 2026-05-05 df15796
include performance section in paper
