### 2026-05-05 554b48b
update results after optimized parallelization strategy
