### 2026-05-05 cb8efcb
bump version to v0.3.2
