+++ paper/paper.md
@@ -46,53 +46,47 @@ bibliography: bibliography.bib
 
 Coronary artery anomalies (CAAs) and coronary artery disease (CAD) require precise morphological and functional assessment for diagnosis and treatment planning. Cardiac computed tomography angiography (CCTA) provides a comprehensive 3D coronary anatomy but lacks the sub-millimeter resolution and dynamic tissue detail available from intravascular imaging, such as intravascular ultrasound (IVUS) and optical coherence tomography (OCT). 
 
-The `multimodars` package is a general-purpose toolkit that registers high-resolution intravascular pullbacks to CCTA-derived centerlines, producing locally enhanced fusion 3D vessel representations. Developed initially to quantify dynamic lumen changes in CAAs, the toolkit produces high-fidelity models suitable for visualization, geometric analysis, and patient-specific modelling. It implements four alignment paradigms (full, double-pair, single-pair, single) to compare pullbacks acquired under different haemodynamic states (e.g., rest vs. pharmacologic stress) or at different clinical timepoints (e.g., pre- vs. post-stenting). `Multimodars` targets deterministic, reproducible multimodal fusion for both specialized CAA research and general CAD applications [@stark2025true].
+The `multimodars` package is a general-purpose toolkit that registers high-resolution intravascular pullbacks to CCTA-derived centerlines, producing locally enhanced fusion 3D vessel representations. Developed initially to quantify dynamic lumen changes in CAAs, the toolkit produces high-fidelity models suitable for visualization, geometric analysis, and patient-specific modeling. It implements four alignment paradigms (full, double-pair, single-pair, single) to compare pullbacks acquired under different hemodynamic states (e.g., rest vs. pharmacologic stress) or at different clinical timepoints (e.g., pre- vs. post-stenting). `Multimodars` targets reproducible multimodal fusion for both specialized CAA research and general CAD applications, producing quantitative lumen metrics (minimum lumen area, stenosis fraction, elliptic ratio, per-frame deformation) and optional interpolated deformation animations alongside the 3D models.
 
 # Statement of Need
 
 Building reliable 3D coronary models requires combining complementary imaging modalities. Intravascular imaging offers exceptional local resolution but lacks whole-vessel context and 3D orientation. CCTA provides the global 3D geometry but suffers from limited spatial resolution and artifacts like blooming. `Multimodars` fills a critical gap for researchers in **cardiac imaging**, **interventional cardiology**, and **biomedical engineering** who require high-fidelity lumen models for:
 
 * Automated quantification of vessel deformation under stress.
-* Computational Fluid Dynamics (CFD) and Fluid-Structure Interaction (FSI) simulations.
-* Digital twin solutions.
+* Patient-specific Computational Fluid Dynamics (CFD) and Fluid-Structure Interaction (FSI) simulations, which require high-fidelity lumen geometry unavailable from either modality alone [@tang2014; @kilic2020evolution].
+* Longitudinal and interventional comparisons (e.g., pre- vs. post-stenting) within a consistent registration framework.
+* Digital twin and 3D-printing workflows for biomechanical device testing.
 
-The package accepts CSV and NumPy inputs, including data formats produced by the [AIVUS-CAA](https://github.com/AI-in-Cardiovascular-Medicine/AIVUS-CAA) software [@stark2025automated], providing a standardized pipeline from raw image segmentation to final 3D fusion.
[…]
