+++ paper/bibliography.bib
@@ -90,6 +90,42 @@
   doi={10.1016/j.cmpb.2025.109065}
 }
 
+@article{formato2023,
+  title={Morphological Changes of Anomalous Coronary Arteries From the Aorta During the Cardiac Cycle Assessed by IVUS in Resting Conditions},
+  author={Formato, Giovanni Maria and Agnifili, Mauro Luca and Arzuffi, Luca and Rosato, Antonio and Ceserani, Valentina and Zuniga Olaya, Karina Geraldina and Secchi, Francesco and Deamici, Miriam and Conti, Michele and Auricchio, Ferdinando and Bedogni, Francesco and Frigiola, Alessandro and Lo Rito, Mauro},
+  journal={Circulation: Cardiovascular Interventions},
+  volume={16},
+  number={7},
+  pages={e012636},
+  year={2023},
+  publisher={American Heart Association},
+  doi={10.1161/CIRCINTERVENTIONS.122.012636}
+}
+
+@article{doulaverakis2013,
+  title={IVUSAngio Tool: A publicly available software for fast and accurate 3D reconstruction of coronary arteries},
+  author={Doulaverakis, Charalampos and Tsampoulatidis, Ioannis and Antoniadis, Antonios P and Chatzizisis, Yiannis S and Giannopoulos, Andreas A and Kompatsiaris, Yiannis and Giannoglou, George D},
+  journal={Computers in Biology and Medicine},
+  volume={43},
+  number={11},
+  pages={1793--1803},
+  year={2013},
+  publisher={Elsevier},
+  doi={10.1016/j.compbiomed.2013.08.013}
+}
+
+@article{tang2014,
+  title={Human coronary plaque wall thickness correlated positively with flow shear stress and negatively with plaque wall stress: an IVUS-based fluid-structure interaction multi-patient study},
+  author={Fan, Rui and Tang, Dalin and Yang, Chun and Zheng, Jie and Bach, Richard and Wang, Liang and Muccigrosso, David and Billiar, Kristen and Zhu, Jian and Ma, Genshan and Maehara, Akiko and Mintz, Gary S},
+  journal={BioMedical Engineering OnLine},
+  volume={13},
+  number={1},
+  pages={32},
+  year={2014},
+  publisher={Springer},
+  doi={10.1186/1475-925X-13-32}
+}
+
 @article{stark202611,
   title={An 11-year-old athlete with dynamic stenosis of coronary artery anomaly with normal FFR-dobutamine but abnormal iFR-dobutamine},
   author={Stark, Anselm W and Mokhtari, Ali and Kakizaki, Ryota and Bigler, Marius R and Siepe, Matthias and Shiri, Isaac and R{\"a}ber, Lorenz and Gr{\"a}ni, Christoph},
