+++ paper/paper.md
@@ -63,7 +63,7 @@ The package accepts CSV and NumPy inputs, including data formats produced by the
 
 # State of the Field
 
-Prior research has established the clinical value of CCTA/intravascular fusion [@van20103d; @wu20203d; @bourantas2013new; @boogers2012automated; @kilic2020evolution], but two barriers persist. First, most existing fusion solutions are tied to proprietary vendor hardware or closed-source commercial workstations, limiting academic transparency; the closest open-source precedent, IVUSAngio [@doulaverakis2013], relies on biplane X-ray angiography rather than CCTA and provides no multi-state support. Second, no existing toolkit is specifically tailored for **multi-state** analysis—comparing rest versus pharmacologic stress or pre- versus post-intervention—while maintaining a consistent alignment across pullbacks. Published IVUS analyses of anomalous aortic origin of coronary arteries (AAOCA) quantify pulsatile deformation only under resting conditions [@formato2023]; extending such analyses to stress conditions requires the cross-state registration logic implemented in `multimodars`. While packages like `trimesh` or `SimpleITK` provide general mesh and registration utilities, they do not offer the domain-specific coronary alignment logic (e.g., cumulative rotation propagation to preserve vessel torsion, ellipticity-weighted inter-pullback harmonization, CCTA centerline morphing) required here.
+Prior research has established the clinical value of CCTA/intravascular fusion [@van20103d; @wu20203d; @bourantas2013new; @boogers2012automated; @kilic2020evolution @ilic2025comprehensive], but two barriers persist. First, most existing fusion solutions are tied to proprietary vendor hardware or closed-source commercial workstations, limiting academic transparency; the closest open-source precedent, IVUSAngio [@doulaverakis2013], relies on biplane X-ray angiography rather than CCTA and provides no multi-state support. Second, no existing toolkit is specifically tailored for **multi-state** analysis—comparing rest versus pharmacologic stress or pre- versus post-intervention—while maintaining a consistent alignment across pullbacks. Published IVUS analyses of anomalous aortic origin of coronary arteries (AAOCA) quantify pulsatile deformation only under resting conditions [@formato2023]; extending such analyses to stress conditions requires the cross-state registration logic implemented in `multimodars`. While packages like `trimesh` or `SimpleITK` provide general mesh and registration utilities, they do not offer the domain-specific coronary alignment logic (e.g., cumulative rotation propagation to preserve vessel torsion, ellipticity-weighted inter-pullback harmonization, CCTA centerline morphing) required here.
 
 # Software Design
 
