### 2026-05-05 8dce190
update citations
