+++ paper/bibliography.bib
@@ -56,18 +56,6 @@
   doi={10.4244/EIJV9I5A94}
 }
 
-@article{boogers2012automated,
-  title={Automated quantification of coronary plaque with computed tomography: comparison with intravascular ultrasound using a dedicated registration algorithm for fusion-based quantification},
-  author={Boogers, Mark J and Broersen, Alexander and Van Velzen, Jo{\"e}lla E and De Graaf, Fleur R and El-Naggar, Heba M and Kitslaar, Pieter H and Dijkstra, Jouke and Delgado, Victoria and Boersma, Eric and de Roos, Albert and others},
-  journal={European heart journal},
-  volume={33},
-  number={8},
-  pages={1007--1016},
-  year={2012},
-  publisher={The University of Chicago Press},
-  doi={10.1093/eurheartj/ehr465}
-}
-
 @article{ilic2025comprehensive,
   title={A comprehensive workflow for CCTA and OCT data fusion with 3D printing validation: advancing patient-specific testing environments for percutaneous coronary intervention devices},
   author={Ilic, Marc and H{\"a}ner, Jonas and Lehmann, Julian and Stark, Anselm W and Illi, Jo{\"e}l and Gr{\"a}ni, Christoph and Bentele, Marc and Baumann-Zumstein, Philine and Busch, Julia and Haeberlin, Andreas},
