### 2026-05-06 b5a35a8
fix citations, missing dependency and UnicodeDecodeError
