### 2026-05-06 ae1e18c
removed chatGPT version number since several were used
