### 2026-05-06 9d75847
cleanup joss submission version
