+++ ROADMAP.MD
@@ -1,93 +1,43 @@
 # Python package that has one purpose: Combining any intravascular image modality with CCTA meshes.
 
+# Hot fixes
+- [] Labeling doesn't work as expected especially for `from_file` variants
+
 # Optimizations
 - [] Optimize the ray intersection algorithm
 - [] Better parallelization strategy for frames alignment by processing in batches (frames only have to be aligned relative to each other not sequentially)
 
-# Centerline alignment module
-(- [] Create Centerline directly from mesh) (v1.0.0)
-
 # Mesh combination (v0.3.0)
+## New/improved features
+- [] "Smooth" the aortic ring of boundary points first, by projecting on plane to shift outliers
 - [] Write a more stable function to find intramural points on aorta (for now works with excluded RCA points)
-- [] fix centerline resampling
 - [] Create an aortic and coronary wall with a consistent connection
+- [] Find bifurcation reference points automatically
+- [] `numpy_to_report` function that allows to read in lists for the four variables of interest (frame, phase, measurement_1, measurement_2)
+- [] write metadata to .obj files so they can be reconstructed completely to geometry/geometrypair
+
+## Fixes/Improvements
+- [] Write a more stable function to find intramural points on aorta (for now works with excluded RCA points)
+- [] Align algorithms only work with geometry pair not only geometry
+- [] fix centerline resampling
+
+## Investigate/check logic
+- [] When aligning geometry with wall has a rotation in the wall -> logic error for centerline alignment?
 
 # CCTA clinical assessment (v0.4.0)
-- [] automatically find the required reference points for alignment
+- [] automatically find the required reference points for anomalous alignment
 - [] automatically assess lumen area, minor-, major axis, mla, intramural length etc. from the CCTA mesh.
 - [] return hausdorff distances per point between aligned ivus mesh and ccta mesh.
 - [] function to automatically assign wall thickness to ccta and combined ivus/ccta mesh
-- [] helper function to automatically export only certain points on mesh (e.g. RCA or LCA)
-- [] function to combine anomalous section only with the distal part
 
-# Prepublish checklist (JOSS review feedback — open until resolved):
-
-## Dependencies / packaging
-- [x] Add `scipy` as an explicit runtime dependency in `pyproject.toml` (required by `mm.label_geometry()`).
-- [x] Convert `optional-dependencies` groups that are not user-facing extras (docs, test, dev)
-  to PEP 735 `dependency-groups` so they are not shipped with the package on PyPI.
-  Note: `trimesh` was **not** moved to dev — it is imported throughout the core package
-  (`_converters.py`, `ccta/` modules) and must remain a runtime dependency.
-  Moved `plotly`, `nbformat`, `ipykernel` into `[dependency-groups.dev]`; `[dependency-groups.test]`
-  and `[dependency-groups.docs]` replace the former optional-dependency groups.
-  `viz` and `meshlab` remain as `[project.optional-dependencies]` (user-facing extras).
-
-## Examples / notebooks
-- [x] Include the `examples/` folder (with data) directly in the repository so users can clone
-  and run without a separate download, and so CI can execute the notebooks as tests.
-- [x] Fix `ivus_to_centerline.ipynb`: create `output/` and `output/unprocessed/` directories
-  on-the-fly with `os.makedirs(..., exist_ok=True)` so the notebook does not error on a fresh clone.
-- [x] Add a unit test that runs the example notebook end-to-end (e.g. using `nbmake` or `pytest-notebook`)
-  so regressions are caught automatically.
-
-## Performance claim
[…]
