### 2026-05-13 157d13c
prepare generic alignment functions
