+++ src/intravascular/centerline_align/align_algorithms.rs
@@ -7,6 +7,50 @@ use crate::intravascular::processing::process_utils::{
 };
 use nalgebra::{Point3, Rotation3, Unit, Vector3};
 
+/// Allows alignment algorithms to operate on either a single [`Geometry`] or a [`GeometryPair`].
+pub trait AlignTarget: Sized + Clone {
+    /// Returns the primary geometry used as the reference for computing transformations.
+    fn primary_geometry(&self) -> &Geometry;
+    /// Applies frame transformations to all contained geometries.
+    fn apply_frame_transforms(self, transformations: &[FrameTransformation]) -> Self;
+    /// Rotates all contained geometries by `angle` radians around each frame's normal.
+    fn rotate_all(self, angle: f64) -> Self;
+}
+
+impl AlignTarget for Geometry {
+    fn primary_geometry(&self) -> &Geometry {
+        self
+    }
+
+    fn apply_frame_transforms(mut self, transformations: &[FrameTransformation]) -> Self {
+        apply_transforms_to_geometry(&mut self, transformations);
+        self
+    }
+
+    fn rotate_all(mut self, angle: f64) -> Self {
+        self.rotate_geometry(angle);
+        self
+    }
+}
+
+impl AlignTarget for GeometryPair {
+    fn primary_geometry(&self) -> &Geometry {
+        &self.geom_a
+    }
+
+    fn apply_frame_transforms(mut self, transformations: &[FrameTransformation]) -> Self {
+        apply_transforms_to_geometry(&mut self.geom_a, transformations);
+        apply_transforms_to_geometry(&mut self.geom_b, transformations);
+        self
+    }
+
+    fn rotate_all(mut self, angle: f64) -> Self {
+        self.geom_a.rotate_geometry(angle);
+        self.geom_b.rotate_geometry(angle);
+        self
+    }
+}
+
 #[derive(Debug, Clone, Copy)]
 pub struct FrameTransformation {
     pub frame_index: u32,
@@ -281,17 +325,17 @@ pub fn best_rotation_three_point(
 }
 
 /// Refines alignment using Hausdorff distance within a limited search space
-pub fn refine_alignment_hausdorff(
-    geom_pair: &GeometryPair,
+pub fn refine_alignment_hausdorff<T: AlignTarget>(
+    target: &T,
     centerline: &Centerline,
     initial_cl_ref_idx: usize,
     initial_rotation: f64,
     mutated_points: &[ContourPoint],
-    angle_search_range: f64, // e.g., 15° in radians
+    angle_search_range: f64,
     angle_step: f64,
-    index_search_range: usize, // e.g., 2 points
+    index_search_range: usize,
 ) -> (f64, usize) {
-    let len_frames = geom_pair.geom_a.frames.len();
+    let len_frames = target.primary_geometry().frames.len();
 
     let mut best_angle = initial_rotation;
     let mut best_cl_ref_idx = initial_cl_ref_idx;
@@ -304,38 +348,32 @@ pub fn refine_alignment_hausdorff(
         initial_cl_ref_idx
     );
 
-    // Search over centerline indices
     for delta_idx in -(index_search_range as isize)..=(index_search_range as isize) {
         let current_cl_ref_idx = (initial_cl_ref_idx as isize + delta_idx) as usize;
 
-        // Ensure we have enough centerline points for all frames
         if current_cl_ref_idx + len_frames >= centerline.points.len() {
             continue;
         }
 
-        // Create centerline segment for this index
         let cl_end_idx = current_cl_ref_idx + len_frames;
         let cl_segment = Centerline {
             points: centerline.points[current_cl_ref_idx..cl_end_idx].to_vec(),
         };
 
-        // Search over rotation angles
[…]
