### 2026-05-13 ed8a86d
provide tests to show same results with pygeometry and pygeometrypair
