### 2026-05-13 9658069
add parallelization to ray intersection algorithm
