### 2026-05-13 e9d11ce
adjusting the stitching algorithm to project boundary on plane
