### 2026-05-13 34daf8a
smoother proximal stitching with laplacian
