+++ multimodars/ccta/manipulating.py
@@ -37,6 +37,34 @@ def _project_to_best_fit_plane(
     return [tuple(p) for p in projected]
 
 
+def _smooth_ring_laplacian(
+    points: list[tuple[float, float, float]],
+    iterations: int = 5,
+    alpha: float = 0.5,
+) -> list[tuple[float, float, float]]:
+    """Laplacian smoothing of a closed boundary ring.
+
+    Each vertex is blended toward the midpoint of its two ring neighbors.
+    Since the input is already coplanar, the result stays on the same plane
+    (a linear combination of coplanar points is coplanar).
+
+    Parameters
+    ----------
+    iterations : int
+        Number of smoothing passes.
+    alpha : float
+        Weight kept on the original position (0 = full Laplacian, 1 = no-op).
+    """
+    if len(points) < 3:
+        return points
+    pts = np.array(points, dtype=np.float64)
+    for _ in range(iterations):
+        prev = pts.copy()
+        neighbor_avg = (np.roll(prev, 1, axis=0) + np.roll(prev, -1, axis=0)) / 2.0
+        pts = alpha * prev + (1.0 - alpha) * neighbor_avg
+    return [tuple(p) for p in pts]
+
+
 def _order_boundary_components(
     boundary_indices: set[int],
     adj_map: Mapping[int, list[int] | set[int]],
@@ -435,17 +463,10 @@ def remove_labeled_points_from_mesh(
         for i in range(n_vertices)
         if keep_mask[i] and any(j in remove_indices for j in adj_map.get(i, []))
     }
-    # Project each ring component onto its own best-fit plane so that a
-    # proximal and distal ring (produced by two separate removals) are each
-    # flattened independently rather than onto a single averaged plane.
     components = _order_boundary_components(boundary_indices, adj_map)
-    ordered_boundary: list[tuple[int, tuple]] = []
-    boundary_points: list[tuple] = []
-    for component in components:
-        raw = [tuple(mesh.vertices[i]) for i in component]
-        projected = _project_to_best_fit_plane(raw)
-        ordered_boundary.extend(zip(component, projected))
-        boundary_points.extend(projected)
+    boundary_points: list[tuple] = [
+        tuple(mesh.vertices[i]) for component in components for i in component
+    ]
 
     # 4. Drop faces that reference any removed vertex
     face_keep_mask = np.all(keep_mask[mesh.faces], axis=1)
@@ -456,13 +477,7 @@ def remove_labeled_points_from_mesh(
     new_index[keep_mask] = np.arange(keep_mask.sum(), dtype=np.int64)
     new_faces = new_index[new_faces]
 
-    # Update boundary vertices in the mesh to their projected positions so that
-    # new_coord_set and all downstream coordinate lookups stay consistent.
-    new_vertices = mesh.vertices[keep_mask].copy()
-    for orig_idx, proj_pt in ordered_boundary:
-        remapped = new_index[orig_idx]
-        if remapped >= 0:
-            new_vertices[remapped] = proj_pt
+    new_vertices = mesh.vertices[keep_mask]
     new_mesh = trimesh.Trimesh(vertices=new_vertices, faces=new_faces, process=False)
 
     # 6. Rebuild the results dict with updated coordinate lists
@@ -557,13 +572,9 @@ def keep_labeled_points_from_mesh(
         i for i in keep_indices if any(j in remove_indices for j in adj_map.get(i, []))
     }
     components = _order_boundary_components(boundary_indices, adj_map)
-    ordered_boundary: list[tuple[int, tuple]] = []
[…]
