### 2026-05-15 0b51cda
add buffer to clamping of boundary points
