### 2026-05-15 5464253
ensure that after clipping to plane no points, are "above" plane for consecutive layers
