### 2026-05-15 110f658
add tests for new functions (ai generated)
