+++ tests/test_ccta.py
@@ -27,7 +27,10 @@ from multimodars.ccta.labeling import (
     _prepare_faces_for_rust,
 )
 from multimodars.ccta.manipulating import (
+    _clamp_to_plane,
+    _enforce_layer_gap_from_plane,
     _fix_ring_direction_by_distance,
+    _prepare_prox_dist_boundary_pts,
     _rotate_to_nearest_iv,
     _stitch_boundary_ring,
     keep_labeled_points_from_mesh,
@@ -655,3 +658,311 @@ class TestStitchBoundaryRing:
         if valid.any():
             avg_normal = patch.face_normals[valid].mean(axis=0)
             assert np.dot(avg_normal, outward) > 0
+
+
+# ---------------------------------------------------------------------------
+# Additional mesh factories for ostium tests
+# ---------------------------------------------------------------------------
+
+def _make_concentric_ring_mesh() -> trimesh.Trimesh:
+    """Three concentric rings of 4 vertices each in the z=0 plane.
+
+    Ring A (inner, radius 1): indices 0-3
+    Ring B (middle, radius 2): indices 4-7
+    Ring C (outer, radius 3): indices 8-11
+
+    Faces connect adjacent rings so adjacency map gives
+    A-neighbours = B and B-neighbours = C.
+    """
+    angles = [0.0, np.pi / 2, np.pi, 3 * np.pi / 2]
+    ring_a = [(np.cos(a), np.sin(a), 0.0) for a in angles]
+    ring_b = [(2.0 * np.cos(a), 2.0 * np.sin(a), 0.0) for a in angles]
+    ring_c = [(3.0 * np.cos(a), 3.0 * np.sin(a), 0.0) for a in angles]
+    verts = np.array(ring_a + ring_b + ring_c, dtype=float)
+    faces = []
+    for i in range(4):
+        j = (i + 1) % 4
+        faces.extend([
+            [i, j, i + 4], [j, j + 4, i + 4],
+            [i + 4, j + 4, i + 8], [j + 4, j + 8, i + 8],
+        ])
+    return trimesh.Trimesh(vertices=verts, faces=np.array(faces), process=False)
+
+
+def _make_annular_xz_mesh() -> trimesh.Trimesh:
+    """Annular mesh lying in the XZ plane (y=0).
+
+    Inner ring (radius 1, 8 verts): indices 0-7  — these become boundary_points.
+    Outer ring (radius 2, 8 verts): indices 8-15 — second aortic layer.
+
+    The ring plane has normal [0,1,0]; the IV ring (XY plane) has normal [0,0,1].
+    The angle between them is 90°, so the anomalous clamping path is triggered.
+    """
+    n = 8
+    angles = np.linspace(0, 2 * np.pi, n, endpoint=False)
+    inner = np.column_stack([np.cos(angles), np.zeros(n), np.sin(angles)])
+    outer = np.column_stack([2.0 * np.cos(angles), np.zeros(n), 2.0 * np.sin(angles)])
+    verts = np.vstack([inner, outer])
+    faces = []
+    for i in range(n):
+        j = (i + 1) % n
+        faces.extend([[i, j, i + n], [j, j + n, i + n]])
+    return trimesh.Trimesh(vertices=verts, faces=np.array(faces), process=False)
+
+
+# ===========================================================================
+# manipulating._clamp_to_plane
+# ===========================================================================
+
+class TestClampToPlane:
+    """Plane: z=0, normal=[0,0,1]. Correct side is z>0."""
+
+    _origin = np.array([0.0, 0.0, 0.0])
+    _normal = np.array([0.0, 0.0, 1.0])
+
[…]
