### 2026-05-15 e45999d
update documentation and changelog
