+++ tests/test_ccta.py
@@ -10,6 +10,7 @@ Covers:
                   _rotate_to_nearest_iv, _fix_ring_direction_by_distance,
                   _stitch_boundary_ring
 """
+
 from __future__ import annotations
 
 import importlib
@@ -19,7 +20,10 @@ import pytest
 import trimesh
 
 from multimodars import PyContourPoint
-from multimodars.ccta.fixing_functions import manual_hole_fill, postprocess_stitched_mesh
+from multimodars.ccta.fixing_functions import (
+    manual_hole_fill,
+    postprocess_stitched_mesh,
+)
 from multimodars.ccta.labeling import (
     _find_aortic_points,
     _find_faces_for_points,
@@ -40,10 +44,10 @@ from multimodars.ccta.manipulating import (
     sync_results_to_mesh,
 )
 
-
 # ---------------------------------------------------------------------------
 # Shared mesh factories
 # ---------------------------------------------------------------------------
+
 
 def _make_grid_mesh() -> trimesh.Trimesh:
     """3x3 grid (9 vertices, 8 triangular faces, z=0 plane).
@@ -108,6 +112,7 @@ def _make_iv_pts(coords) -> list[PyContourPoint]:
 # Fixtures
 # ---------------------------------------------------------------------------
 
+
 @pytest.fixture
 def grid_mesh():
     return _make_grid_mesh()
@@ -124,9 +129,9 @@ def grid_results(grid_mesh):
     verts = [tuple(v) for v in grid_mesh.vertices]
     return {
         "mesh": grid_mesh,
-        "aorta_points": verts[6:9],   # vertices 6, 7, 8
-        "rca_points":   verts[0:3],   # vertices 0, 1, 2
-        "lca_points":   verts[3:6],   # vertices 3, 4, 5
+        "aorta_points": verts[6:9],  # vertices 6, 7, 8
+        "rca_points": verts[0:3],  # vertices 0, 1, 2
+        "lca_points": verts[3:6],  # vertices 3, 4, 5
         "rca_removed_points": [],
         "lca_removed_points": [],
     }
@@ -135,6 +140,7 @@ def grid_results(grid_mesh):
 # ===========================================================================
 # labeling._find_aortic_points
 # ===========================================================================
+
 
 class TestFindAorticPoints:
     def test_basic_set_difference(self, grid_mesh):
@@ -164,6 +170,7 @@ class TestFindAorticPoints:
 # labeling._find_faces_for_points
 # ===========================================================================
 
+
 class TestFindFacesForPoints:
     def test_corner_vertex_finds_its_face(self, grid_mesh):
         # vertex 0 = (0,0,0) belongs to face [0,1,3] → face index 0
@@ -191,6 +198,7 @@ class TestFindFacesForPoints:
 # labeling._prepare_faces_for_rust
 # ===========================================================================
 
+
 class TestPrepareFacesForRust:
     def test_all_faces_when_no_args(self, grid_mesh):
         rust_faces = _prepare_faces_for_rust(grid_mesh)
@@ -212,7 +220,9 @@ class TestPrepareFacesForRust:
 
     def test_subset_via_points(self, grid_mesh):
         # corner vertex 0 only in ~1 face, so subset < all
-        rust_faces = _prepare_faces_for_rust(grid_mesh, points=[(0.0, 0.0, 0.0)], tol=1e-6)
+        rust_faces = _prepare_faces_for_rust(
+            grid_mesh, points=[(0.0, 0.0, 0.0)], tol=1e-6
+        )
         assert 0 < len(rust_faces) < len(grid_mesh.faces)
 
     def test_empty_points_returns_all(self, grid_mesh):
@@ -224,6 +234,7 @@ class TestPrepareFacesForRust:
 # ===========================================================================
 # labeling._final_reclassification
[…]
