### 2026-05-15 cdef5c2
changed "{}", var to "{var}" style
