+++ play.py
@@ -2,6 +2,9 @@ import os
 from pathlib import Path
 import numpy as np
 import multimodars as mm
+import matplotlib
+
+matplotlib.use("TkAgg")  # Use TkAgg backend for interactive plotting
 import matplotlib.pyplot as plt
 
 cwd = Path.cwd()
@@ -61,6 +64,37 @@ print(
 #         ha='center',
 #         va='center',
 #     )
+
+plt.tight_layout()
+plt.show()
+
+lca_cl = lca_cl.calculate_branches(2.0)
+
+list_x = []
+list_y = []
+list_z = []
+point_ids = []
+branch_ids = []
+
+for point in lca_cl.points:
+    list_x.append(point.contour_point.x)
+    list_y.append(point.contour_point.y)
+    list_z.append(point.contour_point.z)
+    branch_ids.append(point.branch_id)
+    point_ids.append(point.contour_point.point_index)
+
+branch_arr = np.array(branch_ids)
+
+fig = plt.figure()
+ax = fig.add_subplot(projection="3d")
+sc = ax.scatter(list_x, list_y, list_z, c=branch_arr, cmap="tab10", s=10)
+fig.colorbar(sc, ax=ax, label="Branch ID")
+ax.set_xlabel("X")
+ax.set_ylabel("Y")
+ax.set_zlabel("Z")
+print(
+    f"Branches: {len(lca_cl.branch_start_indices)}, start indices: {lca_cl.branch_start_indices}"
+)
 
 plt.tight_layout()
 plt.show()
