### 2026-05-18 aac028f
refactor calculate_branches add prototype for split by sharp angle
