+++ src/intravascular/io/input.rs
@@ -1,3 +1,6 @@
+#[allow(unused_imports)]
+use crate::ccta::binding::ccta_py::build_adjacency_map;
+
 use super::geometry::ContourType;
 
 use anyhow::{anyhow, Context, Result};
@@ -434,8 +437,8 @@ impl Centerline {
     pub fn calculate_branches(&mut self, spacing_tolerance: f64) {
         const MIN_BRANCH_SIZE: usize = 5;
 
-        let n = self.points.len();
-        if n == 0 {
+        let n_points = self.points.len();
+        if n_points == 0 {
             self.branch_start_indices = vec![];
             return;
         }
@@ -444,7 +447,7 @@ impl Centerline {
 
         // Identify segment boundaries (large consecutive gaps).
         let mut seg_starts: Vec<usize> = vec![0];
-        for i in 1..n {
+        for i in 1..n_points {
             if self.points[i - 1]
                 .contour_point
                 .distance_to(&self.points[i].contour_point)
@@ -453,14 +456,85 @@ impl Centerline {
                 seg_starts.push(i);
             }
         }
-        seg_starts.push(n); // sentinel
+        seg_starts.push(n_points); // sentinel
+
+        let adj_map = self.build_sparse_tree_adjacency(seg_starts, n_points, threshold);
+
+        let (main_path, side_components) = self.identify_components_with_bfs(&adj_map, n_points);
+
+        // Tiny components are artefacts; merge into branch 0 instead of own branch.
+        let mut artefacts: Vec<Vec<usize>> = Vec::new();
+        let mut real_branches: Vec<Vec<usize>> = Vec::new();
+        for comp in side_components {
+            if comp.len() < MIN_BRANCH_SIZE {
+                artefacts.push(comp);
+            } else {
+                real_branches.push(comp);
+            }
+        }
+        real_branches.sort_by_key(|b| std::cmp::Reverse(b.len()));
+
+        // Order each branch spatially before split_sharp_angles, which requires
+        // consecutive triplets to represent actual spatial neighbours.
+        let ordered_real_branches: Vec<Vec<usize>> = real_branches
+            .into_iter()
+            .map(|b| Self::order_chain(&b, &adj_map))
+            .collect();
+
+        let real_branches = self.split_sharp_angles(ordered_real_branches);
+
+        let mut new_points: Vec<CenterlinePoint> = Vec::with_capacity(n_points);
+        let mut branch_start_indices: Vec<usize> = Vec::new();
+        let mut global_idx: u32 = 0;
+
+        branch_start_indices.push(0);
+        for &idx in &main_path {
+            let mut pt = self.points[idx].clone();
+            pt.branch_id = 0;
+            pt.contour_point.point_index = global_idx;
+            global_idx += 1;
+            new_points.push(pt);
+        }
+        for artefact in &artefacts {
+            for idx in Self::order_chain(artefact, &adj_map) {
+                let mut pt = self.points[idx].clone();
+                pt.branch_id = 0;
+                pt.contour_point.point_index = global_idx;
+                global_idx += 1;
+                new_points.push(pt);
+            }
+        }
+
+        for (i, branch) in real_branches.iter().enumerate() {
+            branch_start_indices.push(new_points.len());
+            // branch is already spatially ordered from the pre-sort above
+            for &idx in branch {
+                let mut pt = self.points[idx].clone();
+                pt.branch_id = (i + 1) as u32;
+                pt.contour_point.point_index = global_idx;
+                global_idx += 1;
+                new_points.push(pt);
+            }
+        }
+
+        self.points = new_points;
+        self.branch_start_indices = branch_start_indices;
+        self.recompute_normals();
+    }
+
[…]
