### 2026-05-18 b1e7d14
fixing splitting error, so main segment can also be split
