### 2026-05-18 0169997
seperate calculate branches logic from sharp angle, splitting and merging
