+++ src/intravascular/io/input.rs
@@ -1,6 +1,3 @@
-#[allow(unused_imports)]
-use crate::ccta::binding::ccta_py::build_adjacency_map;
-
 use super::geometry::ContourType;
 
 use anyhow::{anyhow, Context, Result};
@@ -472,28 +469,8 @@ impl Centerline {
                 real_branches.push(comp);
             }
         }
-        // Split the main path at sharp angles too: BFS diameter goes from one tip
-        // to the other, so a bifurcation (e.g. LAD/LCX) sits as an interior elbow.
-        // The longest resulting segment stays as branch 0; shorter ones join real_branches.
-        let mut main_path_segs = self.split_sharp_angles(vec![main_path]);
-        let new_main_path = if main_path_segs.len() == 1 {
-            main_path_segs.remove(0)
-        } else {
-            let longest_idx = main_path_segs
-                .iter()
-                .enumerate()
-                .max_by_key(|(_, s)| s.len())
-                .map(|(i, _)| i)
-                .unwrap_or(0);
-            let longest = main_path_segs.remove(longest_idx);
-            real_branches.extend(main_path_segs);
-            longest
-        };
-
         real_branches.sort_by_key(|b| std::cmp::Reverse(b.len()));
 
-        // Order each branch spatially before split_sharp_angles, which requires
-        // consecutive triplets to represent actual spatial neighbours.
         let ordered_real_branches: Vec<Vec<usize>> = real_branches
             .into_iter()
             .map(|b| Self::order_chain(&b, &adj_map))
@@ -504,7 +481,7 @@ impl Centerline {
         let mut global_idx: u32 = 0;
 
         branch_start_indices.push(0);
-        for &idx in &new_main_path {
+        for &idx in &main_path {
             let mut pt = self.points[idx].clone();
             pt.branch_id = 0;
             pt.contour_point.point_index = global_idx;
@@ -722,72 +699,6 @@ impl Centerline {
         ordered
     }
 
-    /// Split any branch whose path contains an opening angle < 90° at an interior
-    /// point.  The opening angle at `curr` is measured between the arms
-    /// `prev→curr` and `next→curr`; for a straight segment this is ≈ 180°, so
-    /// `cos > 0` reliably marks an acute elbow that indicates two sub-branches
-    /// were accidentally concatenated.
-    ///
-    /// A work queue handles cascading splits; every sub-segment is re-checked
-    /// until no sharp angle remains.  Returns two or more branch index lists.
-    fn split_sharp_angles(&self, branches: Vec<Vec<usize>>) -> Vec<Vec<usize>> {
-        let mut result: Vec<Vec<usize>> = Vec::new();
-        let mut queue: Vec<Vec<usize>> = branches;
-
-        while let Some(branch) = queue.pop() {
-            let n = branch.len();
-            if n < 3 {
-                result.push(branch);
-                continue;
-            }
-
-            let mut split_at: Option<usize> = None;
-            for i in 1..n - 1 {
-                let prev = &self.points[branch[i - 1]].contour_point;
-                let curr = &self.points[branch[i]].contour_point;
-                let next = &self.points[branch[i + 1]].contour_point;
-
-                let v1 = Vector3::new(prev.x - curr.x, prev.y - curr.y, prev.z - curr.z);
-                let v2 = Vector3::new(next.x - curr.x, next.y - curr.y, next.z - curr.z);
-
-                let norm1 = v1.norm();
-                let norm2 = v2.norm();
-                if norm1 < 1e-10 || norm2 < 1e-10 {
-                    continue;
[…]
