+++ src/ccta/discretizing/creating.rs
@@ -1,0 +1,392 @@
+use crate::intravascular::io::geometry::Contour;
+use crate::intravascular::io::input::ContourPoint;
+use nalgebra::Vector3;
+
+/// Filters and resamples raw projected slices from `walk_centerline_slices`:
+/// - Removes empty slices (CL was outside the vessel).
+/// - Removes slices whose points don't cover all four angular quadrants (partial entry slices).
+/// - Fits a closed Catmull-Rom spline through the remaining points and resamples each to
+///   exactly `n_points` evenly-spaced points along the spline.
+pub fn create_uniform_contours(contours: Vec<Contour>, n_points: usize) -> Vec<Contour> {
+    contours
+        .into_iter()
+        .filter(|c| !c.points.is_empty())
+        .filter(has_full_angular_coverage)
+        .filter_map(|c| resample_spline(c, n_points))
+        .collect()
+}
+
+// Returns true when the contour's points cover all four 90-degree sectors around the centroid.
+fn has_full_angular_coverage(contour: &Contour) -> bool {
+    if contour.points.len() < 4 {
+        return false;
+    }
+    let centroid = match contour.centroid {
+        Some(c) => c,
+        None => return false,
+    };
+    let (u, v) = match local_basis(&contour.points, centroid) {
+        Some(b) => b,
+        None => return false,
+    };
+    let c = Vector3::new(centroid.0, centroid.1, centroid.2);
+    let mut quadrants = [false; 4];
+    for p in &contour.points {
+        let pv = Vector3::new(p.x - c.x, p.y - c.y, p.z - c.z);
+        let qu = pv.dot(&u);
+        let qv = pv.dot(&v);
+        let idx = match (qu >= 0.0, qv >= 0.0) {
+            (true, true) => 0,
+            (false, true) => 1,
+            (false, false) => 2,
+            (true, false) => 3,
+        };
+        quadrants[idx] = true;
+    }
+    quadrants.iter().all(|&q| q)
+}
+
+// Sort by angle in the local plane, fit a closed Catmull-Rom spline, resample to n_points.
+fn resample_spline(contour: Contour, n_points: usize) -> Option<Contour> {
+    if n_points < 2 || contour.points.len() < 3 {
+        return None;
+    }
+    let centroid = contour.centroid?;
+    let (u, v) = local_basis(&contour.points, centroid)?;
+    let c = Vector3::new(centroid.0, centroid.1, centroid.2);
+
+    // Sort control points by angle in the local 2D plane.
+    let mut angle_pts: Vec<(f64, Vector3<f64>)> = contour
+        .points
+        .iter()
+        .map(|p| {
+            let d = Vector3::new(p.x - c.x, p.y - c.y, p.z - c.z);
+            let angle = d.dot(&v).atan2(d.dot(&u));
+            (angle, Vector3::new(p.x, p.y, p.z))
+        })
+        .collect();
+    angle_pts.sort_by(|a, b| a.0.partial_cmp(&b.0).unwrap());
+    let ctrl: Vec<Vector3<f64>> = angle_pts.into_iter().map(|(_, p)| p).collect();
+    let n = ctrl.len();
+
+    // Dense-sample the closed Catmull-Rom spline.
+    const SAMPLES_PER_SEG: usize = 32;
+    let mut curve: Vec<Vector3<f64>> = Vec::with_capacity(n * SAMPLES_PER_SEG + 1);
+    for i in 0..n {
+        let p0 = ctrl[(i + n - 1) % n];
+        let p1 = ctrl[i];
+        let p2 = ctrl[(i + 1) % n];
+        let p3 = ctrl[(i + 2) % n];
+        for j in 0..SAMPLES_PER_SEG {
[…]
