### 2026-05-19 6aaf98c
include Voronoi algorithm to handle curved vessels more reliably.
