+++ src/ccta/discretizing/projecting.rs
@@ -2,16 +2,19 @@ use crate::intravascular::io::geometry::{Contour, ContourType};
 use crate::intravascular::io::input::{Centerline, CenterlinePoint, ContourPoint};
 use nalgebra::Vector3;
 
-/// Walks branch `branch_id` at uniform arc-length steps of `step_size`, collecting a projected
-/// slice at each resampled position. The slab half-thickness is `step_size / 2.0`, so adjacent
-/// slices just touch without overlap. Returns one `Contour` per resampled position.
+/// Walks branch `branch_id` at uniform arc-length steps of `step_size`, assigns each mesh point
+/// to its geometrically closest anchor via Voronoi partitioning, projects it onto that anchor's
+/// perpendicular plane, and returns one `Contour` per sampled position.
+///
+/// Voronoi assignment prevents far-away vessel sections from contaminating a slice: a point that
+/// is physically on a distant part of the vessel will always be closer (in 3-D) to the anchors on
+/// that distant section, so it never ends up in the wrong cross-section.
 pub fn walk_centerline_slices(
     centerline: &Centerline,
     points: &[(f64, f64, f64)],
     branch_id: u32,
     step_size: f64,
 ) -> Vec<Contour> {
-    let half = step_size / 2.0;
     let branch_pts: Vec<&CenterlinePoint> = centerline
         .points
         .iter()
@@ -23,16 +26,90 @@ pub fn walk_centerline_slices(
 
     let cum = branch_cum_arc(&branch_pts);
     let total = *cum.last().unwrap();
-
     let sample_positions = build_sample_positions(total, step_size);
-    sample_positions
+
+    let anchors: Vec<CenterlinePoint> = sample_positions
         .iter()
         .enumerate()
-        .map(|(k, &s)| {
-            let anchor = interpolate_branch_at_s(&branch_pts, &cum, s, k);
-            slice_nearest_points(&anchor, points, half, k as u32)
+        .map(|(k, &s)| interpolate_branch_at_s(&branch_pts, &cum, s, k))
+        .collect();
+
+    if anchors.is_empty() {
+        return vec![];
+    }
+
+    // Voronoi: each mesh point goes to its geometrically closest anchor.
+    let mut buckets: Vec<Vec<ContourPoint>> = vec![vec![]; anchors.len()];
+    for &(px, py, pz) in points {
+        let closest = anchors.iter().enumerate().min_by(|(_, a), (_, b)| {
+            let da = sq_dist3(
+                px,
+                py,
+                pz,
+                a.contour_point.x,
+                a.contour_point.y,
+                a.contour_point.z,
+            );
+            let db = sq_dist3(
+                px,
+                py,
+                pz,
+                b.contour_point.x,
+                b.contour_point.y,
+                b.contour_point.z,
+            );
+            da.partial_cmp(&db).unwrap_or(std::cmp::Ordering::Equal)
+        });
+        if let Some((anchor_idx, anchor)) = closest {
+            let (qx, qy, qz) = project_to_plane((px, py, pz), anchor);
+            let point_index = buckets[anchor_idx].len() as u32;
+            buckets[anchor_idx].push(ContourPoint {
+                frame_index: anchor_idx as u32,
+                point_index,
+                x: qx,
+                y: qy,
+                z: qz,
+                aortic: false,
+            });
[…]
