### 2026-05-19 41813b3
add get_branch method
