+++ multimodars/ccta/labeling.py
@@ -569,3 +569,60 @@ def label_anomalous_region(
         )
 
     return results
+
+
+def label_branches(
+    centerline,
+    results: dict,
+    results_key: str = "rca_points",
+    branch_id: int | list[int] = 0,
+    bounding_sphere_radius_mm: float = 3.0,
+) -> dict:
+    """Partition a coronary region into main branch and side branches based on coronary label.
+
+    Parameters
+    ----------
+    centerline : PyCenterline
+        Centerline of the coronary vessel of interest.
+    results : dict
+        Labelled results dictionary (e.g. from :func:`label_geometry`).
+        Must contain the key specified by *results_key*.
+    results_key : str, optional
+        Key in *results* whose point list is partitioned.
+        Default is ``"rca_points"``.
+    branch_id : int or list of int, optional
+        Branch index or list of branch indices whose combined points form the
+        main branch (e.g. ``[0, 1]`` for LAD + LCx).  Default is ``0``.
+    bounding_sphere_radius_mm : float, optional
+        Radius of the rolling sphere used to collect candidate vertices.
+        Default is ``3.0``.
+
+    Returns
+    -------
+    dict
+        The input *results* dictionary extended with two new keys:
+
+        * ``"{results_key}_main"`` - vertices along the main branch(es).
+        * ``"{results_key}_side"`` - remaining vertices not on any main branch.
+    """
+    branch_ids = [branch_id] if isinstance(branch_id, int) else list(branch_id)
+
+    main_set: set = set()
+    for bid in branch_ids:
+        branch = centerline.get_branch(bid)
+        points_found = find_centerline_bounded_points_simple(
+            branch, results[results_key], bounding_sphere_radius_mm
+        )
+        main_set.update(points_found)
+
+    main_points = [p for p in results[results_key] if p in main_set]
+    side_points = [p for p in results[results_key] if p not in main_set]
+
+    results[f"{results_key}_main"] = main_points
+    results[f"{results_key}_side"] = side_points
+
+    print(f"\nBranch labeling for '{results_key}' (branch_ids={branch_ids}):")
+    print(f"  {results_key}_main: {len(main_points)}")
+    print(f"  {results_key}_side: {len(side_points)}")
+
+    return results
