### 2026-05-19 9858b44
fix walk_centerline_slices to work with all branches
