+++ play.py
@@ -117,7 +117,7 @@ print(
 plt.tight_layout()
 plt.show()
 results = mm.label_branches(rca_cl, results)
-contours = mm.discretize_vessel(rca_cl, results["rca_points_main"], 0, 1.0, 200)
+contours = mm.discretize_vessel(rca_cl, results["rca_points_side"], 1, 1.0, 200)
 print(f"Discretized {len(contours)} contours")
 
 fig = plt.figure()
@@ -153,3 +153,6 @@ ax.set_zlabel("Z")
 ax.set_title(f"Discretized vessel — {len(contours)} contours")
 plt.tight_layout()
 plt.show()
+
+print(contours[0].get_area())
+print(contours[-1].get_area())
