### 2026-05-20 a06e1dd
initial working draft for automatic discretization with ref points
