+++ src/ccta/binding/ccta_py.rs
@@ -10,12 +10,113 @@ use crate::ccta::adjust_mesh::scale_coronary::{
     centerline_based_diameter_optimization, centerline_based_wall_diameter_optimization,
     clean_up_non_section_points, find_points_by_cl_region_rs,
 };
+use crate::ccta::discretizing::data::DiscretizedVesselTree;
 use crate::ccta::discretizing::discretize_vessel_rs;
 use crate::intravascular::binding::classes::{PyCenterline, PyContour, PyFrame};
 use pyo3::prelude::*;
 
 type Point3D = (f64, f64, f64);
 type TriangleTuple = (Point3D, Point3D, Point3D);
+/// (main_ref, counter_clock_ref, clock_ref) — all three as plain xyz tuples.
+type RefTriplet = (Point3D, Point3D, Point3D);
+
+// ─── PyDiscretizedVesselTree ──────────────────────────────────────────────────
+
+/// Fully discretized coronary vessel tree (aorta + RCA + LCA + side branches).
+///
+/// Attributes
+/// ----------
+/// discretized_aorta : list of PyContour
+///     Cross-sectional contours along the aortic centerline.
+/// discretized_rca_main : list of PyContour
+///     Cross-sectional contours along the RCA main vessel.
+/// discretized_lca_main : list of PyContour
+///     Cross-sectional contours along the LCA main vessel.
+/// rca_branches : list of list of PyContour
+///     Per-side-branch contour lists for the RCA.  ``rca_branches[i]``
+///     corresponds to RCA branch_id ``i + 1``.
+/// lca_branches : list of list of PyContour
+///     Per-side-branch contour lists for the LCA.
+/// rca_references : list of tuple
+///     Orientation triplets ``(main_ref, counter_clock_ref, clock_ref)`` along
+///     the RCA, sorted proximal → distal.  Each element is a 3-tuple of
+///     ``(x, y, z)`` coordinate tuples.
+/// lca_references : list of tuple
+///     Same structure for the LCA.
+/// index_ao_rca : int
+///     Index into ``discretized_aorta`` of the slice closest to the RCA ostium.
+/// index_ao_lca : int
+///     Index into ``discretized_aorta`` of the slice closest to the LCA ostium.
+#[pyclass]
+#[derive(Debug, Clone)]
+pub struct PyDiscretizedVesselTree {
+    #[pyo3(get)]
+    pub discretized_aorta: Vec<PyContour>,
+    #[pyo3(get)]
+    pub discretized_rca_main: Vec<PyContour>,
+    #[pyo3(get)]
+    pub discretized_lca_main: Vec<PyContour>,
+    #[pyo3(get)]
+    pub rca_branches: Vec<Vec<PyContour>>,
+    #[pyo3(get)]
+    pub lca_branches: Vec<Vec<PyContour>>,
+    #[pyo3(get)]
+    pub rca_references: Vec<RefTriplet>,
+    #[pyo3(get)]
+    pub lca_references: Vec<RefTriplet>,
+    #[pyo3(get, set)]
+    pub index_ao_rca: usize,
+    #[pyo3(get, set)]
+    pub index_ao_lca: usize,
+}
+
+#[pymethods]
+impl PyDiscretizedVesselTree {
+    fn __repr__(&self) -> String {
+        format!(
+            "DiscretizedVesselTree(ao={}, rca_main={}, lca_main={}, rca_branches={}, lca_branches={}, rca_refs={}, lca_refs={})",
+            self.discretized_aorta.len(),
+            self.discretized_rca_main.len(),
+            self.discretized_lca_main.len(),
+            self.rca_branches.len(),
+            self.lca_branches.len(),
+            self.rca_references.len(),
+            self.lca_references.len(),
+        )
+    }
+}
+
[…]
