+++ src/ccta/discretizing/utils.rs
@@ -1,0 +1,220 @@
+use crate::intravascular::io::input::{Centerline, CenterlinePoint, ContourPoint};
+use nalgebra::Vector3;
+
+/// Smooths centerline positions with a Gaussian kernel (per branch) and recomputes normals.
+///
+/// `sigma` is the half-width in number of centerline points.  A value of 1.0 is a gentle
+/// neighbourhood average; 3–5 removes noise while keeping the overall vessel path; larger
+/// values heavily round corners.  Branches are processed independently so no smoothing
+/// bleeds across the bifurcation.
+pub fn smooth_centerline(centerline: &Centerline, sigma: f64) -> Centerline {
+    if centerline.points.is_empty() || sigma < 1e-12 {
+        return centerline.clone();
+    }
+
+    let n = centerline.points.len();
+    let max_branch = centerline
+        .points
+        .iter()
+        .map(|p| p.branch_id)
+        .max()
+        .unwrap_or(0);
+
+    let mut sx = vec![0.0f64; n];
+    let mut sy = vec![0.0f64; n];
+    let mut sz = vec![0.0f64; n];
+
+    for branch_id in 0..=max_branch {
+        let indices: Vec<usize> = centerline
+            .points
+            .iter()
+            .enumerate()
+            .filter(|(_, p)| p.branch_id == branch_id)
+            .map(|(i, _)| i)
+            .collect();
+
+        if indices.is_empty() {
+            continue;
+        }
+
+        // Truncate kernel at 3σ to avoid O(n²) cost on long vessels.
+        let radius = (3.0 * sigma).ceil() as usize;
+
+        for (li, &gi) in indices.iter().enumerate() {
+            let j_start = li.saturating_sub(radius);
+            let j_end = (li + radius + 1).min(indices.len());
+            let (mut wx, mut wy, mut wz, mut wt) = (0.0f64, 0.0f64, 0.0f64, 0.0f64);
+
+            for (k, &gi_j) in indices[j_start..j_end].iter().enumerate() {
+                let j = j_start + k;
+                let diff = (li as f64) - (j as f64);
+                let w = (-0.5 * diff * diff / (sigma * sigma)).exp();
+                let pt = &centerline.points[gi_j].contour_point;
+                wx += w * pt.x;
+                wy += w * pt.y;
+                wz += w * pt.z;
+                wt += w;
+            }
+
+            if wt > 1e-12 {
+                sx[gi] = wx / wt;
+                sy[gi] = wy / wt;
+                sz[gi] = wz / wt;
+            } else {
+                let pt = &centerline.points[gi].contour_point;
+                sx[gi] = pt.x;
+                sy[gi] = pt.y;
+                sz[gi] = pt.z;
+            }
+        }
+    }
+
+    // Build new points with smoothed positions (normals are placeholder for now).
+    let mut new_points: Vec<CenterlinePoint> = centerline
+        .points
+        .iter()
+        .enumerate()
+        .map(|(i, p)| CenterlinePoint {
+            contour_point: ContourPoint {
+                x: sx[i],
+                y: sy[i],
+                z: sz[i],
+                ..p.contour_point
+            },
+            normal: p.normal,
+            branch_id: p.branch_id,
+        })
+        .collect();
+
+    // Recompute normals from smoothed positions per branch.
+    for branch_id in 0..=max_branch {
+        let indices: Vec<usize> = new_points
+            .iter()
+            .enumerate()
+            .filter(|(_, p)| p.branch_id == branch_id)
+            .map(|(i, _)| i)
+            .collect();
+
+        if indices.is_empty() {
+            continue;
+        }
+
[…]
