+++ src/ccta/discretizing/resampling.rs
@@ -16,7 +16,7 @@ pub fn create_uniform_contours(contours: Vec<Contour>, n_points: usize) -> Vec<C
         .collect()
 }
 
-// Returns true when the contour's points cover all four 90-degree sectors around the centroid.
+/// Returns true when the contour's points cover all four 90-degree sectors around the centroid.
 fn has_full_angular_coverage(contour: &Contour) -> bool {
     if contour.points.len() < 4 {
         return false;
@@ -25,154 +25,187 @@ fn has_full_angular_coverage(contour: &Contour) -> bool {
         Some(c) => c,
         None => return false,
     };
-    let (u, v) = match local_basis(&contour.points, centroid) {
-        Some(b) => b,
+    let (axis_u, axis_v) = match local_basis(&contour.points, centroid) {
+        Some(basis) => basis,
         None => return false,
     };
-    let c = Vector3::new(centroid.0, centroid.1, centroid.2);
+    let centroid_v = Vector3::new(centroid.0, centroid.1, centroid.2);
     let mut quadrants = [false; 4];
     for p in &contour.points {
-        let pv = Vector3::new(p.x - c.x, p.y - c.y, p.z - c.z);
-        let qu = pv.dot(&u);
-        let qv = pv.dot(&v);
-        let idx = match (qu >= 0.0, qv >= 0.0) {
+        let offset = Vector3::new(p.x - centroid_v.x, p.y - centroid_v.y, p.z - centroid_v.z);
+        let proj_u = offset.dot(&axis_u);
+        let proj_v = offset.dot(&axis_v);
+        let quadrant = match (proj_u >= 0.0, proj_v >= 0.0) {
             (true, true) => 0,
             (false, true) => 1,
             (false, false) => 2,
             (true, false) => 3,
         };
-        quadrants[idx] = true;
+        quadrants[quadrant] = true;
     }
     quadrants.iter().all(|&q| q)
 }
 
-// Sort by angle in the local plane, fit a closed Catmull-Rom spline, resample to n_points.
+/// Angle-sort control points, fit a closed Catmull-Rom spline, resample to `n_points`.
 fn resample_spline(contour: Contour, n_points: usize) -> Option<Contour> {
     if n_points < 2 || contour.points.len() < 3 {
         return None;
     }
     let centroid = contour.centroid?;
-    let (u, v) = local_basis(&contour.points, centroid)?;
-    let c = Vector3::new(centroid.0, centroid.1, centroid.2);
+    let basis = local_basis(&contour.points, centroid)?;
 
-    // Sort control points by angle in the local 2D plane.
-    let mut angle_pts: Vec<(f64, Vector3<f64>)> = contour
-        .points
+    let ctrl = sort_by_angle(&contour.points, centroid, basis);
+    let curve = sample_closed_spline(&ctrl);
+    let arc_lengths = cumulative_arc_lengths(&curve);
+
+    let total_length = *arc_lengths.last().unwrap();
+    if total_length < 1e-10 {
+        return None;
+    }
+
+    let resampled = uniform_resample(&curve, &arc_lengths, total_length, n_points);
+    Some(build_output_contour(contour, resampled))
+}
+
+/// Sort contour points by angle in the local 2D plane spanned by `basis`.
+fn sort_by_angle(
+    points: &[ContourPoint],
+    centroid: (f64, f64, f64),
+    (axis_u, axis_v): (Vector3<f64>, Vector3<f64>),
+) -> Vec<Vector3<f64>> {
[…]
