+++ src/ccta/discretizing/mod.rs
@@ -1,10 +1,10 @@
-pub mod creating;
-pub mod data;
 pub mod projecting;
+pub mod resampling;
 mod utils;
+pub mod vessel_tree;
 
-use crate::ccta::discretizing::creating::create_uniform_contours;
 use crate::ccta::discretizing::projecting::walk_centerline_slices;
+use crate::ccta::discretizing::resampling::create_uniform_contours;
 use crate::intravascular::io::geometry::Contour;
 use crate::intravascular::io::input::Centerline;
 
@@ -18,6 +18,9 @@ pub fn discretize_vessel_rs(
     step_size: f64,
     n_points: usize,
 ) -> Vec<Contour> {
-    let raw = walk_centerline_slices(centerline, points, branch_id, step_size);
+    const SMOOTH_SIGMA: f64 = 2.5;
+
+    let cl_smooth = utils::smooth_centerline(centerline, SMOOTH_SIGMA);
+    let raw = walk_centerline_slices(&cl_smooth, points, branch_id, step_size);
     create_uniform_contours(raw, n_points)
 }

[…]
