### 2026-05-21 c9bb9a2
refactor vessel tree and simplify _extrac_wall_from_frames helper
