+++ src/ccta/discretizing/vessel_tree.rs
@@ -200,9 +200,8 @@ impl DiscretizedVesselTree {
     }
 }
 
-// ─── private helpers ────────────────────────────────────────────────────────
-
-/// Build a sorted (proximal→distal) reference-triplet list for one vessel.
+/// Build a sorted (proximal→distal) reference-triplet list for one vessel by
+/// combining one ostium reference and one reference per side-branch bifurcation.
 fn vessel_references(
     ao_centroid: Vector3<f64>,
     main: &[Contour],
@@ -210,144 +209,179 @@ fn vessel_references(
 ) -> Vec<ReferenceTriplet> {
     let main_centroids: Vec<Vector3<f64>> = main.iter().map(contour_centroid).collect();
 
-    // "Up" hint: direction from aorta centroid to first main-vessel centroid,
-    // constant for the whole vessel to keep all refs in the same orientation frame.
+    // Constant "up" hint keeps all triplets in the same orientation frame.
     let up_hint = (main_centroids[0] - ao_centroid)
         .try_normalize(1e-12)
         .unwrap_or(Vector3::z());
 
-    // (sorting_key, ReferenceTriplet) – key = index in main contour list
     let mut tagged: Vec<(usize, ReferenceTriplet)> = Vec::new();
 
-    // ── 1. Ostium reference ──────────────────────────────────────────────────
-    if let Some(first) = main.first() {
-        if first.points.len() > 2 {
-            // Normal at the ostium: toward the second contour
-            let normal = if main.len() > 1 {
-                (main_centroids[1] - main_centroids[0])
-                    .try_normalize(1e-12)
-                    .unwrap_or(Vector3::z())
-            } else {
-                (main_centroids[0] - ao_centroid)
-                    .try_normalize(1e-12)
-                    .unwrap_or(Vector3::z())
-            };
-
-            // main_ref: minor-axis point closer to the aorta
-            let ((pa, pb), _) = first.find_closest_opposite_3d();
-            let pta = Vector3::new(pa.x, pa.y, pa.z);
-            let ptb = Vector3::new(pb.x, pb.y, pb.z);
-            let main_ref_v = if (pta - ao_centroid).norm() <= (ptb - ao_centroid).norm() {
-                pta
-            } else {
-                ptb
-            };
-
-            // counter_clock / clock: major-axis pair
-            let ((p1, p2), _) = first.find_farthest_points();
-            let (cc, cl) = assign_cc_clock(
-                Vector3::new(p1.x, p1.y, p1.z),
-                Vector3::new(p2.x, p2.y, p2.z),
-                main_centroids[0],
-                normal,
-                up_hint,
-            );
-
-            tagged.push((
-                0,
-                ReferenceTriplet {
-                    main_ref: (main_ref_v.x, main_ref_v.y, main_ref_v.z),
-                    counter_clock_ref: (cc.x, cc.y, cc.z),
-                    clock_ref: (cl.x, cl.y, cl.z),
-                },
-            ));
-        }
+    if let Some(entry) = ostium_reference(ao_centroid, main, &main_centroids, up_hint) {
+        tagged.push(entry);
     }
-
-    // ── 2. Side-branch bifurcation references ────────────────────────────────
     for branch_contours in side_branches {
-        if branch_contours.is_empty() {
-            continue;
+        if let Some(entry) =
+            sidebranch_reference(ao_centroid, main, &main_centroids, branch_contours, up_hint)
+        {
[…]
