### 2026-05-21 f79df5f
fix smooth centerline util to use symmetric truncation (before values were not preserved)
