### 2026-05-23 9fe8610
add additional debug plots
