+++ multimodars/ccta/debug_plots.py
@@ -8,7 +8,7 @@ from trimesh.points import PointCloud
 from trimesh.visual import ColorVisuals
 
 if TYPE_CHECKING:
-    from ..multimodars import PyCenterline
+    from ..multimodars import PyCenterline, PyDiscretizedVesselTree
 
 
 def plot_results_key(
@@ -226,3 +226,397 @@ def compare_centerline_scaling(
 
     print("Showing Scene 3: Side-by-side comparison (Blue=Original, Orange=Scaled)")
     scene3.show()
+
+
+def plot_vessel_tree(
+    tree: PyDiscretizedVesselTree,
+    title: str = "Discretized Vessel Tree",
+    pts_per_contour: int = 24,
+) -> None:
+    """Show an interactive 3-D Plotly visualisation of a discretized vessel tree.
+
+    Each vessel segment is rendered as a stack of contour rings (subsampled to
+    ``pts_per_contour`` points for performance) plus centroid dots that trace
+    the vessel centre-line.  Orientation reference triplets are shown as
+    distinct marker symbols.
+
+    Colour coding
+    -------------
+    * Silver         — aorta
+    * Steel-blue     — RCA main
+    * Shades of blue — RCA side branches
+    * Coral          — LCA main
+    * Shades of orange — LCA side branches
+    * Yellow dots    — contour centroids
+    * Red ×          — main reference point
+    * Orange ▲       — counter-clockwise reference
+    * Magenta ▼      — clockwise reference
+
+    Parameters
+    ----------
+    tree:
+        Fully populated vessel tree (output of
+        :func:`~multimodars.ccta.discretization_map.discretize_vessel_tree`).
+    title:
+        Figure title.
+    pts_per_contour:
+        Number of evenly-spaced points sampled from each contour ring.
+        Lower values improve rendering speed; higher values show more detail.
+    """
+    import plotly.graph_objects as go  # lazy import — plotly is optional
+
+    _RCA_BRANCH_COLORS = ["#4fa3e0", "#7ec8e3", "#a8d8ea", "#b8dfed"]
+    _LCA_BRANCH_COLORS = ["#e07f4f", "#e3a87e", "#eac0a8", "#edd0b8"]
+
+    traces: list[go.BaseTraceType] = []
+
+    # ── helpers ──────────────────────────────────────────────────────────────
+
+    def _ring_xyz(contour, n: int):
+        """Subsample a contour ring and close it; return (xs, ys, zs) lists."""
+        pts = contour.points
+        m = len(pts)
+        if m == 0:
+            return [], [], []
+        step = max(1, m // n)
+        sampled = pts[::step]
+        xs = [p.x for p in sampled] + [sampled[0].x]
+        ys = [p.y for p in sampled] + [sampled[0].y]
+        zs = [p.z for p in sampled] + [sampled[0].z]
+        return xs, ys, zs
+
+    def _add_contours(contours, color: str, name: str, opacity: float = 0.7) -> None:
+        """One NaN-separated line trace for all rings + one marker trace for centroids."""
+        rx, ry, rz = [], [], []
+        cx, cy, cz = [], [], []
+        for c in contours:
+            xs, ys, zs = _ring_xyz(c, pts_per_contour)
+            if not xs:
+                continue
+            rx += xs + [None]
+            ry += ys + [None]
+            rz += zs + [None]
+            if c.centroid:
+                cx.append(c.centroid[0])
+                cy.append(c.centroid[1])
+                cz.append(c.centroid[2])
+
+        if rx:
+            traces.append(
+                go.Scatter3d(
+                    x=rx,
+                    y=ry,
+                    z=rz,
+                    mode="lines",
+                    name=name,
+                    line=dict(color=color, width=1),
+                    opacity=opacity,
+                )
[…]
