+++ multimodars/ccta/discretization_map.py
@@ -1,0 +1,288 @@
+from __future__ import annotations
+
+import numpy as np
+from scipy.interpolate import splprep, splev
+
+from ..multimodars import (
+    PyContour,
+    PyContourPoint,
+    PyCenterline,
+    PyDiscretizedVesselTree,
+    discretize_vessel_tree as _discretize_vessel_tree,
+)
+from .labeling import label_branches as _label_branches
+
+
+def _fit_bspline_contour(
+    contour: PyContour,
+    smoothing: float = 0.0,
+    degree: int = 3,
+) -> PyContour:
+    """Return a copy of *contour* whose points lie on a closed B-spline fit.
+
+    Parameters
+    ----------
+    contour:
+        Source contour.
+    smoothing:
+        Smoothing condition ``s`` for ``scipy.interpolate.splprep``.
+        Range: ``[0, ∞)``.
+
+        * ``s = 0``       — exact interpolation; spline passes through every point.
+        * ``s ≈ n``       — gentle smoothing (~1 mm² average residual per point).
+        * ``s ≈ 5 * n``   — strong smoothing.
+
+          where ``n`` is the number of contour points (e.g. ``n = 100`` →
+          start around ``s = 100.0`` and tune from there).
+    degree:
+        B-spline degree (1-5).  3 (cubic) is the typical choice.
+    """
+    pts = contour.points
+    n = len(pts)
+    if n < degree + 1:
+        return contour
+
+    xs = np.array([p.x for p in pts], dtype=np.float64)
+    ys = np.array([p.y for p in pts], dtype=np.float64)
+    zs = np.array([p.z for p in pts], dtype=np.float64)
+
+    try:
+        tck, _ = splprep([xs, ys, zs], s=smoothing, k=degree, per=True)
+    except Exception:
+        return contour
+
+    u_new = np.linspace(0.0, 1.0, n, endpoint=False)
+    xs_new, ys_new, zs_new = splev(u_new, tck)
+
+    centroid = (
+        float(np.mean(xs_new)),
+        float(np.mean(ys_new)),
+        float(np.mean(zs_new)),
+    )
+    new_points = [
+        PyContourPoint(
+            pts[i].frame_index,
+            pts[i].point_index,
+            float(xs_new[i]),
+            float(ys_new[i]),
+            float(zs_new[i]),
+            pts[i].aortic,
+        )
+        for i in range(n)
+    ]
+
+    return PyContour(
+        contour.id,
+        contour.original_frame,
+        new_points,
+        centroid,
+        contour.aortic_thickness,
+        contour.pulmonary_thickness,
+        contour.kind,
+    )
+
+
+def _replace_contours_with_bsplines(
+    tree: PyDiscretizedVesselTree,
+    smoothing: float = 0.0,
+    degree: int = 3,
+) -> PyDiscretizedVesselTree:
+    """Replace every contour in *tree* with a closed B-spline fit (in-place)."""
+
+    def fit(c: PyContour) -> PyContour:
+        return _fit_bspline_contour(c, smoothing, degree)
+
+    tree.discretized_aorta = [fit(c) for c in tree.discretized_aorta]
+    tree.discretized_rca_main = [fit(c) for c in tree.discretized_rca_main]
+    tree.discretized_lca_main = [fit(c) for c in tree.discretized_lca_main]
+    tree.rca_branches = [[fit(c) for c in branch] for branch in tree.rca_branches]
+    tree.lca_branches = [[fit(c) for c in branch] for branch in tree.lca_branches]
+    return tree
+
+
+def _extract_side_branches(results_dict: dict, prefix: str) -> list[list[tuple]]:
[…]

[…]
