+++ src/ccta/binding/ccta_py.rs
@@ -50,15 +50,15 @@ type RefTriplet = (Point3D, Point3D, Point3D);
 #[pyclass]
 #[derive(Debug, Clone)]
 pub struct PyDiscretizedVesselTree {
-    #[pyo3(get)]
+    #[pyo3(get, set)]
     pub discretized_aorta: Vec<PyContour>,
-    #[pyo3(get)]
+    #[pyo3(get, set)]
     pub discretized_rca_main: Vec<PyContour>,
-    #[pyo3(get)]
+    #[pyo3(get, set)]
     pub discretized_lca_main: Vec<PyContour>,
-    #[pyo3(get)]
+    #[pyo3(get, set)]
     pub rca_branches: Vec<Vec<PyContour>>,
-    #[pyo3(get)]
+    #[pyo3(get, set)]
     pub lca_branches: Vec<Vec<PyContour>>,
     #[pyo3(get)]
     pub rca_references: Vec<RefTriplet>,
@@ -83,6 +83,61 @@ impl PyDiscretizedVesselTree {
             self.rca_references.len(),
             self.lca_references.len(),
         )
+    }
+
+    /// Recompute orientation reference triplets and aortic ostium centroids
+    /// from the current contour data.
+    ///
+    /// Call this after replacing contours (e.g. with B-spline fits) so that
+    /// ``rca_references``, ``lca_references``, ``ao_rca``, and ``ao_lca``
+    /// reflect the updated geometry.
+    pub fn calculate_ref_pts(&mut self) -> PyResult<()> {
+        let convert =
+            |contours: &[PyContour]| -> PyResult<Vec<crate::intravascular::io::geometry::Contour>> {
+                contours.iter().map(|c| c.to_rust_contour()).collect()
+            };
+
+        let rust_aorta = convert(&self.discretized_aorta)?;
+        let rust_rca = convert(&self.discretized_rca_main)?;
+        let rust_lca = convert(&self.discretized_lca_main)?;
+        let rust_rca_branches: PyResult<Vec<Vec<_>>> =
+            self.rca_branches.iter().map(|b| convert(b)).collect();
+        let rust_lca_branches: PyResult<Vec<Vec<_>>> =
+            self.lca_branches.iter().map(|b| convert(b)).collect();
+
+        let tree = DiscretizedVesselTree {
+            discretized_aorta: rust_aorta,
+            discretized_rca_main: rust_rca,
+            discretized_lca_main: rust_lca,
+            rca_branches: rust_rca_branches?,
+            lca_branches: rust_lca_branches?,
+            spacing: 1.0,
+            rca_references: vec![],
+            lca_references: vec![],
+            ao_rca: (0.0, 0.0, 0.0),
+            ao_lca: (0.0, 0.0, 0.0),
+            pts_cusp_rcc: None,
+            pts_cusp_lcc: None,
+            pts_cusp_acc: None,
+            index_stj_slice: None,
+            index_aa: None,
+        };
+
+        let updated = tree.calculate_ref_pts();
+        self.rca_references = updated
+            .rca_references
+            .into_iter()
+            .map(|r| (r.main_ref, r.counter_clock_ref, r.clock_ref))
+            .collect();
+        self.lca_references = updated
+            .lca_references
+            .into_iter()
+            .map(|r| (r.main_ref, r.counter_clock_ref, r.clock_ref))
+            .collect();
+        self.ao_rca = updated.ao_rca;
+        self.ao_lca = updated.ao_lca;
+
+        Ok(())
     }
 }
 
@@ -671,7 +726,8 @@ pub fn smooth_mesh_labels(
[…]

[…]
