### 2026-05-24 dbdd765
fix order of superior and inferior reference points
