### 2026-05-25 0b7020e
prepare prototype for better wall alignment
