### 2026-05-25 ff68e3c
fix debug plots to work with trimesh
