+++ fullworkflow.py
@@ -1,0 +1,178 @@
+import os
+from pathlib import Path
+import numpy as np
+import multimodars as mm
+import trimesh
+
+cwd = Path.cwd()
+for candidate in [cwd, cwd.parent, cwd.parent.parent]:
+    if (candidate / "examples" / "data").exists():
+        os.chdir(candidate / "examples" / "data")
+        break
+    elif (candidate / "data").exists():
+        os.chdir(candidate / "data")
+        break
+print(f"Working directory: {os.getcwd()}")
+
+rca_cl_raw = np.genfromtxt("./centerline_rca_short.csv", delimiter=",")
+lca_cl_raw = np.genfromtxt("./centerline_lca.csv", delimiter=",")
+aorta_cl_raw = np.genfromtxt("./centerline_aorta.csv", delimiter=",")
+rca_cl = mm.numpy_to_centerline(rca_cl_raw)
+lca_cl = mm.numpy_to_centerline(lca_cl_raw)
+aorta_cl = mm.numpy_to_centerline(aorta_cl_raw)
+
+results, (rca_cl, lca_cl, ao_cl) = mm.label_geometry(
+    path_ccta_geometry="./NARCO_119_noside.stl",
+    path_centerline_aorta="./centerline_aorta.csv",
+    path_centerline_rca="./centerline_rca_short.csv",
+    path_centerline_lca="./centerline_lca.csv",
+    bounding_sphere_radius_mm=3.0,
+    n_points_intramural=100,
+    anomalous_rca=True,
+    anomalous_lca=False,
+    control_plot=True,
+)
+
+rca_cl, lca_cl, results = mm.prepare_centerlines(
+    rca_cl,
+    lca_cl,
+    results,
+    branch_sigma=2.0,
+    control_plot=True,
+)
+
+list_edges = lca_cl.find_sharp_angles(branch_id=0, cos_threshold=0.0)
+lca_cl = lca_cl.split_branch(0, list_edges[4])
+lca_cl = lca_cl.merge_branches(0, 4)
+lca_cl = lca_cl.check_centerline()
+results = mm.label_branches(lca_cl, results, results_key="lca_points")
+
+tree = mm.discretize_vessel_tree(
+    ao_cl,
+    rca_cl,
+    lca_cl,
+    results,
+    step_size=1.0,
+    n_points=100,
+    b_spline=True,  # set True + tune bspline_smoothing to smooth noisy contours
+    bspline_smoothing=5.0,
+    control_plot=True,
+)
+
+rest, (dia_logs, sys_logs) = mm.from_file_singlepair(
+    input_path="ivus_rest",
+    labels=["aligned_dia", "aligned_sys"],
+    output_path="output/rest",
+)
+
+ref_points = tree.rca_references[0]
+
+rca_cl_main = rca_cl.get_branch(0)  # alignment needs single-branch CL
+aligned, resampled_cl = mm.align_combined(
+    rca_cl_main,
+    rest,
+    ref_points[0],  # aortic reference point
+    ref_points[1],  # superior reference point
+    ref_points[2],  # inferior reference point
+    results["rca_points"],  # CCTA point cloud for Hausdorff refinement
+    angle_range_deg=30.0,
+    write=True,
+    watertight=False,
+    output_dir="test",
+    align_wall_anomalous=True,
+)
+
+results = mm.label_anomalous_region(
+    centerline=rca_cl,
+    frames=aligned.geom_a.frames,
+    results=results,
+    results_key="rca_points",
+    debug_plot=False,
+)
+
+prox_scaling, distal_scaling = mm.find_distal_and_proximal_scaling(
+    frames=aligned.geom_a.frames,
+    centerline=rca_cl,
[…]
