### 2026-05-25 85e0f22
add sharp edges debug plot
