### 2026-05-25 55b1e79
adjust function variable names for reference points
