+++ src/intravascular/centerline_align/align_algorithms.rs
@@ -263,16 +263,17 @@ fn rotate_contour_around_centroid(contour: &mut Contour, angle: f64) {
 pub fn best_rotation_three_point(
     contour: &Contour,
     reference_point: &ContourPoint,
-    aortic_ref_pt: (f64, f64, f64),
-    upper_ref_pt: (f64, f64, f64),
-    lower_ref_pt: (f64, f64, f64),
+    main_ref_pt: (f64, f64, f64),
+    counterclockwise_ref_pt: (f64, f64, f64),
+    clockwise_ref_pt: (f64, f64, f64),
     angle_step: f64,
     centerline_point: &CenterlinePoint,
 ) -> f64 {
     let index_reference = reference_point.point_index;
 
-    let [target_aortic, target_upper, target_lower]: [Point3<f64>; 3] =
-        [aortic_ref_pt, upper_ref_pt, lower_ref_pt].map(|(x, y, z)| Point3::new(x, y, z));
+    let [target_main, target_ccw, target_cw]: [Point3<f64>; 3] =
+        [main_ref_pt, counterclockwise_ref_pt, clockwise_ref_pt]
+            .map(|(x, y, z)| Point3::new(x, y, z));
 
     let mut best_angle = 0.0;
     let mut min_total_error = f64::MAX;
@@ -297,33 +298,32 @@ pub fn best_rotation_three_point(
 
         let n_points = temp_points.len() as u32;
 
-        let p_aortic = temp_points
+        let p_main = temp_points
             .iter()
             .find(|p| p.point_index == index_reference)
             .unwrap();
-        let cont_p_upper = temp_points.iter().find(|p| p.point_index == 0).unwrap();
-        let cont_p_lower = temp_points
+        // point_index == 0 is the highest-Y point → counterclockwise side (viewed proximal → distal)
+        let cont_p_ccw = temp_points.iter().find(|p| p.point_index == 0).unwrap();
+        // point_index == n/2 is the diametrically opposite point → clockwise side
+        let cont_p_cw = temp_points
             .iter()
             .find(|p| p.point_index == (n_points / 2))
             .unwrap();
 
-        let d_aortic = nalgebra::distance(
-            &Point3::new(p_aortic.x, p_aortic.y, p_aortic.z),
-            &target_aortic,
+        let d_main = nalgebra::distance(&Point3::new(p_main.x, p_main.y, p_main.z), &target_main);
+
+        let d_ccw = nalgebra::distance(
+            &Point3::new(cont_p_ccw.x, cont_p_ccw.y, cont_p_ccw.z),
+            &target_ccw,
         );
 
-        let d_upper = nalgebra::distance(
-            &Point3::new(cont_p_upper.x, cont_p_upper.y, cont_p_upper.z),
-            &target_upper,
-        );
-
-        let d_lower = nalgebra::distance(
-            &Point3::new(cont_p_lower.x, cont_p_lower.y, cont_p_lower.z),
-            &target_lower,
+        let d_cw = nalgebra::distance(
+            &Point3::new(cont_p_cw.x, cont_p_cw.y, cont_p_cw.z),
+            &target_cw,
         );
 
         // Calculate sum of squared errors
-        let total_error = d_aortic.powi(2) + d_upper.powi(2) + d_lower.powi(2);
+        let total_error = d_main.powi(2) + d_ccw.powi(2) + d_cw.powi(2);
 
         if total_error < min_total_error {
             min_total_error = total_error;
@@ -896,9 +896,9 @@ mod align_algorithms_tests {
         };
 
         // Set targets to match the current positions (so best rotation should be 0)
-        let aortic_ref_pt = (1.0, 0.0, 0.0);
[…]
