+++ README.md
@@ -77,9 +77,9 @@ cl_raw = np.genfromtxt("examples/data/centerline_raw.csv", delimiter=",")
 centerline = mm.numpy_to_centerline(cl_raw)
 aligned_pair, cl_resampled = mm.align_three_point(
     centerline, rest,
-    aortic_ref_pt=(12.2605, -201.3643, 1751.0554),
-    upper_ref_pt=(11.7567, -202.1920, 1754.7975),
-    lower_ref_pt=(15.6605, -202.1920, 1749.9655),
+    main_ref_pt=(12.2605, -201.3643, 1751.0554),
+    counterclockwise_ref_pt=(11.7567, -202.1920, 1754.7975),
+    clockwise_ref_pt=(15.6605, -202.1920, 1749.9655),
     write=True,
     output_dir="output/aligned",
 )
@@ -134,4 +134,4 @@ Please kindly cite the following paper if you use this repository.
 }
 ```
 
-Stark, Anselm W., Marc Ilic, Ali Mokhtari, Pooya Mohammadi Kazaj, Christoph Graeni, and Isaac Shiri. "multimodars: A Rust-powered toolkit for multi-modality cardiac image fusion and registration." arXiv preprint arXiv:2510.06241 (2025).
+Stark et al., (2026). multimodars: A Rust-powered toolkit for multi-modality cardiac image fusion and registration. Journal of Open Source Software, 11(121), 10131, https://doi.org/10.21105/joss.10131.

[…]
