+++ multimodars/ccta/debug_plots.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from collections import defaultdict
 from typing import TYPE_CHECKING, cast
 
 import numpy as np
@@ -9,6 +10,25 @@ from trimesh.visual import ColorVisuals
 
 if TYPE_CHECKING:
     from ..multimodars import PyCenterline, PyDiscretizedVesselTree
+
+
+def _make_point_cloud(arr: np.ndarray, color: list[int]) -> PointCloud:
+    c = np.array(color, dtype=np.uint8)
+    return PointCloud(arr, colors=np.tile(c, (len(arr), 1)))
+
+
+def _get_cl_arry(cl: PyCenterline) -> np.ndarray:
+    return np.array(
+        [(p.contour_point.x, p.contour_point.y, p.contour_point.z) for p in cl.points],
+        dtype=np.float64,
+    )
+
+
+def _group_by_branch(cl: PyCenterline) -> dict[int, list]:
+    by_branch: dict[int, list] = defaultdict(list)
+    for p in cl.points:
+        by_branch[p.branch_id].append(p.contour_point)
+    return by_branch
 
 
 def plot_results_key(
@@ -90,8 +110,7 @@ def plot_results_key(
         )
         if enabled and pts:
             arr = np.array(pts, dtype=np.float64)
-            colors = np.tile(color, (len(pts), 1))
-            scene_geoms.append(PointCloud(arr, colors=colors))
+            scene_geoms.append(_make_point_cloud(arr, color))
 
     if not scene_geoms:
         print("Nothing to show - all regions are disabled or empty.")
@@ -102,25 +121,15 @@ def plot_results_key(
     scene_geoms.append(mesh_visual)
 
     if cl_rca:
-        rca = _get_cl_arry(cl_rca)
-        scene_geoms.append(PointCloud(rca, colors=[0, 100, 200, 255]))
+        scene_geoms.append(_make_point_cloud(_get_cl_arry(cl_rca), [0, 100, 200, 255]))
     if cl_lca:
-        lca = _get_cl_arry(cl_lca)
-        scene_geoms.append(PointCloud(lca, colors=[0, 150, 0, 255]))
+        scene_geoms.append(_make_point_cloud(_get_cl_arry(cl_lca), [0, 150, 0, 255]))
     if cl_aorta:
-        ao = _get_cl_arry(cl_aorta)
-        scene_geoms.append(PointCloud(ao, colors=[200, 200, 0, 255]))
+        scene_geoms.append(
+            _make_point_cloud(_get_cl_arry(cl_aorta), [200, 200, 0, 255])
+        )
 
-    scene = trimesh.Scene(scene_geoms)
-    scene.show()
-
-
-def _get_cl_arry(cl: PyCenterline) -> np.ndarray:
-    arr = np.array(
-        [(p.contour_point.x, p.contour_point.y, p.contour_point.z) for p in cl.points],
-        dtype=np.float64,
-    )
-    return arr
+    trimesh.Scene(scene_geoms).show()
 
 
 def compare_centerline_scaling(
@@ -151,78 +160,40 @@ def compare_centerline_scaling(
     """
     print("\n=== CENTERLINE SCALING COMPARISON ===")
 
-    region_array = np.array(region_points, dtype=np.float64)
-    region_colors = np.tile([255, 0, 0, 255], (len(region_points), 1))  # Red
-
-    # Scene 1: Original mesh with region highlighted
-    scene1 = trimesh.Scene(
-        [original_mesh, PointCloud(region_array, colors=region_colors)]
+    region_cloud = _make_point_cloud(
+        np.array(region_points, dtype=np.float64), [255, 0, 0, 255]
     )
[…]
