+++ docs/tutorial_ccta.rst
@@ -11,14 +11,15 @@ Tutorial - CCTA Module
 This step-by-step tutorial demonstrates how to:
 
 1. Read in and label a CCTA geometry from an STL file and centerline CSV files
-2. Load an intravascular geometry and fine-align it to the CCTA point cloud
-3. Label the anomalous (intramural) region within the CCTA mesh
-4. Compute radial scaling factors for the proximal, distal, and aortic regions
-5. Morph the CCTA mesh to match the intravascular geometry
-6. Remove the intramural region and stitch the CCTA to the intravascular geometry
-7. Remesh and smooth the stitched geometry
-8. Visualise labeled regions and export section STL files
-9. Re-label the final stitched geometry
+2. Prepare centerlines, detect branches, and discretize the vessel tree
+3. Load an intravascular geometry and fine-align it to the CCTA point cloud
+4. Label the anomalous (intramural) region within the CCTA mesh
+5. Compute radial scaling factors for the proximal, distal, and aortic regions
+6. Morph the CCTA mesh to match the intravascular geometry
+7. Remove the intramural region and stitch the CCTA to the intravascular geometry
+8. Remesh and smooth the stitched geometry
+9. Visualise labeled regions and export section STL files
+10. Re-label the final stitched geometry
 
 The goal of this module is to replace a section on the CCTA geometry with an intravascular geometry.
 
@@ -132,7 +133,121 @@ They can be converted to :class:`multimodars.PyCenterline` objects with normals 
     lca_cl   = mm.numpy_to_centerline(lca_cl_raw)
     aorta_cl = mm.numpy_to_centerline(aorta_cl_raw)
 
-2. Load and align intravascular geometry
+2. Prepare centerlines, detect branches, and discretize the vessel tree
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Before discretizing the vessel geometry, both coronary centerlines must have their branches
+detected and the surface-mesh points labeled by branch.  :func:`multimodars.prepare_centerlines`
+handles all of this in a single call: it runs ``calculate_branches`` and ``check_centerline``
+on each centerline, then calls :func:`multimodars.label_branches` so the ``results`` dictionary
+gains keys ``rca_points_main``, ``rca_points_side_1``, …, ``lca_points_main``,
+``lca_points_side_1``, …:
+
+.. code-block:: python
+
+    rca_cl, lca_cl, results = mm.prepare_centerlines(
+        rca_cl, lca_cl, results,
+        branch_sigma=2.0,
+        control_plot=False,   # True opens a trimesh scene of the branch assignment
+    )
+
+**Parameter reference:**
+
+- ``branch_sigma``: Gaussian smoothing radius (mm) applied during branch detection.  Increase
+  if the algorithm over-segments a noisy centerline; decrease to preserve fine anatomical detail.
+- ``control_plot``: opens an interactive trimesh scene coloured by branch ID and showing the
+  labelled surface points, so you can verify the assignment before discretizing.
+
+Inspecting and correcting sharp angles
+"""""""""""""""""""""""""""""""""""""""
+
+Noisy centerlines sometimes contain kinks that cause branch detection to misfire.
+:func:`multimodars.find_sharp_angles` scans a branch for positions where the local tangent
+reverses by more than the supplied cosine threshold and opens an optional debug scene with each
+flagged position highlighted in a distinct colour:
+
+.. code-block:: python
+
+    # Inspect: red × marks every position that may need a correction
[…]

[…]
