### 2026-05-27 fa42929
bump version sand packages adjust to new pyo3 bindings
