+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "multimodars"
-version = "0.3.5"
+version = "0.4.0"
 edition = "2021"
 license = "MIT"
 
@@ -9,14 +9,14 @@ name = "multimodars"
 crate-type = ["cdylib"]
 
 [dependencies]
-pyo3 = "0.27.2"
+pyo3 = "0.28.3"
 csv = "1.4.0"
 serde = { version = "1.0.228", features = ["derive"]}
-image = "0.25.9"
-nalgebra = "0.34.1"
-toml = "0.9.8"
-rayon = "1.7.1"
+image = "0.25.10"
+nalgebra = "0.35.0"
+toml = "1.1.2+spec-1.1.0"
+rayon = "1.12.0"
 crossbeam = "0.8.4"
-anyhow = "1.0.100"
+anyhow = "1.0.102"
 approx = "0.5"
-indicatif = "0.17"
+indicatif = "0.18.4"
