### 2026-06-01 57e638f
refactor: use parent module idiom for function imports
