+++ src/intravascular/binding/classes.rs
@@ -1,2144 +1,10 @@
-use crate::intravascular::io::geometry::{Contour, ContourType, Frame, Geometry};
-use crate::intravascular::io::input::{
-    Centerline, CenterlinePoint, ContourPoint, InputData, Record,
-};
-use crate::intravascular::processing::align_between::GeometryPair;
-use crate::intravascular::processing::process_utils;
-use anyhow::{anyhow, Result};
-use nalgebra::Vector3;
-use pyo3::prelude::*;
-use std::collections::HashMap;
-
-type GeomSummary = (f64, f64, f64);
-type PairSummary = ((GeomSummary, GeomSummary), Vec<[f64; 6]>);
-use std::convert::TryFrom;
-
-/// Python representation of the intravascular imaging input data for one cardiac phase.
-///
-/// Attributes
-/// ----------
-/// lumen : list of PyContour
-///     Vessel lumen contours.
-/// eem : list of PyContour or None
-///     Vessel EEM (external elastic membrane) contours.
-/// calcification : list of PyContour or None
-///     Vessel calcification contours.
-/// sidebranch : list of PyContour or None
-///     Vessel sidebranch contours.
-/// record : list of PyRecord or None
-///     Metadata records about the input data.
-/// ref_point : PyContourPoint
-///     Reference point used for alignment.
-/// diastole : bool
-///     ``True`` when the data corresponds to the diastolic phase.
-/// label : str
-///     Human-readable label for this input dataset.
-///
-/// Examples
-/// --------
-/// >>> input_data = PyInputData(
-/// ...     lumen=[lumen_contour1, lumen_contour2, ...],
-/// ...     eem=[eem_contour1, eem_contour2, ...],
-/// ...     calcification=[],
-/// ...     sidebranch=[],
-/// ...     record=record,
-/// ...     diastole=True,
-/// ...     lablel="Pat00_diastole_rest"
-/// ... )
-#[pyclass(from_py_object)]
-#[derive(Debug, Clone)]
-pub struct PyInputData {
-    #[pyo3(get, set)]
-    pub lumen: Vec<PyContour>,
-    #[pyo3(get, set)]
-    pub eem: Option<Vec<PyContour>>,
-    #[pyo3(get, set)]
-    pub calcification: Option<Vec<PyContour>>,
-    #[pyo3(get, set)]
-    pub sidebranch: Option<Vec<PyContour>>,
-    #[pyo3(get, set)]
-    pub record: Option<Vec<PyRecord>>,
-    #[pyo3(get, set)]
-    pub ref_point: PyContourPoint,
-    #[pyo3(get, set)]
-    pub diastole: bool,
-    #[pyo3(get, set)]
-    pub label: String,
-}
-
-#[pymethods]
-impl PyInputData {
-    #[new]
-    fn new(
-        lumen: Vec<PyContour>,
-        eem: Option<Vec<PyContour>>,
-        calcification: Option<Vec<PyContour>>,
-        sidebranch: Option<Vec<PyContour>>,
-        record: Option<Vec<PyRecord>>,
-        ref_point: PyContourPoint,
-        diastole: bool,
-        label: String,
-    ) -> Self {
-        Self {
-            lumen,
-            eem,
-            calcification,
-            sidebranch,
-            record,
-            ref_point,
-            diastole,
-            label,
-        }
-    }
-
-    fn __repr__(&self) -> String {
-        format!(
-            "InputData(lumen={}, eem={}, calcification={}, sidebranch={}, record={}, ref_point={}, diastole={}, label='{}')",
-            self.lumen.len(),
-            self.eem.as_ref().map_or(0, |v| v.len()),
[…]
