+++ src/types/native/frame.rs
@@ -1,0 +1,845 @@
+use super::contour::{Contour, ContourType};
+use super::contour_point::ContourPoint;
+use std::collections::HashMap;
+use std::f64::consts::PI;
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct Frame {
+    pub id: u32,
+    pub centroid: (f64, f64, f64),
+    // groundtruth, must exist!
+    pub lumen: Contour,
+    pub extras: HashMap<ContourType, Contour>,
+    pub reference_point: Option<ContourPoint>,
+}
+
+impl Frame {
+    /// Sets a value for all contours and reference point in the frame.
+    /// You can set `id`, `original_frame`, `points`, or `centroid` for all contours and reference point.
+    pub fn set_value(
+        &mut self,
+        id: Option<u32>,
+        lumen_points: Option<Vec<ContourPoint>>,
+        centroid: Option<(f64, f64, f64)>,
+        z_value: Option<f64>,
+    ) {
+        if let Some(new_id) = id {
+            self.id = new_id;
+            self.lumen.id = new_id;
+            for contour in self.extras.values_mut() {
+                contour.id = new_id;
+            }
+        }
+        if let Some(new_points) = lumen_points.clone() {
+            self.lumen.points = new_points.clone();
+            for contour in self.extras.values_mut() {
+                contour.points = new_points.clone();
+            }
+        }
+        if let Some(new_centroid) = centroid {
+            self.lumen.centroid = Some(new_centroid);
+            for contour in self.extras.values_mut() {
+                contour.centroid = Some(new_centroid);
+            }
+            self.centroid = new_centroid;
+        }
+        if let Some(z_value) = z_value {
+            for point in &mut self.lumen.points {
+                point.z = z_value;
+            }
+
+            if let Some(ref mut centroid) = self.lumen.centroid {
+                centroid.2 = z_value;
+            }
+
+            for contour in self.extras.values_mut() {
+                for point in &mut contour.points {
+                    point.z = z_value;
+                }
+
+                if let Some(ref mut centroid) = contour.centroid {
+                    centroid.2 = z_value;
+                }
+            }
+
+            if let Some(ref mut point) = self.reference_point {
+                point.z = z_value;
+            }
+
+            self.centroid.2 = z_value;
+        }
+    }
+
+    pub fn rotate_frame(&mut self, angle: f64) {
+        if angle == 0.0 {
+            return;
+        }
+        let center = (self.centroid.0, self.centroid.1);
+
+        self.lumen.points = self
+            .lumen
+            .points
+            .iter()
+            .map(|p| p.rotate_point(angle, center))
+            .collect();
+
+        for contour in self.extras.values_mut() {
+            contour.points = contour
+                .points
+                .iter()
+                .map(|p| p.rotate_point(angle, center))
+                .collect();
+        }
+
+        if let Some(ref_point) = &mut self.reference_point {
+            *ref_point = ref_point.rotate_point(angle, center);
+        }
+    }
+
+    pub fn translate_frame(&mut self, translation: (f64, f64, f64)) {
+        let (dx, dy, dz) = translation;
+
+        for p in self.lumen.points.iter_mut() {
+            p.x += dx;
+            p.y += dy;
+            p.z += dz;
+        }
+
+        self.lumen.compute_centroid();
+
+        for contour in self.extras.values_mut() {
+            for p in contour.points.iter_mut() {
+                p.x += dx;
+                p.y += dy;
+                p.z += dz;
+            }
+
+            contour.compute_centroid();
+        }
+
+        if let Some(ref_point) = &mut self.reference_point {
+            ref_point.x += dx;
[…]
