+++ src/intravascular/io/build.rs
@@ -1,0 +1,471 @@
+use super::geometry::{Contour, ContourType, Frame, Geometry};
+use super::input::{ContourPoint, InputData};
+use std::collections::{HashMap, HashSet};
+use std::path::Path;
+
+pub fn build_geometry_from_inputdata(
+    input_data: Option<InputData>,
+    path: Option<&Path>,
+    label: &str,
+    diastole: bool,
+    image_center: (f64, f64),
+    radius: f64,
+    n_points: u32,
+) -> anyhow::Result<Geometry> {
+    let input_data = if let Some(input_data) = input_data {
+        input_data
+    } else if let Some(path) = path {
+        // Default mapping for typical AIVUS-CAA structure
+        let mut names = HashMap::new();
+        names.insert(ContourType::Lumen, "lumen");
+        names.insert(ContourType::Eem, "eem");
+        names.insert(ContourType::Calcification, "calcium");
+        names.insert(ContourType::Sidebranch, "branch");
+        names.insert(ContourType::Catheter, "catheter");
+
+        InputData::process_directory(path, names, diastole, label)?
+    } else {
+        return Err(anyhow::anyhow!(
+            "Either input_data or path must be provided"
+        ));
+    };
+    let print_input_data = input_data.clone();
+
+    // First, collect all unique original frame IDs from ALL contour types
+    let mut all_original_frames: HashSet<u32> = HashSet::new();
+
+    for point in &input_data.lumen {
+        all_original_frames.insert(point.frame_index);
+    }
+
+    if let Some(eem_points) = &input_data.eem {
+        for point in eem_points {
+            all_original_frames.insert(point.frame_index);
+        }
+    }
+
+    if let Some(calc_points) = &input_data.calcification {
+        for point in calc_points {
+            all_original_frames.insert(point.frame_index);
+        }
+    }
+
+    if let Some(side_points) = &input_data.sidebranch {
+        for point in side_points {
+            all_original_frames.insert(point.frame_index);
+        }
+    }
+
+    all_original_frames.insert(input_data.ref_point.frame_index);
+
+    let mut sorted_original_frames: Vec<u32> = all_original_frames.into_iter().collect();
+    sorted_original_frames.sort();
+
+    let frame_mapping: HashMap<u32, u32> = sorted_original_frames
+        .iter()
+        .enumerate()
+        .map(|(i, &original_id)| (original_id, i as u32))
+        .collect();
+
+    // Now build contours using the shared mapping
+    let lumen_contours = Contour::build_contour_with_mapping(
+        input_data.lumen,
+        input_data.record.clone(),
+        ContourType::Lumen,
+        &frame_mapping,
+    )?;
+
+    let eem_contours = if let Some(eem_points) = input_data.eem {
+        Contour::build_contour_with_mapping(eem_points, None, ContourType::Eem, &frame_mapping)?
+    } else {
+        Vec::new()
+    };
+
+    let calcification_contours = if let Some(calc_points) = input_data.calcification {
+        Contour::build_contour_with_mapping(
+            calc_points,
+            None,
+            ContourType::Calcification,
+            &frame_mapping,
+        )?
+    } else {
+        Vec::new()
+    };
+
+    let sidebranch_contours = if let Some(side_points) = input_data.sidebranch {
+        Contour::build_contour_with_mapping(
+            side_points,
+            None,
+            ContourType::Sidebranch,
+            &frame_mapping,
+        )?
+    } else {
+        Vec::new()
+    };
+
+    let mut frame_map: HashMap<u32, Frame> = HashMap::new();
+
+    for mut contour in lumen_contours {
+        contour.compute_centroid();
+        let frame_id = contour.id;
+        let mut frame = Frame {
+            id: frame_id,
[…]

[…]
