+++ docs/api/classes/PyContourType.rst
@@ -1,0 +1,10 @@
+PyContourType
+=============
+
+.. currentmodule:: multimodars
+
+.. autoclass:: PyContourType
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :inherited-members:
