### 2026-06-01 b7b0ab1
remove now unncessary classes.rs
