+++ src/intravascular/binding.rs
@@ -1,5 +1,4 @@
 pub mod align;
-pub mod classes;
 pub mod entry;
 pub mod functions;
 
