+++ src/intravascular/binding/classes.rs
@@ -1,7 +1,0 @@
-// Re-exports consumed by align.rs and functions.rs within this binding module.
-// All other types are imported directly from crate::types::binding in lib.rs.
-pub use crate::types::binding::py_centerline::PyCenterline;
-pub use crate::types::binding::py_contour::PyContourType;
-pub use crate::types::binding::py_geometry::PyGeometry;
-pub use crate::types::binding::py_geometry_pair::PyGeometryPair;
-pub use crate::types::binding::py_input_data::PyInputData;
