### 2026-06-04 4a9b71a
move all types into own module
