+++ src/ccta/binding/ccta_py.rs
@@ -5,167 +5,12 @@ use crate::ccta::adjust_mesh::label_coronary;
 use crate::ccta::adjust_mesh::label_coronary::Triangle;
 use crate::ccta::adjust_mesh::scale_coronary;
 use crate::ccta::discretizing;
-use crate::ccta::discretizing::vessel_tree::DiscretizedVesselTree;
-use crate::types::binding::{PyCenterline, PyContour, PyFrame};
+use crate::types::binding::{PyCenterline, PyContour, PyDiscretizedVesselTree, PyFrame};
+use crate::types::native::DiscretizedVesselTree;
 use pyo3::prelude::*;
 
 type Point3D = (f64, f64, f64);
 type TriangleTuple = (Point3D, Point3D, Point3D);
-/// (main_ref, clock_ref, counter_clock_ref) — all three as plain xyz tuples.
-type RefTriplet = (Point3D, Point3D, Point3D);
-
-// ─── PyDiscretizedVesselTree ──────────────────────────────────────────────────
-
-/// Fully discretized coronary vessel tree (aorta + RCA + LCA + side branches).
-///
-/// Attributes
-/// ----------
-/// discretized_aorta : list of PyContour
-///     Cross-sectional contours along the aortic centerline.
-/// discretized_rca_main : list of PyContour
-///     Cross-sectional contours along the RCA main vessel.
-/// discretized_lca_main : list of PyContour
-///     Cross-sectional contours along the LCA main vessel.
-/// rca_branches : list of list of PyContour
-///     Per-side-branch contour lists for the RCA.  ``rca_branches[i]``
-///     corresponds to RCA branch_id ``i + 1``.
-/// lca_branches : list of list of PyContour
-///     Per-side-branch contour lists for the LCA.
-/// rca_references : list of tuple
-///     Orientation triplets ``(main_ref, clock_ref, counter_clock_ref)`` along
-///     the RCA, sorted proximal → distal.  Each element is a 3-tuple of
-///     ``(x, y, z)`` coordinate tuples.
-/// lca_references : list of tuple
-///     Same structure for the LCA.
-/// ao_rca : tuple of float
-///     Centroid ``(x, y, z)`` of the aorta slice closest to the RCA ostium.
-/// ao_lca : tuple of float
-///     Centroid ``(x, y, z)`` of the aorta slice closest to the LCA ostium.
-#[pyclass(skip_from_py_object)]
-#[derive(Debug, Clone)]
-pub struct PyDiscretizedVesselTree {
-    #[pyo3(get, set)]
-    pub discretized_aorta: Vec<PyContour>,
-    #[pyo3(get, set)]
-    pub discretized_rca_main: Vec<PyContour>,
-    #[pyo3(get, set)]
-    pub discretized_lca_main: Vec<PyContour>,
-    #[pyo3(get, set)]
-    pub rca_branches: Vec<Vec<PyContour>>,
-    #[pyo3(get, set)]
-    pub lca_branches: Vec<Vec<PyContour>>,
-    #[pyo3(get)]
-    pub rca_references: Vec<RefTriplet>,
-    #[pyo3(get)]
-    pub lca_references: Vec<RefTriplet>,
-    #[pyo3(get, set)]
-    pub ao_rca: Point3D,
-    #[pyo3(get, set)]
-    pub ao_lca: Point3D,
-}
-
-#[pymethods]
-impl PyDiscretizedVesselTree {
-    fn __repr__(&self) -> String {
-        format!(
-            "DiscretizedVesselTree(ao={}, rca_main={}, lca_main={}, rca_branches={}, lca_branches={}, rca_refs={}, lca_refs={})",
-            self.discretized_aorta.len(),
-            self.discretized_rca_main.len(),
[…]
