+++ src/types/native.rs
@@ -2,6 +2,7 @@ pub mod centerline;
 pub mod centerline_point;
 pub mod contour;
 pub mod contour_point;
+pub mod discretized_tree;
 pub mod frame;
 pub mod geometry;
 pub mod geometry_pair;
@@ -12,6 +13,7 @@ pub use centerline::Centerline;
 pub use centerline_point::CenterlinePoint;
 pub use contour::{Contour, ContourType};
 pub use contour_point::ContourPoint;
+pub use discretized_tree::{DiscretizedVesselTree, ReferenceTriplet};
 pub use frame::Frame;
 pub use geometry::Geometry;
 pub use geometry_pair::GeometryPair;
