+++ src/ccta/discretizing/vessel_tree.rs
@@ -1,73 +1,7 @@
-use crate::types::native::{Centerline, Contour};
-use anyhow::{Ok, Result};
-use nalgebra::Vector3;
-
-pub struct ReferenceTriplet {
-    pub main_ref: (f64, f64, f64),
-    pub counter_clock_ref: (f64, f64, f64), // view from proximal to distal, former upper_ref
-    pub clock_ref: (f64, f64, f64),         // former lower_ref
-}
-
-pub struct DiscretizedVesselTree {
-    pub discretized_aorta: Vec<Contour>,
-    pub discretized_rca_main: Vec<Contour>,
-    pub discretized_lca_main: Vec<Contour>,
-    pub spacing: f64,
-    /// One entry per RCA side branch; index `i` → branch_id `i + 1`.
-    pub rca_branches: Vec<Vec<Contour>>,
-    /// One entry per LCA side branch; index `i` → branch_id `i + 1`.
-    pub lca_branches: Vec<Vec<Contour>>,
-    pub rca_references: Vec<ReferenceTriplet>,
-    pub lca_references: Vec<ReferenceTriplet>,
-    /// Centroid of the aorta slice closest to the RCA ostium.
-    pub ao_rca: (f64, f64, f64),
-    /// Centroid of the aorta slice closest to the LCA ostium.
-    pub ao_lca: (f64, f64, f64),
-    pub pts_cusp_rcc: Option<Vec<(f64, f64, f64)>>,
-    pub pts_cusp_lcc: Option<Vec<(f64, f64, f64)>>,
-    pub pts_cusp_acc: Option<Vec<(f64, f64, f64)>>,
-    pub index_stj_slice: Option<usize>,
-    pub index_aa: Option<usize>,
-}
+use crate::types::native::{Centerline, Contour, DiscretizedVesselTree};
+use anyhow::Result;
 
 impl DiscretizedVesselTree {
-    pub fn new(
-        discretized_aorta: Vec<Contour>,
-        discretized_rca_main: Vec<Contour>,
-        discretized_lca_main: Vec<Contour>,
-        spacing: f64,
-        rca_branches: Vec<Vec<Contour>>,
-        lca_branches: Vec<Vec<Contour>>,
-        rca_references: Vec<ReferenceTriplet>,
-        lca_references: Vec<ReferenceTriplet>,
-        ao_lca: (f64, f64, f64),
-        ao_rca: (f64, f64, f64),
-        pts_cusp_rcc: Option<Vec<(f64, f64, f64)>>,
-        pts_cusp_lcc: Option<Vec<(f64, f64, f64)>>,
-        pts_cusp_acc: Option<Vec<(f64, f64, f64)>>,
-        index_stj_slice: Option<usize>,
-        index_aa: Option<usize>,
-    ) -> anyhow::Result<DiscretizedVesselTree> {
-        let tree = DiscretizedVesselTree {
-            discretized_aorta,
-            discretized_rca_main,
-            discretized_lca_main,
-            spacing,
-            rca_branches,
-            lca_branches,
-            rca_references,
-            lca_references,
-            ao_lca,
-            ao_rca,
-            pts_cusp_rcc,
-            pts_cusp_lcc,
-            pts_cusp_acc,
-            index_stj_slice,
-            index_aa,
-        };
-        Ok(tree)
-    }
-
     /// Discretize the full vessel tree from pre-labelled point sets.
     ///
     /// Smooths all three centerlines with a Gaussian (σ = 2.5 points) before
@@ -111,7 +45,6 @@ impl DiscretizedVesselTree {
             n_points,
         );
 
-        // side_branches_rca[i] carries the surface points for branch_id i+1.
[…]

[…]
