### 2026-06-04 4764acf
move 3D poitn trait to types, small refactor of some functions in types with 3D point
