+++ src/types/native/centerline_point.rs
@@ -1,4 +1,4 @@
-use super::contour_point::ContourPoint;
+use super::contour_point::{ContourPoint, Point3D};
 use nalgebra::Vector3;
 
 #[derive(Debug, Clone, PartialEq)]
@@ -7,4 +7,16 @@ pub struct CenterlinePoint {
     pub normal: Vector3<f64>,
     /// 0 = main vessel, 1+ = side branches ordered by descending length.
     pub branch_id: u32,
+}
+
+impl Point3D for CenterlinePoint {
+    fn x(&self) -> f64 {
+        self.contour_point.x
+    }
+    fn y(&self) -> f64 {
+        self.contour_point.y
+    }
+    fn z(&self) -> f64 {
+        self.contour_point.z
+    }
 }
