### 2026-06-04 e76d279
update changelog fix python binding doc strings
