### 2026-06-04 37d55b8
update changelog fix python binding doc strings
