+++ Cargo.toml
@@ -14,7 +14,7 @@ csv = "1.4.0"
 serde = { version = "1.0.228", features = ["derive"]}
 image = "0.25.10"
 nalgebra = "0.35.0"
-toml = "1.1.2+spec-1.1.0"
+toml = "1.1.2"
 rayon = "1.12.0"
 crossbeam = "0.8.4"
 anyhow = "1.0.102"
