### 2026-06-04 9ad45a0
move downsample_contour_points
