### 2026-06-04 879c224
move downsample_contour_points
