+++ src/ccta/adjust_mesh/scale_coronary.rs
@@ -2,12 +2,12 @@ use crate::types::native::{Centerline, CenterlinePoint, Frame};
 use rayon::prelude::*;
 use std::collections::HashSet;
 
-type Point3D = (f64, f64, f64);
+type Coords3 = (f64, f64, f64);
 
 pub fn centerline_based_wall_diameter_optimization(
     centerline: &Centerline,
-    ref_point_coronary: &Point3D,
-    aortic_points: &[Point3D],
+    ref_point_coronary: &Coords3,
+    aortic_points: &[Coords3],
 ) -> f64 {
     if centerline.points.is_empty() || aortic_points.is_empty() {
         return 0.0;
@@ -25,7 +25,7 @@ pub fn centerline_based_wall_diameter_optimization(
         None => return 0.0,
     };
 
-    let closest_aortic: &Point3D = match aortic_points.iter().min_by(|a, b| {
+    let closest_aortic: &Coords3 = match aortic_points.iter().min_by(|a, b| {
         let dist_a = super::calculate_squared_distance(*a, ref_point_coronary);
         let dist_b = super::calculate_squared_distance(*b, ref_point_coronary);
         dist_a
@@ -66,8 +66,8 @@ pub fn centerline_based_wall_diameter_optimization(
 }
 
 pub fn centerline_based_aortic_diameter_optimization(
-    intramural_points: &[Point3D],
-    reference_points: &[Point3D],
+    intramural_points: &[Coords3],
+    reference_points: &[Coords3],
     centerline: &Centerline,
 ) -> f64 {
     let start = -2.0f64;
@@ -91,12 +91,12 @@ pub fn centerline_based_aortic_diameter_optimization(
 }
 
 pub fn centerline_based_diameter_optimization(
-    anomalous_points: &[Point3D],
+    anomalous_points: &[Coords3],
     n_proximal: usize,
     n_distal: usize,
     centerline: &Centerline,
-    proximal_reference: &[Point3D],
-    distal_reference: &[Point3D],
+    proximal_reference: &[Coords3],
+    distal_reference: &[Coords3],
 ) -> (f64, f64) {
     let (proximal_points, anomalous_points_new) =
         find_region_points(anomalous_points, proximal_reference, n_proximal);
@@ -134,10 +134,10 @@ pub fn centerline_based_diameter_optimization(
 }
 
 fn find_region_points(
-    anomalous_points: &[Point3D],
-    reference_points: &[Point3D],
+    anomalous_points: &[Coords3],
+    reference_points: &[Coords3],
     n_points: usize,
-) -> (Vec<Point3D>, Vec<Point3D>) {
+) -> (Vec<Coords3>, Vec<Coords3>) {
     if anomalous_points.is_empty() || reference_points.is_empty() || n_points == 0 {
         return (Vec::new(), anomalous_points.to_vec());
     }
@@ -165,12 +165,12 @@ fn find_region_points(
 
     let selected_indices: HashSet<usize> = selected_slice.iter().map(|(i, _)| *i).collect();
 
-    let selected_points: Vec<Point3D> = selected_slice
+    let selected_points: Vec<Coords3> = selected_slice
         .iter()
         .map(|(i, _)| anomalous_points[*i])
         .collect();
 
-    let remaining_points: Vec<Point3D> = anomalous_points
+    let remaining_points: Vec<Coords3> = anomalous_points
         .iter()
         .enumerate()
         .filter_map(|(i, pt)| {
@@ -188,7 +188,7 @@ fn find_region_points(
 /// Symmetric average nearest-neighbor distance between two point sets.
[…]
