### 2026-06-04 467b89a
update changelog and bump version
