### 2026-06-04 629bd6f
fix documentation compile errors
