### 2026-06-09 b74e7cc
move traits to native and downsample to contour
