+++ src/types/native/centerline_point.rs
@@ -1,9 +1,9 @@
-use super::contour_point::{ContourPoint, Point3D};
+use super::{contour_point, Point3D};
 use nalgebra::Vector3;
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct CenterlinePoint {
-    pub contour_point: ContourPoint,
+    pub contour_point: contour_point::ContourPoint,
     pub normal: Vector3<f64>,
     /// 0 = main vessel, 1+ = side branches ordered by descending length.
     pub branch_id: u32,
