+++ src/types/native.rs
@@ -11,11 +11,22 @@ pub mod record;
 
 pub use centerline::Centerline;
 pub use centerline_point::CenterlinePoint;
-pub use contour::{Contour, ContourType};
-pub use contour_point::{downsample_contour_points, ContourPoint, Point3D};
+pub use contour::{downsample_contour_points, Contour, ContourType};
+pub use contour_point::ContourPoint;
 pub use discretized_tree::{DiscretizedVesselTree, ReferenceTriplet};
 pub use frame::Frame;
 pub use geometry::Geometry;
 pub use geometry_pair::GeometryPair;
 pub use input_data::InputData;
 pub use record::{read_records, Record};
+
+pub trait Point3D {
+    fn x(&self) -> f64;
+    fn y(&self) -> f64;
+    fn z(&self) -> f64;
+}
+
+pub trait Transform {
+    fn rotation(self) -> Self;
+    fn translation(self) -> Self;
+}
