### 2026-06-09 3ed132c
implement trait for rotation and translation
