### 2026-06-18 ada4a14
refactor: moving input_data to io module since onyl helper struct holding input data
