+++ src/types/native/contour_point.rs
@@ -1,10 +1,5 @@
 use super::{Point3D, Transform};
-use anyhow::{anyhow, Context, Result};
-use csv::ReaderBuilder;
 use serde::Deserialize;
-use std::fs::File;
-use std::io::{BufRead, BufReader};
-use std::path::Path;
 
 impl Point3D for ContourPoint {
     fn x(&self) -> f64 {
@@ -57,34 +52,6 @@ impl Transform for ContourPoint {
     }
 }
 
-/// Utility: detect whether the file uses comma or tab as delimiter.
-pub(crate) fn detect_delimiter<P: AsRef<Path>>(path: P) -> Result<u8> {
-    let file = File::open(&path).with_context(|| {
-        format!(
-            "failed to open file for delimiter sniffing: {:?}",
-            path.as_ref()
-        )
-    })?;
-    let mut reader = BufReader::new(file);
-    let mut first_line = String::new();
-    reader
-        .read_line(&mut first_line)
-        .with_context(|| "failed to read first line for delimiter detection")?;
-
-    // Count occurrences
-    let tabs = first_line.matches('\t').count();
-    let commas = first_line.matches(',').count();
-
-    if tabs > commas {
-        Ok(b'\t')
-    } else if commas > tabs {
-        Ok(b',')
-    } else {
-        // default to comma
-        Ok(b',')
-    }
-}
-
 #[derive(Debug, Deserialize, Clone, Copy, PartialEq)]
 pub struct ContourPoint {
     pub frame_index: u32,
@@ -101,61 +68,6 @@ pub struct ContourPoint {
 }
 
 impl ContourPoint {
-    /// Reads contour points from a CSV file.
-    pub fn read_contour_data<P: AsRef<Path> + std::fmt::Debug + Clone>(
-        path: P,
-    ) -> anyhow::Result<Vec<ContourPoint>> {
-        let delim = detect_delimiter(&path)?;
-        let file = File::open(path)?;
-        let mut rdr = csv::ReaderBuilder::new()
-            .has_headers(false)
-            .delimiter(delim)
-            .from_reader(file);
-
-        let mut points = Vec::new();
-        for result in rdr.records() {
-            match result {
-                Ok(record) => match record.deserialize(None) {
-                    Ok(point) => points.push(point),
-                    Err(e) => eprintln!("Skipping invalid record: {e:?}"),
-                },
-                Err(e) => eprintln!("Skipping invalid row: {e:?}"),
-            }
-        }
-
-        Ok(points)
-    }
-
-    pub fn read_reference_point<P: AsRef<Path>>(path: P) -> Result<ContourPoint> {
-        let delim = detect_delimiter(&path)?;
-        // 1) Open the file, with context on failure
-        let file = File::open(&path)
-            .with_context(|| format!("failed to open reference-point file {:?}", path.as_ref()))?;
-
-        // 2) Build a TSV reader
-        let mut rdr = ReaderBuilder::new()
-            .has_headers(false)
-            .delimiter(delim)
-            .from_reader(file);
-
-        // 3) Grab the first record (if any)…
-        let first = rdr
-            .deserialize()
-            .next()
-            // if there was literally no row at all:
-            .ok_or_else(|| {
-                anyhow!(
-                    "reference-point file {:?} was empty — this data is required",
-                    path.as_ref()
-                )
-            })?;
-
-        // 4) And now propagate any parse / I/O error with its own context:
-        let point: ContourPoint =
-            first.with_context(|| "failed to deserialize first reference-point record")?;
-        Ok(point)
-    }
-
     /// Computes the Euclidean 3-D distance between two contour points.
     pub fn distance_to(&self, other: &ContourPoint) -> f64 {
         let dx = self.x - other.x;
