### 2026-06-18 0b5759b
remove unused geometry mod
