+++ src/intravascular/io/geometry.rs
@@ -1,6 +1,0 @@
-// Re-export all geometry types from the top-level types module.
-// Types now live in crate::types::native but are re-exported here
-// for backward compatibility of internal intravascular imports.
-pub use crate::types::native::contour::{Contour, ContourType};
-pub use crate::types::native::frame::Frame;
-pub use crate::types::native::geometry::Geometry;
