+++ src/intravascular/io/build.rs
@@ -1,5 +1,7 @@
-use super::geometry::{Contour, ContourType, Frame, Geometry};
 use super::input::{ContourPoint, InputData};
+use crate::types::native::contour::{Contour, ContourType};
+use crate::types::native::frame::Frame;
+use crate::types::native::geometry::Geometry;
 use std::collections::{HashMap, HashSet};
 use std::path::Path;
 
