+++ src/intravascular/io.rs
@@ -1,5 +1,4 @@
 pub mod build;
-pub mod geometry;
 pub mod input;
 mod integrity_check;
 pub mod output;
