+++ src/intravascular/binding/entry.rs
@@ -4,7 +4,6 @@ use std::fs::File;
 use std::io::Write;
 use std::path::Path;
 
-use crate::intravascular::io::geometry::{ContourType, Geometry};
 use crate::intravascular::io::input::InputData;
 use crate::intravascular::io::output;
 use crate::intravascular::processing::align_between;
@@ -15,6 +14,8 @@ use crate::intravascular::processing::postprocessing as postprocess;
 use crate::intravascular::processing::preprocessing;
 use crate::intravascular::processing::preprocessing::ProcessingOptions;
 use crate::intravascular::to_object;
+use crate::types::native::contour::ContourType;
+use crate::types::native::geometry::Geometry;
 
 // tolerance of distance between frames [mm], that counts as 0
 const TOLERANCE: f64 = 0.03;
