+++ src/ccta/adjust_mesh/scale_coronary.rs
@@ -437,6 +437,7 @@ mod tests {
                     },
                     tangent: Vector3::new(1.0, 0.0, 0.0),
                     branch_id: 0,
+                    radius: 0.0,
                 },
                 CenterlinePoint {
                     contour_point: ContourPoint {
@@ -449,6 +450,7 @@ mod tests {
                     },
                     tangent: Vector3::new(1.0, 0.0, 0.0),
                     branch_id: 0,
+                    radius: 0.0,
                 },
             ],
             branch_start_indices: vec![0],
@@ -482,6 +484,7 @@ mod tests {
                 },
                 tangent: Vector3::new(1.0, 0.0, 0.0),
                 branch_id: 0,
+                radius: 0.0,
             }],
             branch_start_indices: vec![0],
         };
@@ -533,6 +536,8 @@ mod tests {
     //                     aortic: false,
     //                 },
     //                 tangent: Vector3::new(1.0, 0.0, 0.0).into(),
+    //                 branch_id: 0,
+    //                 radius: 0.0,
     //             },
     //             CenterlinePoint {
     //                 contour_point: ContourPoint {
@@ -544,6 +549,8 @@ mod tests {
     //                     aortic: false,
     //                 },
     //                 tangent: Vector3::new(1.0, 0.0, 0.0).into(),
+    //                 branch_id: 0,
+    //                 radius: 0.0,
     //             },
     //         ],
     //     };
