### 2026-06-18 c15ca43
implement a vtp parser function
