+++ src/intravascular/binding.rs
@@ -4,5 +4,6 @@ pub mod functions;
 
 pub use functions::{
     from_array_doublepair, from_array_full, from_array_single, from_array_singlepair,
-    from_file_doublepair, from_file_full, from_file_single, from_file_singlepair, to_obj,
+    from_file_doublepair, from_file_full, from_file_single, from_file_singlepair,
+    read_centerline_vtp, to_obj,
 };
