+++ src/intravascular/io/input.rs
@@ -256,17 +256,192 @@ pub fn read_records<P: AsRef<Path>>(path: P) -> anyhow::Result<Vec<Record>> {
     Ok(records)
 }
 
-// pub fn read_centerline_vtp<P: AsRef<Path>>(path: P) -> anyhow::Result<Centerline> {
-//     let file = File::open(path)?;
-//     let reader = BufReader::new(file);
-//     let points = Vec::new();
-//     let branch_start_indices = Vec::new();
+pub fn read_centerline_vtp<P: AsRef<Path>>(path: P) -> anyhow::Result<Centerline> {
+    use crate::types::native::centerline_point::CenterlinePoint;
+    use nalgebra::Vector3;
 
-//     for line in reader.deserialize() {
+    const BINARY_PROBE_BYTES: usize = 512;
+    const MIN_TANGENT_NORM: f64 = 1e-12;
 
-//     }
-//     todo!()
-// }
+    let path = path.as_ref();
+
+    let raw = std::fs::read(path).with_context(|| format!("cannot open {:?}", path))?;
+
+    if raw
+        .iter()
+        .take(BINARY_PROBE_BYTES)
+        .any(|&b| b < 0x09 || (b > 0x0d && b < 0x20))
+    {
+        anyhow::bail!(
+            "{:?} appears to be a binary VTP file; only ASCII-format VTP is supported. \
+             Re-export from your software with 'ASCII' data mode.",
+            path
+        );
+    }
+
+    let xml = String::from_utf8(raw).with_context(|| format!("{:?}: not valid UTF-8", path))?;
+
+    for fmt in ["format=\"binary\"", "format=\"appended\""] {
+        if xml.contains(fmt) {
+            anyhow::bail!(
+                "{:?}: binary-encoded DataArrays detected ({}); only ASCII format is \
+                 supported. Re-export with 'ASCII' data mode.",
+                path,
+                fmt
+            );
+        }
+    }
+
+    fn extract_section<'a>(xml: &'a str, tag: &str) -> anyhow::Result<&'a str> {
+        let open = format!("<{}", tag);
+        let close = format!("</{}>", tag);
+        let start = xml
+            .find(&open)
+            .ok_or_else(|| anyhow::anyhow!("VTP: <{}> section not found", tag))?;
+        let rest = &xml[start..];
+        let end_rel = rest
+            .find(&close)
+            .ok_or_else(|| anyhow::anyhow!("VTP: </{}> not found", tag))?;
+        Ok(&rest[..end_rel + close.len()])
+    }
+
+    fn dataarray_text<'a>(section: &'a str, name: &str) -> anyhow::Result<&'a str> {
+        let needle = format!("Name=\"{}\"", name);
+        let pos = section
+            .find(&needle)
+            .ok_or_else(|| anyhow::anyhow!("VTP: DataArray Name=\"{}\" not found", name))?;
+        let before = &section[..pos];
+        let da_start = before
+            .rfind("<DataArray")
+            .ok_or_else(|| anyhow::anyhow!("VTP: no <DataArray before Name=\"{}\"", name))?;
+        let rest = &section[da_start..];
+        let tag_end = rest
+            .find('>')
+            .ok_or_else(|| anyhow::anyhow!("VTP: unclosed <DataArray Name=\"{}\">", name))?;
+        let inner = &rest[tag_end + 1..];
+        let close_pos = inner
+            .find("</DataArray>")
+            .ok_or_else(|| anyhow::anyhow!("VTP: no </DataArray> for Name=\"{}\"", name))?;
+        // <InformationKey> nodes can appear inside the Points array in some exporters.
+        let text = inner[..close_pos].trim();
[…]
