### 2026-06-19 1f2c17a
add prettier print statement for centerline
