+++ src/intravascular/binding.rs
@@ -2,6 +2,7 @@ pub mod align;
 pub mod entry;
 pub mod functions;
 
+pub use align::{align_combined, align_manual, align_three_point};
 pub use functions::{
     from_array_doublepair, from_array_full, from_array_single, from_array_singlepair,
     from_file_doublepair, from_file_full, from_file_single, from_file_singlepair,
