### 2026-06-19 f9400ec
add cleanup function for vtp data (removing duplicates)
