+++ src/types/native/centerline.rs
@@ -294,6 +294,27 @@ impl Centerline {
         path
     }
 
+    /// Mean arc-length spacing between consecutive points of branch 0.
+    ///
+    /// Only intra-branch pairs are considered.  Returns `1.0` if branch 0
+    /// has fewer than two points.
+    pub fn mean_spacing(&self) -> f64 {
+        let end = self
+            .branch_start_indices
+            .get(1)
+            .copied()
+            .unwrap_or(self.points.len());
+        let main = &self.points[..end];
+        if main.len() < 2 {
+            return 1.0;
+        }
+        let sum: f64 = main
+            .windows(2)
+            .map(|w| w[0].contour_point.distance_to(&w[1].contour_point))
+            .sum();
+        sum / (main.len() - 1) as f64
+    }
+
     /// 95th-percentile of consecutive-point spacings — O(n).
     ///
     /// Operates only on adjacent pairs in the original ordering so large
@@ -585,11 +606,112 @@ impl Centerline {
 
         self.rebuild_from_branches(branches);
     }
+
+    /// Remove the run-alongside-main-branch prefix from every side branch and
+    /// optionally strip the inlet region from branch 0.
+    ///
+    /// VTP files export every branch starting from the vessel origin, so side
+    /// branches share a common prefix with branch 0.  For each side branch this
+    /// method trims the contiguous leading prefix whose points all lie within
+    /// one mean inter-point spacing of branch 0 of at least one main-branch
+    /// point.  The last point of the trimmed prefix is kept as the bifurcation
+    /// junction.  Branches whose entire extent lies within that buffer are
+    /// dropped completely.
+    ///
+    /// If `rm_start_mm > 0`, the leading points of branch 0 are also removed
+    /// up to `rm_start_mm` arc-length from its first point.  This is useful
+    /// when the main branch starts at the aortic inlet and the proximal region
+    /// is outside the region of interest.
+    pub fn cleanup_vtp_data(&mut self, rm_start_mm: f64) {
+        if self.branch_start_indices.is_empty() {
+            return;
+        }
+
+        let buffer = self.mean_spacing();
+        let buffer_sq = buffer * buffer;
+
+        let mut branches = self.branches_as_vecs();
+
+        if branches.len() > 1 {
+            let main_pts: Vec<(f64, f64, f64)> = branches[0]
+                .iter()
+                .map(|p| (p.contour_point.x, p.contour_point.y, p.contour_point.z))
+                .collect();
+
+            let close_to_main = |pt: &CenterlinePoint| -> bool {
+                let (x, y, z) = (pt.contour_point.x, pt.contour_point.y, pt.contour_point.z);
+                main_pts.iter().any(|&(mx, my, mz)| {
+                    (x - mx).powi(2) + (y - my).powi(2) + (z - mz).powi(2) <= buffer_sq
+                })
+            };
+
+            for branch in branches.iter_mut().skip(1) {
+                let first_outside = branch.iter().position(|pt| !close_to_main(pt));
+                match first_outside {
+                    None => branch.clear(),
+                    Some(0) => {}
+                    Some(i) => {
+                        branch.drain(..i - 1);
+                    }
+                }
+            }
+
+            branches.retain(|b| !b.is_empty());
+        }
+
+        if rm_start_mm > 0.0 && branches[0].len() > 1 {
+            let mut arc = 0.0;
+            let mut trim_idx = 0;
+            for i in 1..branches[0].len() {
[…]
