+++ src/types/binding/py_centerline.rs
@@ -69,7 +69,7 @@ impl PyCenterline {
         format!(
             "Centerline(len={}, spacing={:.2} mm, branches={:?})",
             self.points.len(),
-            self._spacing(),
+            self.mean_spacing(),
             self.branch_start_indices.len(),
         )
     }
@@ -80,23 +80,6 @@ impl PyCenterline {
 
     fn __len__(&self) -> usize {
         self.points.len()
-    }
-
-    fn _spacing(&self) -> f64 {
-        // if fewer than 2 points there is no spacing
-        if self.points.len() < 2 {
-            return 0.0;
-        }
-
-        // sum distances between consecutive contour points
-        let mut total: f64 = 0.0;
-        let mut count: usize = 0;
-        for pair in self.points.windows(2) {
-            total += pair[0].contour_point.distance(&pair[1].contour_point);
-            count += 1;
-        }
-
-        total / (count as f64)
     }
 
     fn points_as_tuples(&self) -> Vec<(f64, f64, f64)> {
@@ -244,6 +227,33 @@ impl PyCenterline {
         })
     }
 
+    /// Remove the run-alongside-main-branch prefix from every side branch and
+    /// optionally strip the inlet region from branch 0.
+    ///
+    /// VTP files export every branch starting from the vessel origin, so side
+    /// branches share a common prefix with branch 0.  This method trims that
+    /// prefix from each side branch, keeping only the bifurcation junction and
+    /// the diverged portion.  Branches that overlap with branch 0 entirely are
+    /// dropped.  The trim threshold is one mean inter-point spacing of branch 0.
+    ///
+    /// Parameters
+    /// ----------
+    /// rm_start_mm : float, optional
+    ///     Arc-length in mm to remove from the start of branch 0 (the inlet
+    ///     region).  Set to ``0.0`` (default) to leave branch 0 untouched.
+    ///
+    /// Returns
+    /// -------
+    /// PyCenterline
+    ///     New centerline with overlapping prefixes removed from all side
+    ///     branches and the inlet trimmed from branch 0 if requested.
+    #[pyo3(signature = (rm_start_mm = 5.0))]
+    pub fn cleanup_vtp_data(&self, rm_start_mm: f64) -> PyResult<PyCenterline> {
+        let mut cl = self.to_rust_centerline();
+        cl.cleanup_vtp_data(rm_start_mm);
+        Ok(PyCenterline::from(&cl))
+    }
+
     /// Normalise branch ordering so that downstream processing is consistent.
     ///
     /// * **Branch 0** – the point with the highest z-coordinate is moved to
@@ -262,13 +272,16 @@ impl PyCenterline {
     }
 }
 
-// Moved out of pymethods since it's for internal use
 impl PyCenterline {
     pub fn to_rust_centerline(&self) -> Centerline {
         Centerline {
             points: self.points.iter().map(|p| p.into()).collect(),
             branch_start_indices: self.branch_start_indices.clone(),
         }
+    }
+
+    pub(crate) fn mean_spacing(&self) -> f64 {
+        self.to_rust_centerline().mean_spacing()
     }
 }
 
