### 2026-06-19 4d807ec
small refactor
