### 2026-06-19 8f68dfc
fix better handling of index_search_range equals 0
