### 2026-06-19 4c8558a
add smoothing to vtp clean up
