### 2026-07-07 7dbe2d3
update changelog and notebook, add example vtp files
