+++ docs/tutorial_ccta.rst
@@ -81,8 +81,8 @@ To address this limitation, a ray-casting algorithm is employed. A ray is cast f
 *Left:* 3D visualization of all cast rays. *Right:* schematic diagram of the occlusion-detection step.
 
 As a final clean-up, any RCA or LCA vertex whose immediate mesh neighbours carry no
-same-label assignment — an island vertex disconnected from all other coronary-labeled
-vertices — is reclassified as aortic. This adjacency-based elimination ensures that
+same-label assignment - an island vertex disconnected from all other coronary-labeled
+vertices - is reclassified as aortic. This adjacency-based elimination ensures that
 the returned ``rca_points`` and ``lca_points`` sets form topologically connected regions
 on the mesh surface.
 
@@ -109,17 +109,18 @@ on the mesh surface.
 
 The returned ``results`` dictionary contains:
 
-- ``"mesh"`` — the CCTA geometry as a ``trimesh.Trimesh`` object.
-- ``"aorta_points"`` — vertices labeled as aorta: ``[(x, y, z), ...]``.
-- ``"rca_points"`` — vertices labeled as RCA: ``[(x, y, z), ...]``.
-- ``"lca_points"`` — vertices labeled as LCA: ``[(x, y, z), ...]``.
-- ``"rca_removed_points"`` — RCA vertices inside the intramural course that were removed from
+- ``"mesh"`` - the CCTA geometry as a ``trimesh.Trimesh`` object.
+- ``"aorta_points"`` - vertices labeled as aorta: ``[(x, y, z), ...]``.
+- ``"rca_points"`` - vertices labeled as RCA: ``[(x, y, z), ...]``.
+- ``"lca_points"`` - vertices labeled as LCA: ``[(x, y, z), ...]``.
+- ``"rca_removed_points"`` - RCA vertices inside the intramural course that were removed from
   the RCA label: ``[(x, y, z), ...]``.
-- ``"lca_removed_points"`` — same for LCA.
+- ``"lca_removed_points"`` - same for LCA.
 
-The centerline CSV files must contain three columns (no header): ``x``, ``y``, ``z`` in mm.
-They can be converted to :class:`multimodars.PyCenterline` objects with normals using
-:func:`multimodars.numpy_to_centerline`:
+Centerlines can be supplied in two formats.
+
+**CSV files** - three columns (no header): ``x``, ``y``, ``z`` in mm.  Convert to
+:class:`~multimodars.PyCenterline` with :func:`multimodars.numpy_to_centerline`:
 
 .. code-block:: python
 
@@ -133,6 +134,20 @@ They can be converted to :class:`multimodars.PyCenterline` objects with normals 
     lca_cl   = mm.numpy_to_centerline(lca_cl_raw)
     aorta_cl = mm.numpy_to_centerline(aorta_cl_raw)
 
+**VTP files** (recommended) - ASCII VTK PolyData exported by 3D-Slicer or VMTK.
+:func:`multimodars.read_centerline_vtp` parses the file and identifies branches automatically.
+:meth:`~multimodars.PyCenterline.cleanup_vtp_data` then removes the run-alongside-main-branch
+prefix that VTP files attach to every side branch and optionally smooths the result:
+
+.. code-block:: python
+
+    rca_cl   = mm.read_centerline_vtp("data/rca_cl.vtp").cleanup_vtp_data(smooth=True)
+    lca_cl   = mm.read_centerline_vtp("data/lca_cl.vtp").cleanup_vtp_data(smooth=True)
+    aorta_cl = mm.read_centerline_vtp("data/ao_cl.vtp").cleanup_vtp_data(smooth=True)
+
+When using VTP files, pass the :class:`~multimodars.PyCenterline` objects directly to
[…]
