+++ src/intravascular/centerline_align/align_algorithms.rs
@@ -1,10 +1,10 @@
-use crate::intravascular::io::input::ContourPoint;
-use crate::intravascular::io::input::{Centerline, CenterlinePoint};
 use crate::intravascular::processing::align_between::GeometryPair;
 use crate::intravascular::processing::process_utils::hausdorff_distance;
 use crate::types::native;
 use crate::types::native::contour::Contour;
 use crate::types::native::geometry::Geometry;
+use crate::types::native::ContourPoint;
+use crate::types::native::{Centerline, CenterlinePoint};
 use nalgebra::{Point3, Rotation3, Unit, Vector3};
 
 /// Allows alignment algorithms to operate on either a single [`Geometry`] or a [`GeometryPair`].
