+++ src/intravascular/centerline_align/align.rs
@@ -1,8 +1,8 @@
-use crate::intravascular::io::input::{Centerline, ContourPoint};
 use crate::intravascular::processing::align_between::GeometryPair;
 use crate::types::native::contour::{Contour, ContourType};
 use crate::types::native::frame::Frame;
 use crate::types::native::geometry::Geometry;
+use crate::types::native::{Centerline, ContourPoint};
 use anyhow::anyhow;
 use nalgebra::{Point3, Rotation3, Unit, Vector3};
 
