### 2026-07-07 e2a4d48
same for geometrypair
