+++ src/intravascular/binding/entry.rs
@@ -7,7 +7,6 @@ use std::path::Path;
 use crate::intravascular::io::input::InputData;
 use crate::intravascular::io::output;
 use crate::intravascular::processing::align_between;
-use crate::intravascular::processing::align_between::GeometryPair;
 use crate::intravascular::processing::align_within;
 use crate::intravascular::processing::align_within::AlignLog;
 use crate::intravascular::processing::postprocessing as postprocess;
@@ -16,6 +15,7 @@ use crate::intravascular::processing::preprocessing::ProcessingOptions;
 use crate::intravascular::to_object;
 use crate::types::native::contour::ContourType;
 use crate::types::native::geometry::Geometry;
+use crate::types::native::GeometryPair;
 
 // tolerance of distance between frames [mm], that counts as 0
 const TOLERANCE: f64 = 0.03;
