+++ src/intravascular/processing/preprocessing.rs
@@ -204,7 +204,8 @@ pub fn prepare_n_geometries(
 
 #[cfg(test)]
 mod preprocessing_tests {
-    use crate::intravascular::io::input::{ContourPoint, InputData};
+    use crate::intravascular::io::input::InputData;
+    use crate::types::native::ContourPoint;
     use anyhow::Ok;
 
     use super::*;
