+++ src/intravascular/processing/postprocessing.rs
@@ -1,11 +1,11 @@
 use anyhow::Context;
 
-use super::align_between::GeometryPair;
 use super::wall;
 use crate::types::native::contour::{Contour, ContourType};
 use crate::types::native::frame::Frame;
 use crate::types::native::geometry::Geometry;
 use crate::types::native::ContourPoint;
+use crate::types::native::GeometryPair;
 use std::cmp::Ordering;
 use std::collections::HashMap;
 
