### 2026-07-07 26407b4
clean up more input paths
