### 2026-07-07 9b291e4
refactor process_directory function
