### 2026-07-08 84b44e0
move distance_to to point3d and simplify find_reference_cl_point_idx
