### 2026-07-08 31a866e
refactor cleanup_vtp_data
