### 2026-07-08 e28e894
adjust input for new format and updated gitattributes
