+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "multimodars"
-version = "0.4.3"
+version = "0.5.0"
 edition = "2021"
 license = "MIT"
 
