### 2026-07-08 efae8ba
add downloads badge to README
