+++ README.md
@@ -8,6 +8,7 @@
 </p>
 
 [![PyPI](https://img.shields.io/pypi/v/multimodars.svg)](https://pypi.org/project/multimodars)
+[![Downloads](https://static.pepy.tech/badge/multimodars)](https://pepy.tech/project/multimodars)
 [![License](https://img.shields.io/pypi/l/multimodars.svg)](LICENSE)
 [![Docs](https://img.shields.io/readthedocs/multimoda-rs)](https://multimoda-rs.readthedocs.io)
 [![Tests and Build](https://github.com/yungselm/multimoda-rs/actions/workflows/CI.yml/badge.svg)](https://github.com/yungselm/multimoda-rs/actions/workflows/CI.yml)
