### 2026-07-09 a283f51
add step_size to remove_occluded_points_ray_triangle
