### 2026-07-09 d1cfeee
run timing analysis and formulate optimization plan
