+++ ROADMAP.MD
@@ -4,6 +4,84 @@
 - [] Optimize the ray intersection algorithm
 - [] Better parallelization strategy for frames alignment by processing in batches (frames only have to be aligned relative to each other not sequentially)
 
+## `label_geometry` scales badly with vertex count — profiling results (2026-07-09)
+
+**Top-level timings, full `fullworkflow.py` run (189.9s total):**
+
+| Step | Time |
+|---|---|
+| `label_geometry` (final, 57,248 verts) | 95.1s |
+| `label_geometry` (initial, 25,171 verts) | 38.6s |
+| `fix_and_remesh_stitched_mesh` | 17.6-17.8s |
+| `align_combined` | 16.1s |
+| `prepare_centerlines` | 5.3-5.8s |
+| `stitch_ccta_to_intravascular` | 4.5s |
+| `discretize_vessel_tree` | 4.3-4.5s |
+| `label_branches` | 2.4s |
+| `from_file_singlepair` | 1.8s |
+| `find_sharp_angles` | 1.7s |
+| `remove_labeled_points_from_mesh` | 0.5s |
+| `label_anomalous_region` | 0.25s |
+| everything else (scaling/sync steps, exports, etc.) | <0.11s each |
+
+**Internal breakdown of `label_geometry` (initial call → final call):**
+
+| Sub-step | Initial (25,171 verts) | Final (57,248 verts) | Input growth | Time growth |
+|---|---|---|---|---|
+| `find_centerline_bounded_points_simple` (Rust, rca+lca) | 22.4s | 51.1s | ~2.4-2.8x points | 2.3x |
+| `_find_faces_for_points` (Python, via `_prepare_faces_for_rust`) | 3.1s | 19.5s | 2.8x points | 6.2x |
+| `remove_occluded_points_ray_triangle` (Rust) | 4.4s | 11.3s | ~2.8x points | 2.5x |
+| `clean_outlier_points` — lca (Rust) | 0.73s | 3.71s | 2.4x points | 5.1x |
+| `clean_outlier_points` — rca (Rust) | 0.54s | 3.48s | 2.9x points | 6.4x |
+| removed-points diff (`p for p in a if p not in b`, Python) | 0.18s | 1.45s | 2.8x points | 8.0x |
+| `_final_reclassification` (Python, loop over mesh vertices) | 0.57s | 1.36s | 2.3x verts | 2.4x |
+| `plot_results_key` (control_plot rendering) | 5.9s | 2.4s | — | n/a |
+
+Timed a full `fullworkflow.py` run end to end (`examples/fullworkflow copy.py`, timers added
+per top-level call) plus internal timers inside `label_geometry`
+(`multimodars/ccta/labeling.py`). `label_geometry` is called twice —
+once on the raw CCTA mesh (25,171 verts) and once on the remeshed mesh after
+`fix_and_remesh_stitched_mesh` (57,248 verts, ~2.3x more vertices) — and cost jumped from
+38.6s to 95.1s, i.e. worse than linear in vertex count. Breakdown of the internal timers
+(initial call → final call):
+
+- [] `find_centerline_bounded_points_simple` (Rust, rca+lca combined): 22.4s → 51.1s.
+  Largest single cost in both calls; roughly linear with `points_list × centerline points`,
+  but still the top target for algorithmic improvement (spatial index instead of brute rolling
+  sphere check).
+- [] `_find_faces_for_points` (called via `_prepare_faces_for_rust`, pure Python in
+  `labeling.py`): 3.1s → 19.5s — a 6.2x increase for only a 2.8x increase in input points.
+  Root cause: for every occlusion-candidate point it computes `np.linalg.norm(verts - p,
+  axis=1)` over **all** mesh vertices to find the nearest one — O(points_found × n_vertices).
[…]
