+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "multimodars"
-version = "0.5.0"
+version = "0.5.1"
 edition = "2021"
 license = "MIT"
 
@@ -9,7 +9,7 @@ name = "multimodars"
 crate-type = ["cdylib"]
 
 [dependencies]
-pyo3 = "0.28.3"
+pyo3 = "0.29.0"
 csv = "1.4.0"
 serde = { version = "1.0.228", features = ["derive"]}
 image = "0.25.10"
@@ -17,6 +17,6 @@ nalgebra = "0.35.0"
 toml = "1.1.2"
 rayon = "1.12.0"
 crossbeam = "0.8.4"
-anyhow = "1.0.102"
+anyhow = "1.0.103"
 approx = "0.5"
-indicatif = "0.18.4"
+indicatif = "0.18.6"
