### 2026-07-09 6afedf3
use RTree to solve find_centerline_bounded_points
