### 2026-07-09 7261d70
leverage RTree wherever possible
