+++ ROADMAP.MD
@@ -85,6 +85,95 @@ investigated further yet): `fix_and_remesh_stitched_mesh` 17.6-17.8s, `align_com
 `discretize_vessel_tree` 4.3-4.5s, `label_branches` 2.4s, `find_sharp_angles` 1.7s — all far
 below the `label_geometry` cost but candidates if the workflow needs to get faster overall.
 
+## R-tree / Rust-migration audit — `src/ccta` and `multimodars/ccta` (2026-07-09)
+
+Re-profiled `label_geometry` on the 25,171-vertex mesh after the `find_centerline_bounded_points`
+fix above. New bottleneck order:
+
+| Step | Time |
+|---|---|
+| `plot_results_key` | 6.446s |
+| `remove_occluded_points_ray_triangle` (rca) | 4.353s |
+| `_prepare_faces_for_rust` (rca) | 3.184s |
+| `clean_outlier_points` (lca) | 0.728s |
+| `_final_reclassification` | 0.560s |
+| `clean_outlier_points` (rca) | 0.545s |
+| `load mesh` | 0.524s |
+| `find_centerline_bounded_points_simple` (rca+lca) | 0.236s *(was 22.4s)* |
+| `diff rca_removed_points` | 0.177s |
+| `_find_aortic_points` (post/pre-clean) | 0.088s / 0.085s |
+| **TOTAL** | **16.969s** *(was 38.6s on this mesh size before any fixes)* |
+
+Ran a full audit of `src/ccta/**/*.rs` and `multimodars/ccta/**/*.py` for the same
+O(N×M)/O(N²) brute-force pattern that `find_centerline_bounded_points` had, plus a check of
+what pure-Python logic in `multimodars/ccta` should move into Rust. Prioritized by expected
+impact:
+
+### A. Small effort, ready to apply next
+- [] `rca_removed_points` / `lca_removed_points` diff (`labeling.py:219-222, 242-245`,
+  `p for p in a if p not in b`) — `b` is still a plain `list`, so this is still an O(n×m)
+  linear scan. Convert to a `set` before the membership test. Still costing 0.177s at 25k
+  verts; was 1.45s at 57k verts.
+- [] `clean_up_non_section_points` (`src/ccta/adjust_mesh/scale_coronary.rs:357-416`, exposed
+  as `clean_outlier_points`) — for every cleanup point scans all `reference_points` *and* all
+  of `points_to_cleanup` again for self-neighbor counting: O(cleanup×(cleanup+reference)).
+  Same `locate_within_distance` recipe as the already-fixed function. Currently 0.55-0.73s per
+  call, called twice per `label_geometry`.
+- [] `remove_occluded_points_ray_triangle_rust` Pass B (`label_coronary.rs:141-151`) — for
+  every mesh point scans all of `faces_to_exclude` (small, ~150-200) doing point-to-triangle
+  distance: O(points × faces_to_exclude). An R-tree over just the excluded faces, queried per
+  point, is a direct copy of the already-fixed pattern.
+
+### B. Medium effort, high value
+- [] `_find_faces_for_points` / `_prepare_faces_for_rust` (`labeling.py:357-406`) — still pure
+  Python, O(points × n_vertices) via `np.linalg.norm` over all mesh vertices per candidate
+  point. Best Rust-migration candidate in the whole module: mesh vertices/faces already cross
+  the FFI boundary elsewhere, so this can become a Rust binding using an R-tree over mesh
+  vertices, and ideally get fused directly into `remove_occluded_points_ray_triangle_rust` so
+  points don't round-trip through Python just to come back as `Triangle`s. Currently 3.184s per
+  call (was 19.5s at 57k verts).
[…]
