### 2026-07-09 4f26f3a
move find faces to rust
