+++ tests/test_ccta.py
@@ -1,9 +1,10 @@
 """Tests for the multimodars.ccta module.
 
 Covers:
-  - labeling:   _find_aortic_points, _final_reclassification
-  - multimodars.multimodars: find_faces_near_points (Rust binding backing
-                labeling.label_geometry's occlusion-removal face lookup)
+  - multimodars.multimodars: find_faces_near_points, find_aortic_points,
+                final_reclassification (Rust bindings backing
+                labeling.label_geometry's occlusion-removal and
+                adjacency-based label-smoothing steps)
   - fixing_functions: manual_hole_fill, postprocess_stitched_mesh
   - manipulating: remove_labeled_points_from_mesh,
                   keep_labeled_points_from_mesh, sync_results_to_mesh,
@@ -25,11 +26,11 @@ from multimodars.ccta.fixing_functions import (
     manual_hole_fill,
     postprocess_stitched_mesh,
 )
-from multimodars.ccta.labeling import (
-    _find_aortic_points,
-    _final_reclassification,
+from multimodars.multimodars import (
+    find_faces_near_points,
+    find_aortic_points,
+    final_reclassification,
 )
-from multimodars.multimodars import find_faces_near_points
 from multimodars.ccta.manipulating import (
     _clamp_to_plane,
     _enforce_layer_gap_from_plane,
@@ -138,30 +139,34 @@ def grid_results(grid_mesh):
 
 
 # ===========================================================================
-# labeling._find_aortic_points
+# multimodars.multimodars.find_aortic_points
+# (Rust binding that replaced labeling._find_aortic_points)
 # ===========================================================================
 
 
 class TestFindAorticPoints:
     def test_basic_set_difference(self, grid_mesh):
+        verts = [tuple(v) for v in grid_mesh.vertices]
         rca = [(0.0, 0.0, 0.0), (1.0, 0.0, 0.0)]
         lca = [(0.0, 1.0, 0.0)]
-        aortic = _find_aortic_points(grid_mesh.vertices, rca, lca)
+        aortic = find_aortic_points(verts, rca, lca)
         assert len(aortic) == 6  # 9 total − 3 labelled
         assert (0.0, 0.0, 0.0) not in aortic
         assert (0.0, 1.0, 0.0) not in aortic
 
     def test_empty_rca_lca_returns_all(self, grid_mesh):
-        aortic = _find_aortic_points(grid_mesh.vertices, [], [])
-        assert len(aortic) == len(grid_mesh.vertices)
+        verts = [tuple(v) for v in grid_mesh.vertices]
+        aortic = find_aortic_points(verts, [], [])
+        assert len(aortic) == len(verts)
 
     def test_all_labelled_returns_empty(self, grid_mesh):
-        all_verts = [tuple(v) for v in grid_mesh.vertices]
-        aortic = _find_aortic_points(grid_mesh.vertices, all_verts[:5], all_verts[5:])
+        verts = [tuple(v) for v in grid_mesh.vertices]
+        aortic = find_aortic_points(verts, verts[:5], verts[5:])
         assert aortic == []
 
     def test_output_is_list_of_tuples(self, grid_mesh):
-        aortic = _find_aortic_points(grid_mesh.vertices, [], [])
+        verts = [tuple(v) for v in grid_mesh.vertices]
+        aortic = find_aortic_points(verts, [], [])
         assert all(isinstance(p, tuple) for p in aortic)
         assert all(len(p) == 3 for p in aortic)
 
@@ -215,11 +220,29 @@ class TestFindFacesNearPoints:
 
 
 # ===========================================================================
-# labeling._final_reclassification
+# multimodars.multimodars.final_reclassification
+# (Rust binding that replaced labeling._final_reclassification)
 # ===========================================================================
 
 
[…]
