+++ src/lib.rs
@@ -50,6 +50,11 @@ fn multimodars(_py: Python, m: pyo3::prelude::Bound<'_, PyModule>) -> PyResult<(
         ccta_py::find_faces_near_points,
         m.clone()
     )?)?;
+    m.add_function(wrap_pyfunction!(ccta_py::find_aortic_points, m.clone())?)?;
+    m.add_function(wrap_pyfunction!(
+        ccta_py::final_reclassification,
+        m.clone()
+    )?)?;
     m.add_function(wrap_pyfunction!(
         ccta_py::adjust_diameter_centerline_morphing_simple,
         m.clone()
