### 2026-07-09 4aa3a76
replace the slow trimehs version with rust implementation
