+++ multimodars/ccta/manipulating.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+import time
+from contextlib import contextmanager
 from collections.abc import Mapping
 import numpy as np
 import trimesh
@@ -9,12 +11,46 @@ from ..multimodars import (
     PyFrame,
     PyCenterline,
     build_adjacency_map,
+    fix_mesh_winding,
     adjust_diameter_centerline_morphing_simple,
     find_proximal_distal_scaling,
     find_aortic_scaling,
     find_aortic_wall_scaling as _find_aortic_wall_scaling,
 )
 from .._converters import geometry_to_trimesh
+
+
+@contextmanager
+def _timed(label: str, store: list | None = None):
+    start = time.perf_counter()
+    try:
+        yield
+    finally:
+        elapsed = time.perf_counter() - start
+        if store is not None:
+            store.append((label, elapsed))
+        print(f"  [TIMER] {label}: {elapsed:.3f}s")
+
+
+def _print_timing_summary(store: list, title: str):
+    print(f"\n=== {title} (sorted by duration) ===")
+    for label, elapsed in sorted(store, key=lambda x: x[1], reverse=True):
+        print(f"  {elapsed:8.3f}s  {label}")
+    print(f"  {sum(t for _, t in store):8.3f}s  TOTAL")
+
+
+def _fast_fix_normals(mesh: trimesh.Trimesh) -> None:
+    """Drop-in replacement for ``trimesh.Trimesh.fix_normals()``.
+
+    trimesh's own ``fix_winding`` does a Python/NetworkX BFS over the
+    face-adjacency graph with several small numpy allocations per edge -
+    O(n_edges) with heavy per-iteration overhead (e.g. ~3.9s on a ~52k-face
+    mesh). ``fix_mesh_winding`` is a Rust port of the same BFS-consistency
+    algorithm; ``fix_inversion`` (the volume-sign flip check) is already
+    vectorized numpy in trimesh, so it's left as-is.
+    """
+    mesh.faces = np.array(fix_mesh_winding(mesh.faces.tolist()), dtype=mesh.faces.dtype)
+    trimesh.repair.fix_inversion(mesh, multibody=False)
 
 
 def _project_to_best_fit_plane(
@@ -751,40 +787,50 @@ def stitch_ccta_to_intravascular(
     boundary-ring plane and the IV plane form an angle ≥ ``ostium_angle_threshold_deg``
     (default 45°).
     """
-    iv_mesh = iv_mesh.downsample(n_points_iv_cont)
-    iv_mesh_points = [
-        (p.x, p.y, p.z) for frame in iv_mesh.frames for p in frame.lumen.points
-    ]
+    _timings: list = []
+
+    with _timed("downsample iv_mesh", _timings):
+        iv_mesh = iv_mesh.downsample(n_points_iv_cont)
+        iv_mesh_points = [
+            (p.x, p.y, p.z) for frame in iv_mesh.frames for p in frame.lumen.points
+        ]
     proximal_centroid = iv_mesh.frames[0].centroid
     distal_centroid = iv_mesh.frames[-1].centroid
     proximal_points = iv_mesh.frames[0].lumen.points
     distal_points = iv_mesh.frames[-1].lumen.points
 
-    prox_boundary_pts, dist_boundary_pts, mesh = _prepare_prox_dist_boundary_pts(
-        mesh,
-        results,
-        proximal_centroid,
-        distal_centroid,
-        proximal_is_ostium=proximal_is_ostium,
-        proximal_iv_frame_pts=iv_mesh.frames[0].lumen.points,
-        clamp_overshoot=clamp_overshoot,
-    )
+    with _timed("_prepare_prox_dist_boundary_pts (total)", _timings):
[…]
