+++ multimodars/ccta/labeling.py
@@ -1,8 +1,6 @@
 from __future__ import annotations
 
 
-import time
-from contextlib import contextmanager
 from pathlib import Path
 from typing import Any
 import trimesh
@@ -21,25 +19,6 @@ from ..multimodars import (
 from .._converters import numpy_to_centerline
 from ..io.read_geometrical import read_mesh
 from .debug_plots import plot_results_key
-
-
-@contextmanager
-def _timed(label: str, store: list | None = None):
-    start = time.perf_counter()
-    try:
-        yield
-    finally:
-        elapsed = time.perf_counter() - start
-        if store is not None:
-            store.append((label, elapsed))
-        print(f"  [TIMER] {label}: {elapsed:.3f}s")
-
-
-def _print_timing_summary(store: list, title: str = "label_geometry timing summary"):
-    print(f"\n=== {title} (sorted by duration) ===")
-    for label, elapsed in sorted(store, key=lambda x: x[1], reverse=True):
-        print(f"  {elapsed:8.3f}s  {label}")
-    print(f"  {sum(t for _, t in store):8.3f}s  TOTAL")
 
 
 def label_geometry(
@@ -120,86 +99,76 @@ def label_geometry(
         Re-raises any error that occurs while reading the mesh or centerline
         files, after printing a descriptive message.
     """
-    _timings: list = []
-
-    with _timed("load mesh", _timings):
-        if isinstance(path_ccta_geometry, trimesh.Trimesh):
-            mesh = path_ccta_geometry
+    if isinstance(path_ccta_geometry, trimesh.Trimesh):
+        mesh = path_ccta_geometry
+        print(
+            f"Using provided mesh: {len(mesh.vertices)} vertices, {len(mesh.faces)} faces"
+        )
+    else:
+        try:
+            mesh = read_mesh(path_ccta_geometry)
             print(
-                f"Using provided mesh: {len(mesh.vertices)} vertices, {len(mesh.faces)} faces"
+                f"Loaded mesh: {len(mesh.vertices)} vertices, {len(mesh.faces)} faces"
             )
-        else:
-            try:
-                mesh = read_mesh(path_ccta_geometry)
-                print(
-                    f"Loaded mesh: {len(mesh.vertices)} vertices, {len(mesh.faces)} faces"
-                )
-            except Exception as e:
-                print(f"Error reading CCTA mesh from {path_ccta_geometry}: {e}")
-                raise
+        except Exception as e:
+            print(f"Error reading CCTA mesh from {path_ccta_geometry}: {e}")
+            raise
 
-    with _timed("load centerlines", _timings):
-        if isinstance(path_centerline_aorta, PyCenterline):
-            cl_aorta = path_centerline_aorta
-            print(f"Using provided aorta centerline: {len(cl_aorta.points)} points")
-        elif isinstance(path_centerline_aorta, np.ndarray):
-            cl_aorta = numpy_to_centerline(path_centerline_aorta)
-            print(f"Using provided aorta centerline: {len(cl_aorta.points)} points")
-        else:
-            try:
-                cl_aorta_raw = np.genfromtxt(path_centerline_aorta, delimiter=",")
-                cl_aorta = numpy_to_centerline(cl_aorta_raw)
-                print(f"Loaded aorta centerline: {len(cl_aorta.points)} points")
-            except Exception as e:
-                print(
-                    f"Error reading Aorta centerline from {path_centerline_aorta}: {e}"
-                )
-                raise
+    if isinstance(path_centerline_aorta, PyCenterline):
+        cl_aorta = path_centerline_aorta
+        print(f"Using provided aorta centerline: {len(cl_aorta.points)} points")
+    elif isinstance(path_centerline_aorta, np.ndarray):
+        cl_aorta = numpy_to_centerline(path_centerline_aorta)
[…]
