+++ multimodars/ccta/manipulating.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-import time
-from contextlib import contextmanager
 from collections.abc import Mapping
 import numpy as np
 import trimesh
@@ -18,25 +16,6 @@ from ..multimodars import (
     find_aortic_wall_scaling as _find_aortic_wall_scaling,
 )
 from .._converters import geometry_to_trimesh
-
-
-@contextmanager
-def _timed(label: str, store: list | None = None):
-    start = time.perf_counter()
-    try:
-        yield
-    finally:
-        elapsed = time.perf_counter() - start
-        if store is not None:
-            store.append((label, elapsed))
-        print(f"  [TIMER] {label}: {elapsed:.3f}s")
-
-
-def _print_timing_summary(store: list, title: str):
-    print(f"\n=== {title} (sorted by duration) ===")
-    for label, elapsed in sorted(store, key=lambda x: x[1], reverse=True):
-        print(f"  {elapsed:8.3f}s  {label}")
-    print(f"  {sum(t for _, t in store):8.3f}s  TOTAL")
 
 
 def _fast_fix_normals(mesh: trimesh.Trimesh) -> None:
@@ -787,50 +766,40 @@ def stitch_ccta_to_intravascular(
     boundary-ring plane and the IV plane form an angle ≥ ``ostium_angle_threshold_deg``
     (default 45°).
     """
-    _timings: list = []
-
-    with _timed("downsample iv_mesh", _timings):
-        iv_mesh = iv_mesh.downsample(n_points_iv_cont)
-        iv_mesh_points = [
-            (p.x, p.y, p.z) for frame in iv_mesh.frames for p in frame.lumen.points
-        ]
+    iv_mesh = iv_mesh.downsample(n_points_iv_cont)
+    iv_mesh_points = [
+        (p.x, p.y, p.z) for frame in iv_mesh.frames for p in frame.lumen.points
+    ]
     proximal_centroid = iv_mesh.frames[0].centroid
     distal_centroid = iv_mesh.frames[-1].centroid
     proximal_points = iv_mesh.frames[0].lumen.points
     distal_points = iv_mesh.frames[-1].lumen.points
 
-    with _timed("_prepare_prox_dist_boundary_pts (total)", _timings):
-        prox_boundary_pts, dist_boundary_pts, mesh = _prepare_prox_dist_boundary_pts(
-            mesh,
-            results,
-            proximal_centroid,
-            distal_centroid,
-            proximal_is_ostium=proximal_is_ostium,
-            proximal_iv_frame_pts=iv_mesh.frames[0].lumen.points,
-            clamp_overshoot=clamp_overshoot,
-            _timings=_timings,
-        )
+    prox_boundary_pts, dist_boundary_pts, mesh = _prepare_prox_dist_boundary_pts(
+        mesh,
+        results,
+        proximal_centroid,
+        distal_centroid,
+        proximal_is_ostium=proximal_is_ostium,
+        proximal_iv_frame_pts=iv_mesh.frames[0].lumen.points,
+        clamp_overshoot=clamp_overshoot,
+    )
     prox_point_step = len(proximal_points) // len(prox_boundary_pts)
     dist_point_step = len(distal_points) // len(dist_boundary_pts)
 
     # Adjust start point
-    with _timed("rotate/adjust start point", _timings):
-        if prox_start_mode == "highest_z" or dist_start_mode == "highest_z":
-            iv_mesh = iv_mesh.sort_frame_points()
-            proximal_points = iv_mesh.frames[0].lumen.points
-            distal_points = iv_mesh.frames[-1].lumen.points
-        if prox_start_mode == "highest_z":
[…]
