+++ ROADMAP.MD
@@ -4,37 +4,43 @@
 
 Performance pass on `label_geometry` / `stitch_ccta_to_intravascular` (2026-07-09): full
 pipeline dropped from ~190s to ~24s (release build). Full before/after numbers and root causes
-are in the CHANGELOG. Remaining candidates, deferred because they weren't real bottlenecks once
-the release-build baseline was corrected:
+are in the CHANGELOG. Considered this pass **closed** after two follow-up investigations both
+came back negative-or-marginal — see below. A first round of candidates (`voronoi_partition`,
+the `scale_coronary.rs` sweep functions, and the ray-triangle intersection pass) were dropped
+earlier the same way, after re-profiling under a release build showed they were never real
+bottlenecks.
 
-- [] `voronoi_partition` (`src/ccta/discretizing/projecting.rs`) — full linear scan of `anchors`
-  per mesh point; same R-tree-over-the-small-side recipe as the fixes already applied elsewhere.
-- [] `find_closest_centerline_point_optimized`, `find_cl_points_in_range`, `find_region_points`,
-  `symmetric_nn_distance` (`src/ccta/adjust_mesh/scale_coronary.rs`) — rescan a *fixed*
-  centerline/reference set from scratch on every step of a ~40-80 iteration scaling sweep;
-  building the R-tree once outside the loop would compound. Likely the highest-leverage fix left
-  per line of code changed, if these ever show up as hot in a real profile.
-- [] `fix_mesh_winding` (the Rust BFS winding-consistency fix added this pass) could also replace
-  `mesh_remeshed.fix_normals()` inside `fix_and_remesh_stitched_mesh`
-  (`multimodars/ccta/fixing_functions.py`) — same trimesh bottleneck, on an even bigger mesh
-  (114k faces post-remesh). Deliberately not done — out of scope for that pass.
-- [] `plot_results_key` (control-plot rendering) is `label_geometry`'s biggest line item once
-  everything else is fixed, but it's rendering, not spatial search — separate investigation
-  (e.g. downsampling points before scatter plotting).
+- [x] ~~`fix_mesh_winding` reuse in `fix_and_remesh_stitched_mesh`~~ — **INVESTIGATED, NOT A
+  BOTTLENECK (2026-07-09):** timed both `mesh_remeshed.fix_normals()` calls in
+  `fixing_functions.py` against the real dataset — 0.161s out of 18.5s total, negligible.
+  Reason: `trimesh.repair.fix_winding` early-exits when `mesh.is_winding_consistent` is already
+  `True`, and pymeshlab's isotropic remeshing produces globally consistent winding by
+  construction — unlike `stitch_ccta_to_intravascular`'s case, where concatenating several
+  separately-oriented mesh pieces genuinely breaks consistency and triggers the expensive BFS.
+  `fix_and_remesh_stitched_mesh`'s cost is confirmed to be pymeshlab's own C++ work, not a
+  hidden Python bottleneck — the original call to leave it alone was correct.
+- [x] ~~`remove_labeled_points_from_mesh` / `keep_labeled_points_from_mesh`~~ — **INVESTIGATED,
+  NOT WORTH IT (2026-07-09):** confirmed real (same shape of pure-Python full-mesh adjacency
+  loop as `_final_reclassification`) but only 0.211s of ~24s total (<1%), and
+  `keep_labeled_points_from_mesh` isn't even called by the real workflow. Deliberately left
+  alone — diminishing returns past this point in the pipeline.
+- [] `plot_results_key` (control-plot rendering) was `label_geometry`'s biggest line item in the
+  audit table, but that number is from a debug build and `control_plot=False` has been used in
+  every release-build measurement since — its real cost under release is unverified. Lowest
[…]
