### 2026-07-09 dc6d053
use par iter for discretize vessel tre
