### 2026-07-09 df74864
improve anomalous region labeling
