### 2026-07-10 d17ed2a
use nalgebra fro scale_coronary as well
