+++ src/ccta/adjust_mesh/scale_coronary.rs
@@ -1,4 +1,5 @@
 use crate::types::native::{Centerline, CenterlinePoint, Frame};
+use nalgebra::Point3;
 use rayon::prelude::*;
 use std::collections::HashSet;
 
@@ -13,7 +14,6 @@ pub fn centerline_based_wall_diameter_optimization(
         return 0.0;
     }
 
-    // find closest centerline point
     let closest_cl = match centerline.points.iter().min_by(|a, b| {
         let dist_a = super::calculate_squared_distance(*a, ref_point_coronary);
         let dist_b = super::calculate_squared_distance(*b, ref_point_coronary);
@@ -36,31 +36,28 @@ pub fn centerline_based_wall_diameter_optimization(
         None => return 0.0,
     };
 
-    // vector centerline point to coronary point
-    let vector = (
-        ref_point_coronary.0 - closest_cl.contour_point.x,
-        ref_point_coronary.1 - closest_cl.contour_point.y,
-        ref_point_coronary.2 - closest_cl.contour_point.z,
+    let ref_point = Point3::new(
+        ref_point_coronary.0,
+        ref_point_coronary.1,
+        ref_point_coronary.2,
     );
+    let closest_cl_point = Point3::new(
+        closest_cl.contour_point.x,
+        closest_cl.contour_point.y,
+        closest_cl.contour_point.z,
+    );
+    let closest_aortic_point = Point3::new(closest_aortic.0, closest_aortic.1, closest_aortic.2);
 
-    let magnitude = (vector.0 * vector.0 + vector.1 * vector.1 + vector.2 * vector.2).sqrt();
-    if magnitude == 0.0 {
+    // vector centerline point to coronary point
+    let vector = ref_point - closest_cl_point;
+    let Some(unit) = vector.try_normalize(0.0) else {
         return 0.0;
-    }
-    let unit = (
-        vector.0 / magnitude,
-        vector.1 / magnitude,
-        vector.2 / magnitude,
-    );
+    };
 
     // project (ref_point_coronary - closest_aortic) onto unit direction
     // minimizes ||closest_aortic + t*unit - ref_point_coronary||², constrained t >= 0
-    let to_ref = (
-        ref_point_coronary.0 - closest_aortic.0,
-        ref_point_coronary.1 - closest_aortic.1,
-        ref_point_coronary.2 - closest_aortic.2,
-    );
-    let t = to_ref.0 * unit.0 + to_ref.1 * unit.1 + to_ref.2 * unit.2;
+    let to_ref = ref_point - closest_aortic_point;
+    let t = to_ref.dot(&unit);
 
     t.max(0.0)
 }
@@ -227,29 +224,19 @@ pub fn centerline_based_diameter_morphing(
         .par_iter()
         .map(|point| {
             let closest_cl_point = find_closest_centerline_point_optimized(centerline, *point);
-
-            let vector = (
-                point.0 - closest_cl_point.contour_point.x,
-                point.1 - closest_cl_point.contour_point.y,
-                point.2 - closest_cl_point.contour_point.z,
+            let p = Point3::new(point.0, point.1, point.2);
+            let cl_point = Point3::new(
+                closest_cl_point.contour_point.x,
+                closest_cl_point.contour_point.y,
+                closest_cl_point.contour_point.z,
             );
 
-            let magnitude =
-                (vector.0 * vector.0 + vector.1 * vector.1 + vector.2 * vector.2).sqrt();
-
-            if magnitude > 0.0 {
-                let normalized_vector = (
-                    vector.0 / magnitude,
-                    vector.1 / magnitude,
-                    vector.2 / magnitude,
-                );
-                (
-                    point.0 + normalized_vector.0 * diameter_adjustment_mm,
-                    point.1 + normalized_vector.1 * diameter_adjustment_mm,
[…]
