+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "multimodars"
-version = "0.5.1"
+version = "0.5.2"
 edition = "2021"
 license = "MIT"
 
@@ -20,4 +20,4 @@ crossbeam = "0.8.4"
 anyhow = "1.0.103"
 approx = "0.5"
 indicatif = "0.18.6"
-rstar = "0.12"
+rstar = "0.13.0"
