### 2026-07-31 c2d38d1
first test for radius from centerline
