+++ examples/stitching.py
@@ -30,7 +30,8 @@ results, (rca_cl, lca_cl, ao_cl) = mm.label_geometry(
     path_centerline_aorta="../data/centerline_aorta.csv",
     path_centerline_rca="../data/centerline_rca_short.csv",
     path_centerline_lca="../data/centerline_lca.csv",
-    bounding_sphere_radius_mm=3.0,
+    bounding_sphere_radius_mm_rca=3.0,
+    bounding_sphere_radius_mm_lca=3.0,
     n_points_intramural=100,
     anomalous_rca=True,
     anomalous_lca=False,
@@ -164,7 +165,8 @@ results, (rca_cl, lca_cl, ao_cl) = mm.label_geometry(
     path_centerline_aorta="../data/centerline_aorta.csv",
     path_centerline_rca="../data/centerline_rca_short.csv",
     path_centerline_lca="../data/centerline_lca.csv",
-    bounding_sphere_radius_mm=3.0,
+    bounding_sphere_radius_mm_rca=3.0,
+    bounding_sphere_radius_mm_lca=3.0,
     n_points_intramural=100,
     anomalous_rca=True,
     anomalous_lca=False,
