### 2026-07-31 a270f09
add hardcoded maximum for the radius
