### 2026-07-31 73113e6
remove the auto radius from all functions
