+++ src/ccta/adjust_mesh/label_coronary.rs
@@ -192,115 +192,40 @@ pub fn remove_occluded_points_ray_triangle_rust(
     filtered_points
 }
 
-/// Buffer applied on top of each centerline point's own radius in `auto` mode, to
-/// account for the mesh surface sitting slightly outside the maximum inscribed
-/// sphere radius reported by the centerline source (e.g. VTP's
-/// `MaximumInscribedSphereRadius`).
-const AUTO_RADIUS_BUFFER: f64 = 1.05;
-
-/// Upper bound on a coronary vessel radius (mm). Centerlines passed into `auto`
-/// mode may carry aortic points mixed in with coronary ones; aortic radii are far
-/// larger than any real coronary radius, so points above this cutoff are treated
-/// as aorta and excluded from the auto sphere sizing (otherwise a handful of
-/// aortic points would balloon `max_threshold_sq` and get used as bounding-sphere
-/// centers themselves).
-const MAX_CORONARY_RADIUS_MM: f64 = 5.0;
-
-/// Find mesh points that fall within a bounding sphere of any centerline point.
-///
-/// When `auto` is `false`, every centerline point uses the single fixed `radius`
-/// (mm), matching the previous behaviour.
-///
-/// When `auto` is `true`, `radius` is ignored and each centerline point instead
-/// uses its own `CenterlinePoint::radius` (local vessel radius) with a 5% buffer.
-/// This requires every point on `centerline` to carry a real (> 0) radius — e.g.
-/// centerlines loaded via `read_centerline_vtp`, which populates radius from the
-/// VTP's `MaximumInscribedSphereRadius` array. Centerlines built from bare (x, y, z)
-/// data (e.g. `numpy_to_centerline`) always have radius 0.0 and are rejected with
-/// an error rather than silently producing zero-radius (i.e. empty) spheres.
+/// Find mesh points that fall within a bounding sphere of any centerline point,
+/// using a single fixed `radius` (mm) for every centerline point.
 pub fn find_centerline_bounded_points(
     centerline: Centerline,
     points: &[(f64, f64, f64)],
     radius: f64,
-    auto: bool,
 ) -> Result<Vec<(f64, f64, f64)>, String> {
     let checked_centerline = check_centerline(&centerline);
     if points.is_empty() || checked_centerline.points.is_empty() {
-        return Ok(Vec::new());
-    }
-
-    if auto && checked_centerline.points.iter().any(|p| p.radius <= 0.0) {
         return Err(
-            "auto radius requested but centerline has one or more points with radius <= 0.0; \
-             load a centerline that carries real per-point radii (e.g. via read_centerline_vtp) \
-             or pass auto=false with a fixed radius"
-                .to_string(),
+            "find_centerline_bounded_points failed because `Centerline` is empty".to_string(),
         );
     }
 
-    // R-tree over the centerline points (the small side, typically ~1000 points),
-    // tagged with each point's own radius so `auto` mode can size its bounding
-    // sphere per centerline point instead of using one fixed radius for the whole
-    // vessel. Each mesh point then costs O(log M) instead of a scan over all M
-    // centerline points, so the whole query is O(N log M) instead of O(N * M).
-    //
-    // In `auto` mode, points above `MAX_CORONARY_RADIUS_MM` are assumed to be
-    // aortic (mixed into the same centerline) and dropped entirely so they can't
-    // inflate the sphere sizing or serve as bounding-sphere centers themselves.
-    let cl_coords: Vec<rstar::primitives::GeomWithData<[f64; 3], f64>> = checked_centerline
+    // R-tree over the centerline points (the small side, typically ~1000 points).
+    // Each mesh point then costs O(log M) instead of a scan over all M centerline
[…]
