+++ multimodars/ccta/__init__.py
@@ -27,7 +27,6 @@ def label(
     anomalous_lca: bool = False,
     n_points_intramural: int = 120,
     step_size_mm: float = 1.0,
-    auto_radius_vtp: bool = False,
     bounding_sphere_radius_mm_rca: float = 3.0,
     bounding_sphere_radius_mm_lca: float = 3.0,
     tolerance_float: float = 1e-6,
@@ -75,10 +74,6 @@ def label(
     step_size_mm : float, optional
         Step size in mm for iterating over coronary centerline points during
         occlusion removal.  Default is ``1.0`` mm.
-    auto_radius_vtp : bool, optional
-        Determine the rolling sphere radius directly from the ``.vtp`` file
-        instead of using the fixed radii below.  If not loaded from ``.vtp``
-        this raises an error.  Default is ``False``.
     bounding_sphere_radius_mm_rca : float, optional
         Radius in millimetres of the rolling sphere used to collect candidate
         mesh vertices around each centerline point of the RCA.  Default is
@@ -124,7 +119,6 @@ def label(
         anomalous_lca,
         n_points_intramural,
         step_size_mm,
-        auto_radius_vtp,
         bounding_sphere_radius_mm_rca,
         bounding_sphere_radius_mm_lca,
         tolerance_float,
