+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "multimodars"
-version = "0.5.2"
+version = "0.5.3"
 edition = "2021"
 license = "MIT"
 
@@ -11,13 +11,13 @@ crate-type = ["cdylib"]
 [dependencies]
 pyo3 = "0.29.0"
 csv = "1.4.0"
-serde = { version = "1.0.228", features = ["derive"]}
+serde = { version = "1.0.229", features = ["derive"]}
 image = "0.25.10"
 nalgebra = "0.35.0"
-toml = "1.1.2"
+toml = "1.1.4"
 rayon = "1.12.0"
 crossbeam = "0.8.4"
-anyhow = "1.0.103"
+anyhow = "1.0.104"
 approx = "0.5"
 indicatif = "0.18.6"
 rstar = "0.13.0"
