### 2026-08-03 2c312a2
hotfix discretize_vessel_tree
