+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "multimodars"
-version = "0.5.3"
+version = "0.5.4"
 edition = "2021"
 license = "MIT"
 
@@ -9,7 +9,7 @@ name = "multimodars"
 crate-type = ["cdylib"]
 
 [dependencies]
-pyo3 = "0.29.0"
+pyo3 = "0.29.1"
 csv = "1.4.0"
 serde = { version = "1.0.229", features = ["derive"]}
 image = "0.25.10"
