### 2026-08-03 1b422bd
specify that degree for align_manual
