### 2026-08-03 5f7974a
let all alignment functions return their total rotation
