+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "multimodars"
-version = "0.5.4"
+version = "0.5.5"
 edition = "2021"
 license = "MIT"
 

[…]
