### 2026-08-05 4f558a9
add seperate lca and rca removed parameter, fix stitcht boundary ring out of bound
