+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "multimodars"
-version = "0.5.6"
+version = "0.5.7"
 edition = "2021"
 license = "MIT"
 

[…]
