### 2026-08-06 22a2025
small refactor
