+++ multimodars/ccta/labeling.py
@@ -127,50 +127,11 @@ def label_geometry(
             print(f"Error reading CCTA mesh from {path_ccta_geometry}: {e}")
             raise
 
-    if isinstance(path_centerline_aorta, PyCenterline):
-        cl_aorta = path_centerline_aorta
-        print(f"Using provided aorta centerline: {len(cl_aorta.points)} points")
-    elif isinstance(path_centerline_aorta, np.ndarray):
-        cl_aorta = numpy_to_centerline(path_centerline_aorta)
-        print(f"Using provided aorta centerline: {len(cl_aorta.points)} points")
-    else:
-        try:
-            cl_aorta_raw = np.genfromtxt(path_centerline_aorta, delimiter=",")
-            cl_aorta = numpy_to_centerline(cl_aorta_raw)
-            print(f"Loaded aorta centerline: {len(cl_aorta.points)} points")
-        except Exception as e:
-            print(f"Error reading Aorta centerline from {path_centerline_aorta}: {e}")
-            raise
+    cl_aorta = _try_load_cl(path_centerline_aorta, "Aorta")
 
-    if isinstance(path_centerline_lca, PyCenterline):
-        cl_lca = path_centerline_lca
-        print(f"Using provided LCA centerline: {len(cl_lca.points)} points")
-    elif isinstance(path_centerline_lca, np.ndarray):
-        cl_lca = numpy_to_centerline(path_centerline_lca)
-        print(f"Using provided LCA centerline: {len(cl_lca.points)} points")
-    else:
-        try:
-            cl_lca_raw = np.genfromtxt(path_centerline_lca, delimiter=",")
-            cl_lca = numpy_to_centerline(cl_lca_raw)
-            print(f"Loaded LCA centerline: {len(cl_lca.points)} points")
-        except Exception as e:
-            print(f"Error reading LCA centerline from {path_centerline_lca}: {e}")
-            raise
+    cl_lca = _try_load_cl(path_centerline_lca, "LCA")
 
-    if isinstance(path_centerline_rca, PyCenterline):
-        cl_rca = path_centerline_rca
-        print(f"Using provided RCA centerline: {len(cl_rca.points)} points")
-    elif isinstance(path_centerline_rca, np.ndarray):
-        cl_rca = numpy_to_centerline(path_centerline_rca)
-        print(f"Using provided RCA centerline: {len(cl_rca.points)} points")
-    else:
-        try:
-            cl_rca_raw = np.genfromtxt(path_centerline_rca, delimiter=",")
-            cl_rca = numpy_to_centerline(cl_rca_raw)
-            print(f"Loaded RCA centerline: {len(cl_rca.points)} points")
-        except Exception as e:
-            print(f"Error reading RCA centerline from {path_centerline_rca}: {e}")
-            raise
+    cl_rca = _try_load_cl(path_centerline_rca, "RCA")
 
     points_list = [tuple(vertex) for vertex in mesh.vertices.tolist()]
     mesh_faces_list = mesh.faces.tolist()
@@ -194,46 +155,32 @@ def label_geometry(
     lca_removed_points = []
 
     if acute_takeoff_rca:
-        print("Applying occlusion removal for acute-takeoff RCA...")
-        rca_faces_for_rust = find_faces_near_points(
-            points_list, mesh_faces_list, rca_points_found, tolerance_float
+        rca_removed_points, final_rca_points_found = _apply_occlusion_removal(
+            n_points_takeoff_rca,
+            step_size_mm,
+            tolerance_float,
+            cl_aorta,
+            cl_rca,
+            points_list,
+            mesh_faces_list,
+            rca_points_found,
+            "RCA",
         )
-        # Rust implementation, that creates ray between aortic and coronary centerline, and
[…]
