### 2026-08-06 431721b
fix the label_geometry function cleanup
