+++ src/ccta/adjust_mesh/label_coronary.rs
@@ -287,7 +287,7 @@ pub fn find_faces_near_points(
 /// copies of the originating `vertices` entries (no arithmetic in between), so an
 /// exact bit-pattern key is the correct tool here — unlike a radius/nearest-neighbor
 /// query, this is a plain exact-membership test, so no spatial index is needed.
-fn bits_key(p: &(f64, f64, f64)) -> (u64, u64, u64) {
+pub(crate) fn bits_key(p: &(f64, f64, f64)) -> (u64, u64, u64) {
     (p.0.to_bits(), p.1.to_bits(), p.2.to_bits())
 }
 
@@ -384,26 +384,23 @@ pub fn final_reclassification(
         let current_label = labels[i];
         let neighbor_labels: Vec<u8> = neighbors.iter().map(|&n| labels[n]).collect();
 
+        // LOGIC A: isolated RCA/LCA -> aorta
         match current_label {
-            // LOGIC A: isolated RCA/LCA -> aorta
             1 if !neighbor_labels.contains(&1) => new_labels[i] = 0,
             2 if !neighbor_labels.contains(&2) => new_labels[i] = 0,
-            // LOGIC B: removed RCA/LCA point with >70% same-label neighbours -> restored
-            3 => {
-                let rca_neighbors = neighbor_labels.iter().filter(|&&l| l == 1).count();
-                if (rca_neighbors as f64) > (neighbors.len() as f64 * 0.7) {
-                    new_labels[i] = 1;
-                }
-            }
-            4 => {
-                let lca_neighbors = neighbor_labels.iter().filter(|&&l| l == 2).count();
-                if (lca_neighbors as f64) > (neighbors.len() as f64 * 0.7) {
-                    new_labels[i] = 2;
-                }
-            }
             _ => {}
         }
     }
+
+    // LOGIC B: a removed RCA/LCA point is restored if its connected-component's
+    // external boundary is >70% the corresponding coronary label. Evaluated per
+    // component rather than per vertex, so a whole falsely-removed contiguous
+    // patch (whose members mostly neighbour each other, not real RCA/LCA
+    // vertices) is judged against its true external boundary instead of being
+    // masked by its own interior. A single-vertex component reduces to the
+    // original per-vertex check.
+    restore_removed_components(&adjacency, &labels, &mut new_labels, 3, 1);
+    restore_removed_components(&adjacency, &labels, &mut new_labels, 4, 2);
 
     let mut result = ReclassifiedLabels {
         aorta_points: Vec::new(),
@@ -423,6 +420,96 @@ pub fn final_reclassification(
         }
     }
     result
+}
+
+/// Connected components of `subset`, restricted to mesh adjacency: neighbour
+/// traversal only follows vertices that are themselves in `subset` (islands
+/// within the subset stay separate components). Shared by
+/// [`restore_removed_components`] and the `keep_largest_connected_component`
+/// PyO3 binding.
+pub(crate) fn connected_components(
+    adjacency: &HashMap<usize, HashSet<usize>>,
+    subset: &HashSet<usize>,
+) -> Vec<HashSet<usize>> {
+    let mut remaining: HashSet<usize> = subset.clone();
+    let mut components = Vec::new();
+
+    while let Some(&start) = remaining.iter().next() {
+        let mut stack = vec![start];
+        let mut component = HashSet::new();
+        while let Some(i) = stack.pop() {
+            if !component.insert(i) {
+                continue;
+            }
+            if let Some(neighbors) = adjacency.get(&i) {
+                for &n in neighbors {
+                    if remaining.contains(&n) && !component.contains(&n) {
+                        stack.push(n);
+                    }
+                }
+            }
+        }
[…]
