+++ multimodars/ccta/labeling.py
@@ -14,7 +14,7 @@ from ..multimodars import (
     remove_occluded_points_ray_triangle,
     clean_outlier_points,
     find_points_by_cl_region,
-    build_adjacency_map,
+    keep_largest_connected_component,
     PyCenterline,
 )
 from .._converters import numpy_to_centerline
@@ -306,66 +306,6 @@ def _apply_occlusion_removal(
     return rca_removed_points, final_rca_points_found
 
 
-def _keep_largest_connected_component(
-    mesh: trimesh.Trimesh, points: list[tuple[float, float, float]]
-) -> list[tuple[float, float, float]]:
-    """Keep only the largest mesh-connected component of *points*.
-
-    ``find_points_by_cl_region`` classifies points using coordinate-only
-    heuristics (nearest 3-D centerline point, axis-aligned proximal/distal
-    split) with no notion of mesh topology.  That can leave a handful of
-    points assigned to a region despite not being mesh-connected to its main
-    cluster ("islands") - e.g. a point geometrically close to the anomalous
-    segment but on a different, unconnected part of the vessel surface.
-
-    This restricts the mesh's face-adjacency graph to *points* and keeps
-    only the single largest connected component; the rest are dropped
-    (callers typically let dropped points fall through to a different
-    region via a complement/set-difference step, mirroring how
-    :func:`final_reclassification` reassigns isolated vertices to aorta).
-    """
-    if len(points) < 2:
-        return points
-
-    coord_to_idx = {tuple(v): i for i, v in enumerate(mesh.vertices)}
-    point_indices = set()
-    for pt in points:
-        idx = coord_to_idx.get(tuple(pt))
-        if idx is not None:
-            point_indices.add(idx)
-    if not point_indices:
-        return points
-
-    adj_map = build_adjacency_map(mesh.faces.tolist())
-
-    remaining = set(point_indices)
-    components: list[set[int]] = []
-    while remaining:
-        start = next(iter(remaining))
-        stack = [start]
-        component: set[int] = set()
-        while stack:
-            i = stack.pop()
-            if i in component:
-                continue
-            component.add(i)
-            for neighbor in adj_map.get(i, []):
-                if neighbor in remaining and neighbor not in component:
-                    stack.append(neighbor)
-        components.append(component)
-        remaining -= component
-
-    largest = max(components, key=len)
-    if len(components) > 1:
-        dropped = len(point_indices) - len(largest)
-        print(
-            f"  _keep_largest_connected_component: kept {len(largest)}/"
-            f"{len(point_indices)} points ({len(components) - 1} island "
-            f"component(s), {dropped} point(s) dropped)"
-        )
-    return [tuple(mesh.vertices[i]) for i in largest]
-
-
 def label_anomalous_region(
     centerline,
     frames,
@@ -473,6 +413,29 @@ def label_anomalous_region(
         )
 
     return results
+
+
+def _keep_largest_connected_component(
+    mesh: trimesh.Trimesh, points: list[tuple[float, float, float]]
+) -> list[tuple[float, float, float]]:
+    """Keep only the largest mesh-connected component of *points*.
+
+    ``find_points_by_cl_region`` classifies points using coordinate-only
+    heuristics (nearest 3-D centerline point, axis-aligned proximal/distal
+    split) with no notion of mesh topology.  That can leave a handful of
+    points assigned to a region despite not being mesh-connected to its main
+    cluster ("islands") - e.g. a point geometrically close to the anomalous
+    segment but on a different, unconnected part of the vessel surface.
+
+    This restricts the mesh's face-adjacency graph to *points* and keeps
+    only the single largest connected component; the rest are dropped
[…]

[…]
