### 2026-08-06 9528f9b
fix the true bug, where results is not correctly updated
