### 2026-08-06 f7be3fb
ensure that labeling cleanup checks for connectivity
