+++ src/ccta/adjust_mesh/label_coronary.rs
@@ -373,24 +373,16 @@ pub fn final_reclassification(
     let adjacency = build_adjacency_map(faces.to_vec());
 
     let mut new_labels = labels.clone();
-    for i in 0..n_vertices {
-        let Some(neighbors) = adjacency.get(&i) else {
-            continue;
-        };
-        if neighbors.is_empty() {
-            continue;
-        }
 
-        let current_label = labels[i];
-        let neighbor_labels: Vec<u8> = neighbors.iter().map(|&n| labels[n]).collect();
-
-        // LOGIC A: isolated RCA/LCA -> aorta
-        match current_label {
-            1 if !neighbor_labels.contains(&1) => new_labels[i] = 0,
-            2 if !neighbor_labels.contains(&2) => new_labels[i] = 0,
-            _ => {}
-        }
-    }
+    // LOGIC A: minority components of a label - excluding the single largest,
+    // presumed the legitimate main body - are reclassified to a neighbouring
+    // label when their external boundary is >70% that label. Generalizes the
+    // original per-vertex "isolated same-label neighbour" check to whole
+    // mesh-connected islands (a single-vertex island is the size-1 special
+    // case), symmetric to Logic B below.
+    reclassify_minority_components(&adjacency, &labels, &mut new_labels, 0, &[1, 2]); // aorta islands -> RCA/LCA
+    reclassify_minority_components(&adjacency, &labels, &mut new_labels, 1, &[0]); // RCA islands -> aorta
+    reclassify_minority_components(&adjacency, &labels, &mut new_labels, 2, &[0]); // LCA islands -> aorta
 
     // LOGIC B: a removed RCA/LCA point is restored if its connected-component's
     // external boundary is >70% the corresponding coronary label. Evaluated per
@@ -473,6 +465,58 @@ fn component_boundary(
         }
     }
     boundary
+}
+
+/// For every connected component of `subject_label` vertices except the
+/// single largest (presumed the legitimate main body of that label - always
+/// excluded, even when it is the only component, since a real anatomical
+/// region reliably forms one big connected body that must never be
+/// reclassified wholesale just because its aggregate boundary happens to
+/// lean toward one neighbouring label), checks whether the component's
+/// external boundary is >70% one of `target_labels` (checked in order; the
+/// first to clear the threshold wins) and, if so, reclassifies the whole
+/// component to that label.
+fn reclassify_minority_components(
+    adjacency: &HashMap<usize, HashSet<usize>>,
+    labels: &[u8],
+    new_labels: &mut [u8],
+    subject_label: u8,
+    target_labels: &[u8],
+) {
+    let subset: HashSet<usize> = labels
+        .iter()
+        .enumerate()
+        .filter(|&(_, &l)| l == subject_label)
+        .map(|(i, _)| i)
+        .collect();
+    if subset.is_empty() {
+        return;
+    }
+
+    let mut components = connected_components(adjacency, &subset);
+    let largest_idx = components
+        .iter()
+        .enumerate()
+        .max_by_key(|(_, c)| c.len())
+        .map(|(idx, _)| idx)
+        .expect("subset is non-empty, so at least one component exists");
+    components.remove(largest_idx);
+
+    for component in components {
+        let boundary = component_boundary(adjacency, &component);
+        if boundary.is_empty() {
+            continue;
+        }
+        for &target in target_labels {
+            let target_count = boundary.iter().filter(|&&n| labels[n] == target).count();
+            if (target_count as f64) > (boundary.len() as f64 * 0.7) {
+                for &i in &component {
[…]
