+++ src/ccta/adjust_mesh/label_coronary.rs
@@ -384,15 +384,18 @@ pub fn final_reclassification(
     reclassify_minority_components(&adjacency, &labels, &mut new_labels, 1, &[0]); // RCA islands -> aorta
     reclassify_minority_components(&adjacency, &labels, &mut new_labels, 2, &[0]); // LCA islands -> aorta
 
-    // LOGIC B: a removed RCA/LCA point is restored if its connected-component's
-    // external boundary is >70% the corresponding coronary label. Evaluated per
-    // component rather than per vertex, so a whole falsely-removed contiguous
-    // patch (whose members mostly neighbour each other, not real RCA/LCA
-    // vertices) is judged against its true external boundary instead of being
-    // masked by its own interior. A single-vertex component reduces to the
-    // original per-vertex check.
-    restore_removed_components(&adjacency, &labels, &mut new_labels, 3, 1);
-    restore_removed_components(&adjacency, &labels, &mut new_labels, 4, 2);
+    // LOGIC B: a removed RCA/LCA point is restored via majority-vote label
+    // propagation from the mesh boundary - starting from each point's own
+    // directly-adjacent real RCA/LCA vs "other" neighbours and spreading
+    // inward round by round, so points near the true RCA border resolve to
+    // RCA while points on the far, aorta-facing side of the SAME connected
+    // removed patch correctly stay removed, without needing the whole patch
+    // to share one aggregate majority. Ties (a point receiving equal
+    // evidence from both sides in the same round) stay removed - the
+    // conservative default. A single directly-resolved vertex is the
+    // zero-propagation-rounds special case.
+    restore_removed_by_propagation(&adjacency, &labels, &mut new_labels, 3, 1);
+    restore_removed_by_propagation(&adjacency, &labels, &mut new_labels, 4, 2);
 
     let mut result = ReclassifiedLabels {
         aorta_points: Vec::new(),
@@ -519,13 +522,25 @@ fn reclassify_minority_components(
     }
 }
 
-/// Restores whole connected components of `removed_label` vertices to
-/// `target_label` when their external boundary is >70% `target_label`,
-/// mirroring the original per-vertex Logic B threshold but evaluated over the
-/// whole falsely-removed blob at once. Components with an empty boundary
-/// (no external mesh connectivity at all) are left untouched, matching the
-/// original per-vertex `neighbors.is_empty()` guard.
-fn restore_removed_components(
+/// Restores connected `removed_label` vertices to `target_label` via
+/// majority-vote label propagation rather than a single whole-component
+/// vote: a genuinely mixed removed patch (one face hugging the real
+/// coronary lumen, another hugging the aortic wall, mesh-connected through
+/// the same fold) is resolved sub-region by sub-region instead of being
+/// judged as one atomic blob.
+///
+/// Round 0: every removed vertex tallies its directly-adjacent REAL
+/// (non-removed) neighbours - how many are `target_label` vs anything else
+/// ("other"). A strict majority decides it immediately. Every following
+/// round, each vertex decided in the *previous* round contributes one vote
+/// to its still-undecided removed neighbours; a whole round's
+/// contributions are applied together before checking for new majorities,
+/// so a vertex genuinely equidistant from both sides (one vote of each in
+/// the same round) ties and stays undecided - deterministic regardless of
+/// processing order. Vertices that never accumulate a majority (including
+/// a fully mesh-isolated removed component with no real neighbours at all)
+/// are left unchanged.
[…]
