### 2026-08-06 cceb991
add two different orientation methods
