+++ src/types/native/centerline.rs
@@ -565,50 +565,101 @@ impl Centerline {
 
         let mut branches = self.branches_as_vecs();
 
-        // --- Branch 0: highest-z point must be at index 0 ---
-        if !branches[0].is_empty() {
-            let max_z_idx = branches[0]
-                .iter()
-                .enumerate()
-                .max_by(|(_, a), (_, b)| {
-                    a.contour_point
-                        .z
-                        .partial_cmp(&b.contour_point.z)
-                        .unwrap_or(std::cmp::Ordering::Equal)
-                })
-                .map(|(i, _)| i)
-                .unwrap_or(0);
-            if max_z_idx != 0 {
-                branches[0].reverse();
-            }
+        if Self::should_reverse_by_max_z(&branches[0]) {
+            branches[0].reverse();
         }
 
-        // --- Side branches: first point must be the one closest to branch 0 ---
         for k in 1..n {
             if branches[k].is_empty() || branches[0].is_empty() {
                 continue;
             }
-
-            // Clone endpoints so we can borrow branches[0] freely afterwards.
-            let first_pt = branches[k][0].contour_point;
-            let last_pt = branches[k].last().unwrap().contour_point;
-
-            let dist_first = branches[0]
-                .iter()
-                .map(|p| p.contour_point.distance_to(&first_pt))
-                .fold(f64::INFINITY, f64::min);
-
-            let dist_last = branches[0]
-                .iter()
-                .map(|p| p.contour_point.distance_to(&last_pt))
-                .fold(f64::INFINITY, f64::min);
-
-            if dist_last < dist_first {
+            if Self::should_reverse_relative_to(&branches[k], &branches[0]) {
                 branches[k].reverse();
             }
         }
 
         self.rebuild_from_branches(branches);
+    }
+
+    /// Reverse the whole centerline in place if the point with the highest
+    /// z-coordinate is not already at index 0.
+    ///
+    /// Intended for a centerline with no anatomical reference to orient against
+    /// (e.g. the aorta) — use [`Centerline::orient_to_reference`] instead
+    /// whenever one is available. Only correct under the standard CT/DICOM
+    /// convention where z increases toward the head, so the aortic root/valve
+    /// is the highest-z point.
+    pub fn orient_by_max_z(&mut self) {
+        if Self::should_reverse_by_max_z(&self.points) {
+            self.reverse_in_place();
+        }
+    }
+
+    /// Reverse the whole centerline in place if its last point is closer to
+    /// `reference` than its first point is, so that the end nearer `reference`
+    /// becomes index 0.
+    ///
+    /// Distance to `reference` is the minimum distance to any of its points,
+    /// not a single fixed point — e.g. for a coronary centerline, `reference`
+    /// would be the aorta centerline as a whole, not one ostium point.
+    pub fn orient_to_reference(&mut self, reference: &Centerline) {
+        if Self::should_reverse_relative_to(&self.points, &reference.points) {
+            self.reverse_in_place();
+        }
+    }
+
+    /// Reverse `self.points`, reassign sequential `point_index`, and recompute tangents.
+    fn reverse_in_place(&mut self) {
+        self.points.reverse();
+        for (i, p) in self.points.iter_mut().enumerate() {
+            p.contour_point.point_index = i as u32;
+        }
+        self.recompute_tangents();
+    }
+
[…]
