### 2026-08-06 a3bd0e6
initial refactor of centerlines logic
