+++ src/types/native/centerline.rs
@@ -1,7 +1,6 @@
 use super::centerline_point::CenterlinePoint;
 use super::contour_point::ContourPoint;
 use super::Point3D;
-use crate::types::utils;
 use nalgebra::Vector3;
 
 #[derive(Debug, Clone, PartialEq)]
@@ -662,26 +661,16 @@ impl Centerline {
         dist_last < dist_first
     }
 
-    /// Remove the run-alongside-main-branch prefix from every side branch,
-    /// optionally strip the inlet region from branch 0, and optionally smooth.
+    /// Remove the run-alongside-main-branch prefix duplicated by every side branch.
     ///
-    /// VTP files export every branch starting from the vessel origin, so side
-    /// branches share a common prefix with branch 0.  For each side branch this
-    /// method trims the contiguous leading prefix whose points all lie within
-    /// one mean inter-point spacing of branch 0 of at least one main-branch
-    /// point. The last point of the trimmed prefix is kept as the bifurcation
-    /// junction. Branches whose entire extent lies within that buffer are
-    /// dropped completely.
-    ///
-    /// If `rm_start_mm > 0`, the leading points of branch 0 are also removed
-    /// up to `rm_start_mm` arc-length from its first point.  This is useful
-    /// when the main branch starts at the aortic inlet and the proximal region
-    /// is outside the region of interest.
-    ///
-    /// If `smooth` is `true`, a Gaussian kernel with half-width `smooth_sigma`
-    /// (in number of points) is applied per branch after all trimming is done.
-    /// See [`utils::smooth_centerline`] for kernel details.
-    pub fn cleanup_vtp_data(&mut self, rm_start_mm: f64, smooth: bool, smooth_sigma: f64) {
+    /// Some centerline export formats (e.g. VTP) write every branch starting from
+    /// the vessel origin, so side branches share a common prefix with branch 0.
+    /// For each side branch this trims the contiguous leading prefix whose points
+    /// all lie within one mean inter-point spacing of branch 0 of at least one
+    /// main-branch point. The last point of the trimmed prefix is kept as the
+    /// bifurcation junction. Branches whose entire extent lies within that buffer
+    /// are dropped completely.
+    pub fn remove_branch_overlap(&mut self) {
         if self.branch_start_indices.is_empty() {
             return;
         }
@@ -690,14 +679,182 @@ impl Centerline {
         let mut branches = self.branches_as_vecs();
 
         Self::remove_overlapping(&mut branches, buffer * buffer);
-        Self::remove_trailing_start(&mut branches, rm_start_mm);
 
         self.rebuild_from_branches(branches);
+    }
 
-        if smooth {
-            let new_cl_smooth = utils::smooth_centerline(self, smooth_sigma);
-            *self = new_cl_smooth;
+    /// Trim `mm` of arc length off the start of branch 0.
+    ///
+    /// Useful when the main branch starts at the aortic inlet and the proximal
+    /// region is outside the region of interest.
+    pub fn trim_start(&mut self, mm: f64) {
+        if mm <= 0.0 || self.branch_start_indices.is_empty() {
+            return;
         }
+
+        let mut branches = self.branches_as_vecs();
+
+        Self::remove_trailing_start(&mut branches, mm);
+
+        self.rebuild_from_branches(branches);
+    }
+
+    /// Resample every branch independently to even arc-length spacing.
+    ///
[…]
