+++ src/types.rs
@@ -1,3 +1,2 @@
 pub mod binding;
 pub mod native;
-pub(crate) mod utils;
