+++ multimodars/ccta/labeling.py
@@ -15,6 +15,7 @@ from ..multimodars import (
     clean_outlier_points,
     find_points_by_cl_region,
     keep_largest_connected_component,
+    read_centerline_vtp,
     PyCenterline,
 )
 from .._converters import numpy_to_centerline
@@ -24,9 +25,9 @@ from .debug_plots import plot_results_key
 
 def label_geometry(
     path_ccta_geometry: Path | str | trimesh.Trimesh,
-    path_centerline_aorta: Path | str | PyCenterline,
-    path_centerline_rca: Path | str | PyCenterline,
-    path_centerline_lca: Path | str | PyCenterline,
+    path_centerline_aorta: Path | str | PyCenterline | np.ndarray,
+    path_centerline_rca: Path | str | PyCenterline | np.ndarray,
+    path_centerline_lca: Path | str | PyCenterline | np.ndarray,
     acute_takeoff_rca: bool = False,
     acute_takeoff_lca: bool = False,
     n_points_takeoff_rca: int = 120,
@@ -55,13 +56,17 @@ def label_geometry(
     path_ccta_geometry : Path or str
         Path to the CCTA surface mesh file (any format supported by
         :func:`multimodars.io.read_geometrical.read_mesh`).
-    path_centerline_aorta : Path or str
-        Path to a CSV file containing the aortic centerline (comma-delimited,
-        columns: x, y, z, …).
-    path_centerline_rca : Path or str
-        Path to a CSV file containing the RCA centerline.
-    path_centerline_lca : Path or str
-        Path to a CSV file containing the LCA centerline.
+    path_centerline_aorta : PyCenterline, Path, str, or numpy.ndarray
+        Aortic centerline: a ``.vtp`` file path, a CSV file path (comma-delimited,
+        columns: x, y, z, …), an existing ``PyCenterline``, or an array of points.
+        Orientation is normalised automatically (``orient_by_max_z`` for the aorta,
+        ``orient_to_reference`` against the aorta for RCA/LCA); prepare trimming,
+        resampling, branch extraction, and smoothing beforehand, e.g. with
+        :func:`load_and_prepare_centerline`.
+    path_centerline_rca : PyCenterline, Path, str, or numpy.ndarray
+        RCA centerline, same accepted formats as *path_centerline_aorta*.
+    path_centerline_lca : PyCenterline, Path, str, or numpy.ndarray
+        LCA centerline, same accepted formats as *path_centerline_aorta*.
     acute_takeoff_rca : bool, optional
         When ``True`` applies ray-triangle occlusion removal to the RCA region
         to strip points overlapping the aortic wall near an acute-angle takeoff
@@ -265,13 +270,22 @@ def label_geometry(
     return new_results, (cl_rca, cl_lca, cl_aorta)
 
 
-def _try_load_cl(path_centerline_aorta: PyCenterline | Path | str, name: str):
+def _try_load_cl(
+    path_centerline_aorta: PyCenterline | Path | str | np.ndarray, name: str
+):
     if isinstance(path_centerline_aorta, PyCenterline):
         cl_aorta = path_centerline_aorta
         print(f"Using provided {name} centerline: {len(cl_aorta.points)} points")
     elif isinstance(path_centerline_aorta, np.ndarray):
         cl_aorta = numpy_to_centerline(path_centerline_aorta)
         print(f"Using provided {name} centerline: {len(cl_aorta.points)} points")
+    elif str(path_centerline_aorta).lower().endswith(".vtp"):
+        try:
+            cl_aorta = read_centerline_vtp(str(path_centerline_aorta))
+            print(f"Loaded {name} centerline from VTP: {len(cl_aorta.points)} points")
+        except Exception as e:
[…]
