+++ multimodars/__init__.py
@@ -47,6 +47,7 @@ from .ccta.labeling import (
     label_geometry,
     label_anomalous_region,
     label_branches,
+    load_and_prepare_centerline,
 )
 from .ccta.manipulating import (
     scale_region_centerline_morphing,
@@ -131,6 +132,7 @@ __all__ = [
     "discretize_vessel_tree",
     "find_sharp_angles",
     "label_branches",
+    "load_and_prepare_centerline",
 ]
 
 # def _print_banner():

[…]
