### 2026-08-06 1d2083b
cleanup the remaining centerline cleanup in other functions
