+++ examples/stitching.py
@@ -39,8 +39,8 @@ results = mm.label_geometry(
     path_centerline_lca=lca_cl,
     bounding_sphere_radius_mm_rca=3.0,
     bounding_sphere_radius_mm_lca=3.0,
-    n_points_takeoff_rca=100,
-    n_points_takeoff_lca=100,
+    range_mm_takeoff_rca=60.0,  # mm, was a point count before
+    range_mm_takeoff_lca=40.0,  # mm, was a point count before
     acute_takeoff_rca=True,
     acute_takeoff_lca=False,
     control_plot=False,
@@ -180,8 +180,8 @@ results = mm.label_geometry(
     path_centerline_lca="../data/centerline_lca.csv",
     bounding_sphere_radius_mm_rca=3.0,
     bounding_sphere_radius_mm_lca=3.0,
-    n_points_takeoff_rca=100,
-    n_points_takeoff_lca=100,
+    range_mm_takeoff_rca=60.0,  # mm, was a point count before
+    range_mm_takeoff_lca=40.0,  # mm, was a point count before
     acute_takeoff_rca=True,
     acute_takeoff_lca=False,
     control_plot=True,
