### 2026-08-07 1b6e6d3
remove legacy check centerline
