+++ src/types/native/centerline.rs
@@ -549,73 +549,57 @@ impl Centerline {
         self.rebuild_from_branches(branches);
     }
 
-    /// Ensure consistent ordering across all branches:
+    /// Reverse branch 0 in place if its highest-z point is not already at its start,
+    /// then apply the same "closer end goes first" rule to every side branch,
+    /// using branch 0 (post-reversal) as the reference.
     ///
-    /// * **Branch 0** – the point with the highest z-coordinate must be at index 0.
-    ///   If it is not, the entire branch is reversed.
-    /// * **Side branches (1 … n-1)** – the endpoint that is closest to any point on
-    ///   branch 0 must be the *first* point of the branch.  If the last point is
-    ///   closer to branch 0 than the first, the branch is reversed.
-    pub fn check_centerline(&mut self) {
-        let n = self.branch_start_indices.len();
-        if n == 0 {
-            return;
-        }
-
-        let mut branches = self.branches_as_vecs();
-
-        if Self::should_reverse_by_max_z(&branches[0]) {
-            branches[0].reverse();
-        }
-
-        for k in 1..n {
-            if branches[k].is_empty() || branches[0].is_empty() {
-                continue;
-            }
-            if Self::should_reverse_relative_to(&branches[k], &branches[0]) {
-                branches[k].reverse();
-            }
-        }
-
-        self.rebuild_from_branches(branches);
-    }
-
-    /// Reverse branch 0 in place if its highest-z point is not already at its start.
-    ///
-    /// Only branch 0 is touched — side branches keep their own order, since this
-    /// answers "which end of the main vessel is proximal", not a per-branch concern.
     /// Intended for a centerline with no anatomical reference to orient against
     /// (e.g. the aorta) — use [`Centerline::orient_to_reference`] instead whenever
     /// one is available. Only correct under the standard CT/DICOM convention where
     /// z increases toward the head, so the aortic root/valve is the highest-z point.
     pub fn orient_by_max_z(&mut self) {
-        if self.branch_start_indices.is_empty() {
+        let n = self.branch_start_indices.len();
+        if n == 0 {
             return;
         }
         let mut branches = self.branches_as_vecs();
         if Self::should_reverse_by_max_z(&branches[0]) {
             branches[0].reverse();
         }
+        if let Some((first, rest)) = branches.split_first_mut() {
+            for branch in rest.iter_mut() {
+                if Self::should_reverse_relative_to(branch, first.as_slice()) {
+                    branch.reverse();
+                }
+            }
+        }
         self.rebuild_from_branches(branches);
     }
 
     /// Reverse branch 0 in place if its last point is closer to `reference`'s
-    /// branch 0 than its first point is, so the end nearer `reference` becomes
-    /// branch 0's start.
+    /// branch 0 than its first point is, then apply the same rule to every side
+    /// branch — so every branch of `self` starts at whichever end is nearer
+    /// `reference`.
     ///
-    /// Only branch 0 is touched on both sides — side branches of `self` keep
-    /// their own order, and any side branches `reference` has are ignored so a
-    /// stray one can't skew the distance check. Distance to `reference` is the
-    /// minimum distance to any point of its branch 0, not a single fixed point —
-    /// e.g. for a coronary centerline, `reference` would be the aorta centerline,
-    /// not one ostium point.
+    /// Any side branches `reference` has are ignored so a stray one can't skew
[…]
