+++ src/types/binding/py_centerline.rs
@@ -311,36 +311,19 @@ impl PyCenterline {
         Ok(PyCenterline::from(&cl))
     }
 
-    /// Normalise branch ordering so that downstream processing is consistent.
+    /// Reverse branch 0 if its highest-z point is not already at its start,
+    /// then apply the same "closer end goes first" rule to every side branch,
+    /// using branch 0 (post-reversal) as the reference.
     ///
-    /// * **Branch 0** – the point with the highest z-coordinate is moved to
-    ///   index 0 (the whole branch is reversed if necessary).
-    /// * **Side branches** – the endpoint closest to branch 0 becomes index 0
-    ///   (the branch is reversed if necessary).
+    /// For centerlines with no anatomical reference to orient against, e.g. the
+    /// aorta — use ``orient_to_reference`` instead whenever one is available.
+    /// Only correct under the standard CT/DICOM convention where z increases
+    /// toward the head, so the aortic root/valve is the highest-z point.
     ///
     /// Returns
     /// -------
     /// PyCenterline
     ///     New centerline with all branches in canonical order.
-    pub fn check_centerline(&self) -> PyResult<PyCenterline> {
-        let mut cl = self.to_rust_centerline();
-        cl.check_centerline();
-        Ok(PyCenterline::from(&cl))
-    }
-
-    /// Reverse branch 0 if its highest-z point is not already at its start.
-    ///
-    /// Only branch 0 is touched — side branches keep their own order, since this
-    /// answers "which end of the main vessel is proximal", not a per-branch
-    /// concern. For centerlines with no anatomical reference to orient against,
-    /// e.g. the aorta — use ``orient_to_reference`` instead whenever one is
-    /// available. Only correct under the standard CT/DICOM convention where z
-    /// increases toward the head, so the aortic root/valve is the highest-z point.
-    ///
-    /// Returns
-    /// -------
-    /// PyCenterline
-    ///     New centerline with branch 0 reversed if necessary.
     pub fn orient_by_max_z(&self) -> PyResult<PyCenterline> {
         let mut cl = self.to_rust_centerline();
         cl.orient_by_max_z();
@@ -348,15 +331,14 @@ impl PyCenterline {
     }
 
     /// Reverse branch 0 if its last point is closer to `reference`'s branch 0
-    /// than its first point is, so the end nearer `reference` becomes branch 0's
-    /// start.
+    /// than its first point is, then apply the same rule to every side branch —
+    /// so every branch starts at whichever end is nearer `reference`.
     ///
-    /// Only branch 0 is touched on both sides — side branches of `self` keep
-    /// their own order, and any side branches `reference` has are ignored so a
-    /// stray one can't skew the distance check. Distance to `reference` is the
-    /// minimum distance to any point of its branch 0, not a single fixed point —
-    /// e.g. for a coronary centerline, `reference` would be the aorta centerline,
-    /// not one ostium point.
+    /// Any side branches `reference` has are ignored so a stray one can't skew
+    /// the distance check — only `reference`'s branch 0 is ever measured
+    /// against. Distance to `reference` is the minimum distance to any point of
+    /// its branch 0, not a single fixed point — e.g. for a coronary centerline,
+    /// `reference` would be the aorta centerline, not one ostium point.
     ///
     /// Parameters
     /// ----------
@@ -366,7 +348,7 @@ impl PyCenterline {
     /// Returns
     /// -------
     /// PyCenterline
-    ///     New centerline with branch 0 reversed if necessary.
+    ///     New centerline with all branches in canonical order.
     pub fn orient_to_reference(&self, reference: &PyCenterline) -> PyResult<PyCenterline> {
         let mut cl = self.to_rust_centerline();
         cl.orient_to_reference(&reference.to_rust_centerline());
