### 2026-08-07 ee41b56
update notebook, changelog and bump version
