+++ src/types/native/centerline.rs
@@ -429,8 +429,9 @@ impl Centerline {
         self.recompute_tangents();
     }
 
-    /// Return local positions (0-indexed within the branch) of interior points
-    /// where the opening angle satisfies `cos_angle > cos_threshold`.
+    /// Return global `point_index` values (as in `ContourPoint::point_index`, i.e.
+    /// indices into the flat `points` Vec) of interior points on `branch_id` where
+    /// the opening angle satisfies `cos_angle > cos_threshold`.
     /// Use `cos_threshold = 0.0` for < 90°, `0.5` for < 60°, etc.
     pub fn find_sharp_angles(&self, branch_id: u32, cos_threshold: f64) -> Vec<usize> {
         let idx = branch_id as usize;
@@ -460,19 +461,34 @@ impl Centerline {
                 }
                 v1.dot(&v2) / (n1 * n2) > cos_threshold
             })
+            .map(|local_pos| start + local_pos)
             .collect()
     }
 
-    /// Split the branch at `local_pos` (0-indexed within the branch).
-    /// Both resulting segments include the split point.
-    /// If `branch_id == 0` the longer segment stays as branch 0.
-    /// For side branches the first segment keeps its slot; the second is appended.
-    pub fn split_branch(&mut self, branch_id: u32, local_pos: usize) {
-        let mut branches = self.branches_as_vecs();
+    /// Split `branch_id` at `point_index` (a global index into `points`, as
+    /// returned by [`Centerline::find_sharp_angles`] - must fall within
+    /// `branch_id`'s own range). Both resulting segments include the split
+    /// point. Branches are re-sorted by descending length afterwards, so
+    /// branch 0 is always the longest overall, matching the invariant
+    /// established by [`Centerline::calculate_branches`].
+    pub fn split_branch(&mut self, branch_id: u32, point_index: usize) {
         let idx = branch_id as usize;
-        if idx >= branches.len() {
+        let n = self.branch_start_indices.len();
+        if idx >= n {
             return;
         }
+        let start = self.branch_start_indices[idx];
+        let end = if idx + 1 < n {
+            self.branch_start_indices[idx + 1]
+        } else {
+            self.points.len()
+        };
+        if point_index < start || point_index >= end {
+            return;
+        }
+        let local_pos = point_index - start;
+
+        let mut branches = self.branches_as_vecs();
         let branch = branches.remove(idx);
         if local_pos == 0 || local_pos >= branch.len().saturating_sub(1) {
             branches.insert(idx, branch);
@@ -481,26 +497,18 @@ impl Centerline {
 
         let seg_a = branch[..=local_pos].to_vec();
         let seg_b = branch[local_pos..].to_vec();
+        branches.push(seg_a);
+        branches.push(seg_b);
 
-        if branch_id == 0 {
-            let (main_seg, other_seg) = if seg_a.len() >= seg_b.len() {
-                (seg_a, seg_b)
-            } else {
-                (seg_b, seg_a)
-            };
-            branches.insert(0, main_seg);
-            branches.push(other_seg);
-        } else {
-            branches.insert(idx, seg_a);
-            branches.push(seg_b);
-        }
-
+        Self::sort_branches_by_length(&mut branches);
         self.rebuild_from_branches(branches);
     }
 
     /// Merge two branches into one. Endpoints are matched by minimum distance
-    /// so the segments are concatenated in the correct spatial order.
[…]
