### 2026-08-07 8ea21c4
fix failing notebook tests
