+++ docs/tutorial_intravascular.rst
@@ -328,9 +328,9 @@ and the spacing is adjusted to match the z-spacing of the :class:`PyGeometry`.
 
 .. code-block:: python
 
-    aligned_geometry, resampled_cl = mm.align_three_point(
+    aligned_geometry, spacing_mm, total_rotation_deg = mm.align_three_point(
         centerline=centerline,
-        geometry_pair=rest,
+        geometry=rest,
         main_ref_pt=(12.2605, -201.3643, 1751.0554),
         counterclockwise_ref_pt=(11.7567, -202.1920, 1754.7975),
         clockwise_ref_pt=(15.6605, -202.1920, 1749.9655),
@@ -344,7 +344,7 @@ Hausdorff distances between the pointcloud and the geometry, :func:`multimodars.
 
 .. code-block:: python
 
-    aligned, resampled_cl = mm.align_combined(
+    aligned, spacing_mm, total_rotation_deg = mm.align_combined(
         centerline,
         rest,
         (12.2605, -201.3643, 1751.0554),
