### 2026-08-13 f513362
refactor extract stitching into it's own module
