+++ multimodars/ccta/__init__.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from . import manipulating
+from . import stitching
+from . import scaling
 from . import labeling
 from . import centerline_prep
 from . import debug_plots as debug_plots
@@ -216,45 +217,43 @@ def scale(
         Updated *results* with ``"mesh"`` replaced by the fully scaled mesh
         and all coordinate lists remapped to the new vertex positions.
     """
-    prox_scaling, distal_scaling = manipulating.find_distal_and_proximal_scaling(
+    prox_scaling, distal_scaling = scaling.find_distal_and_proximal_scaling(
         frames=aligned_frames,
         centerline=cl_vessel,
         results=results,
     )
 
-    aortic_scaling = manipulating.find_aorta_scaling(
+    aortic_scaling = scaling.find_aorta_scaling(
         frames=aligned_frames,
         cl_aorta=cl_aorta,
         results=results,
     )
 
-    scaled_distal = manipulating.scale_region_centerline_morphing(
+    scaled_distal = scaling.scale_region_centerline_morphing(
         mesh=results["mesh"],
         region_points=results["distal_points"],
         centerline=cl_vessel,
         diameter_adjustment_mm=distal_scaling,
     )
-    results = manipulating.sync_results_to_mesh(results, results["mesh"], scaled_distal)
+    results = stitching.sync_results_to_mesh(results, results["mesh"], scaled_distal)
 
-    scaled_distal_aortic = manipulating.scale_region_centerline_morphing(
+    scaled_distal_aortic = scaling.scale_region_centerline_morphing(
         mesh=results["mesh"],
         region_points=results["aorta_points"] + results["rca_removed_points"],
         centerline=cl_aorta,
         diameter_adjustment_mm=aortic_scaling,
     )
-    results = manipulating.sync_results_to_mesh(
+    results = stitching.sync_results_to_mesh(
         results, results["mesh"], scaled_distal_aortic
     )
 
-    scaled_proximal = manipulating.scale_region_centerline_morphing(
+    scaled_proximal = scaling.scale_region_centerline_morphing(
         mesh=results["mesh"],
         region_points=results["proximal_points"],
         centerline=cl_vessel,
         diameter_adjustment_mm=prox_scaling,
     )
-    results = manipulating.sync_results_to_mesh(
-        results, results["mesh"], scaled_proximal
-    )
+    results = stitching.sync_results_to_mesh(results, results["mesh"], scaled_proximal)
 
     return results
 
@@ -318,11 +317,9 @@ def stitch(
             "Install it with: pip install 'multimodars[meshlab]'"
         )
 
-    updated_results = manipulating.remove_labeled_points_from_mesh(
-        results, region_remove
-    )
+    updated_results = stitching.remove_labeled_points_from_mesh(results, region_remove)
 
-    stitched = manipulating.stitch_ccta_to_intravascular(
+    stitched = stitching.stitch_ccta_to_intravascular(
         geometry,
         updated_results["mesh"],
         updated_results,
@@ -418,7 +415,7 @@ def export_section_stl(
     elif type in _REGION_KEYS:
         region_points = results.get(_REGION_KEYS[type], [])
         if type == "aorta":
-            sub_mesh_dict = manipulating.keep_labeled_points_from_mesh(
+            sub_mesh_dict = stitching.keep_labeled_points_from_mesh(
                 results, ["aorta_points", "rca_removed_points", "lca_removed_points"]
             )
             sub_mesh = sub_mesh_dict["mesh"]
[…]
