+++ multimodars/ccta/scaling.py
@@ -1,0 +1,298 @@
+from __future__ import annotations
+
+import numpy as np
+import trimesh
+
+from ..multimodars import (
+    PyFrame,
+    PyCenterline,
+    adjust_diameter_centerline_morphing_simple,
+    find_proximal_distal_scaling,
+    find_aortic_scaling,
+    find_aortic_wall_scaling as _find_aortic_wall_scaling,
+)
+
+
+def scale_region_centerline_morphing(
+    mesh: trimesh.Trimesh,
+    region_points: list,
+    centerline,
+    diameter_adjustment_mm: float,
+) -> trimesh.Trimesh:
+    """Scale a mesh region radially around its centerline.
+
+    Each vertex in *region_points* is displaced along the direction from the
+    nearest centerline point outward (positive *diameter_adjustment_mm*) or
+    inward (negative).
+
+    Parameters
+    ----------
+    mesh : trimesh.Trimesh
+        The original mesh.  A copy is returned; the input is not modified.
+    region_points : list of tuple
+        ``(x, y, z)`` coordinates of the vertices to be scaled.  Only vertices
+        present in this list are moved.
+    centerline : PyCenterline
+        Centerline of the vessel region used as the morphing axis.
+    diameter_adjustment_mm : float
+        Diameter change in millimetres.  Positive values expand the lumen;
+        negative values contract it.
+
+    Returns
+    -------
+    trimesh.Trimesh
+        A new mesh with the selected region scaled.
+
+    Warns
+    -----
+    If no vertices matching *region_points* are found in the mesh, a warning
+    is printed and the unmodified copy is returned.
+    """
+    scaled_mesh = mesh.copy()
+
+    region_vertex_indices_list: list[int] = []
+    region_set = set(region_points)
+
+    for idx, vertex in enumerate(scaled_mesh.vertices):
+        if tuple(vertex) in region_set:
+            region_vertex_indices_list.append(idx)
+
+    region_vertex_indices = np.array(region_vertex_indices_list)
+
+    if len(region_vertex_indices) == 0:
+        print("Warning: No vertices found for scaling region")
+        return scaled_mesh
+
+    print(f"\nScaling {len(region_vertex_indices)} vertices around {centerline}")
+    print(f"Diameter adjustment: {np.round(diameter_adjustment_mm, 2)} mm")
+
+    region_vertices_list = [
+        tuple(vertex) for vertex in scaled_mesh.vertices[region_vertex_indices]
+    ]
+    adjusted_points = adjust_diameter_centerline_morphing_simple(
+        centerline=centerline,
+        points=region_vertices_list,
+        diameter_adjustment_mm=diameter_adjustment_mm,
+    )
+
+    scaled_mesh.vertices[region_vertex_indices] = np.array(
+        adjusted_points, dtype=np.float64
+    )
+
+    # Clear mesh cache since we modified vertices directly
+    scaled_mesh.vertices.flags["WRITEABLE"] = False
+
+    return scaled_mesh
+
+
+def find_distal_and_proximal_scaling(
+    frames,
+    centerline,
+    results: dict,
+    dist_range: int = 3,
+    prox_range: int = 2,
+) -> tuple[float, float]:
+    """Compute the optimal radial scaling factors for the proximal and distal segments.
+
+    Collects lumen wall points from the first *prox_range* and last *dist_range*
+    imaging frames as reference geometry, then calls the Rust
+    ``find_proximal_distal_scaling`` routine to find the scaling factors that
+    best match the anomalous segment endpoints to those references.
+
+    Parameters
+    ----------
+    frames : list of PyFrame
+        Ordered intravascular imaging frames for the vessel.
+    centerline : PyCenterline
+        Centerline of the vessel region.
+    results : dict
[…]

[…]
