### 2026-08-13 a5b0bd3
rough refactor of stitching
