+++ multimodars/ccta/stitching.py
@@ -12,161 +12,6 @@ from ..multimodars import (
 from .._converters import geometry_to_trimesh
 
 
-def _fast_fix_normals(mesh: trimesh.Trimesh) -> None:
-    """Drop-in replacement for ``trimesh.Trimesh.fix_normals()``.
-
-    trimesh's own ``fix_winding`` does a Python/NetworkX BFS over the
-    face-adjacency graph with several small numpy allocations per edge -
-    O(n_edges) with heavy per-iteration overhead (e.g. ~3.9s on a ~52k-face
-    mesh). ``fix_mesh_winding`` is a Rust port of the same BFS-consistency
-    algorithm; ``fix_inversion`` (the volume-sign flip check) is already
-    vectorized numpy in trimesh, so it's left as-is.
-    """
-    mesh.faces = np.array(fix_mesh_winding(mesh.faces.tolist()), dtype=mesh.faces.dtype)
-    trimesh.repair.fix_inversion(mesh, multibody=False)
-
-
-def _project_to_best_fit_plane(
-    points: list[tuple[float, float, float]],
-) -> list[tuple[float, float, float]]:
-    """Project a ring of boundary points onto their best-fit plane.
-
-    Fits a plane via SVD (the plane normal is the direction of minimum variance)
-    and orthogonally projects every point onto it, flattening noise perpendicular
-    to the ring.
-    """
-    if len(points) < 3:
-        return points
-    pts = np.array(points, dtype=np.float64)
-    centroid = pts.mean(axis=0)
-    _, _, Vt = np.linalg.svd(pts - centroid, full_matrices=False)
-    normal = Vt[-1]
-    distances = (pts - centroid) @ normal
-    projected = pts - np.outer(distances, normal)
-    return [tuple(p) for p in projected]
-
-
-def _plane_normal_svd(pts: np.ndarray) -> np.ndarray:
-    """Best-fit plane normal for a point cloud via SVD (minimum-variance axis)."""
-    centroid = pts.mean(axis=0)
-    _, _, Vt = np.linalg.svd(pts - centroid, full_matrices=False)
-    return Vt[-1]
-
-
-def _angle_between_planes_deg(n1: np.ndarray, n2: np.ndarray) -> float:
-    """Acute angle in degrees between two planes given their normals."""
-    cos = np.clip(np.abs(np.dot(n1, n2)), 0.0, 1.0)
-    return float(np.degrees(np.arccos(cos)))
-
-
-def _clamp_to_plane(
-    points: list[tuple[float, float, float]],
-    plane_origin: np.ndarray,
-    plane_normal: np.ndarray,
-    overshoot: float = 0.0,
-) -> list[tuple[float, float, float]]:
-    """Clamp wrong-side points to the IV plane, then enforce a minimum gap.
-
-    Step 1: project any point on the wrong side of the plane onto it.
-    Step 2: if ``overshoot`` > 0, every point (including freshly clamped ones
-    that now sit exactly on the plane) that is within ``overshoot`` mm of the
-    plane on the correct side is pushed further away until it is exactly
-    ``overshoot`` mm from the plane.  This creates a clean buffer zone between
-    the aortic boundary ring and the IV ostium plane, avoiding the sharp angle
-    that would otherwise form.
-    """
-    pts = np.array(points, dtype=np.float64)
-    dists = (pts - plane_origin) @ plane_normal
-    correct_sign = np.sign(np.median(dists))
-
-    # Step 1: project wrong-side points onto the plane
-    wrong = (np.sign(dists) != correct_sign) & (dists != 0.0)
-    pts[wrong] -= np.outer(dists[wrong], plane_normal)
-
-    if overshoot > 0.0:
[…]
