### 2026-08-13 4aa93c6
add debug plot for boundary ring identification
