+++ multimodars/ccta/debug_plots.py
@@ -8,6 +8,8 @@ import trimesh
 from trimesh.points import PointCloud
 from trimesh.visual import ColorVisuals
 
+from ..multimodars import build_adjacency_map
+
 if TYPE_CHECKING:
     from ..multimodars import PyCenterline, PyDiscretizedVesselTree
 
@@ -506,5 +508,152 @@ def plot_sharp_angles(
             continue
         arr = np.array([(p.x, p.y, p.z) for p in segment], dtype=np.float64)
         scene_geoms.append(_make_point_cloud(arr, _PALETTE[i % len(_PALETTE)]))
+
+    trimesh.Scene(scene_geoms).show()
+
+
+def _order_boundary_rings(
+    boundary_indices: set[int],
+    adj_map,
+) -> list[list[int]]:
+    """Walk each connected boundary component into an ordered ring of vertex indices.
+
+    Local copy of the ordering logic used by
+    :func:`~multimodars.ccta.stitching._order_boundary_components` so the debug
+    plot stays self-contained.  Adjacency is restricted to boundary vertices,
+    then each component is traced edge-by-edge; disconnected regions (e.g. two
+    separate removals) come back as separate rings.
+    """
+    if not boundary_indices:
+        return []
+    if len(boundary_indices) == 1:
+        return [list(boundary_indices)]
+
+    ring_adj = {
+        i: [j for j in adj_map.get(i, []) if j in boundary_indices]
+        for i in boundary_indices
+    }
+
+    remaining = set(boundary_indices)
+    rings: list[list[int]] = []
+    while remaining:
+        start = next(iter(remaining))
+        ring = [start]
+        remaining.discard(start)
+        prev, current = -1, start
+        while True:
+            nxt = next(
+                (n for n in ring_adj.get(current, []) if n != prev and n in remaining),
+                None,
+            )
+            if nxt is None:
+                break
+            ring.append(nxt)
+            remaining.discard(nxt)
+            prev, current = current, nxt
+        rings.append(ring)
+    return rings
+
+
+def plot_boundary_edges(
+    results: dict,
+    key: str = "boundary_points",
+    show_mesh: bool = True,
+) -> None:
+    """Open an interactive trimesh scene of a stored boundary ring set and its edges.
+
+    Reconstructs the boundary as one or more ordered rings (matching
+    :func:`~multimodars.ccta.stitching._order_boundary_components`), draws each
+    ring's edges as connected line segments, and colours the ring vertices
+    red -> blue by their walk order.  This makes both the seam direction and any
+    split into multiple disconnected rings immediately visible - useful after
+    :func:`~multimodars.ccta.stitching.remove_labeled_points_from_mesh`, whose
+    ``"boundary_points"`` list is a flat concatenation that hides the ring
+    structure.
+
+    Colour coding
+    -------------
+    * Mesh - light gray, semi-transparent.
+    * Ring vertices - red (ring start) -> blue (ring end) gradient.
+    * Ring edges - one distinct colour per connected ring.
+
+    Parameters
+    ----------
+    results : dict
+        Must contain ``"mesh"`` and *key*.
+    key : str
+        Which stored point list to treat as the boundary.  Defaults to
+        ``"boundary_points"``; pass ``"prox_boundary_points"`` or
+        ``"dist_boundary_points"`` to inspect the stitching seams instead.
+    show_mesh : bool
+        Include the mesh in the scene for context.
+    """
+    print("\n=== BOUNDARY EDGES PLOT ===")
+
+    mesh: trimesh.Trimesh = results["mesh"]
[…]
