+++ multimodars/__init__.py
@@ -70,7 +70,12 @@ from .ccta.discretization_map import (
     find_sharp_angles,
 )
 from .ccta.fixing_functions import fix_and_remesh_stitched_mesh, manual_hole_fill
-from .ccta.debug_plots import plot_results_key, plot_centerline_edges, plot_sharp_angles
+from .ccta.debug_plots import (
+    plot_results_key,
+    plot_centerline_edges,
+    plot_sharp_angles,
+    plot_boundary_edges,
+)
 
 __all__ = [
     # Core classes
@@ -132,6 +137,7 @@ __all__ = [
     "plot_results_key",
     "plot_centerline_edges",
     "plot_sharp_angles",
+    "plot_boundary_edges",
     "discretize_vessel",
     "discretize_vessel_tree",
     "find_sharp_angles",
