### 2026-08-13 bafa9e3
cleanup functions for boundary rings after creation
