+++ multimodars/ccta/stitching.py
@@ -15,6 +15,7 @@ from .._converters import geometry_to_trimesh
 def remove_labeled_points_from_mesh(
     results: dict,
     region_keys: list[str] | str = "anomalous_points",
+    target_boundaries: int = 1,
 ) -> dict:
     """Remove one or more labeled regions of vertices from the mesh.
 
@@ -34,6 +35,11 @@ def remove_labeled_points_from_mesh(
     region_keys : str or list of str
         Key(s) in *results* whose point lists should be removed from the mesh.
         Defaults to ``"anomalous_points"`` for backwards compatibility.
+    target_boundaries : int
+        Number of open-boundary rings the removal is expected to create.
+        Removing a single blob leaves one ring (default); removing a region
+        that splits the surface (e.g. aorta + intramural wall) can leave two.
+        The rim is cleaned and reduced to this many rings before it is stored.
 
     Returns
     -------
@@ -71,23 +77,35 @@ def remove_labeled_points_from_mesh(
     keep_mask = np.ones(n_vertices, dtype=bool)
     keep_mask[list(remove_indices)] = False
 
-    # 3. Build adjacency map and find boundary vertices before removing faces:
-    #    kept vertices that had at least one removed neighbour will form the
-    #    open boundary ring used for stitching.
+    # 3. Mark the removal rim: kept vertices that had at least one removed
+    #    neighbour.  This seeds which open boundaries to clean, so the mesh's
+    #    unrelated rims (aorta inlet, vessel ends) are left alone.
     adj_map = build_adjacency_map(mesh.faces.tolist())
     boundary_indices = {
         i
         for i in range(n_vertices)
         if keep_mask[i] and any(j in remove_indices for j in adj_map.get(i, []))
     }
-    components = _order_boundary_components(boundary_indices, adj_map)
+
+    # 4. Drop faces that reference any removed vertex, then clean the exposed
+    #    rim.  Vertices that cannot form a clean ring are deleted from the mesh
+    #    too - not just skipped in the ring - so the stored boundary really is
+    #    the mesh's open edge.
+    face_keep_mask = np.all(keep_mask[mesh.faces], axis=1)
+    extra_drop, components = _clean_open_boundary(
+        mesh.faces[face_keep_mask],
+        mesh.vertices,
+        boundary_indices,
+        target_n=target_boundaries,
+    )
+    if extra_drop:
+        keep_mask[np.fromiter(extra_drop, dtype=np.int64)] = False
+        face_keep_mask = np.all(keep_mask[mesh.faces], axis=1)
+    new_faces = mesh.faces[face_keep_mask]
+
     boundary_points: list[tuple] = [
         tuple(mesh.vertices[i]) for component in components for i in component
     ]
-
-    # 4. Drop faces that reference any removed vertex
-    face_keep_mask = np.all(keep_mask[mesh.faces], axis=1)
-    new_faces = mesh.faces[face_keep_mask]
 
     # 5. Remap vertex indices in the kept faces
     new_index = np.full(n_vertices, -1, dtype=np.int64)
@@ -106,7 +124,12 @@ def remove_labeled_points_from_mesh(
 
     print(f"Applying removal of '{region_keys}'")
     print(f"Removed {len(points_to_remove)}")
-    print(f"Created {len(updated['boundary_points'])} boundary points")
+    if extra_drop:
+        print(f"Culled {len(extra_drop)} unclean boundary vertices from the mesh")
+    print(
+        f"Created {len(updated['boundary_points'])} boundary points "
+        f"in {len(components)} ring(s): {[len(c) for c in components]}"
+    )
 
     for key in region_keys:
         updated[key] = []
@@ -129,6 +152,7 @@ def remove_labeled_points_from_mesh(
 def keep_labeled_points_from_mesh(
     results: dict,
[…]
