### 2026-08-13 1f9d420
allow for two boundary rings after removing
