+++ multimodars/ccta/stitching.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-from collections.abc import Mapping
 import numpy as np
 import trimesh
 
@@ -10,6 +9,7 @@ from ..multimodars import (
     fix_mesh_winding,
 )
 from .._converters import geometry_to_trimesh
+from .boundary import clean_open_boundary
 
 
 def remove_labeled_points_from_mesh(
@@ -92,7 +92,7 @@ def remove_labeled_points_from_mesh(
     #    too - not just skipped in the ring - so the stored boundary really is
     #    the mesh's open edge.
     face_keep_mask = np.all(keep_mask[mesh.faces], axis=1)
-    extra_drop, components = _clean_open_boundary(
+    extra_drop, components = clean_open_boundary(
         mesh.faces[face_keep_mask],
         mesh.vertices,
         boundary_indices,
@@ -219,7 +219,7 @@ def keep_labeled_points_from_mesh(
     # Drop faces that reference any removed vertex, then clean the exposed rim,
     # deleting unclean boundary vertices from the mesh as well as the ring.
     face_keep_mask = np.all(keep_mask[mesh.faces], axis=1)
-    extra_drop, components = _clean_open_boundary(
+    extra_drop, components = clean_open_boundary(
         mesh.faces[face_keep_mask],
         mesh.vertices,
         boundary_indices,
@@ -262,304 +262,6 @@ def keep_labeled_points_from_mesh(
             updated[key] = [p for p in updated[key] if tuple(p) in new_coord_set]
 
     return updated
-
-
-def _open_boundary_edges(faces: np.ndarray) -> np.ndarray:
-    """Return the open-boundary edges of *faces* (edges used by exactly one face).
-
-    Parameters
-    ----------
-    faces : (F, 3) int array
-        Triangle vertex indices.
-
-    Returns
-    -------
-    (E, 2) int array
-        Sorted vertex-index pairs, one per edge lying on the open rim.
-    """
-    if len(faces) == 0:
-        return np.empty((0, 2), dtype=np.int64)
-    edges = faces[:, [0, 1, 1, 2, 2, 0]].reshape(-1, 2)
-    edges = np.sort(edges, axis=1)
-    uniq, counts = np.unique(edges, axis=0, return_counts=True)
-    return uniq[counts == 1]
-
-
-def _despike_ring(
-    order: list[int],
-    vertices: np.ndarray,
-    graph: Mapping[int, set[int]],
-    cos_thresh: float,
-) -> list[int]:
-    """Remove "bump" spikes from an ordered ring.
-
-    A bump spike is a vertex the rim detours out to and immediately back from,
-    so the two edge directions leaving it point nearly the same way (cosine
-    close to ``+1``); a normal rim vertex has them pointing nearly opposite
-    (cosine close to ``-1``).  Any vertex whose cosine exceeds *cos_thresh* is
-    dropped, which reconnects its two neighbours directly.  Iterated because
-    removing one tip can expose the next.
-    """
-    if len(order) < 3:
-        return order
-    closed = order[0] in graph.get(order[-1], set())
-    pts = list(order)
-    changed = True
-    while changed and len(pts) > 3:
-        changed = False
-        m = len(pts)
-        for i in range(m):
-            if not closed and (i == 0 or i == m - 1):
-                continue
-            d1 = vertices[pts[(i - 1) % m]] - vertices[pts[i]]
-            d2 = vertices[pts[(i + 1) % m]] - vertices[pts[i]]
-            n1 = float(np.linalg.norm(d1))
-            n2 = float(np.linalg.norm(d2))
-            if n1 == 0.0 or n2 == 0.0:
-                continue
[…]
