### 2026-08-13 6cb6039
directly store the different boundary rings
