+++ multimodars/ccta/stitching.py
@@ -9,7 +9,102 @@ from ..multimodars import (
     fix_mesh_winding,
 )
 from .._converters import geometry_to_trimesh
-from .boundary import clean_open_boundary
+from .boundary import clean_open_boundary, order_boundary_rings
+
+# Per-ring boundary keys: "boundary_points_1", "boundary_points_2", ...
+_BOUNDARY_RING_PREFIX = "boundary_points_"
+
+
+def _store_boundary_rings(
+    updated: dict,
+    vertices: np.ndarray,
+    rings: list[list[int]],
+) -> None:
+    """Store *rings* in *updated* as one list per ring, plus the flat total.
+
+    Each ring becomes ``boundary_points_1``, ``boundary_points_2``, ... in walk
+    order, so downstream code can treat a rim as a unit instead of guessing the
+    split back out of a flat list.  ``boundary_points`` still holds the
+    concatenation of all rings for callers that just want every boundary vertex.
+    Per-ring keys from an earlier call are cleared first, so a stale ring from a
+    previous removal cannot linger and be mistaken for a current one.
+    """
+    for key in [k for k in updated if k.startswith(_BOUNDARY_RING_PREFIX)]:
+        del updated[key]
+    per_ring = [[tuple(vertices[i]) for i in ring] for ring in rings]
+    for n, pts in enumerate(per_ring, start=1):
+        updated[f"{_BOUNDARY_RING_PREFIX}{n}"] = pts
+    updated["boundary_points"] = [pt for pts in per_ring for pt in pts]
+
+
+def _boundary_rings(results: dict, mesh: trimesh.Trimesh) -> list[list[tuple]]:
+    """Return the boundary rings stored in *results*, each in walk order.
+
+    Prefers the per-ring ``boundary_points_<n>`` lists written by
+    :func:`_store_boundary_rings`.  If only the flat ``boundary_points`` is
+    present - a results dict built before ring grouping, or by hand - the rings
+    are recovered from the mesh's open edges instead, so callers always get
+    properly ordered rings rather than an arbitrary point order.
+    """
+    rings: list[list[tuple]] = []
+    n = 1
+    while (key := f"{_BOUNDARY_RING_PREFIX}{n}") in results:
+        if results[key]:
+            rings.append([tuple(p) for p in results[key]])
+        n += 1
+    if rings:
+        return rings
+
+    flat = results.get("boundary_points") or []
+    if not flat:
+        return []
+    coord_to_idx = {tuple(v): i for i, v in enumerate(mesh.vertices)}
+    seeds = {idx for pt in flat if (idx := coord_to_idx.get(tuple(pt))) is not None}
+    return [
+        [tuple(mesh.vertices[i]) for i in ring]
+        for ring in order_boundary_rings(mesh.faces, mesh.vertices, seeds)
+    ]
+
+
+def _assign_rings_to_ends(
+    rings: list[list[tuple]],
+    prox_centroid: tuple[float, float, float],
+    dist_centroid: tuple[float, float, float],
+) -> tuple[list[tuple], list[tuple], list[int]]:
+    """Choose which whole ring stitches to each end of the intravascular mesh.
+
+    Of every way to hand two distinct rings to the proximal and distal ends, the
+    pairing with the smallest total centroid distance wins.  Assigning whole
+    rings - rather than each boundary point independently - means a stray point
+    that happens to sit nearer the far centroid can no longer tear one ring
+    across both seams.
+
+    Returns
+    -------
+    (list, list, list[int])
+        The proximal ring, the distal ring, and the indices of any rings left
+        over.
+    """
[…]
