### 2026-08-13 b7a0803
implement new stitching algorithm
