+++ multimodars/ccta/stitching.py
@@ -361,6 +361,7 @@ def stitch_ccta_to_intravascular(
     dist_start_mode: str = "nearest_iv",
     proximal_is_ostium: bool = True,
     clamp_overshoot: float = 0.5,
+    boundary_point_ratio: float = 1.0,
 ) -> dict:
     """Stitch an aligned intravascular mesh to a CCTA mesh.
 
@@ -394,6 +395,16 @@ def stitch_ccta_to_intravascular(
     proximal_points = iv_mesh.frames[0].lumen.points
     distal_points = iv_mesh.frames[-1].lumen.points
 
+    # Vessel axis: outward for the proximal patch points toward frames[0], and
+    # vice-versa for the distal patch.  Needed before the boundary prep, since
+    # the ostium plane is slid along the proximal outward direction.
+    prox_c = np.array(iv_mesh.frames[0].centroid)
+    dist_c = np.array(iv_mesh.frames[-1].centroid)
+    prox_outward = prox_c - dist_c  # points toward the proximal end
+    dist_outward = dist_c - prox_c  # points toward the distal end
+
+    target_n = max(3, round(boundary_point_ratio * len(proximal_points)))
+
     prox_boundary_pts, dist_boundary_pts, mesh = _prepare_prox_dist_boundary_pts(
         mesh,
         results,
@@ -402,9 +413,11 @@ def stitch_ccta_to_intravascular(
         proximal_is_ostium=proximal_is_ostium,
         proximal_iv_frame_pts=iv_mesh.frames[0].lumen.points,
         clamp_overshoot=clamp_overshoot,
+        target_n=target_n,
+        prox_outward=prox_outward,
     )
-    prox_point_step = len(proximal_points) // len(prox_boundary_pts)
-    dist_point_step = len(distal_points) // len(dist_boundary_pts)
+    prox_point_step = max(1, len(proximal_points) // len(prox_boundary_pts))
+    dist_point_step = max(1, len(distal_points) // len(dist_boundary_pts))
 
     # Adjust start point
     if prox_start_mode == "highest_z" or dist_start_mode == "highest_z":
@@ -419,16 +432,6 @@ def stitch_ccta_to_intravascular(
         dist_boundary_pts = _adjust_start_point_by_z(dist_boundary_pts)
     else:
         dist_boundary_pts = _rotate_to_nearest_iv(dist_boundary_pts, distal_points[0])
-
-    # Compute the vessel axis so each patch can be flipped to face outward,
-    # and also used as the consistent reference normal for direction checking.
-    # frames[0] is the "proximal" end, frames[-1] the "distal" end.
-    # The outward direction for the proximal patch points away from the mesh
-    # interior (toward frames[0]), and vice-versa for the distal patch.
-    prox_c = np.array(iv_mesh.frames[0].centroid)
-    dist_c = np.array(iv_mesh.frames[-1].centroid)
-    prox_outward = prox_c - dist_c  # points toward the proximal end
-    dist_outward = dist_c - prox_c  # points toward the distal end
 
     # Check / fix winding direction of each boundary ring vs its IV ring
     # independently, using the method that matches the start-point strategy.
@@ -451,12 +454,8 @@ def stitch_ccta_to_intravascular(
         )
 
     # Step 3: stitch each boundary ring to its IV ring
-    prox_patch = _stitch_boundary_ring(
-        prox_boundary_pts, proximal_points, prox_point_step, prox_outward
-    )
-    dist_patch = _stitch_boundary_ring(
-        dist_boundary_pts, distal_points, dist_point_step, dist_outward
-    )
+    prox_patch = _stitch_rings(prox_boundary_pts, proximal_points, prox_outward)
+    dist_patch = _stitch_rings(dist_boundary_pts, distal_points, dist_outward)
[…]
