### 2026-08-13 95aff7f
fix shrinking after smoothign
