### 2026-08-13 e7c4c94
update documentation and CHANGELOG
