+++ tests/test_ccta.py
@@ -8,11 +8,17 @@ Covers:
   - labeling: _keep_largest_connected_component (island-point filter for
                 find_points_by_cl_region's proximal/distal/anomalous output)
   - fixing_functions: manual_hole_fill, postprocess_stitched_mesh
-  - manipulating: remove_labeled_points_from_mesh,
-                  keep_labeled_points_from_mesh, sync_results_to_mesh,
-                  order_points_list, scale_region_centerline_morphing,
-                  _rotate_to_nearest_iv, _fix_ring_direction_by_distance,
-                  _stitch_boundary_ring
+  - stitching: remove_labeled_points_from_mesh,
+               keep_labeled_points_from_mesh, order_points_list,
+               _rotate_to_nearest_iv, _fix_ring_direction_by_distance,
+               _stitch_rings, _prepare_prox_dist_boundary_pts,
+               _condition_ostium_ring, _clamp_to_plane,
+               _enforce_layer_gap_from_plane, and the ring conditioning
+               helpers (_redistribute_ring_evenly,
+               _smooth_ring_preserving_size, _densify_boundary,
+               _assign_rings_to_ends, _shift_plane_clear_of, _toward_aorta)
+  - boundary: open_boundary_edges, order_boundary_rings, clean_open_boundary
+  - scaling: scale_region_centerline_morphing, sync_results_to_mesh
 """
 
 from __future__ import annotations
@@ -34,14 +40,27 @@ from multimodars.multimodars import (
     find_aortic_points,
     final_reclassification,
 )
+from multimodars.ccta.boundary import (
+    clean_open_boundary,
+    open_boundary_edges,
+    order_boundary_rings,
+)
 from multimodars.ccta.stitching import (
+    _assign_rings_to_ends,
     _clamp_to_plane,
+    _condition_ostium_ring,
+    _densify_boundary,
     _enforce_layer_gap_from_plane,
     _fast_fix_normals,
     _fix_ring_direction_by_distance,
     _prepare_prox_dist_boundary_pts,
+    _redistribute_ring_evenly,
+    _ring_calibre,
     _rotate_to_nearest_iv,
-    _stitch_boundary_ring,
+    _shift_plane_clear_of,
+    _smooth_ring_preserving_size,
+    _stitch_rings,
+    _toward_aorta,
     keep_labeled_points_from_mesh,
     order_points_list,
     remove_labeled_points_from_mesh,
@@ -914,54 +933,70 @@ class TestFixRingDirectionByDistance:
 
 
 # ===========================================================================
-# manipulating._stitch_boundary_ring
+# stitching._stitch_rings
 # ===========================================================================
 
 
-class TestStitchBoundaryRing:
+def _ring_coords(n: int, radius: float = 1.0, z: float = 0.0):
+    angles = np.linspace(0, 2 * np.pi, n, endpoint=False)
+    return [(radius * float(np.cos(a)), radius * float(np.sin(a)), z) for a in angles]
+
+
+class TestStitchRings:
     def _ring_pts(self, n: int, radius: float = 1.0, z: float = 0.0):
-        angles = np.linspace(0, 2 * np.pi, n, endpoint=False)
-        return [
-            (radius * float(np.cos(a)), radius * float(np.sin(a)), z) for a in angles
-        ]
+        return _ring_coords(n, radius, z)
 
     def test_creates_trimesh(self):
-        n_b, n_iv = 6, 12
-        boundary_pts = self._ring_pts(n_b, radius=1.0)
-        iv_pts = _make_iv_pts(self._ring_pts(n_iv, radius=1.2))
-        patch = _stitch_boundary_ring(boundary_pts, iv_pts, n_iv // n_b)
[…]
